### 2026-01-16 #9790 @editorialbot
[PRE REVIEW]: Virtual Brain Inference (VBI) toolkit on EBRAINS
**Submitting author:** @ziaeemehr (<a href="http://orcid.org/0000-0002-4696-9947">Abolfazl Ziaeemehr</a>)
**Repository:** https://github.com/ins-amu/vbi
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3
