### 2026-01-16 860ec17
Add Fig2.png and update references in paper.bib with DOIs
