+++ paper/paper.bib
@@ -1,9 +1,14 @@
-@Article{Betancourt2014,
-author="M. Betancourt  and S. Byrne and S. Livingstone and M. Girolami",
-title="The geometric foundations of Hamiltonian Monte Carlo",
-journal="arXiv:1410.5110",
-year="2014",
+@article{betancourt2017geometric,
+  author  = {Betancourt, Michael and Byrne, Simon and Livingstone, Sam and Girolami, Mark},
+  title   = {The geometric foundations of Hamiltonian Monte Carlo},
+  journal = {Bernoulli},
+  year    = {2017},
+  volume  = {23},
+  number  = {4A},
+  pages   = {2257--2298},
+  doi     = {10.3150/16-BEJ810}
 }
+
 
 @Article{Betancourt2013b,
 author="M. Betancourt  and M. Girolami",
@@ -19,6 +24,7 @@ year="2013",
   volume={217},
   pages={116839},
   year={2020},
+  doi={10.1016/j.neuroimage.2020.116839},
   publisher={Elsevier}
 }
 @article{Sisson2007,
@@ -29,6 +35,7 @@ year="2013",
   number={6},
   pages={1760--1765},
   year={2007},
+  doi={10.1073/pnas.0607208104},
   publisher={National Acad Sciences}
 }
 
@@ -40,6 +47,7 @@ year="2013",
   number={10},
   pages={e1000196},
   year={2008},
+  doi={10.1371/journal.pcbi.1000196},
   publisher={Public Library of Science San Francisco, USA}
 }
 
@@ -51,6 +59,7 @@ year="2013",
   number={2153},
   pages={20180132},
   year={2019},
+  doi={10.1098/rsta.2018.0132},
   publisher={The Royal Society Publishing}
 }
 
@@ -61,6 +70,7 @@ year="2013",
   volume={111},
   pages={385--430},
   year={2015},
+  doi={10.1016/j.neuroimage.2015.01.002},
   publisher={Elsevier}
 }
 
@@ -71,6 +81,7 @@ year="2013",
   volume={132},
   pages={155--165},
   year={2020},
+  doi={10.1016/j.neunet.2020.08.003},
   publisher={Elsevier}
 }
 
@@ -78,6 +89,7 @@ year="2013",
   title={Masked autoregressive flow for density estimation},
   author={Papamakarios, George and Pavlakou, Theo and Murray, Iain},
   journal={Advances in Neural Information Processing Systems},
+  doi={10.48550/arXiv.1705.07057},
   year={2017}
 }
 
@@ -87,6 +99,7 @@ year="2013",
   journal={Advances in Neural Information Processing Systems},
   volume={32},
   pages={7511--7522},
+  doi={10.48550/arXiv.1906.04032},
   year={2019}
 }
 
@@ -97,6 +110,7 @@ year="2013",
   volume={9},
   pages={e56261},
   year={2020},
+  doi={10.7554/eLife.56261},
   publisher={eLife Sciences Publications, Ltd}
 }
 
@@ -108,16 +122,8 @@ year="2013",
   number={3},
   pages={035019},
   year={2024},
+  doi={10.1088/2632-2153/ad6230},
   publisher={IOP Publishing}
-}
-
-@article{betancourt2017geometric,
-  title={The geometric foundations of hamiltonian monte carlo},
-  author={Betancourt, Michael and Byrne, Simon and Livingstone, Sam and Girolami, Mark},
-  journal={Bernoulli},
-  pages={2257--2298},
-  year={2017},
-  publisher={JSTOR}
 }
 
 @article{cranmer2020frontier,
@@ -128,6 +134,7 @@ year="2013",
   number={48},
   pages={30055--30062},
[…]
