### 2026-01-16 aea42c0
Add DOI to Betancourt and Girolami article in paper.bib
