+++ paper/paper.bib
@@ -14,6 +14,7 @@
 author="M. Betancourt  and M. Girolami",
 title="Hamiltonian Monte Carlo for hierarchical models",
 journal="arXiv:1312.0906",
+doi={10.48550/arXiv.1312.0906},
 year="2013",
 }
 
