### 2026-01-16 #9790 @crvernon
@Ziaeemehr - Note that JOSS papers now require the following additional sections:
1. Software Design (~200 words): Explain the trade-offs you weighed, the design/architecture you settled on, and why it matters. If other packages exist in this space, you must include an explicit "build vs. contribute" justification explaining why you created new software rather than contributing to existing projects.
2. Research Impact Statement (~200 words): Provide concise evidence of either realized impact (e.g., external use, integrations, enabled results) or credible near-term significance (novel capability with benchmarks, reproducible materials, and community-readiness signals such as documentation, tests, license, releases, and contribution process).
3. AI Usage Disclosure: Disclose any use of generative AI in your submission, including software creation, documentation generation, or paper authoring. If you did not use AI tools, please state this explicitly.

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

> **Paper file info**:
>
> 📄 Wordcount for `paper.md` is **1272**
>
> ✅ The paper includes a `Statement of need` section
>
> ```
> 🔴 Failed to discover a `Software Design` section in paper
>
> 🔴 Failed to discover a `Research Impact Statement` section in paper
>
> 🔴 Failed to discover an `AI usage disclosure` section in paper
> ```
