### 2026-01-21 6c752ca
Update paper.md
