### 2026-02-06 6421827
chore: update CHANGELOG for version 0.4.1 with new models, enhancements, and fixes
