### 2026-02-12 3b677ce
Refactor JR_sde and JR_sdde classes to auto-generate valid_params from default parameters; add comprehensive unit tests for C++ implementations.
