+++ vbi/tests/test_jansen_rit_cpp.py
@@ -1,0 +1,409 @@
+"""Tests for C++ Jansen-Rit model implementations.
+
+This module tests the Python wrappers for C++ compiled Jansen-Rit models.
+Tests are automatically skipped if C++ modules are not compiled.
+"""
+
+import unittest
+import numpy as np
+import networkx as nx
+import pytest
+import os
+from copy import deepcopy
+
+# Check if C++ modules are available
+CPP_JR_AVAILABLE = True
+CPP_IMPORT_ERROR = None
+try:
+    from vbi.models.cpp.jansen_rit import JR_sde, JR_sdde
+except ImportError as e:
+    CPP_JR_AVAILABLE = False
+    CPP_IMPORT_ERROR = str(e)
+
+# Skip decorator for when C++ is not available
+skip_if_cpp_unavailable = pytest.mark.skipif(
+    not CPP_JR_AVAILABLE,
+    reason=f"C++ modules not compiled or not available: {CPP_IMPORT_ERROR}"
+)
+
+
+class TestJRSDE(unittest.TestCase):
+    """Test cases for JR_sde C++ implementation."""
+    
+    def setUp(self):
+        """Set up test fixtures."""
+        np.random.seed(42)
+        self.nn = 4
+        self.weights = nx.to_numpy_array(nx.complete_graph(self.nn))
+        
+        self.base_params = {
+            "weights": self.weights,
+            "G": 0.5,
+            "A": 3.25,
+            "B": 22.0,
+            "a": 0.1,
+            "b": 0.05,
+            "dt": 0.05,
+            "t_end": 100.0,
+            "t_transition": 50.0,
+            "seed": 42,
+        }
+    
+    @skip_if_cpp_unavailable
+    def test_initialization(self):
+        """Test that JR_sde can be initialized with default parameters."""
+        model = JR_sde(self.base_params)
+        
+        self.assertEqual(model.N, self.nn)
+        self.assertEqual(model.num_nodes, self.nn)
+        self.assertEqual(model.G, 0.5)
+        self.assertEqual(model.A, 3.25)
+        self.assertIsNotNone(model.weights)
+    
+    @skip_if_cpp_unavailable
+    def test_valid_params_auto_generation(self):
+        """Test that valid_params are automatically generated from get_default_parameters."""
+        model = JR_sde(self.base_params)
+        
+        # Check that valid_params is a list
+        self.assertIsInstance(model.valid_params, list)
+        
+        # Check that it contains expected parameters
+        expected_params = ['A', 'B', 'G', 'weights', 'dt', 't_end', 'seed']
+        for param in expected_params:
+            self.assertIn(param, model.valid_params)
+        
+        # Check that the number of valid params matches defaults
+        defaults = model.get_default_parameters()
+        self.assertEqual(len(model.valid_params), len(defaults))
+    
+    @skip_if_cpp_unavailable
+    def test_invalid_parameter_rejection(self):
+        """Test that invalid parameters are properly rejected."""
+        params = self.base_params.copy()
+        params['invalid_param'] = 123
+        
+        with self.assertRaises(ValueError) as context:
+            JR_sde(params)
+        
+        self.assertIn('Invalid parameter', str(context.exception))
+    
+    @skip_if_cpp_unavailable
+    def test_get_default_parameters(self):
+        """Test that get_default_parameters returns a complete dict."""
+        model = JR_sde(self.base_params)
+        defaults = model.get_default_parameters()
+        
+        # Check it's a dictionary
+        self.assertIsInstance(defaults, dict)
+        
+        # Check key parameters are present
+        key_params = ['G', 'A', 'B', 'a', 'b', 'dt', 't_end', 'weights', 'seed']
[…]
