### 2026-02-12 83ca108
Fix indentation in set_initial_state function and add unit tests for VEP_sde model
