+++ vbi/tests/test_vep_cpp.py
@@ -1,0 +1,343 @@
+"""Tests for C++ VEP (Virtual Epileptic Patient) model implementation.
+
+This module tests the Python wrapper for C++ compiled VEP model.
+Tests are automatically skipped if C++ modules are not compiled.
+"""
+
+import unittest
+import numpy as np
+import pytest
+import os
+from copy import deepcopy
+
+# Check if C++ modules are available
+CPP_VEP_AVAILABLE = True
+CPP_IMPORT_ERROR = None
+try:
+    from vbi.models.cpp.vep import VEP_sde
+except ImportError as e:
+    CPP_VEP_AVAILABLE = False
+    CPP_IMPORT_ERROR = str(e)
+
+# Skip decorator for when C++ is not available
+skip_if_cpp_unavailable = pytest.mark.skipif(
+    not CPP_VEP_AVAILABLE,
+    reason=f"C++ VEP module not compiled or not available: {CPP_IMPORT_ERROR}"
+)
+
+
+class TestVEPSDE(unittest.TestCase):
+    """Test suite for VEP_sde C++ implementation."""
+
+    def setUp(self):
+        """Set up test fixtures."""
+        self.nn = 4
+        self.weights = np.random.rand(self.nn, self.nn)
+        
+        self.base_params = {
+            'weights': self.weights,
+            'G': 1.0,
+            'tau': 10.0,
+            'eta': -1.5,
+            'noise_sigma': 0.1,
+            'iext': 0.0,
+            'dt': 0.01,
+            'tend': 50.0,
+            'tcut': 0.0,
+            'noise_seed': 0,
+            'record_step': 1,
+            'method': 'euler',
+            'output': 'test_output_vep'
+        }
+
+    def tearDown(self):
+        """Clean up after tests."""
+        import shutil
+        if os.path.exists('test_output_vep'):
+            shutil.rmtree('test_output_vep')
+
+    @skip_if_cpp_unavailable
+    def test_initialization(self):
+        """Test that VEP_sde initializes correctly with valid parameters."""
+        model = VEP_sde(self.base_params)
+        
+        # Check that parameters are set
+        self.assertEqual(model.G, 1.0)
+        self.assertEqual(model.tau, 10.0)
+        self.assertEqual(model.eta, -1.5)
+        self.assertEqual(model.noise_sigma, 0.1)
+        self.assertIsNone(model.seed)
+        
+    @skip_if_cpp_unavailable
+    def test_get_default_parameters(self):
+        """Test that default parameters are returned correctly."""
+        model = VEP_sde()
+        defaults = model.get_default_parameters()
+        
+        # Check that all expected parameters are present
+        expected_keys = [
+            'G', 'seed', 'initial_state', 'weights', 'tau', 'eta',
+            'noise_sigma', 'iext', 'dt', 'tend', 'tcut', 'noise_seed',
+            'record_step', 'method', 'output'
+        ]
+        
+        for key in expected_keys:
+            self.assertIn(key, defaults)
+    
+    @skip_if_cpp_unavailable
+    def test_valid_params_auto_generation(self):
+        """Test that valid_params is automatically generated from get_default_parameters."""
+        model = VEP_sde(self.base_params)
+        defaults = model.get_default_parameters()
+        
+        # valid_params should match the keys from get_default_parameters
+        self.assertEqual(set(model.valid_params), set(defaults.keys()))
+        
+    @skip_if_cpp_unavailable
+    def test_invalid_parameter_rejection(self):
+        """Test that invalid parameters are rejected."""
+        params = self.base_params.copy()
+        params['invalid_param'] = 123
+        
+        with self.assertRaises(ValueError) as context:
+            VEP_sde(params)
+        
[…]
