### 2026-02-12 a66858b
Refactor C++ model classes to inherit from BaseModel and unify parameter management
- Updated JR_sde, KM_sde, MPR_sde, VEP_sde, and WC_ode classes to inherit from BaseModel.
- Implemented consistent parameter handling methods across models, including get_default_parameters, check_parameters, and parameter validation.
- Enhanced error messages for parameter import issues.
- Added unit tests for BaseModel interface compliance across all C++ models.
- Improved code readability by standardizing docstring formats and removing commented-out code.
- Ensured backward compatibility with existing parame
[…]
