+++ vbi/models/cpp/damp_oscillator.py
@@ -2,23 +2,24 @@ import os
 from typing import Any
 import numpy as np
 
+from vbi.models.cpp.base import BaseModel
+
 try:
     from vbi.models.cpp._src.do import DO as _DO
 except ImportError as e:
-    print(f"Could not import modules: {e}, probably C++ code is not compiled or properly linked.")
+    print(
+        f"Could not import modules: {e}, probably C++ code is not compiled or properly linked."
+    )
 
-class DO:
 
-    '''
+class DO(BaseModel):
+    """
     Damp Oscillator model class.
-    '''
-
-    valid_params = ["a", "b", "dt", "t_start", "t_end", "t_cut",
-                    "initial_state", "method", "output"]
+    """
 
     # ---------------------------------------------------------------
-    def __init__(self, par={}):
-        '''
+    def __init__(self, par: dict = {}):
+        """
         Parameters
         ----------
         par : dictionary
@@ -28,7 +29,9 @@ class DO:
             - **t_end** [double] final time for simulation.
             - **initial_state** [list] initial state of the system.
 
-        '''
+        """
+        super().__init__()
+        self.valid_params = list(self.get_default_parameters().keys())
         self.check_parameters(par)
         self._par = self.get_default_parameters()
         self._par.update(par)
@@ -41,7 +44,7 @@ class DO:
     def __str__(self) -> str:
         """
         Return a string representation of the model parameters.
-        
+
         Returns
         -------
         str
@@ -49,11 +52,11 @@ class DO:
         """
         print("Damped Oscillator Model (C++)")
         print("-----------------------------")
-        
+
         # Model parameters
         print(f"a = {self.a}")
         print(f"b = {self.b}")
-        
+
         # Simulation parameters
         print(f"dt = {self.dt}")
         print(f"t_start = {self.t_start}")
@@ -62,25 +65,17 @@ class DO:
         print(f"method = {self.method}")
         print(f"output = {self.output}")
         print(f"initial_state = {self.initial_state}")
-        
+
         return ""
 
     def __call__(self, *args: Any, **kwds: Any) -> Any:
         print("Damp Oscillator model")
         return self._par
 
-    def check_parameters(self, par):
-        '''
-        check if the parameters are valid.
-        '''
-        for key in par.keys():
-            if key not in self.valid_params:
-                raise ValueError("Invalid parameter: " + key)
-
     def get_default_parameters(self):
-        '''
+        """
         return default parameters for damp oscillator model.
-        '''
+        """
 
         params = {
             "a": 0.1,
@@ -99,7 +94,7 @@ class DO:
     def update_par(self, par={}):
         """
         Update model parameters.
-        
+
         Parameters
         ----------
         par : dict, optional
@@ -111,9 +106,9 @@ class DO:
                 setattr(self, key, par[key])
 
     def prepare_input(self):
-        '''
+        """
         prepare input for cpp model.
-        '''
+        """
         self.t_start = float(self.t_start)
         self.t_end = float(self.t_end)
         self.dt = float(self.dt)
@@ -130,7 +125,7 @@ class DO:
         self.initial_state = np.asarray(self.initial_state, dtype=np.float64)
 
     # ---------------------------------------------------------------
-    def run(self, par={}, x0=None, verbose=False):
+    def run(self, par: dict = {}, x0: np.ndarray = None, verbose: bool = False):
         """
         Run the damped oscillator simulation.
 
@@ -140,7 +135,7 @@ class DO:
             Dictionary of parameters to update for this simulation run.
             Any parameter from the class documentation can be updated.
         x0 : array-like, optional
-            Initial state vector [x0, y0] of length 2. If None, uses the 
[…]

[…]
