### 2026-02-12 2f5279d
Enhance C++ model initialization and error handling in tests; improve documentation for JR_sde model
