+++ vbi/models/cpp/__init__.py
@@ -1,0 +1,10 @@
+"""
+VBI C++ Models Module
+
+This module provides C++ implementations of various neural mass models
+with a unified interface through the BaseModel class.
+"""
+
+from .base import BaseModel
+
+__all__ = ['BaseModel']
