### 2026-02-12 3a04b32
Remove explicit C++ compilation step from workflow; clarify installation notes for Python dependencies
