### 2026-02-12 06b8dee
Improve error handling for C++ module imports in Jansen-Rit and VEP tests; ensure proper availability checks
