### 2026-02-12 a477d2c
Update installation steps in workflow; clarify C++ compilation dependencies and remove outdated notes
