### 2026-02-12 17b0c73
Add parameter description methods to various C++ models; enhance documentation
