### 2026-02-12 7d16905
Remove get_parameter_descriptions method from VEP_sde class; streamline code
