### 2026-02-12 8a2b748
Enhance model documentation to support heterogeneous parameters
- Updated the documentation for the Damp Oscillator, Jansen-Rit, Kuramoto, Montbrio-Pazo-Roxin, Virtual Epileptic Patient, and Wilson-Cowan models to clarify that they support heterogeneous parameters across brain regions.
- Specified that parameters marked as "scalar|vector" can be provided as either single values or arrays, allowing for more flexible model configurations.
