+++ vbi/models/cpp/mpr.py
@@ -13,7 +13,12 @@ except ImportError as e:
 
 class MPR_sde(BaseModel):
     """
-    MPR model
+    Montbrio-Pazo-Roxin (MPR) model.
+    
+    This model supports heterogeneous parameters across brain regions.
+    Parameters marked as "scalar|vector" in the parameter descriptions can be 
+    specified as either single values (applied to all regions) or arrays 
+    (one value per region).
     """
 
     def __init__(self, par: dict = {}, parbold={}) -> None:
