### 2026-02-12 b632c25
Enhance API documentation for C++ models; add module details and improve formatting
