### 2026-02-12 a5dcc6e
Add test script for parameter table display functionality across C++ models
