### 2026-02-12 5bb3eaa
Refactor parameter names for backward compatibility in JR_sde, VEP_sde, and WC_ode classes; update tests accordingly
