### 2026-02-13 c1b6594
Refactor model parameter handling in Jansen-Rit, VEP, and Wilson-Cowan models
- Introduced default parameter dictionaries (JR_DEFAULTS, VEP_DEFAULTS, WC_DEFAULTS) for single source of truth.
- Updated constructors to use default values from the respective dictionaries.
- Added methods to retrieve default parameters and their descriptions for each model.
- Enhanced parameter validation and listing functionality.
- Improved string representation of model parameters.
- Refactored parameter checks to ensure consistency across models.
