+++ vbi/models/numba/vep.py
@@ -1,11 +1,13 @@
 import warnings
 import numpy as np
+from typing import Dict, Any
 from numba import njit, jit
 from numba.experimental import jitclass
 from numba.extending import register_jitable
 from numba import float64, int64, types
 from vbi.utils import print_valid_parameters
 from numba.core.errors import NumbaPerformanceWarning
+from vbi.models.numba.base import BaseNumbaModel
 
 warnings.simplefilter("ignore", category=NumbaPerformanceWarning)
 
@@ -50,6 +52,24 @@ vep_spec = [
     ("output", types.string),
 ]
 
+# Default parameter values - single source of truth
+VEP_DEFAULTS = {
+    'G': 1.0,
+    'dt': 0.01,
+    'tau': 10.0,
+    'eta': -1.5,
+    'iext': 0.0,
+    'weights': None,  # Required - must be provided by user
+    't_cut': 0.0,
+    't_end': 100.0,
+    'method': 'euler',
+    'seed': -1,
+    'initial_state': None,  # Auto-generated if not provided
+    'sigma': 0.1,
+    'record_step': 1,
+    'output': 'output',
+}
+
 
 @jitclass(vep_spec)
 class ParVEP:
@@ -60,53 +80,27 @@ class ParVEP:
     compiled by Numba for efficient access during integration. It automatically
     calculates the number of nodes (nn) from the weights matrix shape.
     
-    Parameters
-    ----------
-    G : float, optional
-        Global coupling strength (default: 1.0)
-    dt : float, optional  
-        Integration time step in seconds (default: 0.01)
-    tau : float, optional
-        Time constant for slow variable dynamics (default: 10.0)
-    eta : array_like, optional
-        Epileptogenicity per region (default: [-1.5])
-    iext : array_like, optional
-        External input per region (default: [0.0])
-    weights : array_like
-        Connectivity matrix (nn x nn)
-    t_cut : float, optional
-        Time to cut initial transient (default: 0.0)
-    t_end : float, optional
-        End time of simulation (default: 100.0)
-    method : str, optional
-        Integration method: "euler" or "heun" (default: "euler")
-    seed : int, optional
-        Random seed for noise generation (default: -1)
-    initial_state : array_like, optional
-        Initial state vector (default: [0.0, 0.0])
-    sigma : float, optional
-        Noise amplitude (default: 0.1)
-    record_step : int, optional
-        Recording decimation factor (default: 1)
-    output : str, optional
-        Output directory (default: "output")
+    Note: This is an internal class used by the VEP_sde class. Users should
+    not instantiate this class directly. Use VEP_sde instead.
+    
+    Default values are defined in VEP_DEFAULTS dictionary.
     """
     def __init__(
         self,
-        G=1.0,
-        dt=0.01,
-        tau=10.0,
-        eta=np.array([-1.5]),
-        iext=np.array([0.0]),
+        G=VEP_DEFAULTS['G'],
+        dt=VEP_DEFAULTS['dt'],
+        tau=VEP_DEFAULTS['tau'],
+        eta=np.array([VEP_DEFAULTS['eta']]),
+        iext=np.array([VEP_DEFAULTS['iext']]),
         weights=np.array([[], []]),
-        t_cut=0.0,
-        t_end=100.0,
-        method="euler",
-        seed=-1,
+        t_cut=VEP_DEFAULTS['t_cut'],
+        t_end=VEP_DEFAULTS['t_end'],
+        method=VEP_DEFAULTS['method'],
+        seed=VEP_DEFAULTS['seed'],
         initial_state=np.array([0.0, 0.0]),
-        sigma=0.1,
-        record_step=1,
[…]
