+++ vbi/models/numba/base.py
@@ -1,0 +1,374 @@
+"""
+Base class for VBI Numba models providing a unified interface for parameter management.
+
+This base class provides a consistent API across all Numba-based neural mass models,
+similar to the BaseModel class used for C++ models. However, it accounts for the
+specific architecture of Numba models which use jitclass for efficient parameter storage.
+"""
+from abc import ABC, abstractmethod
+from typing import Dict, List, Any
+import numpy as np
+
+
+class BaseNumbaModel(ABC):
+    """
+    Abstract base class for all VBI Numba models.
+    
+    This class provides a unified interface for model parameter management,
+    ensuring consistency across different Numba model implementations.
+    
+    Unlike the C++ BaseModel, Numba models typically use a jitclass (e.g., ParVEP, ParJR)
+    for storing parameters that will be used in JIT-compiled functions. This base class
+    provides a Python-level interface while the actual parameters may be stored in the
+    jitclass instance (typically self.P or self.par).
+    
+    Attributes
+    ----------
+    P : jitclass instance
+        Internal jitclass instance storing all model parameters for Numba compilation.
+    valid_params : list
+        List of valid parameter names for this model.
+    """
+    
+    def __init__(self):
+        """Initialize the base Numba model."""
+        self.P = None  # Will be set by subclass to jitclass instance
+        self.valid_params = []
+        self._checked = False
+    
+    @abstractmethod
+    def get_default_parameters(self) -> Dict[str, Any]:
+        """
+        Get the default parameters for the model.
+        
+        Returns
+        -------
+        dict
+            Dictionary containing default parameter values.
+            
+        Examples
+        --------
+        >>> model = JR_sde({"weights": np.eye(2)})
+        >>> defaults = model.get_default_parameters()
+        >>> print(defaults['G'])
+        1.0
+        """
+        pass
+    
+    @abstractmethod
+    def get_parameter_descriptions(self) -> Dict[str, str]:
+        """
+        Get descriptions for all model parameters.
+        
+        Returns
+        -------
+        dict
+            Dictionary mapping parameter names to their descriptions.
+            Can also return tuples of (description, type) where type is one of:
+            'scalar', 'vector', 'matrix', 'string', 'int', 'bool'.
+            
+        Examples
+        --------
+        >>> model = JR_sde({"weights": np.eye(2)})
+        >>> descriptions = model.get_parameter_descriptions()
+        >>> print(descriptions['G'])
+        ('Global coupling strength', 'scalar')
+        """
+        pass
+    
+    def check_parameters(self, par: Dict[str, Any]) -> None:
+        """
+        Validate that all provided parameters are valid for this model.
+        
+        Parameters
+        ----------
+        par : dict
+            Dictionary of parameters to validate.
+            
+        Raises
+        ------
+        ValueError
+            If any parameter name is not in the valid_params list.
+            
+        Examples
+        --------
+        >>> model = JR_sde({"weights": np.eye(2)})
+        >>> model.check_parameters({'G': 2.0})  # Valid
+        >>> model.check_parameters({'invalid_param': 1.0})  # Raises ValueError
+        """
+        invalid_params = []
+        for key in par.keys():
+            if key not in self.valid_params:
+                invalid_params.append(key)
+        
+        if invalid_params:
+            print(f"Invalid parameter(s): {', '.join(invalid_params)}")
+            self.print_valid_parameters()
+            raise ValueError(f"Invalid parameter(s): {', '.join(invalid_params)}")
+    
+    def get_parameters(self) -> Dict[str, Any]:
+        """
+        Get all model parameters as a dictionary.
+        
+        Returns
+        -------
+        dict
[…]

[…]
