### 2026-02-16 21fd4c7
Refactor Stuart-Landau model and unify parameter access methods
- Introduced a unified parameter access API in SL_sde class, including methods for retrieving parameters, default values, and descriptions.
- Added default parameters as a single source of truth using SL_DEFAULTS dictionary.
- Updated parameter validation to ensure consistency and clarity.
- Enhanced documentation for parameter descriptions and usage examples.
- Modified tests to cover new parameter access methods and ensure functionality remains intact.
- Refactored existing code to improve readability and maintainability.
