+++ vbi/models/numba/ghb.py
@@ -4,6 +4,7 @@ from numba import njit
 from numba.experimental import jitclass
 from numba.core.errors import NumbaPerformanceWarning
 from numba import float64, int64
+from vbi.models.numba.base import BaseNumbaModel
 
 warnings.simplefilter("ignore", category=NumbaPerformanceWarning)
 
@@ -38,7 +39,7 @@ def run(P, times):
     omega = P.omega
     tcut = P.tcut
     decimate = P.decimate
-    init_state = P.init_state
+    initial_state = P.initial_state
 
     epsilon = 0.5
     itaus = 1.25
@@ -63,7 +64,7 @@ def run(P, times):
     # act = np.zeros((nn, nt))
 
     # initial conditions (similar value for all regions)
-    x_init, y_init = init_state[:nn], init_state[nn:]
+    x_init, y_init = initial_state[:nn], initial_state[nn:]
     x[:] = x_init
     y[:] = y_init
 
@@ -125,7 +126,7 @@ def run(P, times):
     return t_bold, bold
 
 
-class GHB_sde(object):
+class GHB_sde(BaseNumbaModel):
     """
     Numba implementation of the Generic Hopf Bifurcation model with BOLD signal generation.
     
@@ -179,7 +180,7 @@ class GHB_sde(object):
         * - `decimate`
           - Decimation factor for output time series (every `decimate`-th point is saved).
           - 1
-        * - `init_state`
+        * - `initial_state`
           - Initial state vector of shape (2*nn,) containing [x₀, y₀]. If None, random initial conditions are generated.
           - np.array([])
         * - `seed`
@@ -227,9 +228,64 @@ class GHB_sde(object):
             available parameters. The 'weights', 'eta', and 'omega' parameters 
             are required for proper functionality.
         """
+        super().__init__()
         self.valid_par = [par_spec[i][0] for i in range(len(par_spec))]
+        # Set valid_params for BaseNumbaModel (excluding derived parameters if any)
+        self.valid_params = [k for k in GHB_DEFAULTS.keys()]
         self.check_parameters(par)
         self.P = self.get_par_obj(par)
+    
+    def get_default_parameters(self):
+        """
+        Get the default parameters for the Generic Hopf Bifurcation model.
+        
+        Returns a copy of the GHB_DEFAULTS dictionary.
+        
+        Returns
+        -------
+        dict
+            Dictionary containing default parameter values.
+            
+        Examples
+        --------
+        >>> from vbi.models.numba.ghb import GHB_sde
+        >>> import numpy as np
+        >>> model = GHB_sde({"weights": np.eye(2), "eta": np.array([0.1, 0.1]), "omega": np.array([40.0, 40.0])})
+        >>> defaults = model.get_default_parameters()
+        >>> print(defaults['G'])
+        1.0
+        """
+        return GHB_DEFAULTS.copy()
+    
+    def get_parameter_descriptions(self):
+        """
+        Get descriptions for all Generic Hopf Bifurcation model parameters.
+        
+        Returns
+        -------
+        dict
+            Dictionary mapping parameter names to tuples of (description, type).
+            
+        Examples
+        --------
+        >>> model = GHB_sde({"weights": np.eye(2), "eta": np.array([0.1, 0.1]), "omega": np.array([40.0, 40.0])})
+        >>> descriptions = model.get_parameter_descriptions()
+        >>> print(descriptions['G'])
+        ('Global coupling strength', 'scalar')
+        """
+        return {
+            'G': ('Global coupling strength scaling network connections', 'scalar'),
+            'dt': ('Integration time step (seconds)', 'scalar'),
+            'sigma': ('Noise amplitude (standard deviation)', 'scalar'),
[…]
