+++ vbi/models/numba/rww.py
@@ -7,7 +7,7 @@ from numba.extending import register_jitable
 # from vbi.models.numba.utils import set_seed_compat
 from vbi.models.numba.bold import ParBold, bold_spec, do_bold_step
 from vbi.models.numba.utils import check_vec_size_1d, check_parameters
-
+from vbi.models.numba.base import BaseNumbaModel
 # ----------------------------------------
 # Reduced Wong–Wang spec (excitatory only)
 # ----------------------------------------
@@ -36,32 +36,67 @@ rww_spec = [
     ("s_decimate", int64),
 ]
 
+# Default parameter values - single source of truth
+RWW_DEFAULTS = {
+    'a': 270.0 / 1000.0,
+    'b': 108.0 / 1000.0,
+    'd': 0.154 * 1000.0,
+    'tau_s': 100.0,  # ms
+    'gamma': 0.641,
+    'w': 1.0,
+    'J': 0.2609,
+    'G': 0.0,
+    'I_ext': 0.382,
+    'weights': None,  # Required - must be provided by user
+    'sigma': 0.05,
+    'dt': 1.0,
+    't_end': 1000.0,
+    't_cut': 0.0,
+    'tr': 300.0,
+    'seed': -1,
+    'initial_state': None,  # Optional - auto-generated if not provided
+    'RECORD_S': True,
+    'RECORD_BOLD': True,
+    's_decimate': 1,
+}
+
 
 @jitclass(rww_spec)
 class ParRWW:
+    """
+    Numba jitclass container for Reduced Wong-Wang model parameters.
+    
+    This class holds all parameters needed for the Reduced Wong-Wang model
+    in a format optimized for Numba compilation.
+    
+    Note: This is an internal class used by the RWW_sde class. Users should
+    not instantiate this class directly.
+    
+    Default values are defined in RWW_DEFAULTS dictionary.
+    """
     def __init__(
         self,
-        a=270.0/1000.0,
-        b=108.0/1000.0,
-        d=0.154 * 1000.0,
-        tau_s=100.0,  # ms
-        gamma=0.641,
-        w=1.0,
-        J=0.2609,
-        G=0.0,
-        I_ext=np.array([0.382]),
+        a=RWW_DEFAULTS['a'],
+        b=RWW_DEFAULTS['b'],
+        d=RWW_DEFAULTS['d'],
+        tau_s=RWW_DEFAULTS['tau_s'],
+        gamma=RWW_DEFAULTS['gamma'],
+        w=RWW_DEFAULTS['w'],
+        J=RWW_DEFAULTS['J'],
+        G=RWW_DEFAULTS['G'],
+        I_ext=np.array([RWW_DEFAULTS['I_ext']]),
         weights=np.array([[0.0]]),
-        sigma=0.05,
-        dt=1.0,
-        t_end=1000.0,
-        t_cut=0.0,
-        tr=300.0,
+        sigma=RWW_DEFAULTS['sigma'],
+        dt=RWW_DEFAULTS['dt'],
+        t_end=RWW_DEFAULTS['t_end'],
+        t_cut=RWW_DEFAULTS['t_cut'],
+        tr=RWW_DEFAULTS['tr'],
         nn=1,
-        seed=-1,
+        seed=RWW_DEFAULTS['seed'],
         initial_state=np.array([0.0]),
-        RECORD_S=True,
-        RECORD_BOLD=True,
-        s_decimate=1,
+        RECORD_S=RWW_DEFAULTS['RECORD_S'],
+        RECORD_BOLD=RWW_DEFAULTS['RECORD_BOLD'],
+        s_decimate=RWW_DEFAULTS['s_decimate'],
     ):
         self.a = a
         self.b = b
@@ -132,10 +167,20 @@ def heun_sde(S, t, P):
 # -----------------------------
[…]
