### 2026-02-16 18f0f0e
Add unit tests for Reduced Wong-Wang (RWW) model, covering parameter access, validation, and BOLD signal generation
