+++ vbi/tests/test_rww.py
@@ -1,0 +1,428 @@
+"""
+Unit tests for Reduced Wong-Wang (RWW) model.
+
+These tests verify:
+1. Basic integration runs without errors
+2. Output shapes and types are correct
+3. Numerical stability (no NaN/Inf values)
+4. Parameter validation
+5. BOLD signal generation
+6. Unified parameter access methods
+"""
+
+import unittest
+import numpy as np
+import pytest
+from copy import deepcopy
+from vbi.models.numba.rww import RWW_sde
+
+
+class TestRWWParameterAccess(unittest.TestCase):
+    """Test unified parameter access methods for RWW model."""
+
+    def setUp(self):
+        """Set up common test parameters."""
+        self.weights = np.array([
+            [0.0, 0.1],
+            [0.1, 0.0],
+        ])
+        self.custom_params = {
+            "weights": self.weights,
+            "G": 0.5,
+            "sigma": 0.02,
+            "dt": 1.0,
+            "t_end": 500.0,
+        }
+
+    def test_get_parameters_returns_dict(self):
+        """Test that get_parameters() returns a dictionary."""
+        model = RWW_sde(self.custom_params)
+        params = model.get_parameters()
+        
+        self.assertIsInstance(params, dict)
+        self.assertGreater(len(params), 0)
+
+    def test_get_parameters_includes_custom_values(self):
+        """Test that get_parameters() returns actual custom values."""
+        model = RWW_sde(self.custom_params)
+        params = model.get_parameters()
+        
+        self.assertAlmostEqual(params['G'], 0.5)
+        self.assertAlmostEqual(params['sigma'], 0.02)
+        self.assertAlmostEqual(params['dt'], 1.0)
+
+    def test_get_parameters_includes_defaults(self):
+        """Test that get_parameters() includes default values for non-custom params."""
+        model = RWW_sde(self.custom_params)
+        params = model.get_parameters()
+        
+        # These should have default values
+        self.assertIn('tau_s', params)
+        self.assertIn('gamma', params)
+        self.assertIn('J', params)
+        self.assertIn('w', params)
+
+    def test_get_parameter_single_value(self):
+        """Test that get_parameter(name) returns single parameter value."""
+        model = RWW_sde(self.custom_params)
+        
+        g_value = model.get_parameter('G')
+        self.assertAlmostEqual(g_value, 0.5)
+        
+        tau_value = model.get_parameter('tau_s')
+        self.assertAlmostEqual(tau_value, 100.0)
+
+    def test_get_parameter_derived_param(self):
+        """Test that get_parameter() can access derived parameters like 'nn'."""
+        model = RWW_sde(self.custom_params)
+        
+        nn_value = model.get_parameter('nn')
+        self.assertEqual(nn_value, 2)
+
+    def test_get_parameter_invalid_raises_error(self):
+        """Test that get_parameter() raises error for invalid parameter."""
+        model = RWW_sde(self.custom_params)
+        
+        with self.assertRaises(AttributeError) as context:
+            model.get_parameter('invalid_parameter_name')
+        
+        self.assertIn('not found', str(context.exception).lower())
+
+    def test_get_default_parameters_returns_dict(self):
+        """Test that get_default_parameters() returns dictionary."""
+        model = RWW_sde(self.custom_params)
+        defaults = model.get_default_parameters()
+        
+        self.assertIsInstance(defaults, dict)
+        self.assertGreater(len(defaults), 0)
+
+    def test_get_default_parameters_has_correct_values(self):
+        """Test that get_default_parameters() returns correct default values."""
+        model = RWW_sde(self.custom_params)
[…]
