### 2026-02-16 8b4b4a1
Refactor models to inherit from BaseCupyModel, adding unified parameter management and descriptions for GHB, JR, KM, MPR, WC, and WW models.
