### 2026-02-16 083fa4d
Add unit tests for unified parameter API in CuPy models
