+++ vbi/tests/test_cupy_unified_api.py
@@ -1,0 +1,460 @@
+"""
+Test unified parameter API for all CuPy models.
+
+This test suite verifies that all CuPy models inherit from BaseCupyModel
+and implement the unified parameter management API correctly.
+"""
+
+import unittest
+import numpy as np
+import pytest
+
+
+# Check if CuPy is available
+CUPY_AVAILABLE = True
+try:
+    import cupy
+    if cupy.cuda.is_available():
+        ENGINE = "gpu"
+    else:
+        ENGINE = "cpu"
+except ImportError:
+    CUPY_AVAILABLE = False
+    ENGINE = "cpu"
+
+
+# Import all CuPy models
+GHB_AVAILABLE = True
+try:
+    from vbi.models.cupy.ghb import GHB_sde
+    from vbi.models.cupy.base import BaseCupyModel
+except ImportError:
+    GHB_AVAILABLE = False
+
+MPR_AVAILABLE = True
+try:
+    from vbi.models.cupy.mpr import MPR_sde
+except ImportError:
+    MPR_AVAILABLE = False
+
+WC_AVAILABLE = True
+try:
+    from vbi.models.cupy.wilson_cowan import WC_sde
+except ImportError:
+    WC_AVAILABLE = False
+
+WW_AVAILABLE = True
+try:
+    from vbi.models.cupy.ww import WW_sde
+except ImportError:
+    WW_AVAILABLE = False
+
+JR_AVAILABLE = True
+try:
+    from vbi.models.cupy.jansen_rit import JR_sde
+except ImportError:
+    JR_AVAILABLE = False
+
+KM_AVAILABLE = True
+try:
+    from vbi.models.cupy.km import KM_sde
+except ImportError:
+    KM_AVAILABLE = False
+
+
+class TestCupyUnifiedAPI(unittest.TestCase):
+    """Test unified parameter API for CuPy models."""
+
+    @classmethod
+    def setUpClass(cls):
+        """Set up test fixtures."""
+        cls.seed = 42
+        np.random.seed(cls.seed)
+        
+        # Create test connectivity matrices
+        cls.weights_2x2 = np.random.rand(2, 2)
+        cls.weights_3x3 = np.random.rand(3, 3)
+        
+        # For Kuramoto model
+        cls.omega_3 = np.random.rand(3)
+
+    @pytest.mark.skipif(not GHB_AVAILABLE, reason="GHB model not available")
+    def test_ghb_inherits_from_base(self):
+        """Test that GHB_sde inherits from BaseCupyModel."""
+        par = {'weights': self.weights_2x2, 'seed': self.seed}
+        model = GHB_sde(par=par)
+        self.assertIsInstance(model, BaseCupyModel)
+
+    @pytest.mark.skipif(not GHB_AVAILABLE, reason="GHB model not available")
+    def test_ghb_get_parameters(self):
+        """Test GHB_sde.get_parameters() method."""
+        par = {'weights': self.weights_2x2, 'seed': self.seed}
+        model = GHB_sde(par=par)
+        
+        params = model.get_parameters()
+        self.assertIsInstance(params, dict)
+        self.assertGreater(len(params), 0)
+        self.assertIn('G', params)
+        self.assertIn('eta', params)
+        self.assertIn('omega', params)
+
+    @pytest.mark.skipif(not GHB_AVAILABLE, reason="GHB model not available")
+    def test_ghb_get_parameter(self):
+        """Test GHB_sde.get_parameter() method."""
+        par = {'weights': self.weights_2x2, 'G': 30.0, 'seed': self.seed}
+        model = GHB_sde(par=par)
+        
+        G = model.get_parameter('G')
+        self.assertEqual(G, 30.0)
+
[…]
