### 2026-02-16 814aeb7
Add automodule directives for base models in CuPy and Numba sections of API documentation
