### 2026-02-16 6831c30
Update documentation theme to Furo and adjust requirements
