### 2026-02-16 0bc5365
Add GitHub Actions workflow to auto-update codemeta.json and update version in codemeta.json
