+++ .github/workflows/update-codemeta.yml
@@ -1,0 +1,56 @@
+name: Update Codemeta
+
+on:
+  # Trigger when version tags are pushed
+  push:
+    tags:
+      - 'v*'
+  
+  # Also allow manual trigger
+  workflow_dispatch:
+
+jobs:
+  update-codemeta:
+    runs-on: ubuntu-latest
+    
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.10'
+      
+      - name: Install VBI package
+        run: |
+          pip install -e .
+      
+      - name: Update codemeta.json
+        run: python -m vbi.utils
+      
+      - name: Check if codemeta.json was modified
+        id: check_changes
+        run: |
+          if git diff --quiet codemeta.json; then
+            echo "changed=false" >> $GITHUB_OUTPUT
+          else
+            echo "changed=true" >> $GITHUB_OUTPUT
+          fi
+      
+      - name: Commit and push changes
+        if: steps.check_changes.outputs.changed == 'true'
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          git add codemeta.json
+          git commit -m "Auto-update codemeta.json for version $(python -c 'from vbi._version import __version__; print(__version__)')"
+          git push
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      
+      - name: No changes detected
+        if: steps.check_changes.outputs.changed == 'false'
+        run: echo "codemeta.json is already up to date"
