### 2026-02-16 91bc21a
Refactor CuPy model imports and add function to update codemeta.json with current version
