+++ docs/examples/parameter_tables_cpp.py
@@ -1,161 +1,71 @@
-#!/usr/bin/env python
-"""
-Test script demonstrating the new parameter table display functionality
-for all C++ models in vbi.models.cpp
+"""Print parameter tables for all C++ models.
 
-This script shows:
-1. Uniform table display across all models
-2. Dynamic parameter state (before/after prepare_input)
-3. Proper handling of scalar vs array parameters
+This module mirrors `parameter_tables_numba.py`: it imports selected C++
+model modules, discovers `BaseModel` subclasses, instantiates each model
+with minimal inputs, and prints the formatted parameter table.
 """
+
+from __future__ import annotations
+
+import importlib
+import inspect
 
 import numpy as np
-from vbi.models.cpp.vep import VEP_sde
-from vbi.models.cpp.jansen_rit import JR_sde, JR_sdde
-from vbi.models.cpp.mpr import MPR_sde
-from vbi.models.cpp.wc import WC_ode
-from vbi.models.cpp.km import KM_sde
-from vbi.models.cpp.damp_oscillator import DO
+
+from vbi.models.cpp.base import BaseModel
+
+MODULES = [
+    "vbi.models.cpp.vep",
+    "vbi.models.cpp.jansen_rit",
+    "vbi.models.cpp.mpr",
+    "vbi.models.cpp.wc",
+    "vbi.models.cpp.km",
+    "vbi.models.cpp.damp_oscillator",
+]
 
 
-def test_vep_dynamic_display():
-    """Test VEP model showing scalar to array conversion"""
-    print("\n" + "=" * 90)
-    print("TEST 1: VEP_sde - Dynamic Parameter Display")
-    print("=" * 90)
-    
-    weights = np.random.rand(5, 5)
-    model = VEP_sde({'weights': weights, 'eta': -1.8, 'iext': 0.5})
-    
-    print("\nBEFORE prepare_input() - parameters as initialized:")
-    print("-" * 90)
-    print(model)
-    
-    model.prepare_input()
-    
-    print("\n\nAFTER prepare_input() - scalars converted to arrays:")
-    print("-" * 90)
-    print(model)
-    
-    print("\n✓ Notice: eta and iext changed from scalars to shape (5,)")
+def _default_par(cls: type[BaseModel], nn: int = 2) -> dict:
+    if cls.__name__ == "DO":
+        return {}
+
+    par = {"weights": np.eye(nn)}
+    if cls.__name__ == "JR_sdde":
+        par["delays"] = np.ones((nn, nn))
+    if cls.__name__ == "KM_sde":
+        par["omega"] = np.ones(nn)
+    return par
 
 
-def test_jr_sde_array_params():
-    """Test JR_sde with mixed scalar and array parameters"""
-    print("\n" + "=" * 90)
-    print("TEST 2: JR_sde - Mixed Scalar and Array Parameters")
-    print("=" * 90)
-    
-    weights = np.random.rand(3, 3)
-    A_values = np.array([3.25, 3.5, 3.0])
-    
-    model = JR_sde({'weights': weights, 'A': A_values, 'G': 0.7})
-    
-    print("\nInitial state (A already an array, C parameters as scalars):")
-    print("-" * 90)
-    print(model)
-    
-    model.prepare_input()
-    
-    print("\n\nAfter prepare_input() (C parameters converted to arrays):")
-    print("-" * 90)
-    print(model)
-    
-    print("\n✓ Notice: C0, C1, C2, C3 converted from scalars to shape (3,)")
+def instantiate_and_print(module_name: str, nn: int = 2) -> None:
+    try:
+        mod = importlib.import_module(module_name)
+    except Exception as e:
+        print(f"-- Failed to import {module_name}: {e}")
+        return
+
+    classes = [
+        obj
[…]
