### 2026-02-17 016a7fc
Update section title for clarity and reorganize inference methods in documentation
