### 2026-02-17 3aa1cb5
Update GitHub Actions workflow to enhance codemeta.json auto-update logic and improve commit handling
