+++ .github/workflows/update-codemeta.yml
@@ -18,6 +18,8 @@ jobs:
         uses: actions/checkout@v4
         with:
           fetch-depth: 0
+          ref: ${{ github.event.repository.default_branch || 'main' }}
+          token: ${{ secrets.GITHUB_TOKEN }}
       
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -46,8 +48,9 @@ jobs:
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
           git add codemeta.json
-          git commit -m "Auto-update codemeta.json for version $(python -c 'from vbi._version import __version__; print(__version__)')"
-          git push
+          VERSION=$(python -c 'from vbi._version import __version__; print(__version__)')
+          git commit -m "Auto-update codemeta.json for version $VERSION"
+          git push origin HEAD:${{ github.event.repository.default_branch || 'main' }}
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       
