+++ .github/workflows/update-codemeta.yml
@@ -5,6 +5,8 @@ on:
   push:
     tags:
       - 'v*'
+    branches:
+      - 'main'
   
   # Also allow manual trigger
   workflow_dispatch:
