### 2026-03-11 7f08d33
docs: Add Colab setup and EBRAINS script for easier onboarding
Closes #43
