### 2026-03-16 2c3a1d1
Reorganize documentation structure and enhance test runner script for clarity
