### 2026-03-16 0a83b69
Enhance get_fcd function with new parameters for flexibility and add JAX-compatible version with comprehensive tests for backward compatibility and overlap functionality.
