+++ vbi/feature_extraction/features_utils.py
@@ -267,11 +267,12 @@ def get_fc(ts, masks=None, positive=False, fc_fucntion="corrcoef"):
 
 def get_fcd(
     ts,
-    TR=1,
-    win_len=30,
-    positive=False,
+    tr: float | int | None = None,
+    win_len: float | int = 30,
+    positive: bool = False,
     masks=None,
-    #!TODO: add overlap
+    overlap: float | None = None,
+    TR: float | None = None,  # Deprecated, kept for backward compatibility
 ):
     """
     Compute dynamic functional connectivity.
@@ -281,43 +282,139 @@ def get_fcd(
 
     ts: numpy.ndarray [n_regions, n_timepoints]
         Input signal
-    win_len: int
-        sliding window length in samples, default is 30
-    TR: int
-        repetition time. It refers to the amount of time that
-        passes between consecutive acquired brain volumes during
-        functional magnetic resonance imaging (fMRI) scans.
+    tr: int, float, optional
+        Repetition time in any time unit (milliseconds, seconds, etc.).
+        This parameter sets the unit for `win_len` and `overlap`.
+        Default is 1 (dimensionless, treating window parameters as sample counts).
+    win_len: int, float
+        Sliding window length in the same units as `tr`.
+        For example, if tr=1ms and win_len=30, the window is 30ms.
+        If tr=1 (default), win_len is in sample units.
+        Default is 30.
     positive: bool
-        if True, only positive values of FC are considered.
-        default is False
-    masks: dict
-        dictionary of masks to compute FCD on.
-        default is None, which means that FCD is computed on the full matrix.
-        see also `hbt.utility.make_mask` and `hbt.utility.get_masks`.
+        If True, only positive values of FC are considered (negative correlations set to 0).
+        Default is False.
+    masks: dict, optional
+        Dictionary of masks to compute FCD on.
+        Default is None, which means that FCD is computed on the full matrix.
+        See also `hbt.utility.make_mask` and `hbt.utility.get_masks`.
+    overlap: float, optional
+        Overlap between consecutive windows as a fraction (0.0 to 1.0).
+        - overlap=0.0: no overlap (stride = win_len, non-overlapping windows)
+        - overlap=0.5: 50% overlap (stride = win_len/2)
+        - overlap=0.9: 90% overlap (stride = win_len/10, high overlap)
+        - overlap=1.0: maximum overlap (stride = 1 sample)
+        - If None (default): uses stride of 1 sample (equivalent to overlap=1.0, backward compatible)
+        
+        **Intuitive rule**: Higher overlap value = MORE overlap between windows
+        
+        Examples:
+            - overlap=0.0 with win_len=30 → stride=30 (no overlap)
+            - overlap=0.5 with win_len=30 → stride=15 (50% overlap)
+            - overlap=0.95 with win_len=30 → stride≈2 (95% overlap)
+    TR: int, optional
+        **Deprecated**. Use `tr` instead. Kept for backward compatibility.
 
     Returns
     -------
-        FCD: ndarray
-            matrix of functional connectivity dynamics
+        FCD: dict
+            Dictionary of functional connectivity dynamics matrices, one for each mask.
+            Each FCD is a square matrix of size (n_windows, n_windows) representing
+            correlations between windowed FC patterns over time.
+    
+    Notes
+    -----
+    All temporal parameters (`tr`, `win_len`) must use the same time unit:
[…]
