### 2026-03-16 0103b2d
Add JAX implementation for features_utils to API documentation
