### 2026-03-16 32f4875
Refactor docstring for get_fcd function to improve clarity and formatting
