### 2026-03-17 686f404
Remove test file for features_utils to streamline the test suite
