+++ vbi/tests/test_features_utils_ext.py
@@ -1,648 +1,0 @@
-﻿"""
-Tests for vbi.feature_extraction.features_utils
-
-Covers signal processing, spectral analysis, normalization,
-and statistical utility functions.
-
-Run with:
-    pytest vbi/tests/test_features_utils_ext.py -v
-"""
-
-import unittest
-import numpy as np
-import pytest
-
-from vbi.feature_extraction.features_utils import (
-    autocorr_norm,
-    create_symmetric_matrix,
-    lpc,
-    create_xx,
-    kde,
-    gaussian,
-    calc_ecdf,
-    compute_time,
-    calc_fft,
-    fundamental_frequency,
-    spectral_distance,
-    max_frequency,
-    max_psd,
-    median_frequency,
-    spectral_centroid,
-    spectral_spread,
-    km_order,
-    normalize_signal,
-    seizure_onset_indicator,
-    nat2bit,
-)
-
-
-# ---- helpers ----
-
-def _sine_signal(freq=10.0, fs=100.0, duration=1.0):
-    t = np.arange(0, duration, 1.0 / fs)
-    return np.sin(2.0 * np.pi * freq * t), fs
-
-
-def _multi_region_signal(n_regions=3, n_points=200):
-    rng = np.random.default_rng(42)
-    return rng.standard_normal((n_regions, n_points))
-
-
-# ---- nat2bit ----
-
-@pytest.mark.short
-@pytest.mark.fast
-class TestNat2Bit(unittest.TestCase):
-
-    def test_zero(self):
-        self.assertAlmostEqual(nat2bit(0), 0.0)
-
-    def test_one_nat(self):
-        result = nat2bit(1.0)
-        self.assertAlmostEqual(result, 1.4426950408889634)
-
-    def test_negative_value(self):
-        result = nat2bit(-2.0)
-        self.assertAlmostEqual(result, -2.0 * 1.4426950408889634)
-
-
-# ---- compute_time ----
-
-@pytest.mark.short
-@pytest.mark.fast
-class TestComputeTime(unittest.TestCase):
-
-    def test_length_matches_signal(self):
-        ts = np.zeros(100)
-        t = compute_time(ts, fs=1000)
-        self.assertEqual(len(t), 100)
-
-    def test_first_element_zero(self):
-        t = compute_time(np.zeros(50), fs=500)
-        self.assertAlmostEqual(t[0], 0.0)
-
-    def test_spacing_matches_fs(self):
-        t = compute_time(np.zeros(10), fs=100)
-        dt = t[1] - t[0]
-        self.assertAlmostEqual(dt, 1.0 / 100.0)
-
-    def test_last_element(self):
-        fs = 200
-        n = 100
-        t = compute_time(np.zeros(n), fs=fs)
-        self.assertAlmostEqual(t[-1], (n - 1) / fs)
-
-
-# ---- create_xx ----
-
-@pytest.mark.short
-@pytest.mark.fast
-class TestCreateXx(unittest.TestCase):
-
-    def test_length_matches_input(self):
-        data = np.array([1, 2, 3, 4, 5])
-        xx = create_xx(data)
-        self.assertEqual(len(xx), len(data))
-
-    def test_range_covers_data(self):
-        data = np.array([-3.0, 0.0, 5.0, 2.0])
-        xx = create_xx(data)
-        self.assertLessEqual(xx[0], min(data))
-        self.assertGreaterEqual(xx[-1], max(data))
-
-    def test_constant_signal(self):
-        data = np.ones(10) * 3.0
-        xx = create_xx(data)
-        self.assertEqual(len(xx), 10)
-
-
-# ---- calc_ecdf ----
-
-@pytest.mark.short
-@pytest.mark.fast
[…]
