### 2026-04-10 #9790 @ujjwalkarn
@yasahi-hpc @abhishektiwari @sampottinger @umawfb-create would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
