### 2026-04-20 #9790 @ujjwalkarn
@venkat1617 @ymahlau @sirpalp @mikesha2 @tradertanmay would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
