### 2026-04-20 #9790 @ymahlau
@ujjwalkarn I would be happy to review :)

A quick question, I saw that there is already a peer-reviewed and published paper on this software, which is linked in the repository. Is that fine by JOSS-guidelines and the copyright of the other paper?
