### 2026-04-26 #9790 @ujjwalkarn
@Ziaeemehr I see a [peer reviewed VBI paper](https://elifesciences.org/articles/106194) linked on the [repository](https://github.com/ins-amu/vbi), wanted to understand if the VBI software was reviewed as a part of this reviewed preprint?
