+++ vbi/models/jax/__init__.py
@@ -8,21 +8,21 @@ Requirements:
     pip install vbi[jax]
 """
 
-try:
-    from . import neural_mass
+# try:
+#     from . import neural_mass
     
-    # Also import and re-export models from vbjax
-    try:
-        import vbjax
-        # Re-export vbjax neural_mass models
-        from vbjax import neural_mass as vbjax_neural_mass
-    except ImportError:
-        vbjax_neural_mass = None
+#     # Also import and re-export models from vbjax
+#     try:
+#         import vbjax
+#         # Re-export vbjax neural_mass models
+#         from vbjax import neural_mass as vbjax_neural_mass
+#     except ImportError:
+#         vbjax_neural_mass = None
     
-    __all__ = ['neural_mass', 'vbjax_neural_mass']
+#     __all__ = ['neural_mass', 'vbjax_neural_mass']
     
-except ImportError as e:
-    raise ImportError(
-        "JAX models require JAX to be installed. "
-        "Install with: pip install vbi[jax]"
-    ) from e
+# except ImportError as e:
+#     raise ImportError(
+#         "JAX models require JAX to be installed. "
+#         "Install with: pip install vbi[jax]"
+#     ) from e
