### 2026-04-30 #9790 @mhashemi0873
Hi @ujjwalkarn
Thanks for your message.
No, the software itself was not independently reviewed in that publication.
The paper focused on the concept, methodology, outcomes, and detailed use cases of VBI across whole-brain models and modalities, rather than a formal software review or code audit.
Best regards,
Meysam
