### 2026-05-12 1a19432
Refactor code structure for improved readability and maintainability
