+++ notes/milestones.md
@@ -1,0 +1,1740 @@
+# Detailed Milestone Plan — `vbi/simulator/`
+
+> **Scope note:** Unlike TVB hybrid there is no subnetwork mixing here. Every
+> simulation runs one neural-mass model type across N homogeneous (or
+> node-heterogeneous) nodes. The TVB hybrid ideas that *do* carry over are:
+> expression-string dfuns, a separate coupling class, a ring-buffer history,
+> pluggable integrators, and monitor objects.
+
+---
+
+## Primary Use Case: Simulation-Based Inference (SBI)
+
+The simulator is a **generative model for SBI**. The dominant workload is not
+single runs but **parameter sweeps** that produce training data:
+
+1. Sample thousands of parameter sets `θ` (G, eta, J, noise_amp, …).
+2. For each `θ`, simulate a brain network and extract summary statistics
+   (FC matrix, FCD, power spectrum).
+3. Feed `(θ, features)` pairs into an inference engine (SNPE, SNLE, ABC, …).
+
+### Consequences for design
+
+| Consequence | Design response |
+|---|---|
+| Single simulation is never the bottleneck; **the sweep is** | `Sweeper` is a first-class public class, not an afterthought |
+| Storing raw time series for 10 000 simulations is prohibitive | Feature monitors compute statistics **inline** inside the sim loop; raw time series is optional |
+| CPU and GPU parallelism must be automatic | Each backend implements its own batch strategy (prange / vmap / CUDA grid) |
+| JAX backend enables gradient-based inference (SNPE with reparameterization) | `jax.vmap` + `jax.grad` through the full simulation loop |
+| Feature extraction already exists in `vbi/feature_extraction/` | Feature monitors **reuse** those functions where possible; provide Numba/JAX/CUDA variants where needed |
+| Parameter sets can be huge grids or Latin hypercube samples | `SweepSpec` accepts arbitrary arrays of parameter values, not just grids |
+
+---
+
+## API Contract — enforced across every milestone
+
+**The goal is to minimize user learning time.** A user should be able to go
+from `import` to a running simulation in ≤ 10 lines. Every design decision
+should be measured against this.
+
+### Canonical usage — single run (exploration, debugging)
+
+```python
+from vbi.simulator import Simulator
+from vbi.simulator.models.mpr import mpr
+from vbi.simulator.spec import SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec
+
+spec = SimulationSpec(
+    model        = mpr,
+    integrator   = IntegratorSpec(method="heun", dt=0.01),
+    coupling     = CouplingSpec(kind="linear", a=1.0),
+    monitors     = (MonitorSpec(kind="subsample", period=1.0),),
+    weights      = W,          # (N, N) connectivity matrix
+    tract_lengths= D,          # (N, N) delay distances in mm
+)
+
+sim = Simulator(spec, backend="numba")   # swap to "numpy", "cpp", "jax", "cuda"
+result = sim.run(duration=5000.0)        # ms
+t, state = result["subsample"]
+```
+
+### Canonical usage — parameter sweep (primary SBI use case)
+
+```python
+from vbi.simulator import Sweeper
+from vbi.simulator.spec import SweepSpec
+
+sweep_spec = SweepSpec(
+    params = {
+        "G":   np.linspace(0.5, 5.0, 50),
+        "eta": np.linspace(-5.5, -3.0, 50),
+    },
[…]
