+++ notes/implementation_plan.md
@@ -1,0 +1,355 @@
+# Multi-Backend Simulator — Implementation Plan
+
+## Vision
+
+Build a spec-driven, multi-backend neural mass simulator alongside the existing
+`vbi/models/` code. The model equations are written **once** in a Python spec
+(later also YAML); a code generator emits backend-specific implementations for
+NumPy, Numba (CPU), Numba-CUDA (GPU), JAX, and C++. The public API stays
+identical across backends.
+
+---
+
+## Migration Strategy: Keeping Existing Users Safe
+
+### Guiding principle
+The existing `vbi/models/{numba,cpp,cupy,pytorch,tvbk}/` tree is **not
+touched**. New work lives exclusively under `vbi/simulator/`. Both trees are
+importable simultaneously.
+
+### Phase 0 — Soft-deprecation notice (now, one release)
+- Add `DeprecationWarning` to `vbi/models/pytorch/__init__.py` and
+  `vbi/models/cupy/__init__.py` (heaviest maintenance burden).
+- Do the same for `vbi/models/tvbk/` (hard TVB dependency).
+- CHANGELOG entry: these sub-packages are deprecated, kept through vX.Y, then
+  removed in vX.(Y+1).
+- `vbi/models/numba/` and `vbi/models/cpp/` stay **stable** until the new
+  backends reach feature parity (tracked by a milestone below).
+
+### Phase 1 — New backend lives alongside (this plan)
+All new code goes into `vbi/simulator/`. Existing tests still pass; no breakage
+for users of `vbi/models/numba/` or `vbi/models/cpp/`.
+
+### Phase 2 — Parity declared
+When the new Numba and C++ backends reproduce all existing models, add a
+migration guide and bump the minor version.
+
+### Phase 3 — Hard removal
+Remove `pytorch`, `cupy`, `tvbk` sub-packages. Keep `numba` and `cpp` stubs
+that import from `vbi/simulator/` for one more release, then remove.
+
+---
+
+## Development Order
+
+```
+Python (NumPy) → Numba CPU → C++ → Numba-CUDA → JAX
+```
+
+**Why this order:**
+- NumPy baseline is easiest to validate (compare vs TVB reference).
+- Numba shares Python code structure; the dfun can be the same string, just
+  `@njit`-compiled. Catches spec design issues early.
+- C++ is most critical for large-scale sweeps; doing it third means the spec is
+  already stable.
+- Numba-CUDA and JAX are high-value but more specialist; added last so they
+  don't block neuroimaging users.
+
+---
+
+## Architecture
+
+### Inspiration from `tvb-root-hybrid-cpp`
+The reference project (`tvb/simulator/backend_cpp/`) separates concerns cleanly:
+- **Spec** (`spec.py`) — frozen dataclasses describing what to simulate, not
+  how. `SimulationSpec`, `SubnetworkSpec`, `IntegratorSpec`, `ProjectionSpec`,
+  `MonitorSpec`. Hashed for build-cache keying.
+- **Backend** — reads the spec, generates/compiles code, returns callable.
+- **Hybrid** (`hybrid/`) — Python-level network/coupling/monitor logic that
+  wraps the compiled inner loop.
+
+This project follows the same pattern but owns all layers (no TVB simulator
+dependency at runtime).
+
+---
+
+## Directory Layout
+
+```
+vbi/
+├── models/             # EXISTING — do not change
+│   ├── numba/          # stable
+│   ├── cpp/            # stable
[…]
