### 2026-05-13 7bfd7b2
Add NumPy backend for simulator and sweeper
- Implemented NumpySimulator class for running simulations using NumPy.
- Created NumpySweeper class for parameter sweeps with support for pipelines.
- Defined model specifications for MontbrioPopulationRate (MPR) including state variables and parameters.
- Established simulation specifications including integrator, coupling, and monitor configurations.
- Added support for various monitor types (raw, subsample, tavg, gavg, bold) to capture simulation data.
- Developed tests for validating the functionality of the NumPy backend, including deterministi
[…]
