+++ vbi/simulator/backend/numpy_/simulator.py
@@ -1,0 +1,171 @@
+from __future__ import annotations
+from typing import Callable
+import numpy as np
+
+from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.spec.model import ModelSpec
+from .history import History
+from .coupling import build_coupling
+from .integrators import build_integrator
+from .monitors import build_monitor
+
+
+# ---------------------------------------------------------------------------
+# dfun builder
+# ---------------------------------------------------------------------------
+
+def build_dfun(spec: ModelSpec) -> Callable:
+    """
+    Compile dfun_str expressions into a vectorized NumPy function.
+
+    Generated signature:
+        fn(state, coupling, params) -> np.ndarray  shape (n_sv, n_nodes)
+
+    where:
+        state    : (n_sv, n_nodes)
+        coupling : (n_cvar, n_nodes)  — one row per coupling variable
+        params   : dict[str, scalar | (n_nodes,)]
+
+    For each cvar name, injects  c_{name}  as a local (n_nodes,) variable.
+    Also injects  c = coupling[0]  for single-cvar backward compatibility.
+
+    Uses exec() on our own spec strings — not user-supplied input.
+    Compiled once at build() time, not per step.
+    """
+    sv = spec.sv_names
+    param_names = spec.param_names
+    cvar_names = spec.cvar
+
+    lines = [
+        "import numpy as _np",
+        "from numpy import pi, exp, log, sin, cos, tanh, sqrt",
+        "def _dfun(state, coupling, params):",
+    ]
+    # Unpack state variables
+    for i, name in enumerate(sv):
+        lines.append(f"    {name} = state[{i}]")
+    # Unpack params
+    for name in param_names:
+        lines.append(f"    {name} = params['{name}']")
+    # Inject coupling by cvar name: c_r, c_V, etc.
+    for i, cname in enumerate(cvar_names):
+        lines.append(f"    c_{cname} = coupling[{i}]")
+    # Also expose c = first coupling term for single-cvar models
+    lines.append("    c = coupling[0]")
+    # Compute derivatives
+    lines.append("    _out = _np.empty_like(state)")
+    for i, name in enumerate(sv):
+        lines.append(f"    _out[{i}] = {spec.dfun_str[name]}")
+    lines.append("    return _out")
+
+    src = "\n".join(lines)
+    globs: dict = {}
+    exec(compile(src, "<dfun>", "exec"), globs)
+    return globs["_dfun"]
+
+
+# ---------------------------------------------------------------------------
+# Parameter helpers
+# ---------------------------------------------------------------------------
+
+def _build_params(spec: SimulationSpec) -> dict:
+    params = dict(spec.model.default_params)
+    params.update(spec.node_params)
+    return params
+
+
+def _build_initial_state(spec: SimulationSpec) -> np.ndarray:
+    n_nodes = spec.n_nodes
+    state = np.zeros((spec.model.n_sv, n_nodes))
+    for i, sv in enumerate(spec.model.state_variables):
+        state[i] = sv.default_init
+    return state
+
+
+def _apply_bounds(state: np.ndarray, spec: SimulationSpec) -> None:
+    for i, sv in enumerate(spec.model.state_variables):
+        if sv.lower_bound is not None:
+            np.maximum(state[i], sv.lower_bound, out=state[i])
+        if sv.upper_bound is not None:
+            np.minimum(state[i], sv.upper_bound, out=state[i])
+
+
+# ---------------------------------------------------------------------------
+# Main simulator
+# ---------------------------------------------------------------------------
+
+class NumpySimulator:
+    """Pure-NumPy reference backend — used for correctness validation."""
+
+    def build(self, spec: SimulationSpec) -> None:
+        self.spec = spec
[…]

[…]
