### 2026-05-13 5dc0074
Update milestone documentation and enhance feature extraction imports
- Mark M0 milestone as complete and detail implemented components.
- Defer backend-specific feature extraction variants to respective milestones.
- Refactor imports in `__init__.py` and `utils.py` to handle exceptions more gracefully.
- Add validation tests for TVB reference in `test_mpr_numpy.py`.
