### 2026-05-13 3681151
Add test for ring buffer delay to match TVB convention
