### 2026-05-13 883b1bb
Add helper functions and demo script for VBI simulator and TVB MPR comparison
