+++ docs/examples/simulator_models/mpr_demo.py
@@ -1,0 +1,209 @@
+"""Compare VBI simulator and TVB MPR trajectories.
+
+The network and working-point parameters mirror
+``docs/examples/mpr_sde_numba.ipynb``:
+
+    nn = 6
+    weights = complete_graph(nn)
+    G = 0.33
+    I = 2.0
+    delay = 1.0 ms
+    dt = 0.01 ms
+    tau = 1.0
+    eta = -4.6
+
+The notebook also uses ``noise_amp`` for the Numba SDE model. This comparison is
+deterministic so that TVB and the new simulator can be compared point by point.
+"""
+
+from __future__ import annotations
+
+import argparse
+from pathlib import Path
+
+import numpy as np
+
+from helpers import (
+    comparison_metrics,
+    complete_graph_weights,
+    constant_tract_lengths,
+    ensure_repo_on_path,
+    homogeneous_node_params,
+    make_tvb_connectivity,
+    quiet_optional_imports,
+    quiet_tvb,
+    save_state_comparison_plot,
+)
+
+ensure_repo_on_path(__file__)
+
+# Importing the package root may probe optional dependencies; keep this example's
+# output focused on the simulator run.
+with quiet_optional_imports():
+    from vbi.simulator import Simulator
+    from vbi.simulator.models.mpr import mpr
+    from vbi.simulator.spec.coupling import CouplingSpec
+    from vbi.simulator.spec.integrator import IntegratorSpec
+    from vbi.simulator.spec.monitor import MonitorSpec
+    from vbi.simulator.spec.simulation import SimulationSpec
+
+
+NOTEBOOK_PARAMS = {
+    "nn": 6,
+    "G": 0.33,
+    "dt": 0.01,
+    "tau": 1.0,
+    "eta": -4.6,
+    "Delta": 0.7,
+    "J": 14.5,
+    "I": 2.0,
+    "Gamma": 0.0,
+    "cr": 1.0,
+    "cv": 0.0,
+    "rv_decimate": 10,
+    "tract_length": 4.0,
+    "speed": 4.0,
+}
+
+
+def build_vbi_spec(method: str) -> SimulationSpec:
+    weights = complete_graph_weights(NOTEBOOK_PARAMS["nn"])
+    tract_lengths = constant_tract_lengths(weights, NOTEBOOK_PARAMS["tract_length"])
+    node_param_names = ("tau", "I", "Delta", "J", "eta", "Gamma", "cr", "cv", "G")
+    node_params = homogeneous_node_params(
+        n_nodes=weights.shape[0],
+        params={name: NOTEBOOK_PARAMS[name] for name in node_param_names},
+    )
+
+    return SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method=method, dt=NOTEBOOK_PARAMS["dt"]),
+        coupling=CouplingSpec(kind="linear", a=1.0, b=0.0),
+        monitors=(MonitorSpec(kind="raw"),),
+        weights=weights,
+        tract_lengths=tract_lengths,
+        speed=NOTEBOOK_PARAMS["speed"],
+        node_params=node_params,
+    )
+
+
+def run_vbi(duration: float, method: str) -> tuple[np.ndarray, np.ndarray]:
+    return Simulator(build_vbi_spec(method), backend="numpy").run(duration)["raw"]
+
+
+def run_tvb(duration: float, method: str) -> np.ndarray:
+    try:
+        from tvb.simulator.coupling import Linear
+        from tvb.simulator.integrators import EulerDeterministic, HeunDeterministic
+        from tvb.simulator.models.infinite_theta import MontbrioPazoRoxin
+        from tvb.simulator.monitors import Raw
[…]
