### 2026-05-13 7160abf
Enhance documentation and output path in simulator model examples
