### 2026-05-13 98c5c10
Refactor docstring in mpr_demo.py for clarity and conciseness
