### 2026-05-13 312d33d
Implement additive white noise function and update integrators to use noise amplitude; add test for noise amplitude resolution based on style
