+++ vbi/simulator/spec/integrator.py
@@ -9,5 +9,9 @@ class IntegratorSpec:
     method: Literal["euler", "heun"] = "heun"
     dt: float = 0.01                        # ms
     stochastic: bool = False
-    noise_nsig: np.ndarray | None = None   # shape (n_noise_vars,), one per noise sv
+    # One value per noisy state variable.
+    # "amplitude": noise_nsig * sqrt(dt) * N(0, 1)
+    # "tvb": sqrt(2 * noise_nsig) * sqrt(dt) * N(0, 1)
+    noise_nsig: np.ndarray | None = None
+    noise_style: Literal["amplitude", "tvb"] = "amplitude"
     noise_seed: int = 42
