### 2026-05-13 6a72de3
Add JansenRit model specification and update init file
