+++ vbi/simulator/models/__init__.py
@@ -1,0 +1,2 @@
+from .mpr import mpr
+from .jansen_rit import jansen_rit
