### 2026-05-13 5b253e0
Add Jansen-Rit demo script and validation tests; enhance model specifications
