+++ vbi/simulator/models/__init__.py
@@ -1,2 +1,5 @@
 from .mpr import mpr
 from .jansen_rit import jansen_rit
+from .wilson_cowan import wilson_cowan
+from .reduced_wong_wang import reduced_wong_wang
+from .wong_wang_exc_inh import wong_wang_exc_inh
