+++ docs/examples/simulator_models/jansen_rit_demo.py
@@ -1,0 +1,228 @@
+"""Compare VBI simulator and TVB Jansen-Rit trajectories."""
+
+from __future__ import annotations
+
+import argparse
+from pathlib import Path
+import sys
+
+sys.dont_write_bytecode = True
+
+import numpy as np
+
+from helpers import (
+    comparison_metrics,
+    complete_graph_weights,
+    ensure_repo_on_path,
+    homogeneous_node_params,
+    make_tvb_connectivity,
+    quiet_optional_imports,
+    quiet_tvb,
+    save_state_comparison_plot,
+)
+
+ensure_repo_on_path(__file__)
+
+with quiet_optional_imports():
+    from vbi.simulator import Simulator
+    from vbi.simulator.models.jansen_rit import jansen_rit
+    from vbi.simulator.spec.coupling import CouplingSpec
+    from vbi.simulator.spec.integrator import IntegratorSpec
+    from vbi.simulator.spec.monitor import MonitorSpec
+    from vbi.simulator.spec.simulation import SimulationSpec
+
+
+JR_PARAMS = {
+    "nn": 6,
+    "coupling_strength": 0.01,
+    "dt": 0.05,
+    "A": 3.25,
+    "B": 22.0,
+    "a": 0.1,
+    "b": 0.05,
+    "v0": 5.52,
+    "nu_max": 0.0025,
+    "r": 0.56,
+    "J": 135.0,
+    "a_1": 1.0,
+    "a_2": 1.0,
+    "a_3": 0.25,
+    "a_4": 0.25,
+    "mu": 0.04, # 0.24
+    "decimate": 20,
+}
+
+
+def demo_weights(n_nodes: int) -> np.ndarray:
+    weights = complete_graph_weights(n_nodes)
+    row_sums = weights.sum(axis=1, keepdims=True)
+    row_sums[row_sums == 0.0] = 1.0
+    return weights / row_sums
+
+
+def build_vbi_spec(method: str) -> SimulationSpec:
+    weights = demo_weights(JR_PARAMS["nn"])
+    param_names = (
+        "A", "B", "a", "b", "v0", "nu_max", "r", "J",
+        "a_1", "a_2", "a_3", "a_4", "mu",
+    )
+    node_params = homogeneous_node_params(
+        n_nodes=weights.shape[0],
+        params={name: JR_PARAMS[name] for name in param_names},
+        scalar_names=(),
+    )
+
+    # To run a stochastic simulation, noise_variables=("y4",) is already
+    # declared in the model (y4 is the p(t) entry point from JR 1995).
+    # Just switch the integrator:
+    #
+    #   IntegratorSpec(
+    #       method=method,
+    #       dt=JR_PARAMS["dt"],
+    #       stochastic=True,
+    #       noise_nsig=np.array([0.001]),  # σ, one value per noise_variable
+    #       noise_style="amplitude",       # dW = σ * sqrt(dt) * N(0,1)
+    #       noise_seed=42,
+    #   )
+    return SimulationSpec(
+        model=jansen_rit,
+        integrator=IntegratorSpec(method=method, dt=JR_PARAMS["dt"]),
+        coupling=CouplingSpec(kind="linear", a=JR_PARAMS["coupling_strength"]),
+        monitors=(MonitorSpec(kind="raw"),),
+        weights=weights,
+        tract_lengths=np.zeros_like(weights),
+        speed=1.0,
+        node_params=node_params,
+    )
+
+
+def run_vbi(duration: float, method: str) -> tuple[np.ndarray, np.ndarray]:
+    return Simulator(build_vbi_spec(method), backend="numpy").run(duration)["raw"]
+
+
+def run_tvb(duration: float, method: str) -> np.ndarray:
+    try:
+        from tvb.simulator.coupling import Linear
[…]
