### 2026-05-15 0273f23
Add instant computation methods for coupling classes and update simulator logic for zero-delay cases
