### 2026-05-15 79291cd
Enhance comments in build_vbi_spec function to clarify stochastic simulation setup for VBI and TVB comparisons
