### 2026-05-15 35f9fa8
Add Wilson-Cowan demo script to compare VBI and TVB trajectories
