+++ docs/examples/simulator_models/wilson_cowan_demo.py
@@ -1,0 +1,227 @@
+"""Compare VBI simulator and TVB Wilson-Cowan trajectories."""
+
+from __future__ import annotations
+
+import argparse
+from pathlib import Path
+import sys
+
+sys.dont_write_bytecode = True
+
+import numpy as np
+
+from helpers import (
+    comparison_metrics,
+    complete_graph_weights,
+    ensure_repo_on_path,
+    homogeneous_node_params,
+    make_tvb_connectivity,
+    quiet_optional_imports,
+    quiet_tvb,
+    save_state_comparison_plot,
+)
+
+ensure_repo_on_path(__file__)
+
+with quiet_optional_imports():
+    from vbi.simulator import Simulator
+    from vbi.simulator.models.wilson_cowan import wilson_cowan
+    from vbi.simulator.spec.coupling import CouplingSpec
+    from vbi.simulator.spec.integrator import IntegratorSpec
+    from vbi.simulator.spec.monitor import MonitorSpec
+    from vbi.simulator.spec.simulation import SimulationSpec
+
+
+WC_PARAMS = {
+    "nn": 6,
+    "coupling_strength": 0.15,
+    "dt": 0.01,
+    "c_ee": 12.0,
+    "c_ei": 4.0,
+    "c_ie": 13.0,
+    "c_ii": 11.0,
+    "tau_e": 10.0,
+    "tau_i": 10.0,
+    "a_e": 1.2,
+    "b_e": 2.8,
+    "c_e": 1.0,
+    "theta_e": 0.0,
+    "a_i": 1.0,
+    "b_i": 4.0,
+    "c_i": 1.0,
+    "theta_i": 0.0,
+    "r_e": 1.0,
+    "r_i": 1.0,
+    "k_e": 1.0,
+    "k_i": 1.0,
+    "P": 0.5,
+    "Q": 0.0,
+    "alpha_e": 1.0,
+    "alpha_i": 1.0,
+    "decimate": 100,
+}
+
+
+def demo_weights(n_nodes: int) -> np.ndarray:
+    """Return row-normalized complete-graph weights for a compact demo."""
+    weights = complete_graph_weights(n_nodes)
+    row_sums = weights.sum(axis=1, keepdims=True)
+    row_sums[row_sums == 0.0] = 1.0
+    return weights / row_sums
+
+
+def build_vbi_spec(method: str) -> SimulationSpec:
+    weights = demo_weights(WC_PARAMS["nn"])
+    param_names = (
+        "c_ee", "c_ei", "c_ie", "c_ii", "tau_e", "tau_i",
+        "a_e", "b_e", "c_e", "theta_e", "a_i", "b_i", "c_i",
+        "theta_i", "r_e", "r_i", "k_e", "k_i", "P", "Q",
+        "alpha_e", "alpha_i",
+    )
+    node_params = homogeneous_node_params(
+        n_nodes=weights.shape[0],
+        params={name: WC_PARAMS[name] for name in param_names},
+        scalar_names=(),
+    )
+
+    return SimulationSpec(
+        model=wilson_cowan,
+        integrator=IntegratorSpec(method=method, dt=WC_PARAMS["dt"]),
+        coupling=CouplingSpec(kind="linear", a=WC_PARAMS["coupling_strength"]),
+        monitors=(MonitorSpec(kind="raw"),),
+        weights=weights,
+        tract_lengths=np.zeros_like(weights),
+        speed=1.0,
+        node_params=node_params,
+    )
+
+
+def run_vbi(duration: float, method: str) -> tuple[np.ndarray, np.ndarray]:
+    return Simulator(build_vbi_spec(method), backend="numpy").run(duration)["raw"]
+
+
[…]
