### 2026-05-15 6287369
Refactor Wilson-Cowan demo and model specifications; enhance validation tests for NumPy backend
