+++ docs/examples/simulator_models/wilson_cowan_demo.py
@@ -1,4 +1,4 @@
-"""Compare VBI simulator and TVB Wilson-Cowan trajectories."""
+"""Reproduce the first Wilson-Cowan SDE notebook visualization."""
 
 from __future__ import annotations
 
@@ -11,14 +11,9 @@ sys.dont_write_bytecode = True
 import numpy as np
 
 from helpers import (
-    comparison_metrics,
-    complete_graph_weights,
     ensure_repo_on_path,
     homogeneous_node_params,
-    make_tvb_connectivity,
     quiet_optional_imports,
-    quiet_tvb,
-    save_state_comparison_plot,
 )
 
 ensure_repo_on_path(__file__)
@@ -33,61 +28,71 @@ with quiet_optional_imports():
 
 
 WC_PARAMS = {
-    "nn": 6,
-    "coupling_strength": 0.15,
-    "dt": 0.01,
-    "c_ee": 12.0,
-    "c_ei": 4.0,
-    "c_ie": 13.0,
-    "c_ii": 11.0,
-    "tau_e": 10.0,
-    "tau_i": 10.0,
-    "a_e": 1.2,
-    "b_e": 2.8,
+    "nn": 2,
+    "coupling_strength": 0.0,
+    "dt": 0.05,
+    "t_cut": 101.0,
+    "noise_amp": 0.0005,
+    "noise_seed": 42,
+    "c_ee": 16.0,
+    "c_ei": 12.0,
+    "c_ie": 15.0,
+    "c_ii": 3.0,
+    "tau_e": 8.0,
+    "tau_i": 8.0,
+    "a_e": 1.3,
+    "b_e": 4.0,
     "c_e": 1.0,
     "theta_e": 0.0,
-    "a_i": 1.0,
-    "b_i": 4.0,
+    "a_i": 2.0,
+    "b_i": 3.7,
     "c_i": 1.0,
     "theta_i": 0.0,
     "r_e": 1.0,
     "r_i": 1.0,
-    "k_e": 1.0,
-    "k_i": 1.0,
-    "P": 0.5,
+    "k_e": 0.994,
+    "k_i": 0.999,
+    "P": 1.025,
     "Q": 0.0,
     "alpha_e": 1.0,
     "alpha_i": 1.0,
-    "decimate": 100,
+    "shift_sigmoid": 0.0,
+    "decimate": 1,
 }
 
 
-def demo_weights(n_nodes: int) -> np.ndarray:
-    """Return row-normalized complete-graph weights for a compact demo."""
-    weights = complete_graph_weights(n_nodes)
-    row_sums = weights.sum(axis=1, keepdims=True)
-    row_sums[row_sums == 0.0] = 1.0
-    return weights / row_sums
-
-
-def build_vbi_spec(method: str) -> SimulationSpec:
-    weights = demo_weights(WC_PARAMS["nn"])
+def build_vbi_spec(
+    method: str,
+    stochastic: bool = True,
+    param_overrides: dict[str, float] | None = None,
+) -> SimulationSpec:
+    weights = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.float64)
+    params = dict(WC_PARAMS)
+    if param_overrides:
+        params.update(param_overrides)
     param_names = (
         "c_ee", "c_ei", "c_ie", "c_ii", "tau_e", "tau_i",
         "a_e", "b_e", "c_e", "theta_e", "a_i", "b_i", "c_i",
         "theta_i", "r_e", "r_i", "k_e", "k_i", "P", "Q",
-        "alpha_e", "alpha_i",
+        "alpha_e", "alpha_i", "shift_sigmoid",
     )
     node_params = homogeneous_node_params(
         n_nodes=weights.shape[0],
[…]

[…]
