### 2026-05-15 6132e71
Refactor SweepSpec to use a generic object for pipeline; update tests to utilize MeanStdPipeline
