### 2026-05-15 c5e9be8
Fix inhibitory population gain parameter in WW_DEFAULTS and class ParWW; update related references
