### 2026-05-15 13e10fb
Implement Numba CPU backend for simulation and parameter sweeping
- Added code generation for model-specific Numba functions in `codegen.py`.
- Implemented the `NumbaSimulator` class for single-run simulations in `simulator.py`.
- Created `NumbaSweeperCPU` for parallel parameter sweeps in `sweeper.py`.
- Developed validation tests for the Numba backend in `test_mpr_numba.py`, ensuring results match the NumPy backend to specified tolerances.
- Included support for deterministic and stochastic simulations, as well as handling of delays and monitor outputs.
