+++ vbi/simulator/backend/numba_/simulator.py
@@ -1,0 +1,195 @@
+"""
+NumbaSimulator — single-run Numba CPU backend.
+
+The @njit simulation loop returns a raw subsampled array.
+Python post-processing converts it to the same {kind: (t, data)} dict
+that the NumPy backend returns, so all existing tests and pipelines work
+unchanged.
+"""
+from __future__ import annotations
+
+import numpy as np
+
+from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.backend.numpy_.monitors import _resolve_voi, _bw_init, _bw_step, _bw_bold, _BW_DEFAULTS
+
+from .codegen import (
+    build_numba_dfun,
+    build_params,
+    build_srcbuf,
+    build_initial_state,
+    get_G_idx,
+    get_bounds,
+    get_noise_params,
+)
+from ._nb_sim import nb_simulate_det, nb_simulate_stoch
+
+
+def _count_records(n_steps: int, t_cut_step: int, record_period: int) -> int:
+    if n_steps <= t_cut_step:
+        return 0
+    return (n_steps - t_cut_step + record_period - 1) // record_period
+
+
+def _apply_monitor(kind, m_spec, raw_data, raw_times, model):
+    """
+    Convert raw subsampled array to a single monitor's (t, data).
+
+    raw_data  : (n_record, n_sv, n_nodes)
+    raw_times : (n_record,) — time of each recorded step in ms
+    """
+    voi = _resolve_voi(m_spec.variables, model)
+
+    if kind == "raw":
+        return raw_times.copy(), raw_data[:, voi, :]
+
+    elif kind == "subsample":
+        return raw_times.copy(), raw_data[:, voi, :]
+
+    elif kind == "tavg":
+        n_record, n_sv, n_nodes = raw_data.shape
+        dt_raw = raw_times[1] - raw_times[0] if len(raw_times) > 1 else 1.0
+        period = m_spec.period if m_spec.period else dt_raw
+        istep  = max(1, round(period / dt_raw))
+        n_win  = n_record // istep
+        if n_win == 0:
+            return np.array([]), np.empty((0, len(voi), n_nodes))
+        win = raw_data[:n_win * istep].reshape(n_win, istep, n_sv, n_nodes)
+        avg = win.mean(axis=1)[:, voi, :]
+        # midpoint timing matching NumPy TemporalAverageMonitor
+        t0  = raw_times[0]
+        t_avg = t0 + (np.arange(n_win) * istep + (istep - 1) * 0.5) * dt_raw
+        return t_avg, avg
+
+    elif kind == "gavg":
+        n_record, n_sv, n_nodes = raw_data.shape
+        dt_raw = raw_times[1] - raw_times[0] if len(raw_times) > 1 else 1.0
+        period = m_spec.period if m_spec.period else dt_raw
+        istep  = max(1, round(period / dt_raw))
+        data   = raw_data[::istep, voi, :].mean(axis=2, keepdims=True)
+        t      = raw_times[::istep]
+        return t, data
+
+    elif kind == "bold":
+        # BW integration applied post-hoc to the first VOI (raw per step needed)
+        n_record, n_sv, n_nodes = raw_data.shape
+        dt_raw = raw_times[1] - raw_times[0] if len(raw_times) > 1 else 1.0
+        tr      = m_spec.tr
+        tr_steps = max(1, round(tr / dt_raw))
+        bw_state = _bw_init(n_nodes)
+        times_bold, data_bold = [], []
+        for i in range(n_record):
+            neural = raw_data[i, voi[0] if len(voi) > 0 else 0]
+            bw_state = _bw_step(bw_state, neural, dt_raw, _BW_DEFAULTS)
[…]
