+++ vbi/simulator/backend/numba_/_nb_sim.py
@@ -1,0 +1,375 @@
+"""
+Fixed @njit utility functions for the Numba CPU backend.
+
+The model-specific dfun is passed as a first-class @njit function argument,
+following the pattern used in the TVB hybrid Numba backend.  Everything else
+(coupling, integrators, simulate loop, parallel sweep) lives here so Numba
+can load these functions from disk (cache=True is possible for the pieces
+that do not take a dynamic dfun argument).
+
+Ring-buffer layout  : srcbuf[cv, node, step % horizon]
+Delayed-read index  : (step - 1 - delay + horizon) % horizon
+This matches TVB DenseHistory semantics exactly (validated in test_mpr_numpy.py).
+
+Parameter layout    : params[param_idx, node]  float64  (n_params, n_nodes)
+All params are broadcast to n_nodes columns; scalar params have equal columns.
+"""
+from __future__ import annotations
+
+import numpy as np
+from numba import njit, prange
+
+
+# ---------------------------------------------------------------------------
+# Coupling — Linear (instant, no delays)
+# ---------------------------------------------------------------------------
+
+@njit(cache=True)
+def _coup_linear_instant(cvar_state, weights, G, a, b):
+    """
+    c[cv, tgt] = G * a * sum_src(w[tgt, src] * cvar_state[cv, src]) + b
+
+    cvar_state : (n_cvar, n_nodes)
+    weights    : (n_nodes, n_nodes)   weights[tgt, src]
+    Returns    : (n_cvar, n_nodes)
+    """
+    n_cvar = cvar_state.shape[0]
+    n_nodes = cvar_state.shape[1]
+    out = np.zeros((n_cvar, n_nodes))
+    for cv in range(n_cvar):
+        for tgt in range(n_nodes):
+            s = 0.0
+            for src in range(n_nodes):
+                s += weights[tgt, src] * cvar_state[cv, src]
+            out[cv, tgt] = G * a * s + b
+    return out
+
+
+# ---------------------------------------------------------------------------
+# Coupling — Linear (delayed, ring-buffer read)
+# ---------------------------------------------------------------------------
+
+@njit(cache=True)
+def _coup_linear_delayed(srcbuf, step, horizon, weights, delay_steps, G, a, b):
+    """
+    c[cv, tgt] = G * a * sum_src(w[tgt,src] * srcbuf[cv, src, delayed_idx]) + b
+
+    srcbuf      : (n_cvar, n_nodes, horizon)
+    delay_steps : (n_nodes, n_nodes) int32   delay_steps[src, tgt]
+    step        : current loop counter (read happens before write of this step)
+    Returns     : (n_cvar, n_nodes)
+    """
+    n_cvar = srcbuf.shape[0]
+    n_nodes = srcbuf.shape[1]
+    out = np.zeros((n_cvar, n_nodes))
+    t = step - 1   # last-written logical step
+    for cv in range(n_cvar):
+        for tgt in range(n_nodes):
+            s = 0.0
+            for src in range(n_nodes):
+                d = delay_steps[src, tgt]
+                idx = (t - d + horizon) % horizon
+                s += weights[tgt, src] * srcbuf[cv, src, idx]
+            out[cv, tgt] = G * a * s + b
+    return out
+
+
+# ---------------------------------------------------------------------------
+# Integrators  (dfun_fn passed as first-class @njit function)
+# ---------------------------------------------------------------------------
+
+@njit(cache=False)
+def _heun_det(state, coupling, params, dt, dfun_fn):
+    k1 = dfun_fn(state, coupling, params)
+    k2 = dfun_fn(state + dt * k1, coupling, params)
+    return state + 0.5 * dt * (k1 + k2)
+
+
[…]
