### 2026-05-15 3013c9b
Refactor simulation duration in tests and update pipeline parameters for consistency across Numba and NumPy backends
