+++ vbi/tests/validation/test_sweep_numpy.py
@@ -42,7 +42,7 @@ class TestSweepNoPipeline:
     def test_returns_list_of_dicts(self):
         spec = make_mpr_spec(n_nodes=4, monitors=(MonitorSpec("tavg", period=1.0),))
         sw_spec = SweepSpec(params={"G": np.array([1.0, 2.0, 3.0])})
-        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=200.0)
+        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=100.0)
         assert len(results) == 3
         assert all("tavg" in r for r in results)
 
@@ -52,14 +52,14 @@ class TestSweepNoPipeline:
             "G":   np.linspace(1.0, 3.0, 4),
             "eta": np.linspace(-5.0, -3.0, 3),
         })
-        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=200.0)
+        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=100.0)
         assert len(results) == 4 * 3   # outer product
 
     def test_arbitrary_samples(self):
         spec = make_mpr_spec(n_nodes=2, monitors=(MonitorSpec("tavg", period=1.0),))
         theta = np.array([[1.0, -4.0], [2.0, -5.0], [3.0, -3.5]])
         sw_spec = SweepSpec(params=theta, param_names=("G", "eta"))
-        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=200.0)
+        results = Sweeper(spec, sw_spec, backend="numpy").run(duration=100.0)
         assert len(results) == 3
 
 
@@ -76,7 +76,7 @@ class TestSweepWithPipeline:
         self.pipeline = MeanStdPipeline(
             features=["mean", "std"],
             signal="tavg",
-            t_cut=100.0,
+            t_cut=50.0,
         )
         self.sw_spec = SweepSpec(
             params={"G": np.array([1.0, 2.0, 3.0])},
@@ -86,7 +86,7 @@ class TestSweepWithPipeline:
     def test_returns_labels_and_values(self):
         labels, values = Sweeper(
             self.spec, self.sw_spec, backend="numpy"
-        ).run(duration=300.0)
+        ).run(duration=100.0)
         assert isinstance(labels, list)
         assert isinstance(values, np.ndarray)
         # columns: G + mean_0..mean_3 + std_0..std_3 = 1+4+4=9
@@ -96,7 +96,7 @@ class TestSweepWithPipeline:
     def test_param_values_recorded_correctly(self):
         labels, values = Sweeper(
             self.spec, self.sw_spec, backend="numpy"
-        ).run(duration=300.0)
+        ).run(duration=100.0)
         np.testing.assert_array_equal(
             values[:, 0], [1.0, 2.0, 3.0],
             err_msg="G column must match sweep values"
@@ -104,7 +104,7 @@ class TestSweepWithPipeline:
 
     def test_run_df_shape(self):
         pd = pytest.importorskip("pandas")
-        df = Sweeper(self.spec, self.sw_spec, backend="numpy").run_df(300.0)
+        df = Sweeper(self.spec, self.sw_spec, backend="numpy").run_df(100.0)
         assert df.shape == (3, 9)
         assert "G" in df.columns
 
@@ -122,7 +122,7 @@ class TestSweepConsistency:
             monitors=(MonitorSpec("tavg", period=1.0),),
         )
         pipeline = MeanStdPipeline(
-            features=["mean"], signal="tavg", t_cut=200.0
+            features=["mean"], signal="tavg", t_cut=50.0
         )
 
         # Sweep: G=2.0 only
@@ -130,7 +130,7 @@ class TestSweepConsistency:
             params={"G": np.array([2.0])},
[…]
