### 2026-05-15 526a914
Reduce simulation duration in tests to improve performance and consistency across Numba and NumPy backends
