### 2026-05-15 6106d0c
Update error message tolerance in deterministic test for Heun and Euler methods
