### 2026-05-18 7a5c6a6
Add FeaturePipeline for cfg-dict integration and validation tests
