+++ vbi/feature_extraction/__init__.py
@@ -1,1 +1,17 @@
 # from .tests.all_tests import tests
+
+from .pipeline import FeaturePipeline
+from .features_settings import (
+    get_features_by_domain,
+    get_features_by_given_names,
+    update_cfg,
+    Data_F,
+)
+
+__all__ = [
+    "FeaturePipeline",
+    "get_features_by_domain",
+    "get_features_by_given_names",
+    "update_cfg",
+    "Data_F",
+]
