+++ vbi/feature_extraction/pipeline.py
@@ -1,0 +1,120 @@
+from __future__ import annotations
+
+import numpy as np
+
+
+class FeaturePipeline:
+    """
+    Connects vbi.simulator monitor output to the existing cfg-dict feature
+    extraction machinery.
+
+    Per-feature parameters (window sizes, band ranges, etc.) belong in the
+    cfg dict — built with get_features_by_domain / get_features_by_given_names /
+    update_cfg as usual.  This class only adds the three sim-level concerns:
+    which monitor signal to consume, how much burn-in to discard, and (in
+    future milestones) which JIT backend to dispatch to.
+
+    Parameters
+    ----------
+    cfg : dict
+        Feature configuration from get_features_by_domain or similar.
+    signal : str
+        Monitor kind to consume: "tavg", "subsample", "raw", "bold".
+    t_cut : float
+        Burn-in to discard in ms before extracting features.
+        When used with NumbaSweeperCPU the time array already starts at
+        t_cut, so this has no additional effect (searchsorted returns 0).
+
+    Examples
+    --------
+    cfg = get_features_by_domain("connectivity")
+    cfg = get_features_by_given_names(cfg, ["calc_fc", "calc_fcd"])
+    cfg = update_cfg(cfg, "calc_fcd", {"window_size": 30})
+
+    pipeline = FeaturePipeline(cfg, signal="tavg", t_cut=500.0)
+
+    # Single run
+    result = Simulator(spec, backend="numpy").run(5000.0)
+    labels, values = pipeline.extract(result)
+    df = pipeline.extract_df(result)
+
+    # Sweep
+    sweep_spec = SweepSpec(params={"G": np.linspace(1, 4, 50)}, pipeline=pipeline)
+    df = Sweeper(spec, sweep_spec, backend="numba").run_df(5000.0)
+    """
+
+    def __init__(self, cfg: dict, signal: str = "tavg", t_cut: float = 500.0):
+        self.cfg = cfg
+        self.signal = signal
+        self.t_cut = t_cut
+
+    # ------------------------------------------------------------------
+    # Internal helpers
+    # ------------------------------------------------------------------
+
+    def _prepare(
+        self, t: np.ndarray, ts: np.ndarray
+    ) -> tuple[np.ndarray, float]:
+        """
+        Apply burn-in cut and reshape to (n_nodes, n_steps) for feature fns.
+
+        Returns
+        -------
+        ts_2d : (n_nodes, n_steps) float64
+        fs    : sampling frequency in Hz, inferred from the time array (ms)
+        """
+        t_cut_idx = int(np.searchsorted(t, self.t_cut))
+        ts_cut = ts[t_cut_idx:]
+
+        # fs from the time array (t is in ms → Hz = 1000 / dt_ms)
+        dt_ms = float(t[1] - t[0]) if len(t) > 1 else 1.0
+        fs = 1000.0 / dt_ms
+
+        # Normalise shape to (n_nodes, n_steps)
+        if ts_cut.ndim == 3:
+            # (n_steps, n_voi, n_nodes) — tavg / subsample / raw monitors
+            ts_2d = ts_cut[:, 0, :].T
+        elif ts_cut.ndim == 2:
+            # (n_steps, n_nodes) — bold monitor
+            ts_2d = ts_cut.T
+        else:
+            raise ValueError(
+                f"Unexpected time-series shape {ts_cut.shape}; "
+                "expected (n_steps, n_voi, n_nodes) or (n_steps, n_nodes)."
+            )
+
+        return ts_2d.astype(np.float64, copy=False), fs
+
+    # ------------------------------------------------------------------
+    # Public API
+    # ------------------------------------------------------------------
+
+    def extract(
+        self, monitor_result: dict
+    ) -> tuple[list[str], np.ndarray]:
[…]
