### 2026-05-18 83091b4
Refactor FeaturePipeline integration in sweep test to use updated feature configuration
