### 2026-05-18 bdd5dda
Refactor global coupling strength parameter and update comments for clarity
