### 2026-05-18 d59f0cc
Add Tier-2 JIT feature extraction with Numba for mean, std, and correlation metrics
