+++ vbi/feature_extraction/features_utils_nb.py
@@ -1,0 +1,249 @@
+"""
+Tier-2 JIT feature kernels for the Numba CPU backend.
+
+These @njit functions are called inside prange, so they must be pure numeric
+kernels with no Python-level dispatch, dicts, or pandas.
+
+Supported core set
+------------------
+calc_mean   per-node temporal mean           → (n_nodes,)
+calc_std    per-node temporal std            → (n_nodes,)
+fc_flat     upper-triangle Pearson FC        → (n_nodes*(n_nodes-1)//2,)
+fcd_mean    mean of FCD upper triangle       → scalar
+
+Input convention: ts_2d is always (n_steps, n_nodes) float64.
+This matches nb_simulate_det output ts_i[:, voi, :].
+
+NbExtractorSpec
+---------------
+A plain Python dataclass (not a numba jitclass) that the pipeline and sweeper
+use to communicate which features to compute and how large the output will be.
+The actual number-crunching is done by nb_extract(), which is @njit.
+"""
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+
+import numpy as np
+from numba import njit
+
+
+# ---------------------------------------------------------------------------
+# Core kernels
+# ---------------------------------------------------------------------------
+
+@njit(cache=True)
+def nb_mean(ts):
+    """Per-node temporal mean.  ts: (n_steps, n_nodes) → (n_nodes,)"""
+    n_steps, n_nodes = ts.shape
+    out = np.zeros(n_nodes)
+    for t in range(n_steps):
+        for n in range(n_nodes):
+            out[n] += ts[t, n]
+    for n in range(n_nodes):
+        out[n] /= n_steps
+    return out
+
+
+@njit(cache=True)
+def nb_std(ts):
+    """Per-node temporal std (ddof=1).  ts: (n_steps, n_nodes) → (n_nodes,)"""
+    n_steps, n_nodes = ts.shape
+    mean = nb_mean(ts)
+    out = np.zeros(n_nodes)
+    for t in range(n_steps):
+        for n in range(n_nodes):
+            d = ts[t, n] - mean[n]
+            out[n] += d * d
+    for n in range(n_nodes):
+        out[n] = (out[n] / (n_steps - 1)) ** 0.5 if n_steps > 1 else 0.0
+    return out
+
+
+@njit(cache=True)
+def nb_fc_flat(ts):
+    """
+    Upper triangle (k=1) of Pearson correlation matrix.
+    ts: (n_steps, n_nodes) → (n_nodes*(n_nodes-1)//2,)
+    """
+    n_steps, n_nodes = ts.shape
+    mean = nb_mean(ts)
+
+    n_upper = n_nodes * (n_nodes - 1) // 2
+    out = np.zeros(n_upper)
+    idx = 0
+    for i in range(n_nodes):
+        for j in range(i + 1, n_nodes):
+            cov = 0.0
+            var_i = 0.0
+            var_j = 0.0
+            for t in range(n_steps):
+                xi = ts[t, i] - mean[i]
+                xj = ts[t, j] - mean[j]
+                cov   += xi * xj
+                var_i += xi * xi
+                var_j += xj * xj
+            denom = (var_i * var_j) ** 0.5
+            out[idx] = cov / denom if denom > 0.0 else 0.0
+            idx += 1
+    return out
+
+
+@njit(cache=True)
+def nb_fcd_mean(ts, window_size):
+    """
+    Mean of the FCD (functional connectivity dynamics) upper triangle.
+
+    Slides a window of `window_size` steps over the time series, computes the
+    FC vector for each window, then returns the mean pairwise Pearson
[…]

[…]
