### 2026-05-18 5f1763e
Refactor feature extraction parameters from 'G' to 'cr' in tests and update assertions accordingly
