### 2026-05-18 d3934ff
feat: Implement C++ backend for VBI simulation
- Added runtime.hpp for shared helpers including RingBuffer for delayed coupling history.
- Created sim_module.cpp.mako template for generating C++ simulation module.
- Developed build.py for managing CMake build process and caching compiled modules.
- Introduced codegen.py for translating Python expressions to C++ and rendering templates.
- Implemented simulator.py to provide a C++ pybind11 backend for VBI, mirroring the interface of existing simulators.
- Added test_mpr_cpp.py to validate C++ backend against NumPy reference, including determinis
[…]
