+++ vbi/simulator/backend/cpp/build.py
@@ -1,0 +1,244 @@
+"""
+Build-cache and CMake/compiler invocation for the VBI C++ backend.
+
+Cache key = SHA-256 of SimulationSpec.cache_key() — same model+integrator+n_nodes
+→ same .so.  Parameter sweeps that change only G/eta reuse the compiled binary.
+"""
+from __future__ import annotations
+
+import importlib.machinery
+import importlib.util
+import os
+import shlex
+import shutil
+import subprocess
+import sys
+import sysconfig
+import types
+from pathlib import Path
+
+from vbi.simulator.spec.simulation import SimulationSpec
+
+from .codegen import render_sim_module, render_bindings, render_cmake
+
+_TEMPLATES_DIR = Path(__file__).resolve().parent / "_src"
+_CACHE_DIR = Path(os.environ.get("VBI_CPP_CACHE",
+                                  Path.home() / ".cache" / "vbi" / "cpp"))
+
+# In-process loaded-module cache  {cache_key: module}
+_MODULE_CACHE: dict[str, types.ModuleType] = {}
+
+
+# ---------------------------------------------------------------------------
+# Public API
+# ---------------------------------------------------------------------------
+
+def build_or_load(spec: SimulationSpec, verbose: bool = False) -> types.ModuleType:
+    """
+    Return the compiled pybind11 extension module for `spec`.
+
+    - On first call: generate C++ sources, compile, cache the .so.
+    - On subsequent calls (same spec): load from in-process cache or disk.
+    """
+    key = spec.cache_key()
+
+    if key in _MODULE_CACHE:
+        return _MODULE_CACHE[key]
+
+    build_dir = _CACHE_DIR / key
+    so_path   = _discover_so(build_dir, key)
+
+    if so_path is None:
+        _write_sources(spec, key, build_dir, verbose=verbose)
+        so_path = _compile(build_dir, key, verbose=verbose)
+
+    mod = _load_so(so_path, key)
+    _MODULE_CACHE[key] = mod
+    return mod
+
+
+# ---------------------------------------------------------------------------
+# Source generation
+# ---------------------------------------------------------------------------
+
+def _write_sources(spec: SimulationSpec, key: str,
+                   build_dir: Path, verbose: bool = False) -> None:
+    build_dir.mkdir(parents=True, exist_ok=True)
+
+    module_name      = f"vbi_cpp_{key[:16]}"
+    cpp_filename     = f"{module_name}.cpp"
+    bindings_filename= f"{module_name}_bindings.cpp"
+    runtime_dst      = build_dir / "runtime.hpp"
+
+    # Copy shared runtime header
+    shutil.copy2(_TEMPLATES_DIR / "runtime.hpp", runtime_dst)
+
+    # Render and write generated sources
+    (build_dir / cpp_filename).write_text(
+        render_sim_module(spec, key), encoding="utf-8")
+    (build_dir / bindings_filename).write_text(
+        render_bindings(spec, module_name, cpp_filename), encoding="utf-8")
+    (build_dir / "CMakeLists.txt").write_text(
+        render_cmake(module_name, bindings_filename), encoding="utf-8")
+
+    if verbose:
+        print(f"[vbi-cpp] Sources written → {build_dir}")
+
+
+# ---------------------------------------------------------------------------
+# Compilation
+# ---------------------------------------------------------------------------
+
+def _compile(build_dir: Path, key: str, verbose: bool = False) -> Path:
+    try:
+        return _cmake_build(build_dir, verbose=verbose)
+    except (subprocess.CalledProcessError, FileNotFoundError):
+        if verbose:
[…]
