+++ vbi/simulator/backend/cpp/__init__.py
@@ -1,0 +1,4 @@
+from .simulator import CppSimulator
+from .build import build_or_load
+
+__all__ = ["CppSimulator", "build_or_load"]

[…]
