### 2026-05-18 f5f2ca5
feat: Add C++ backend support with CppSweeper for parameter sweeps and validation tests
