+++ vbi/simulator/backend/cpp/__init__.py
@@ -1,4 +1,5 @@
 from .simulator import CppSimulator
+from .sweeper import CppSweeper
 from .build import build_or_load
 
-__all__ = ["CppSimulator", "build_or_load"]
+__all__ = ["CppSimulator", "CppSweeper", "build_or_load"]
