+++ docs/examples/simulator_models/benchmark_cpp_vs_numpy_numba.py
@@ -1,0 +1,205 @@
+"""
+Benchmark: NumPy vs Numba vs C++ (serial and parallel) for parameter sweeps.
+
+NumPy runs sequentially in Python.
+Numba uses @njit(parallel=True) / prange — true thread parallelism within one process.
+C++ serial   — Python loop calling pybind11 run_simulation() per point.
+C++ parallel — ThreadPoolExecutor; each thread calls pybind11 (GIL released in C++).
+
+Run from this directory:
+    python benchmark_cpp_vs_numpy_numba.py
+    python benchmark_cpp_vs_numpy_numba.py --n-nodes 80 --duration 2000
+    python benchmark_cpp_vs_numpy_numba.py --no-plot
+"""
+from __future__ import annotations
+
+import argparse
+import os
+import time
+from pathlib import Path
+import sys
+
+sys.dont_write_bytecode = True
+
+import numpy as np
+
+from helpers import ensure_repo_on_path, complete_graph_weights
+
+ensure_repo_on_path(__file__)
+
+from vbi.simulator.api import Sweeper
+from vbi.simulator.models.mpr import mpr
+from vbi.simulator.spec import (
+    CouplingSpec,
+    IntegratorSpec,
+    MonitorSpec,
+    SimulationSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+from vbi.simulator.backend.cpp.sweeper import CppSweeper
+
+
+# ---------------------------------------------------------------------------
+# Configuration
+# ---------------------------------------------------------------------------
+
+DEFAULT_N_NODES  = 40
+DEFAULT_DURATION = 500.0   # ms  (short enough for NumPy to finish)
+DEFAULT_DT       = 0.1
+ETA_RANGE        = (-5.5, -3.5)
+N_CPU            = os.cpu_count() // 2 or 4
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def make_spec(n_nodes: int, dt: float) -> SimulationSpec:
+    W = complete_graph_weights(n_nodes)
+    W /= W.sum(axis=1, keepdims=True).clip(1e-8)
+    return SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method="heun", dt=dt, stochastic=False),
+        coupling=CouplingSpec(kind="linear", a=0.1, b=0.0),
+        monitors=(MonitorSpec(kind="raw"),),
+        weights=W,
+    )
+
+
+def make_sweep(n_samples: int) -> SweepSpec:
+    return SweepSpec(params={"eta": np.linspace(*ETA_RANGE, n_samples)})
+
+
+def timed(fn, *args, **kwargs) -> float:
+    t0 = time.perf_counter()
+    fn(*args, **kwargs)
+    return time.perf_counter() - t0
+
+
+# ---------------------------------------------------------------------------
+# Warmup (compile / JIT)
+# ---------------------------------------------------------------------------
+
+def warmup(spec: SimulationSpec) -> None:
+    print("Warming up (Numba JIT + C++ compile)…", flush=True)
+    tiny = SweepSpec(params={"eta": np.array([-4.6, -4.0])})
+    Sweeper(spec, tiny, backend="numba").run(100.0)
+    CppSweeper(spec, tiny).run_serial(100.0)
+    print("  done.\n", flush=True)
+
+
+# ---------------------------------------------------------------------------
[…]
