+++ vbi/simulator/backend/cpp/sweeper.py
@@ -1,0 +1,283 @@
+"""
+CppSweeper — serial and parallel parameter sweep using the C++ backend.
+
+The compiled .so is loaded once per SimulationSpec (cache key).  Each sweep
+point runs run_simulation() with its own params copy; the C++ function releases
+the GIL, so Python ThreadPoolExecutor gives true parallelism.
+
+Interface matches NumpySweeper and NumbaSweeperCPU:
+    run(duration)   → list[dict] | (labels, values)
+    run_df(duration)→ pd.DataFrame
+"""
+from __future__ import annotations
+
+import concurrent.futures
+import numpy as np
+
+from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.spec.sweep import SweepSpec
+from vbi.simulator.backend.numba_.simulator import _apply_monitor
+
+from .build import build_or_load
+from .codegen import build_params_array, get_G, get_noise_data
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _patch_params(base_params: np.ndarray,
+                  param_names: list[str],
+                  model_param_names: list[str],
+                  theta: np.ndarray,
+                  n_nodes: int) -> np.ndarray:
+    """Return a copy of base_params with swept parameters overridden."""
+    p = base_params.copy()
+    for name, val in zip(param_names, theta):
+        idx = model_param_names.index(name)
+        p[idx * n_nodes: idx * n_nodes + n_nodes] = float(val)
+    return p
+
+
+def _run_one(
+    mod,
+    params_flat: np.ndarray,
+    state0: np.ndarray,
+    weights_flat: np.ndarray,
+    idelays_flat: np.ndarray,
+    horizon: int,
+    coup_a: float,
+    coup_b: float,
+    has_delays: bool,
+    noise_flat: np.ndarray,
+    noise_sv_idx: np.ndarray,
+    n_steps: int,
+    record_every: int,
+    t_cut_steps: int,
+) -> np.ndarray:
+    """Run one simulation point and return raw_data (n_record, n_sv, n_nodes)."""
+    return mod.run_simulation(
+        state0, weights_flat, idelays_flat, horizon,
+        params_flat, coup_a, coup_b, has_delays,
+        noise_flat, noise_sv_idx,
+        n_steps, record_every, t_cut_steps,
+    )
+
+
+# ---------------------------------------------------------------------------
+# CppSweeper
+# ---------------------------------------------------------------------------
+
+class CppSweeper:
+    """
+    C++ parameter sweep — serial or parallel via threads.
+
+    Parameters
+    ----------
+    spec        : SimulationSpec   base simulation
+    sweep_spec  : SweepSpec        which params to vary
+    n_workers   : int | None       thread-pool size; None = serial;
+                                   0 = os.cpu_count() threads
+    """
+
+    def __init__(self, spec: SimulationSpec, sweep_spec: SweepSpec,
+                 n_workers: int | None = None):
+        if spec.coupling.kind != "linear":
+            raise NotImplementedError(
+                f"C++ sweeper supports 'linear' coupling; got {spec.coupling.kind!r}.")
+
+        self.spec       = spec
+        self.sweep      = sweep_spec
+        self.n_workers  = n_workers
+
+        dt      = spec.integrator.dt
+        n_nodes = spec.n_nodes
+
+        self._mod = build_or_load(spec)
+
+        # Flat (n_params * n_nodes,) base params — sweep overrides individual rows
+        self._base_params = np.ascontiguousarray(
+            build_params_array(spec).ravel(), dtype=np.float64)
+
+        G = get_G(spec)
[…]

[…]
