### 2026-05-18 212ace4
feat: Enhance benchmarking script to include Numba serial and parallel performance metrics
