+++ docs/examples/simulator_models/benchmark_cpp_vs_numpy_numba.py
@@ -22,6 +22,7 @@ import sys
 sys.dont_write_bytecode = True
 
 import numpy as np
+import numba
 
 from helpers import ensure_repo_on_path, complete_graph_weights
 
@@ -47,7 +48,10 @@ DEFAULT_N_NODES  = 40
 DEFAULT_DURATION = 500.0   # ms  (short enough for NumPy to finish)
 DEFAULT_DT       = 0.1
 ETA_RANGE        = (-5.5, -3.5)
-N_CPU            = os.cpu_count() // 2 or 4
+
+# Both parallel backends (Numba prange + C++ ThreadPoolExecutor) are pinned
+# to the same thread count so the comparison is apples-to-apples.
+N_WORKERS = int(os.environ.get("VBI_BENCH_THREADS", os.cpu_count() // 2 or 4))
 
 
 # ---------------------------------------------------------------------------
@@ -81,10 +85,13 @@ def timed(fn, *args, **kwargs) -> float:
 # ---------------------------------------------------------------------------
 
 def warmup(spec: SimulationSpec) -> None:
-    print("Warming up (Numba JIT + C++ compile)…", flush=True)
+    print(f"Warming up (Numba JIT + C++ compile, {N_WORKERS} threads each)…", flush=True)
+    # Pin Numba's thread pool before the first JIT compile so the compiled
+    # code uses exactly N_WORKERS threads for every subsequent prange call.
+    numba.set_num_threads(N_WORKERS)
     tiny = SweepSpec(params={"eta": np.array([-4.6, -4.0])})
     Sweeper(spec, tiny, backend="numba").run(100.0)
-    CppSweeper(spec, tiny).run_serial(100.0)
+    CppSweeper(spec, tiny, n_workers=N_WORKERS).run_serial(100.0)
     print("  done.\n", flush=True)
 
 
@@ -92,14 +99,25 @@ def warmup(spec: SimulationSpec) -> None:
 # Main benchmark
 # ---------------------------------------------------------------------------
 
+def timed_numba_serial(sweeper, duration: float) -> float:
+    """Run Numba sweep with 1 thread (forces prange to be sequential)."""
+    numba.set_num_threads(1)
+    try:
+        return timed(sweeper.run, duration)
+    finally:
+        numba.set_num_threads(N_WORKERS)
+
+
 def run_benchmark(n_nodes: int, duration: float, dt: float,
                   sweep_sizes: list[int], plot: bool) -> None:
     spec = make_spec(n_nodes, dt)
     warmup(spec)
 
-    cols = ["n_samples", "numpy(s)", "numba(s)", "cpp-ser(s)", f"cpp-par×{N_CPU}(s)",
-            "nb/np", f"cpp-par/np"]
-    col_w = [10, 10, 10, 12, 16, 8, 12]
+    T = N_WORKERS
+    cols = ["n_samples", "np-ser(s)", "nb-ser(s)", f"nb-par×{T}(s)",
+            "cpp-ser(s)", f"cpp-par×{T}(s)", "nb-ser/np", f"nb-par/np",
+            "cpp-ser/np", f"cpp-par/np"]
+    col_w = [10, 10, 10, 14, 11, 14, 11, 11, 12, 11]
     header = "  ".join(c.rjust(w) for c, w in zip(cols, col_w))
     sep    = "-" * len(header)
     print(header)
@@ -107,27 +125,35 @@ def run_benchmark(n_nodes: int, duration: float, dt: float,
 
     rows = []
     for n in sweep_sizes:
[…]
