### 2026-05-18 d013050
feat: Simplify C++ backend loading by removing ImportError handling and add prerequisite checks for dependencies
