+++ vbi/simulator/backend/base.py
@@ -23,14 +23,8 @@ def load_backend(name: str) -> type:
                 "Numba backend requires numba: pip install numba"
             ) from exc
     if name == "cpp":
-        try:
-            from vbi.simulator.backend.cpp.simulator import CppSimulator
-            return CppSimulator
-        except ImportError as exc:
-            raise ImportError(
-                "C++ backend requires pybind11 and mako: "
-                "pip install pybind11 mako"
-            ) from exc
+        from vbi.simulator.backend.cpp.simulator import CppSimulator
+        return CppSimulator
     raise ImportError(
         f"Backend {name!r} is not available. "
         "Available: 'numpy', 'numba', 'cpp'. Coming: 'cuda', 'jax'."
@@ -50,14 +44,8 @@ def load_sweep_backend(name: str) -> type:
                 "Numba backend requires numba: pip install numba"
             ) from exc
     if name == "cpp":
-        try:
-            from vbi.simulator.backend.cpp.sweeper import CppSweeper
-            return CppSweeper
-        except ImportError as exc:
-            raise ImportError(
-                "C++ backend requires pybind11 and mako: "
-                "pip install pybind11 mako"
-            ) from exc
+        from vbi.simulator.backend.cpp.sweeper import CppSweeper
+        return CppSweeper
     raise ImportError(
         f"Sweep backend {name!r} is not available. "
         "Available: 'numpy', 'numba', 'cpp'. Coming: 'cuda', 'jax'."
