### 2026-05-18 c739636
feat: Add CMake template for C++ backend integration and update code generation to use the new template
