### 2026-05-18 facfdb1
feat: Update CppSweeper to enhance serial and parallel execution descriptions and improve benchmark output
