### 2026-05-18 e2e0fea
feat: Update CppSweeper to support batch processing and enhance run methods for improved memory management
