### 2026-05-18 b950627
feat: Update warmup and benchmark functions to accept dynamic worker count for improved flexibility
