+++ docs/examples/simulator_models/benchmark_cpp_vs_numpy_numba.py
@@ -84,14 +84,14 @@ def timed(fn, *args, **kwargs) -> float:
 # Warmup (compile / JIT)
 # ---------------------------------------------------------------------------
 
-def warmup(spec: SimulationSpec) -> None:
-    print(f"Warming up (Numba JIT + C++ compile, {N_WORKERS} threads each)…", flush=True)
+def warmup(spec: SimulationSpec, n_workers: int) -> None:
+    print(f"Warming up (Numba JIT + C++ compile, {n_workers} threads each)…", flush=True)
     # Pin Numba's thread pool before the first JIT compile so the compiled
-    # code uses exactly N_WORKERS threads for every subsequent prange call.
-    numba.set_num_threads(N_WORKERS)
+    # code uses exactly n_workers threads for every subsequent prange call.
+    numba.set_num_threads(n_workers)
     tiny = SweepSpec(params={"eta": np.array([-4.6, -4.0])})
     Sweeper(spec, tiny, backend="numba").run(500.0)
-    cpp_tiny = CppSweeper(spec, tiny, n_workers=N_WORKERS)
+    cpp_tiny = CppSweeper(spec, tiny, n_workers=n_workers)
     cpp_tiny.run_serial(500.0)    # compiles .so and warms ThreadPoolExecutor
     cpp_tiny.run_parallel(500.0)
     print("  done.\n", flush=True)
@@ -101,21 +101,21 @@ def warmup(spec: SimulationSpec) -> None:
 # Main benchmark
 # ---------------------------------------------------------------------------
 
-def timed_numba_serial(sweeper, duration: float) -> float:
+def timed_numba_serial(sweeper, duration: float, n_workers: int) -> float:
     """Run Numba sweep with 1 thread (forces prange to be sequential)."""
     numba.set_num_threads(1)
     try:
         return timed(sweeper.run, duration)
     finally:
-        numba.set_num_threads(N_WORKERS)
+        numba.set_num_threads(n_workers)
 
 
 def run_benchmark(n_nodes: int, duration: float, dt: float,
-                  sweep_sizes: list[int], plot: bool) -> None:
+                  sweep_sizes: list[int], n_workers: int, plot: bool) -> None:
     spec = make_spec(n_nodes, dt)
-    warmup(spec)
+    warmup(spec, n_workers)
 
-    T = N_WORKERS
+    T = n_workers
     cols = ["n_samples", "np-ser(s)", "nb-ser(s)", f"nb-par×{T}(s)",
             "cpp-ser(s)", f"cpp-par×{T}(s)",
             "nb-ser/np", f"nb-par/np", f"cpp-ser/np", f"cpp-par/np"]
@@ -129,10 +129,10 @@ def run_benchmark(n_nodes: int, duration: float, dt: float,
     for n in sweep_sizes:
         sw  = make_sweep(n)
         nb_sweeper = Sweeper(spec, sw, backend="numba")
-        cpp        = CppSweeper(spec, sw, n_workers=N_WORKERS)
+        cpp        = CppSweeper(spec, sw, n_workers=n_workers)
 
         t_np  = timed(Sweeper(spec, sw, backend="numpy").run, duration)
-        t_nbs = timed_numba_serial(nb_sweeper,                duration)
+        t_nbs = timed_numba_serial(nb_sweeper,                duration, n_workers)
         t_nbp = timed(nb_sweeper.run,                         duration)
         t_cs  = timed(cpp.run_serial,                         duration)
         t_cp  = timed(cpp.run_parallel,                       duration)
[…]
