### 2026-05-19 c1704b5
feat: Add tests for ReducedWongWang and WongWangExcInh models in C++ validation suite
