### 2026-05-19 b643ead
feat: Add validation tests for Jansen-Rit and Wilson-Cowan models using Numba CPU backend
