### 2026-05-19 b6ec8ef
feat: Update milestone status and enhance C++ sweep tests for cross-backend validation
