+++ vbi/tests/validation/test_cpp_models.py
@@ -155,10 +155,6 @@ def test_mpr_euler_cpp_vs_numpy():
 
 
 # ---------------------------------------------------------------------------
-# Delayed coupling (DDE)
-# ---------------------------------------------------------------------------
-
-# ---------------------------------------------------------------------------
 # ReducedWongWang (1 SV, bounded [0,1], piecewise-safe transfer function)
 # ---------------------------------------------------------------------------
 
