### 2026-05-19 0c0c8d9
Add new models and validation tests for VBI simulator
- Introduced several new models: CoombesByrne2D, GastSchmidtKnosche_SD, GastSchmidtKnosche_SF, Generic2dOscillator, Kuramoto, LarterBreakspear, SupHopf, and Linear.
- Implemented model specifications including state variables, parameters, and differential function strings for each model.
- Added validation tests for the new models using the NumPy backend, ensuring they run without errors and produce finite outputs.
- Included comparisons with TVB simulations to verify the accuracy of the new models.
