+++ vbi/simulator/models/larter_breakspear.py
@@ -1,0 +1,80 @@
+from vbi.simulator.spec.model import ModelSpec, StateVar, Parameter
+
+# All five intermediates (m_Ca, m_Na, m_K, QV, QZ) are inlined so
+# the dfun_str expressions contain only state-vars, parameters and coupling.
+# TVB parameter 'b' is retained; it does not conflict with the coupling
+# alias 'c' (parameter 'C' is capital, also no conflict).
+
+larter_breakspear = ModelSpec(
+    name="LarterBreakspear",
+    state_variables=(
+        StateVar("V", default_init=0.0,  noise=True, lower_bound=-1.5, upper_bound=1.5),
+        StateVar("W", default_init=0.0,  noise=False, lower_bound=-1.5, upper_bound=1.5),
+        StateVar("Z", default_init=0.0,  noise=False, lower_bound=-1.5, upper_bound=1.5),
+    ),
+    parameters=(
+        Parameter("gCa",   1.1,    "Ca channel conductance"),
+        Parameter("gK",    2.0,    "K channel conductance"),
+        Parameter("gL",    0.5,    "Leak channel conductance"),
+        Parameter("phi",   0.7,    "Temperature scaling factor for K gate"),
+        Parameter("gNa",   6.7,    "Na channel conductance"),
+        Parameter("TK",    0.0,    "K channel half-activation voltage"),
+        Parameter("TCa",  -0.01,   "Ca channel half-activation voltage"),
+        Parameter("TNa",   0.3,    "Na channel half-activation voltage"),
+        Parameter("VCa",   1.0,    "Ca reversal (Nernst) potential"),
+        Parameter("VK",   -0.7,    "K reversal potential"),
+        Parameter("VL",   -0.5,    "Leak reversal potential"),
+        Parameter("VNa",   0.53,   "Na reversal potential"),
+        Parameter("d_K",   0.3,    "K half-activation width"),
+        Parameter("tau_K", 1.0,    "K relaxation time constant (ms)"),
+        Parameter("d_Na",  0.15,   "Na half-activation width"),
+        Parameter("d_Ca",  0.15,   "Ca half-activation width"),
+        Parameter("aei",   2.0,    "Excitatory→inhibitory coupling strength"),
+        Parameter("aie",   2.0,    "Inhibitory→excitatory coupling strength"),
+        Parameter("b",     0.1,    "Time constant scaling for Z equation"),
+        Parameter("C",     0.1,    "Long-range excitatory coupling weight (0 < C < a_ee)"),
+        Parameter("ane",   1.0,    "Non-specific→excitatory input strength"),
+        Parameter("ani",   0.4,    "Non-specific→inhibitory input strength"),
+        Parameter("aee",   0.4,    "Excitatory→excitatory coupling strength"),
+        Parameter("Iext",  0.3,    "Subcortical (thalamic) input current"),
+        Parameter("rNMDA", 0.25,   "NMDA/AMPA receptor ratio"),
+        Parameter("VT",    0.0,    "Excitatory firing threshold"),
+        Parameter("d_V",   0.65,   "Excitatory threshold variance"),
+        Parameter("ZT",    0.0,    "Inhibitory firing threshold"),
+        Parameter("d_Z",   0.7,    "Inhibitory threshold variance"),
+        Parameter("QV_max", 1.0,   "Maximum excitatory firing rate"),
[…]
