### 2026-05-19 fbf54d8
Add validation tests for new models using Numba and structural correctness checks
- Implemented `test_new_models_numba.py` to validate Numba backend against NumPy for various models, ensuring deterministic outputs match within a tolerance of rtol=1e-4.
- Added tests for stochastic behavior, parameter heterogeneity, and state variable bounds for Numba.
- Created `test_new_models_spec.py` to verify structural correctness of ModelSpec objects, checking dimensions, indices, and bounds without running the simulator.
- Included comprehensive checks for model metadata consistency and dfun evaluation 
[…]
