+++ vbi/tests/validation/test_new_models_numba.py
@@ -1,0 +1,249 @@
+"""
+M1 validation: Numba CPU backend for the new models.
+
+Gold standard: NumPy baseline (validated in test_new_models_numpy.py).
+All Numba results must match NumPy to rtol=1e-4 (deterministic).
+"""
+import numpy as np
+import pytest
+
+from vbi.simulator import Simulator, Sweeper
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+
+from vbi.simulator.models.generic_2d_oscillator import generic_2d_oscillator
+from vbi.simulator.models.kuramoto import kuramoto
+from vbi.simulator.models.sup_hopf import sup_hopf
+from vbi.simulator.models.linear import linear
+from vbi.simulator.models.larter_breakspear import larter_breakspear
+from vbi.simulator.models.coombes_byrne_2d import coombes_byrne_2d
+from vbi.simulator.models.gast_sd import gast_sd
+from vbi.simulator.models.gast_sf import gast_sf
+
+numba = pytest.importorskip("numba", reason="numba not installed")
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _weights(n, seed=0):
+    rng = np.random.default_rng(seed)
+    W = np.abs(rng.standard_normal((n, n)))
+    np.fill_diagonal(W, 0.0)
+    W /= W.sum(axis=1, keepdims=True).clip(min=1e-8)
+    return W
+
+
+def _spec(model, n_nodes=4, dt=0.01, coup_a=0.05, method="heun",
+          stochastic=False, node_params=None, tract_lengths=None):
+    W = _weights(n_nodes)
+    if tract_lengths is None:
+        tract_lengths = np.zeros_like(W)
+    noise_nsig = None
+    if stochastic:
+        n_noise = len(model.noise_indices)
+        noise_nsig = np.full(n_noise, 1e-4)
+    return SimulationSpec(
+        model=model,
+        integrator=IntegratorSpec(method=method, dt=dt,
+                                  stochastic=stochastic, noise_nsig=noise_nsig),
+        coupling=CouplingSpec("linear", a=coup_a),
+        monitors=(MonitorSpec("raw"),),
+        weights=W,
+        tract_lengths=tract_lengths,
+        node_params=node_params or {},
+    )
+
+
+def _np(spec, duration):
+    return Simulator(spec, backend="numpy").run(duration)["raw"]
+
+
+def _nb(spec, duration):
+    return Simulator(spec, backend="numba").run(duration)["raw"]
+
+
+# ---------------------------------------------------------------------------
+# Parametrised deterministic match: Numba must reproduce NumPy to rtol=1e-4
+# ---------------------------------------------------------------------------
+
+@pytest.mark.parametrize("model,dt,coup_a,duration", [
+    (generic_2d_oscillator, 0.01, 0.05,  50.0),
+    (kuramoto,              0.01, 0.1,   30.0),
+    (sup_hopf,              0.01, 0.1,   30.0),
+    (linear,                0.01, 0.05,  30.0),
+    (larter_breakspear,     0.1,  0.05,  50.0),
+    (coombes_byrne_2d,      0.01, 0.05,  30.0),
+    (gast_sd,               0.01, 0.02,  30.0),
+    (gast_sf,               0.01, 0.02,  30.0),
+])
+class TestNumbaDeterministicMatchesNumPy:
+
[…]
