+++ vbi/tests/validation/test_new_models_dynamics.py
@@ -1,0 +1,588 @@
+"""
+Dynamical behaviour tests for the new models.
+
+These tests verify that each model produces the correct physics:
+  - known fixed points are reached (or not, depending on regime)
+  - limit cycles have the correct amplitude/frequency
+  - bounds are respected throughout
+  - node-heterogeneous parameters work
+  - key bifurcation parameters shift the regime as expected
+
+All tests use the NumPy backend (backend="numpy") so they are independent
+of Numba/C++ compilation.
+"""
+import dataclasses
+
+import numpy as np
+import pytest
+
+from vbi.simulator import Simulator
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+)
+from vbi.simulator.models.generic_2d_oscillator import generic_2d_oscillator
+from vbi.simulator.models.kuramoto import kuramoto
+from vbi.simulator.models.sup_hopf import sup_hopf
+from vbi.simulator.models.linear import linear
+from vbi.simulator.models.larter_breakspear import larter_breakspear
+from vbi.simulator.models.coombes_byrne_2d import coombes_byrne_2d
+from vbi.simulator.models.gast_sd import gast_sd
+from vbi.simulator.models.gast_sf import gast_sf
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _isolated_spec(model, n_nodes=1, dt=0.01, node_params=None):
+    """Fully disconnected (W=0) single or multi-node spec."""
+    return SimulationSpec(
+        model=model,
+        integrator=IntegratorSpec(method="heun", dt=dt, stochastic=False),
+        coupling=CouplingSpec("linear", a=0.0),
+        monitors=(MonitorSpec("raw"),),
+        weights=np.zeros((n_nodes, n_nodes)),
+        node_params=node_params or {},
+    )
+
+
+def _run(spec, duration):
+    return Simulator(spec, backend="numpy").run(duration)["raw"]
+
+
+def _last(t, d):
+    """Return the last recorded state (n_sv, n_nodes)."""
+    return d[-1]
+
+
+# ---------------------------------------------------------------------------
+# Linear model — exact analytical solution
+# ---------------------------------------------------------------------------
+
+class TestLinearDynamics:
+    def test_exponential_decay(self):
+        """x(t) = x₀ · exp(γ·t) — numerically exact for small dt."""
+        gamma = -5.0
+        x0 = 0.5    # within bounds [-1, 1]; 2.0 would be clamped to 1.0
+        dt = 0.001
+        duration = 1.0
+        spec = dataclasses.replace(
+            _isolated_spec(linear),
+            integrator=IntegratorSpec(method="heun", dt=dt),
+            node_params={"gamma": gamma},
+        )
+        # Force initial state by overriding default_init
+        from vbi.simulator.models.linear import linear as lin_model
+        spec2 = SimulationSpec(
+            model=dataclasses.replace(
+                lin_model,
+                state_variables=(
+                    dataclasses.replace(lin_model.state_variables[0], default_init=x0),
+                ),
+            ),
+            integrator=IntegratorSpec(method="heun", dt=dt),
+            coupling=CouplingSpec("linear", a=0.0),
+            monitors=(MonitorSpec("raw"),),
+            weights=np.zeros((1, 1)),
+            node_params={"gamma": gamma},
+        )
+        t, d = _run(spec2, duration)
+        x_numerical = d[-1, 0, 0]
+        x_analytical = x0 * np.exp(gamma * duration)
[…]
