### 2026-05-19 6b2bd2e
Add Numba-CUDA backend for GPU simulations
- Implemented code generation for CUDA kernels in `codegen.py`, including deterministic and stochastic sweep kernels.
- Created `CudaSimulator` class in `simulator.py` for single-run simulations using the CUDA backend.
- Developed `CudaSweeperGPU` class in `sweeper.py` for parallel parameter sweeps on the GPU.
- Added validation tests for the CUDA backend in `test_mpr_cuda.py`, ensuring compatibility with the Numba CPU backend and verifying output correctness.
- Included helper functions for building parameters, initial states, and noise generation ta
[…]
