+++ vbi/simulator/backend/numba_cuda/sweeper.py
@@ -1,0 +1,265 @@
+"""
+CudaSweeperGPU — parallel parameter sweep on GPU.
+
+One GPU thread = one complete simulation.
+float32 precision on GPU; results returned as float64 numpy arrays.
+
+Usage
+-----
+    sweeper = CudaSweeperGPU(spec, sweep_spec)
+    labels, values = sweeper.run(duration=5000.0)
+    # or
+    results = sweeper.run(duration=5000.0)  # list of dicts when no pipeline
+"""
+from __future__ import annotations
+
+import numpy as np
+
+from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.spec.sweep import SweepSpec
+from vbi.simulator.backend.numba_.simulator import _apply_monitor  # reuse post-processing
+
+from .codegen import (
+    build_cuda_module,
+    build_params_matrix,
+    build_initial_state,
+    build_ring_buffer,
+    get_bounds_arrays,
+    generate_noise,
+)
+
+try:
+    from numba import cuda as _cuda
+    CUDA_AVAILABLE = _cuda.is_available()
+except Exception:
+    CUDA_AVAILABLE = False
+
+_TPB = 256   # threads per block
+
+
+def _require_cuda():
+    if not CUDA_AVAILABLE:
+        raise RuntimeError(
+            "CUDA backend unavailable. Ensure:\n"
+            "  1. An NVIDIA GPU is present and the CUDA toolkit is installed.\n"
+            "  2. `numba[cuda]` is installed: pip install numba[cuda]\n"
+            "  3. CUDA_HOME or PATH includes nvcc.\n"
+            "Alternatively, use backend='numba' (CPU) or backend='cpp'."
+        )
+
+
+def _count_records(n_steps: int, t_cut: int, period: int) -> int:
+    if n_steps <= t_cut:
+        return 0
+    return (n_steps - t_cut + period - 1) // period
+
+
+class CudaSweeperGPU:
+    """
+    GPU parameter sweep backend.
+
+    Parameters
+    ----------
+    spec        : SimulationSpec   base simulation spec
+    sweep_spec  : SweepSpec        which params to vary
+    """
+
+    def __init__(self, spec: SimulationSpec, sweep_spec: SweepSpec):
+        _require_cuda()
+        if spec.coupling.kind != "linear":
+            raise NotImplementedError(
+                f"CUDA backend supports 'linear' coupling; got {spec.coupling.kind!r}."
+            )
+
+        self.spec  = spec
+        self.sweep = sweep_spec
+
+        dt      = spec.integrator.dt
+        n_nodes = spec.n_nodes
+        model   = spec.model
+
+        # Compile CUDA module (dfun + kernels generated together)
+        self._mod = build_cuda_module(model)
+
+        # Connectivity
+        self._weights_f32 = np.ascontiguousarray(spec.weights, dtype=np.float32)
+        ds = spec.delay_steps(dt).astype(np.int32)
+        self._delay_steps  = np.ascontiguousarray(ds)
+        self._horizon      = spec.horizon(dt)
+        self._has_delays   = bool(spec.has_delays)
+
+        # Coupling scale
+        G = float(np.asarray(spec.node_params.get("G",
+                  model.default_params.get("G", 1.0))).mean())
+        self._coup_a = np.float32(spec.coupling.a * G)
+        self._coup_b = np.float32(spec.coupling.b)
+
+        # Bounds
+        self._lo, self._hlo, self._hi, self._hhi = get_bounds_arrays(model)
+
+        # Initial state
+        self._state0 = build_initial_state(spec)
+
+        # Integration flags
+        self._use_heun   = (spec.integrator.method == "heun")
[…]
