+++ vbi/simulator/backend/numba_cuda/codegen.py
@@ -1,0 +1,431 @@
+"""
+Code generation for the Numba-CUDA backend.
+
+In Numba CUDA, device functions must be in the same compilation unit as the
+kernel that calls them.  Therefore this module generates a *complete* Python
+source string containing:
+
+    1. A @cuda.jit(device=True) dfun  — scalar per-node, float32
+    2. The deterministic sweep kernel
+    3. The stochastic sweep kernel
+
+The source string is exec()‑ed as one module so the kernel sees the dfun.
+
+This mirrors TVB's approach (Mako template → exec); VBI uses plain string
+generation instead of Mako to avoid an extra dependency.
+
+Thread / block layout (same as TVB nb_hybrid_cuda_sweep.py)
+------------------------------------------------------------
+    grid  = ((n_samples + TPB - 1) // TPB,)
+    block = (TPB,)
+    tid   = cuda.grid(1)   →   1 thread = 1 complete simulation
+
+Each thread integrates ALL nodes of ONE parameter set sequentially.
+This is simple, avoids inter-thread communication, and scales naturally
+to thousands of sweep points.
+
+Memory layout (float32 throughout)
+------------------------------------
+state  : (n_samples, n_sv,   n_nodes)
+buf    : (n_samples, n_cvar, n_nodes, horizon)  — ring buffer (dense)
+params : (n_samples, n_params)
+ts_out : (n_samples, n_record, n_sv, n_nodes)   — time-series output
+noise  : (n_samples, n_steps, n_sv, n_nodes)    — pre-generated dW (stoch)
+"""
+from __future__ import annotations
+
+import hashlib
+import importlib.util
+import re
+import sys
+import types
+from pathlib import Path
+
+import numpy as np
+
+from vbi.simulator.spec.model import ModelSpec
+from vbi.simulator.spec.simulation import SimulationSpec
+
+_CACHE_DIR = Path.home() / ".cache" / "vbi" / "cuda"
+_MODULE_CACHE: dict[str, types.ModuleType] = {}
+
+# Python expression → CUDA-compatible math.* equivalent
+_MATH_MAP = [
+    (r'\bexp\b',   'math.exp'),
+    (r'\blog2\b',  'math.log2'),
+    (r'\blog\b',   'math.log'),
+    (r'\bsin\b',   'math.sin'),
+    (r'\bcos\b',   'math.cos'),
+    (r'\btanh\b',  'math.tanh'),
+    (r'\bsinh\b',  'math.sinh'),
+    (r'\bcosh\b',  'math.cosh'),
+    (r'\bsqrt\b',  'math.sqrt'),
+    (r'\bfabs\b',  'math.fabs'),
+    (r'\babs\b',   'math.fabs'),
+    (r'\batan2\b', 'math.atan2'),
+    (r'\batan\b',  'math.atan'),
+    (r'\bfloor\b', 'math.floor'),
+    (r'\bceil\b',  'math.ceil'),
+    (r'\bpi\b',    'math.pi'),
+]
+
+
+def _cuda_expr(expr: str) -> str:
+    """Translate a dfun_str expression to CUDA math.* form."""
+    for pattern, repl in _MATH_MAP:
+        expr = re.sub(pattern, repl, expr)
+    return expr
+
+
+def _generate_source(spec: ModelSpec) -> str:
+    """
+    Generate a complete Python module source containing:
+      - _dfun_cuda   : @cuda.jit(device=True) scalar dfun
+      - cuda_sweep_det   : deterministic sweep kernel
+      - cuda_sweep_stoch : stochastic sweep kernel
+    """
+    sv_names    = spec.sv_names
+    cvar_names  = spec.cvar
+    param_names = spec.param_names
+    n_sv        = spec.n_sv
+    n_cvar      = len(cvar_names)
[…]
