### 2026-05-19 5d376e6
feat: Add benchmark and validation scripts for cross-backend sweep testing
