+++ docs/examples/simulator_models/validate_sweep_backends.py
@@ -1,0 +1,228 @@
+"""
+Cross-backend sweep validation.
+
+Checks that NumPy, Numba CPU, C++, and CUDA (if available) all produce
+numerically consistent sweep results for the same model and parameters.
+
+Usage
+-----
+    python validate_sweep_backends.py
+    python validate_sweep_backends.py --model sup_hopf --n-nodes 10 --n-samples 8
+    python validate_sweep_backends.py --rtol 1e-3   # CUDA float32 tolerance
+
+Each available backend is compared against the NumPy reference.
+Exit code 0 = all checks pass; non-zero = at least one mismatch.
+"""
+from __future__ import annotations
+
+import argparse
+import sys
+from pathlib import Path
+
+import warnings
+import numpy as np
+
+# Suppress low-occupancy warnings for tiny CUDA batches in validation
+warnings.filterwarnings(
+    "ignore",
+    message="Grid size.*will likely result in GPU under-utilization",
+)
+
+# ---- repo path ----
+sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+
+from helpers import ensure_repo_on_path
+ensure_repo_on_path(__file__)
+
+from vbi.simulator import Simulator, Sweeper
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+
+# Available models for quick sweeps
+from vbi.simulator.models.sup_hopf            import sup_hopf
+from vbi.simulator.models.generic_2d_oscillator import generic_2d_oscillator
+from vbi.simulator.models.kuramoto            import kuramoto
+from vbi.simulator.models.mpr                 import mpr
+
+MODELS = {
+    "sup_hopf":             (sup_hopf,             "a",   np.linspace(-0.5, 0.5, 8), 0.01, 0.1),
+    "generic_2d_oscillator":(generic_2d_oscillator,"I",   np.linspace(-1.0, 1.0, 8), 0.01, 0.05),
+    "kuramoto":             (kuramoto,             "omega",np.linspace(0.5, 2.0, 8), 0.01, 0.1),
+    "mpr":                  (mpr,                  "eta", np.linspace(-5.5, -4.0, 8),0.01, 0.1),
+}
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _weights(n: int, seed: int = 0) -> np.ndarray:
+    rng = np.random.default_rng(seed)
+    W = np.abs(rng.standard_normal((n, n)))
+    np.fill_diagonal(W, 0.0)
+    W /= W.sum(axis=1, keepdims=True).clip(1e-8)
+    return W
+
+
+def _make_spec(model, n_nodes: int, dt: float, coup_a: float) -> SimulationSpec:
+    return SimulationSpec(
+        model=model,
+        integrator=IntegratorSpec(method="heun", dt=dt, stochastic=False),
+        coupling=CouplingSpec("linear", a=coup_a),
+        monitors=(MonitorSpec("raw"),),
+        weights=_weights(n_nodes),
+    )
+
+
+def _run_sweep(spec, sweep_spec, backend: str, duration: float) -> list[np.ndarray]:
+    """Return list of raw data arrays (one per sweep point)."""
+    results = Sweeper(spec, sweep_spec, backend=backend).run(duration)
+    return [res["raw"][1] for res in results]
+
+
+def _check_match(reference: list[np.ndarray], candidate: list[np.ndarray],
+                 backend: str, rtol: float, atol: float) -> tuple[bool, str]:
[…]
