+++ docs/examples/simulator_models/benchmark_all_backends.py
@@ -1,0 +1,482 @@
+"""
+Backend sweep benchmark: NumPy · Numba (1T / NT) · C++ (serial / parallel) · CUDA.
+
+Strategy
+--------
+* CPU backends (NumPy, Numba, C++) use small-to-medium sweep sizes because
+  their time scales linearly with n_samples.  Running them at thousands of
+  samples would make the benchmark very slow.
+
+* CUDA is only launched for ≥ 64 samples where GPU scheduling overhead is
+  amortised; it shines at ≥ 1024 samples.
+
+* All CPU backends are timed at the same sweep sizes so speedups are
+  directly comparable on the same x-axis.  The CUDA curve continues to
+  larger n_samples on a separate panel (or same log-scale axis).
+
+Model choice
+------------
+SupHopf (2 SVs, 2 params) — fast equations, good for benchmarking.
+Can be overridden to Generic2dOscillator or MPR via CLI.
+
+Usage
+-----
+    python benchmark_all_backends.py
+    python benchmark_all_backends.py --model generic_2d_oscillator --n-nodes 40
+    python benchmark_all_backends.py --no-cuda --no-plot
+    python benchmark_all_backends.py --n-workers 8 --cpu-sizes 4 8 16 32 64 128
+"""
+from __future__ import annotations
+
+import argparse
+import json
+import os
+import sys
+import time
+from pathlib import Path
+
+import warnings
+import numpy as np
+
+# Suppress Numba CUDA low-occupancy warnings for small batches during warmup
+warnings.filterwarnings(
+    "ignore",
+    message="Grid size.*will likely result in GPU under-utilization",
+)
+
+sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+from helpers import ensure_repo_on_path, complete_graph_weights
+ensure_repo_on_path(__file__)
+
+import numba
+
+from vbi.simulator import Sweeper
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+from vbi.simulator.backend.cpp.sweeper import CppSweeper
+
+from vbi.simulator.models.sup_hopf              import sup_hopf
+from vbi.simulator.models.generic_2d_oscillator import generic_2d_oscillator
+from vbi.simulator.models.kuramoto              import kuramoto
+from vbi.simulator.models.mpr                   import mpr
+
+# ---------------------------------------------------------------------------
+# Configuration
+# ---------------------------------------------------------------------------
+
+MODELS = {
+    "sup_hopf":              (sup_hopf,              "a",   (-0.5, 0.5), 0.01, 0.1),
+    "generic_2d_oscillator": (generic_2d_oscillator, "I",   (-1.0, 1.0), 0.01, 0.05),
+    "kuramoto":              (kuramoto,              "omega",(0.5,  2.0), 0.01, 0.05),
+    "mpr":                   (mpr,                  "eta",  (-5.5,-4.0), 0.01, 0.1),
+}
+
+DEFAULT_MODEL    = "sup_hopf"
+DEFAULT_N_NODES  = 20
+DEFAULT_DURATION = 500.0   # ms  — kept short so CPU runs complete in seconds
+DEFAULT_DT       = 0.01
+
+# CPU sweep sizes — linear cost, so keep small; overlap with CUDA at 32–128
+DEFAULT_CPU_SIZES  = [4, 8, 16, 32, 64, 128]
+# CUDA sweep sizes — starts where CPU overlap begins; GPU shines at >= 512
[…]
