### 2026-05-19 6214862
feat: Adjust benchmark parameters for CPU and CUDA sweep sizes
