### 2026-05-19 c37cf48
Refactor CudaSimulator and CudaSweeperGPU for improved connectivity handling and performance
- Updated CudaSimulator to support coalesced memory layout and optional sparse connectivity.
- Introduced automatic detection of sparse vs dense connectivity based on matrix density.
- Enhanced parameter handling in CudaSweeperGPU to streamline device memory transfers.
- Improved documentation for clarity on connectivity options and usage.
- Adjusted kernel launch parameters to accommodate new data structures and layouts.
- Added memory guard to warn about potential GPU memory issues during simulation.
