+++ vbi/simulator/backend/numba_cuda/codegen.py
@@ -1,36 +1,38 @@
 """
 Code generation for the Numba-CUDA backend.
 
-In Numba CUDA, device functions must be in the same compilation unit as the
-kernel that calls them.  Therefore this module generates a *complete* Python
-source string containing:
+Memory layout — coalesced (tid is always the LAST dimension)
+-------------------------------------------------------------
+Old layout (bad for GPU):
+    state[n_samples, n_sv,   n_nodes]   thread tid reads state[tid, sv, node]
+    buf  [n_samples, n_cvar, n_nodes, horizon]
+    → adjacent threads are n_sv*n_nodes elements apart → NO coalescing
 
-    1. A @cuda.jit(device=True) dfun  — scalar per-node, float32
-    2. The deterministic sweep kernel
-    3. The stochastic sweep kernel
+New layout (coalesced):
+    state[n_sv,   n_nodes, n_samples]   thread tid reads state[sv, node, tid]
+    buf  [n_cvar, n_nodes, horizon, n_samples]
+    → adjacent threads access consecutive addresses → ONE memory transaction per warp
 
-The source string is exec()‑ed as one module so the kernel sees the dfun.
+The Python wrappers (sweeper.py / simulator.py) transpose host arrays before
+transferring to device and transpose back after copy_to_host.
 
-This mirrors TVB's approach (Mako template → exec); VBI uses plain string
-generation instead of Mako to avoid an extra dependency.
+Connectivity
+------------
+Two modes are generated depending on the ``sparse`` flag:
 
-Thread / block layout (same as TVB nb_hybrid_cuda_sweep.py)
-------------------------------------------------------------
+Dense (``sparse=False``):
+    Coupling: weights[node, src] — full (n_nodes × n_nodes) float32 matrix.
+    Best for small N or fully-connected topologies.
+
+CSR sparse (``sparse=True``):
+    Coupling: w_data[nnz], w_indices[nnz], w_indptr[n_nodes+1]
+    Optional: idelays_csr[nnz]  — delay per non-zero edge (int32)
+    Best for large N with real connectomes (typical density 20–40 %).
+
+Thread layout (same for both):
     grid  = ((n_samples + TPB - 1) // TPB,)
-    block = (TPB,)
-    tid   = cuda.grid(1)   →   1 thread = 1 complete simulation
-
-Each thread integrates ALL nodes of ONE parameter set sequentially.
-This is simple, avoids inter-thread communication, and scales naturally
-to thousands of sweep points.
-
-Memory layout (float32 throughout)
-------------------------------------
-state  : (n_samples, n_sv,   n_nodes)
-buf    : (n_samples, n_cvar, n_nodes, horizon)  — ring buffer (dense)
-params : (n_samples, n_params)
-ts_out : (n_samples, n_record, n_sv, n_nodes)   — time-series output
-noise  : (n_samples, n_steps, n_sv, n_nodes)    — pre-generated dW (stoch)
+    block = (TPB,)                              TPB = 256
+    tid   = cuda.grid(1)   →  1 thread = 1 complete simulation
 """
 from __future__ import annotations
 
@@ -49,7 +51,6 @@ from vbi.simulator.spec.simulation import SimulationSpec
 _CACHE_DIR = Path.home() / ".cache" / "vbi" / "cuda"
 _MODULE_CACHE: dict[str, types.ModuleType] = {}
 
-# Python expression → CUDA-compatible math.* equivalent
 _MATH_MAP = [
     (r'\bexp\b',   'math.exp'),
     (r'\blog2\b',  'math.log2'),
@@ -71,82 +72,84 @@ _MATH_MAP = [
 
 
[…]
