+++ vbi/simulator/backend/numba_cuda/simulator.py
@@ -1,19 +1,13 @@
 """
-CudaSimulator — single-run GPU backend (batch=1 wrapper around CudaSweeperGPU).
-
-The CUDA backend is batch-first by design: the sweep is the primary use case.
-A single-run simulator is provided for debugging and single-trajectory use;
-it simply calls the sweeper with a 1-element sweep over a dummy parameter.
+CudaSimulator — single-run GPU backend (batch=1 wrapper using coalesced layout).
 """
 from __future__ import annotations
 
 import numpy as np
 
 from vbi.simulator.spec.simulation import SimulationSpec
-from vbi.simulator.spec.sweep import SweepSpec
 from vbi.simulator.backend.numba_.simulator import _apply_monitor
 
-from .sweeper import CudaSweeperGPU, _require_cuda
 from .codegen import (
     build_cuda_module,
     build_params_matrix,
@@ -21,13 +15,9 @@ from .codegen import (
     build_ring_buffer,
     get_bounds_arrays,
     generate_noise,
+    to_csr,
 )
-
-
-def _count_records(n_steps: int, t_cut: int, period: int) -> int:
-    if n_steps <= t_cut:
-        return 0
-    return (n_steps - t_cut + period - 1) // period
+from .sweeper import _require_cuda, _auto_sparse
 
 try:
     from numba import cuda as _cuda
@@ -38,40 +28,55 @@ except Exception:
 _TPB = 256
 
 
+def _count_records(n_steps: int, t_cut: int, period: int) -> int:
+    if n_steps <= t_cut:
+        return 0
+    return (n_steps - t_cut + period - 1) // period
+
+
 class CudaSimulator:
-    """
-    Single-run Numba-CUDA backend.
+    """Single-run Numba-CUDA backend (coalesced, optional sparse)."""
 
-    Interface mirrors NumpySimulator / NumbaSimulator:
-        build(spec)   → None
-        run(duration) → dict[str, (t, data)]
-    """
-
-    def build(self, spec: SimulationSpec, verbose: bool = False) -> None:
+    def build(self, spec: SimulationSpec,
+              connectivity: str = "auto",
+              verbose: bool = False) -> None:
         _require_cuda()
         if spec.coupling.kind != "linear":
             raise NotImplementedError(
                 f"CUDA backend supports 'linear' coupling; got {spec.coupling.kind!r}."
             )
         self.spec = spec
-        self._mod = build_cuda_module(spec.model)
+
+        if connectivity == "auto":
+            self._sparse = _auto_sparse(spec.weights)
+        else:
+            self._sparse = (connectivity == "sparse")
+
+        self._mod = build_cuda_module(spec.model, sparse=self._sparse)
 
         dt      = spec.integrator.dt
         model   = spec.model
         n_nodes = spec.n_nodes
 
         self._weights_f32  = np.ascontiguousarray(spec.weights, dtype=np.float32)
-        self._delay_steps  = np.ascontiguousarray(spec.delay_steps(dt), dtype=np.int32)
+        ds = spec.delay_steps(dt).astype(np.int32)
+        self._delay_steps  = np.ascontiguousarray(ds)
         self._horizon      = spec.horizon(dt)
         self._has_delays   = bool(spec.has_delays)
 
-        G = float(np.asarray(spec.node_params.get(
-            "G", model.default_params.get("G", 1.0))).mean())
+        if self._sparse:
+            (self._w_data, self._w_indices, self._w_indptr,
+             self._idelays_csr, _, _) = to_csr(
+                 spec.weights, ds if self._has_delays else None,
+                 has_delays=self._has_delays)
+
[…]
