### 2026-05-19 3668c2b
feat: Enhance benchmark plotting with dedicated helper functions and improved configuration
