+++ docs/examples/simulator_models/helpers.py
@@ -11,6 +11,8 @@ from typing import Iterator
 
 import numpy as np
 
+import numpy as np
+
 
 def ensure_repo_on_path(file: str, parents_up: int = 3) -> Path:
     """Add the repository root inferred from ``file`` to ``sys.path``."""
@@ -134,3 +136,158 @@ def save_state_comparison_plot(
     fig.tight_layout()
     fig.savefig(out_path, dpi=160)
     plt.close(fig)
+
+
+# ---------------------------------------------------------------------------
+# Backend sweep benchmark plotting
+# ---------------------------------------------------------------------------
+
+def extrapolate_wall(ns_measured: list[int], times: list[float],
+                     ns_extrap: list[int]) -> tuple[np.ndarray, np.ndarray]:
+    """
+    Fit  t = slope * n  on the last ≤3 measured CPU points and evaluate at
+    ns_extrap.  CPU sweep time is linear in n_samples after JIT warmup.
+    """
+    ns    = np.array(ns_measured[-3:], dtype=float)
+    ts    = np.array(times[-3:],       dtype=float)
+    slope = float(np.mean(ts / ns))
+    return np.array(ns_extrap, dtype=float), slope * np.array(ns_extrap, dtype=float)
+
+
+def plot_sweep_benchmark(
+    cpu_rows: list[dict],
+    cuda_rows: list[dict],
+    model_name: str,
+    n_nodes: int,
+    duration: float,
+    dt: float,
+    n_workers: int,
+    repeats: int,
+    out_dir: Path,
+) -> None:
+    """
+    Three-panel benchmark figure.
+
+    Panel 1 — Wall time (log-log):
+        Measured CPU lines + dotted extrapolation to the CUDA x-range.
+        CUDA measured points.
+
+    Panel 2 — CPU speedup vs NumPy (linear scale).
+
+    Panel 3 — Throughput in samples/s (log-log):
+        CPU plateau extrapolated with dotted lines.
+        CUDA measured; crossover vs C++ parallel annotated.
+
+    Parameters
+    ----------
+    cpu_rows  : list of dicts with keys n, t_np, t_nb1, t_nbN, t_cs, t_cp
+    cuda_rows : list of dicts with keys n, t_cuda, rate
+    """
+    try:
+        import matplotlib.pyplot as plt
+    except ImportError:
+        print("matplotlib not available — skipping plot.")
+        return
+
+    cpu_max_n  = max(r["n"] for r in cpu_rows)  if cpu_rows  else 64
+    cuda_max_n = max(r["n"] for r in cuda_rows) if cuda_rows else cpu_max_n
+    extrap_ns  = sorted({n for n in [128, 256, 512, 1024, 2048, 4096, 8192]
+                          if cpu_max_n < n <= cuda_max_n * 2})
+
+    STYLES = [
+        # (row_key, label, color, linestyle, marker, markerfacecolor)
+        ("t_np",  "NumPy serial",          "tab:blue",   "-",  "o", None),
+        ("t_nb1", "Numba 1T",              "tab:orange", "--", "s", None),
+        ("t_nbN", f"Numba {n_workers}T",   "tab:orange", "-",  "s", "white"),
+        ("t_cs",  "C++ serial",            "tab:green",  "--", "^", None),
+        ("t_cp",  f"C++ par {n_workers}T", "tab:green",  "-",  "D", "white"),
+    ]
+
+    fig, axes = plt.subplots(1, 3, figsize=(18, 5))
+    fig.suptitle(
+        f"Sweep benchmark — {model_name}  |  {n_nodes} nodes  |  "
[…]
