+++ docs/examples/simulator_models/benchmark_all_backends.py
@@ -45,7 +45,11 @@ warnings.filterwarnings(
 )
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
-from helpers import ensure_repo_on_path, complete_graph_weights
+from helpers import (
+    ensure_repo_on_path,
+    complete_graph_weights,
+    plot_sweep_benchmark,
+)
 ensure_repo_on_path(__file__)
 
 import numba
@@ -73,18 +77,21 @@ MODELS = {
     "mpr":                   (mpr,                  "eta",  (-5.5,-4.0), 0.01, 0.1),
 }
 
+# ---- Last validated configuration ----
+# SupHopf, 20 nodes, dt=0.1, duration=500ms  →  runs in ~3 min (repeats=2)
+# For brain-scale: --model mpr --n-nodes 80 --cuda-sizes 64 128 512 2048 4096 --repeats 1
 DEFAULT_MODEL    = "sup_hopf"
 DEFAULT_N_NODES  = 20
-DEFAULT_DURATION = 200.0   # ms  — short enough that NumPy finishes in seconds
-DEFAULT_DT       = 0.1     # dt=0.1 ms → 2000 steps per sim; fast on all backends
+DEFAULT_DURATION = 500.0   # ms  — 5000 steps per simulation (dt=0.1)
+DEFAULT_DT       = 0.1     # dt=0.1 ms → fast on all backends
 
-# CPU sweep sizes — linear cost; NumPy gets expensive quickly, so keep ≤ 64
+# CPU: kept ≤ 64 so NumPy finishes in seconds (cost is linear in n_samples)
 DEFAULT_CPU_SIZES  = [4, 8, 16, 32, 64]
-# CUDA sweep sizes — start at 32 (overlap), go to 2048 where GPU dominates
+# CUDA: starts at 32 (overlaps CPU), extends to 2048 where GPU starts winning
 DEFAULT_CUDA_SIZES = [32, 64, 128, 256, 512, 1024, 2048]
 
 N_WORKERS = int(os.environ.get("VBI_BENCH_THREADS", str(max(2, (os.cpu_count() or 4) // 2))))
-N_REPEATS = 3
+N_REPEATS = 2
 
 
 # ---------------------------------------------------------------------------
@@ -288,99 +295,25 @@ def throughput_comparison(cpu_rows: list[dict], cuda_rows: list[dict],
 
 
 # ---------------------------------------------------------------------------
-# Plotting
+# Plotting  — delegated to helpers.plot_sweep_benchmark
 # ---------------------------------------------------------------------------
 
 def plot_results(cpu_rows: list[dict], cuda_rows: list[dict],
                  model_name: str, n_nodes: int, duration: float,
-                 n_workers: int, out_dir: Path) -> None:
-    try:
-        import matplotlib.pyplot as plt
-    except ImportError:
-        print("matplotlib not available — skipping plot.")
-        return
-
-    fig, axes = plt.subplots(1, 3, figsize=(18, 5))
-    fig.suptitle(
-        f"Sweep benchmark — model: {model_name}  |  {n_nodes} nodes  |  "
-        f"{duration} ms  |  CPU {n_workers} threads",
-        fontsize=11,
+                 n_workers: int, repeats: int, dt: float,
+                 out_dir: Path) -> None:
+    """Delegate to helpers.plot_sweep_benchmark."""
+    plot_sweep_benchmark(
+        cpu_rows=cpu_rows,
+        cuda_rows=cuda_rows,
+        model_name=model_name,
+        n_nodes=n_nodes,
+        duration=duration,
+        dt=dt,
+        n_workers=n_workers,
+        repeats=repeats,
[…]
