### 2026-05-20 fefc8ce
feat: Add benchmark script for model sweep and improve plotting functionality
