+++ docs/examples/simulator_models/benchmark_all_models.py
@@ -1,0 +1,231 @@
+"""
+Run the sweep benchmark for every model defined in MODELS and save one
+figure per model.
+
+Each model is benchmarked with the same network size and sweep ranges but
+its own coupling strength (as configured in MODELS).  The Numba JIT and C++
+compilation happen once per model; CUDA JIT is cached to disk after the first
+model.
+
+Usage
+-----
+    python benchmark_all_models.py                      # all defaults
+    python benchmark_all_models.py --n-nodes 40         # larger network
+    python benchmark_all_models.py --no-cuda            # skip GPU
+    python benchmark_all_models.py --models sup_hopf mpr  # subset
+
+Outputs
+-------
+    outputs/benchmark_all_backends_{model_name}.json
+    outputs/benchmark_all_backends_{model_name}.png
+    outputs/benchmark_all_models_summary.json           # cross-model table
+"""
+from __future__ import annotations
+
+import argparse
+import json
+import sys
+import time
+from pathlib import Path
+
+import numpy as np
+
+# ---- repo path ----
+sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+from helpers import ensure_repo_on_path, plot_sweep_benchmark
+ensure_repo_on_path(__file__)
+
+# Import benchmark primitives from the sibling script
+from benchmark_all_backends import (
+    MODELS,
+    DEFAULT_N_NODES,
+    DEFAULT_DURATION,
+    DEFAULT_DT,
+    DEFAULT_CPU_SIZES,
+    DEFAULT_CUDA_SIZES,
+    N_WORKERS,
+    _make_spec,
+    warmup,
+    benchmark_cpu,
+    benchmark_cuda,
+    throughput_comparison,
+    save_json,
+)
+
+OUTPUT_DIR = Path(__file__).parent / "outputs"
+
+
+# ---------------------------------------------------------------------------
+# Per-model run
+# ---------------------------------------------------------------------------
+
+def run_one_model(
+    model_name: str,
+    n_nodes: int,
+    duration: float,
+    dt: float,
+    cpu_sizes: list[int],
+    cuda_sizes: list[int],
+    n_workers: int,
+    repeats: int,
+    no_cuda: bool,
+    out_dir: Path,
+) -> dict:
+    """Benchmark a single model, save JSON + figure, return summary dict."""
+    print()
+    print("=" * 70)
+    print(f"  Model: {model_name}")
+    print("=" * 70)
+
+    spec = _make_spec(model_name, n_nodes, dt)
+    warmup(spec, model_name, n_workers)
+
+    # CPU
+    print(f"\n{'='*70}")
+    print("  CPU backends")
+    print(f"{'='*70}")
+    cpu_rows = benchmark_cpu(spec, model_name, cpu_sizes, duration,
+                             n_workers, repeats)
+
+    # CUDA
+    cuda_rows: list[dict] = []
+    if not no_cuda:
+        print(f"\n{'='*70}")
+        print("  CUDA backend")
+        print(f"{'='*70}")
+        cuda_rows = benchmark_cuda(spec, model_name, cuda_sizes,
+                                   duration, repeats)
+
+    # Summary table
+    throughput_comparison(cpu_rows, cuda_rows, n_workers)
+
+    # Save JSON
+    meta = dict(
+        model=model_name, n_nodes=n_nodes,
+        duration=duration, dt=dt,
+        n_workers=n_workers, repeats=repeats,
+    )
+    save_json(cpu_rows, cuda_rows, meta,
+              out_dir / f"benchmark_all_backends_{model_name}.json")
+
[…]
