### 2026-05-20 6174f49
feat: Expand model coverage in CUDA tests and add bounds compliance checks
