### 2026-05-20 18aea59
feat: Add custom pytest markers for CUDA tests and enhance test descriptions
