+++ vbi/tests/validation/test_mpr_cuda.py
@@ -4,17 +4,30 @@ M3 validation: Numba-CUDA backend.
 All tests are skipped when no CUDA device is present.
 
 Gold standard: Numba CPU backend (validated against NumPy in test_mpr_numba.py).
-GPU uses float32; CPU uses float64.  Accepted tolerance: rtol=1e-3.
+GPU uses float32; CPU uses float64.  Accepted tolerance: rtol=5e-3.
+
+Speed tiers
+-----------
+Normal (default): full coverage, longer durations, more sweep points.
+
+Fast mode  (``pytest -m cuda_fast``):
+    Deterministic only, short durations (≤10 ms), ≤8 sweep points.
+    Skips stochastic, throughput, and pipeline tests.
+    Useful for quick CI checks.
 
 Tests
 -----
-test_cuda_single_matches_numba    — single run, deterministic
-test_cuda_stoch_runs              — stochastic run, finite output
-test_cuda_single_with_delays      — non-zero tract lengths
-test_cuda_sweep_matches_numba     — sweep, deterministic (multi-model)
-test_cuda_sweep_stoch_unique      — stochastic sweep, unique trajectories
-test_cuda_sweep_pipeline_shape    — pipeline mode, shape + finite
-test_cuda_sweep_throughput        — benchmark, not a pass/fail test
+test_cuda_single_matches_numba     — single run, deterministic, MPR
+test_cuda_single_model_isolated    — all models, isolated (coup_a=0)
+test_cuda_single_model_coupled_finite — all models, coupled, finite check
+test_cuda_stoch_runs               — stochastic run, finite output
+test_cuda_single_with_delays       — non-zero tract lengths
+test_cuda_sweep_matches_numba      — sweep, deterministic (selected models)
+test_cuda_sweep_new_models_finite  — sweep, chaotic/unbounded models
+test_cuda_sweep_stoch_unique       — stochastic sweep, unique trajectories
+test_cuda_sweep_80_nodes_finite    — brain-scale N=80
+test_cuda_sweep_pipeline_shape     — pipeline mode, shape + finite
+test_cuda_sweep_throughput         — benchmark, informational only
 """
 import time
 
@@ -27,9 +40,9 @@ try:
 except Exception:
     CUDA_AVAILABLE = False
 
-cuda_only = pytest.mark.skipif(
-    not CUDA_AVAILABLE, reason="no CUDA device available"
-)
+cuda_only  = pytest.mark.skipif(not CUDA_AVAILABLE, reason="no CUDA device available")
+cuda_fast  = pytest.mark.cuda_fast    # mark for -m cuda_fast selection
+cuda_slow  = pytest.mark.cuda_slow    # mark for exclusion in fast mode
 
 from vbi.simulator import Simulator, Sweeper
 from vbi.simulator.spec import (
@@ -88,10 +101,11 @@ def _spec(model=mpr, n_nodes=10, dt=0.01, coup_a=0.1,
 # ---------------------------------------------------------------------------
 
 @cuda_only
+@cuda_fast
 def test_cuda_single_matches_numba():
     """CUDA single run must match Numba CPU to rtol=1e-3 (float32 vs float64)."""
     spec     = _spec(mpr, n_nodes=10, dt=0.01)
-    duration = 100.0
+    duration = 10.0   # short: enough for correctness, fast to run
 
     _, d_nb   = Simulator(spec, backend="numba").run(duration)["raw"]
[…]
