### 2026-05-20 3b46e1a
fix: Update mako import handling in C++ codegen to ensure proper error messaging
