+++ pyproject.toml
@@ -1,142 +1,197 @@
 [build-system]
 requires = [
-    "hatchling",
-    "setuptools", 
+    "setuptools>=61",
     "wheel",
-    "swig>=4.0",
-    "setuptools>=45"
+    "setuptools-scm",
 ]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "vbi"
 dynamic = ["version"]
-description = "Virtual brain inference."
-requires-python = ">=3.8"
+description = "Virtual brain inference — fast simulation and parameter estimation."
+requires-python = ">=3.10"
 license = { text = "MIT" }
 authors = [
-    { name = "Abolfazl Ziaeemehr", email = "a.ziaeemehr@gmail.com" },
-    { name = "Meysam Hashemi", email = "meysam.hashemi@gmail.com" },
-    { name = "Marmaduke Woodman", email = "marmaduke.woodman@gmail.com" }
+    { name = "Abolfazl Ziaeemehr",  email = "a.ziaeemehr@gmail.com" },
+    { name = "Meysam Hashemi",      email = "meysam.hashemi@gmail.com" },
+    { name = "Marmaduke Woodman",   email = "marmaduke.woodman@gmail.com" },
 ]
 readme = "README.md"
-dependencies = [
-    "wheel",
-    "numpy",
-    "scipy",
-    "numba",
-    "h5py",
-    "pandas",
-    "networkx",
-    "nbconvert",
-    "matplotlib",
-    "setuptools-scm",
-    "tqdm",
-    "autograd",
-    "parameterized",
-    "scikit-learn",
-    "pytest",
-    "swig",
-    "rich",
-    "corner"
-]
 classifiers = [
     "Programming Language :: Python :: 3",
     "Topic :: Scientific/Engineering :: Information Analysis",
-    "License :: OSI Approved :: Apache Software License",
+    "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
 
+# ---------------------------------------------------------------------------
+# Core dependencies — always required (light, CPU-only simulation)
+# ---------------------------------------------------------------------------
+# Install:  pip install vbi
+#
+# Provides: NumPy + Numba backends (no C++, no GPU, no inference)
+# ---------------------------------------------------------------------------
+dependencies = [
+    "numpy>=1.23",
+    "scipy>=1.9",
+    "numba>=0.57",
+    "h5py",
+    "pandas",
+    "matplotlib",
+    "tqdm",
+    "scikit-learn",
+]
+
 [project.urls]
-homepage = "https://github.com/ins-amu/vbi"
+homepage   = "https://github.com/ins-amu/vbi"
 repository = "https://github.com/ins-amu/vbi"
+
+# ---------------------------------------------------------------------------
+# Optional dependency groups
+# ---------------------------------------------------------------------------
+
+[project.optional-dependencies]
+
+# ---- Backend accelerators (pick what you have) ----
+
+# C++ pybind11 simulator backend (auto-compiled at first use; requires g++/cmake)
+# Install:  pip install vbi[cpp]
+cpp = [
+    "mako>=1.2",
+    "pybind11>=2.10",
+]
+
+# JAX backend (CPU-only by default; for GPU: pip install jax[cuda12_pip] separately)
+# Install:  pip install vbi[jax]
+jax = [
+    "jax>=0.4",
+    "jaxlib>=0.4",
+]
+
+# CUDA / Numba-CUDA backend (numba[cuda] is the same package as numba)
[…]
