### 2026-05-20 b32e1fe
feat: Update GitHub Actions workflow and pytest markers for improved CI testing
