+++ vbi/tests/validation/test_new_models_numba.py
@@ -22,6 +22,9 @@ from vbi.simulator.models.coombes_byrne_2d import coombes_byrne_2d
 from vbi.simulator.models.gast_sd import gast_sd
 from vbi.simulator.models.gast_sf import gast_sf
 
+pytestmark = pytest.mark.slow
+
+
 numba = pytest.importorskip("numba", reason="numba not installed")
 
 
