### 2026-05-20 2b29ad6
feat: Add new dependencies for enhanced functionality and testing support
