### 2026-05-20 e28c2ea
feat: Implement JAX backend with simulator and sweeper, including validation tests
