+++ vbi/simulator/backend/jax_/simulator.py
@@ -1,0 +1,544 @@
+"""
+JAX backend — differentiable simulator using jax.lax.scan + jax.vmap.
+
+Design notes
+------------
+Ring buffer
+    Shape (horizon, n_cvar, n_nodes); stored in lax.scan carry.
+    Write: buf.at[step % horizon].set(new_cvar)
+    Read:  flat-index trick → buf[:, cv, :].reshape(-1)[idx_time*N + j]
+           avoids Python loops inside traced code.
+
+Time loop
+    Nested scan:
+        outer scan  → n_record iterations, one monitor sample per iteration
+        inner scan  → record_period steps each (subsample / tavg)
+    For raw monitor: record_period = 1, inner scan is trivial.
+    All carry shapes are static (JAX requirement).
+
+Noise
+    jax.random.fold_in(master_key, step) — no key splitting, cheap,
+    reproducible across runs; adapted from bold_delay_gpu.py.
+    master_key is static (not part of JIT-recompile trigger).
+
+Sweep (JaxSweeper)
+    Swept params are collected into a dict of shape-(n_samples,) arrays.
+    jax.vmap maps axis-0 of those arrays → one scalar per run inside the
+    vmapped function.  jax.jit wraps the vmapped call.
+    Non-swept data (weights, delays, base_params, etc.) goes into a
+    static pytree shared across all runs.
+"""
+from __future__ import annotations
+
+from typing import Callable
+
+import jax
+import jax.numpy as jnp
+import numpy as np
+import pandas as pd
+
+from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.spec.sweep import SweepSpec
+from .codegen import build_jax_dfun
+
+
+# ---------------------------------------------------------------------------
+# Parameter helpers
+# ---------------------------------------------------------------------------
+
+def _build_jax_params(spec: SimulationSpec) -> dict:
+    """Return a plain dict of scalar/array params (numpy → python floats)."""
+    params = dict(spec.model.default_params)
+    params.update(spec.node_params)
+    return params
+
+
+def _build_initial_state(spec: SimulationSpec) -> jnp.ndarray:
+    n_nodes = spec.n_nodes
+    state = np.zeros((spec.model.n_sv, n_nodes), dtype=np.float32)
+    for i, sv in enumerate(spec.model.state_variables):
+        state[i] = sv.default_init
+    return jnp.array(state)
+
+
+def _build_bounds(spec: SimulationSpec) -> tuple[jnp.ndarray, jnp.ndarray]:
+    lo = np.array(
+        [sv.lower_bound if sv.lower_bound is not None else -np.inf
+         for sv in spec.model.state_variables],
+        dtype=np.float32,
+    )
+    hi = np.array(
+        [sv.upper_bound if sv.upper_bound is not None else np.inf
+         for sv in spec.model.state_variables],
+        dtype=np.float32,
+    )
+    return jnp.array(lo), jnp.array(hi)
+
+
+def _build_noise_amp(spec: SimulationSpec) -> jnp.ndarray:
+    """(n_sv,) amplitude array — zero for non-noisy state vars."""
+    n_sv = spec.model.n_sv
+    amp = np.zeros(n_sv, dtype=np.float32)
+    ni = list(spec.model.noise_indices)
+    nsig = spec.integrator.noise_nsig
+    if nsig is None:
+        nsig = np.ones(len(ni), dtype=np.float32) * 1e-3
+    nsig = np.asarray(nsig, dtype=np.float32)
+    style = getattr(spec.integrator, "noise_style", "amplitude")
+    if style == "tvb":
+        nsig = np.sqrt(2.0 * nsig)
+    amp[ni] = nsig
+    return jnp.array(amp)
+
+
[…]
