+++ vbi/tests/validation/test_mpr_jax.py
@@ -1,0 +1,306 @@
+"""
+M4 validation: JAX backend — single-run and sweep.
+
+Gold standard: NumPy baseline (validated against TVB in test_mpr_numpy.py).
+
+Tolerance note: JAX uses float32 by default (matching the GPU/CUDA backend),
+while NumPy uses float64.  After N steps the accumulated rounding error is
+O(N * eps_float32) ≈ O(N * 1e-7).  Tests here use short durations (≤ 200 ms)
+to keep accumulated error well below rtol=1e-2.
+
+For long-run comparisons the stochastic moment tests (mean/std) are used
+instead of trajectory matching.
+"""
+import time
+
+import numpy as np
+import pytest
+
+from vbi.simulator import Simulator, Sweeper
+from vbi.simulator.spec import (
+    CouplingSpec, IntegratorSpec, MonitorSpec, SimulationSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+from vbi.simulator.models.mpr import mpr
+from .conftest import make_mpr_spec, make_weights
+
+pytestmark = pytest.mark.slow
+
+jax = pytest.importorskip("jax", reason="jax not installed")
+jnp = jax.numpy
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _np_sim(spec, duration):
+    return Simulator(spec, backend="numpy").run(duration)
+
+
+def _jx_sim(spec, duration):
+    return Simulator(spec, backend="jax").run(duration)
+
+
+# ---------------------------------------------------------------------------
+# Single-run: deterministic — trajectory comparison
+# ---------------------------------------------------------------------------
+
+class TestDeterministicMatchesNumPy:
+    """
+    JAX (float32) vs NumPy (float64) on short runs.
+    rtol=1e-2 accommodates float32 accumulation over 200 ms / 0.1 ms = 2 000 steps.
+    """
+
+    @pytest.mark.parametrize("n_nodes", [1, 2, 10])
+    @pytest.mark.parametrize("method", ["heun", "euler"])
+    def test_raw_matches_numpy(self, n_nodes, method):
+        spec = make_mpr_spec(
+            n_nodes=n_nodes, dt=0.1, method=method,
+            monitors=(MonitorSpec("raw"),),
+        )
+        _, d_np = _np_sim(spec, 200.0)["raw"]
+        _, d_jx = _jx_sim(spec, 200.0)["raw"]
+        assert d_np.shape == d_jx.shape, "shape mismatch"
+        np.testing.assert_allclose(
+            d_jx, d_np, rtol=1e-2,
+            err_msg=f"JAX raw diverges from NumPy (n_nodes={n_nodes}, {method})",
+        )
+
+    def test_r_stays_nonnegative(self):
+        spec = make_mpr_spec(
+            n_nodes=10, dt=0.1,
+            monitors=(MonitorSpec("raw"),),
+        )
+        _, d_jx = _jx_sim(spec, 200.0)["raw"]
+        assert np.all(d_jx[:, 0, :] >= 0.0), "r violated lower bound (r >= 0)"
+
+    def test_subsample_matches_numpy(self):
+        spec = make_mpr_spec(
+            n_nodes=4, dt=0.1,
+            monitors=(MonitorSpec("subsample", period=1.0),),
+        )
+        _, d_np = _np_sim(spec, 200.0)["subsample"]
+        _, d_jx = _jx_sim(spec, 200.0)["subsample"]
+        assert d_np.shape == d_jx.shape
+        np.testing.assert_allclose(d_jx, d_np, rtol=1e-2)
+
[…]
