### 2026-05-20 4059d25
feat: Add Balloon-Windkessel model and BOLD monitor tests for JAX backend
