### 2026-05-20 1de7a63
Add JAX backend support for sweep validation and model simulations
- Updated `validate_sweep_backends.py` to include JAX as a backend for numerical consistency checks, with adjustable tolerance levels for JAX.
- Enhanced JAX integration in `jax_/simulator.py` to support float64 precision and noise handling.
- Modified `IntegratorSpec` to allow specification of JAX data types (float32 or float64).
- Updated `SweepSpec` to include a `same_noise` parameter for controlling noise realization consistency across JAX runs.
- Added comprehensive tests in `test_mpr_jax.py` and `test_new_models_jax.py` t
[…]
