+++ vbi/simulator/backend/jax_/simulator.py
@@ -86,37 +86,50 @@ def _build_jax_params(spec: SimulationSpec) -> dict:
     return params
 
 
+def _resolve_dtype(spec: SimulationSpec):
+    """Return the numpy/jax dtype from IntegratorSpec.jax_dtype."""
+    jax_dtype = getattr(spec.integrator, "jax_dtype", "float32")
+    if jax_dtype == "float64":
+        import jax
+        jax.config.update("jax_enable_x64", True)
+        return np.float64
+    return np.float32
+
+
 def _build_initial_state(spec: SimulationSpec) -> jnp.ndarray:
+    dtype = _resolve_dtype(spec)
     n_nodes = spec.n_nodes
-    state = np.zeros((spec.model.n_sv, n_nodes), dtype=np.float32)
+    state = np.zeros((spec.model.n_sv, n_nodes), dtype=dtype)
     for i, sv in enumerate(spec.model.state_variables):
         state[i] = sv.default_init
     return jnp.array(state)
 
 
 def _build_bounds(spec: SimulationSpec) -> tuple[jnp.ndarray, jnp.ndarray]:
+    dtype = _resolve_dtype(spec)
     lo = np.array(
         [sv.lower_bound if sv.lower_bound is not None else -np.inf
          for sv in spec.model.state_variables],
-        dtype=np.float32,
+        dtype=dtype,
     )
     hi = np.array(
         [sv.upper_bound if sv.upper_bound is not None else np.inf
          for sv in spec.model.state_variables],
-        dtype=np.float32,
+        dtype=dtype,
     )
     return jnp.array(lo), jnp.array(hi)
 
 
 def _build_noise_amp(spec: SimulationSpec) -> jnp.ndarray:
     """(n_sv,) amplitude array — zero for non-noisy state vars."""
+    dtype = _resolve_dtype(spec)
     n_sv = spec.model.n_sv
-    amp = np.zeros(n_sv, dtype=np.float32)
+    amp = np.zeros(n_sv, dtype=dtype)
     ni = list(spec.model.noise_indices)
     nsig = spec.integrator.noise_nsig
     if nsig is None:
-        nsig = np.ones(len(ni), dtype=np.float32) * 1e-3
-    nsig = np.asarray(nsig, dtype=np.float32)
+        nsig = np.ones(len(ni), dtype=dtype) * 1e-3
+    nsig = np.asarray(nsig, dtype=dtype)
     style = getattr(spec.integrator, "noise_style", "amplitude")
     if style == "tvb":
         nsig = np.sqrt(2.0 * nsig)
@@ -339,8 +352,8 @@ def _run_monitor(
         # One output per integration step — simple scan
         _, all_states = jax.lax.scan(
             step_fn, init_carry, None, length=n_steps)
-        # all_states: (n_steps, n_sv, n_nodes)
-        times = jnp.arange(n_steps, dtype=jnp.float32) * dt
+        # all_states: (n_steps, n_sv, n_nodes) — dtype follows state0
+        times = jnp.arange(n_steps, dtype=all_states.dtype) * dt
         return times, all_states
 
     # subsample / tavg / gavg — outer scan, inner scan of istep steps
@@ -389,7 +402,8 @@ class JaxSimulator:
         self._lo, self._hi = _build_bounds(spec)
 
         G = float(self._params.get("G", 1.0))
-        self._weights = jnp.array(spec.weights, dtype=jnp.float32)
+        self._dtype = _resolve_dtype(spec)
+        self._weights = jnp.array(spec.weights, dtype=self._dtype)
         self._delay_steps = jnp.array(spec.delay_steps(dt), dtype=jnp.int32)
[…]
