### 2026-05-21 c9f696a
feat: Enhance benchmarking scripts to include JAX support for performance comparison
