+++ docs/examples/simulator_models/helpers.py
@@ -162,35 +162,50 @@ def plot_sweep_benchmark(
     n_workers: int,
     repeats: int,
     out_dir: Path,
+    jax_rows: list[dict] | None = None,
+    all_cpu_rows: list[dict] | None = None,
 ) -> None:
     """
     Three-panel benchmark figure.
 
     Panel 1 — Wall time (log-log):
-        Measured CPU lines + dotted extrapolation to the CUDA x-range.
-        CUDA measured points.
+        Measured CPU lines + dotted extrapolation to the GPU x-range.
+        CUDA and JAX measured points.
 
     Panel 2 — CPU speedup vs NumPy (linear scale).
 
     Panel 3 — Throughput in samples/s (log-log):
         CPU plateau extrapolated with dotted lines.
-        CUDA measured; crossover vs C++ parallel annotated.
+        CUDA and JAX measured; crossover vs C++ parallel annotated.
 
     Parameters
     ----------
-    cpu_rows  : list of dicts with keys n, t_np, t_nb1, t_nbN, t_cs, t_cp
-    cuda_rows : list of dicts with keys n, t_cuda, rate
+    cpu_rows     : rows with t_np present — used for panels 1 and 2
+    cuda_rows    : list of dicts with keys n, t_cuda, rate
+    jax_rows     : list of dicts with keys n, t_jax, rate_jax, platform (optional)
+    all_cpu_rows : full CPU rows including extended sizes without t_np —
+                   used for panel 3 throughput.  Falls back to cpu_rows if None.
     """
+    jax_rows    = jax_rows    or []
+    # Panel 3 uses all_cpu_rows (may include large-n rows missing t_np)
+    thru_rows   = all_cpu_rows if all_cpu_rows is not None else cpu_rows
     try:
         import matplotlib.pyplot as plt
     except ImportError:
         print("matplotlib not available — skipping plot.")
         return
 
-    cpu_max_n  = max(r["n"] for r in cpu_rows)  if cpu_rows  else 64
-    cuda_max_n = max(r["n"] for r in cuda_rows) if cuda_rows else cpu_max_n
+    cpu_max_n  = max(r["n"] for r in cpu_rows)   if cpu_rows  else 64
+    gpu_max_n  = max(
+        (max(r["n"] for r in cuda_rows) if cuda_rows else 0),
+        (max(r["n"] for r in jax_rows)  if jax_rows  else 0),
+    ) or cpu_max_n
     extrap_ns  = sorted({n for n in [128, 256, 512, 1024, 2048, 4096, 8192]
-                          if cpu_max_n < n <= cuda_max_n * 2})
+                          if cpu_max_n < n <= gpu_max_n * 2})
+
+    # JAX platform label for legend
+    jax_platform = jax_rows[0].get("platform", "GPU") if jax_rows else "GPU"
+    jax_label    = f"JAX float32 ({jax_platform.upper()})"
 
     STYLES = [
         # (row_key, label, color, linestyle, marker, markerfacecolor)
@@ -225,6 +240,10 @@ def plot_sweep_benchmark(
         gn = [r["n"] for r in cuda_rows]
         ax.plot(gn, [r["t_cuda"] for r in cuda_rows], "P-",
                 label="CUDA (float32)", color="tab:red", lw=2, zorder=4)
+    if jax_rows:
+        jn = [r["n"] for r in jax_rows]
+        ax.plot(jn, [r["t_jax"] for r in jax_rows], "h-",
+                label=jax_label, color="tab:purple", lw=2, zorder=4)
     if extrap_ns:
         ax.text(0.97, 0.03, "··· = CPU extrapolation (linear fit)",
                 transform=ax.transAxes, ha="right", va="bottom",
[…]
