### 2026-05-21 8b6faac
feat: Refactor JAX backend to import JaxSweeper from the correct module and add JaxSweeper class implementation
