### 2026-05-21 5e4a809
feat: Add VEP, GHB, SL, and DampedOscillator models with Laplacian coupling to simulator
