### 2026-05-21 81e8286
Add equations reference notebook and enhance model specifications with LaTeX support
- Created a new Jupyter notebook `model_equations.ipynb` to describe equations for all models in `vbi.simulator.models`.
- Enhanced `ModelSpec` to include `dfun_latex` for LaTeX representations of differential equations.
- Updated existing models to provide LaTeX strings for their equations, improving documentation and usability in Jupyter notebooks.
- Added supplementary notes in LaTeX format for clarity on model parameters and equations.
