+++ vbi/simulator/spec/model.py
@@ -20,6 +20,15 @@ class Parameter:
     description: str = ""
 
 
+def _fmt_default(v) -> str:
+    """Format a parameter default value for display."""
+    if isinstance(v, np.ndarray):
+        return f"array({v.shape})" if v.ndim > 0 else f"{float(v):.4g}"
+    if isinstance(v, float):
+        return f"{v:.4g}"
+    return str(v)
+
+
 @dataclass(frozen=True)
 class ModelSpec:
     """
@@ -30,6 +39,10 @@ class ModelSpec:
     (coupling input, scalar per node), and the math functions:
         exp, log, sin, cos, tanh, sqrt, abs, pi
     No 'np.' prefix — the code generator injects the namespace.
+
+    dfun_latex maps each state variable name to a LaTeX string for the RHS
+    of its differential equation.  Used by describe() to render equations in
+    Jupyter notebooks.  Falls back to dfun_str if not provided.
     """
     name: str
     state_variables: tuple[StateVar, ...]
@@ -38,6 +51,9 @@ class ModelSpec:
     dfun_str: dict[str, str]      # {sv_name: expression_string}
     noise_variables: tuple[str, ...] = ()
     reference: str = ""
+    dfun_latex: dict[str, str] = field(default_factory=dict)
+    # Optional LaTeX block rendered below the equations (definitions, where-clauses).
+    latex_notes: str = ""
 
     @property
     def sv_names(self) -> tuple[str, ...]:
@@ -64,3 +80,89 @@ class ModelSpec:
     @property
     def default_params(self) -> dict[str, float | np.ndarray]:
         return {p.name: p.default for p in self.parameters}
+
+    def describe(self) -> None:
+        """
+        Render model equations, parameters, and state variables.
+
+        In Jupyter notebooks the output is typeset with MathJax (Markdown +
+        LaTeX).  In a plain Python session it prints a text summary instead.
+
+        LaTeX equations use the strings stored in ``dfun_latex``; models that
+        do not provide ``dfun_latex`` fall back to the raw ``dfun_str``
+        expressions wrapped in a code block.
+
+        Example
+        -------
+        >>> from vbi.simulator.models import mpr
+        >>> mpr.describe()
+        """
+        def _in_notebook() -> bool:
+            try:
+                from IPython import get_ipython
+                shell = get_ipython()
+                return shell is not None and shell.__class__.__name__ in (
+                    "ZMQInteractiveShell", "TerminalInteractiveShell"
+                )
+            except ImportError:
+                return False
+
+        if _in_notebook():
+            from IPython.display import display, Markdown
+            def _render(md: str) -> None:
+                display(Markdown(md))
+        else:
+            def _render(md: str) -> None:
+                # Plain-text fallback: strip markdown/LaTeX decoration
+                import re
+                txt = md
+                txt = re.sub(r"\$\$(.+?)\$\$", lambda m: f"  {m.group(1)}", txt, flags=re.S)
+                txt = txt.replace("**", "").replace("`", "")
+                print(txt)
+
+        lines: list[str] = []
+
+        # --- Title & reference ---
+        lines += [f"## {self.name}", ""]
+        if self.reference:
+            lines += [f"*{self.reference}*", ""]
+
+        # --- Equations ---
+        lines += ["---", "**Equations**", ""]
+        has_latex = bool(self.dfun_latex)
+        for sv_name in self.sv_names:
+            if has_latex and sv_name in self.dfun_latex:
+                rhs = self.dfun_latex[sv_name]
+                lines.append(f"$$\\dot{{{sv_name}}} = {rhs}$$")
+            else:
[…]
