### 2026-05-21 c2375a1
feat: Enhance equation rendering in Jupyter notebooks with aligned LaTeX support
