### 2026-05-21 f324197
feat: Implement Kuramoto coupling support in Numba and Numpy backends
