### 2026-05-21 d58bb21
feat: Add global coupling strength parameter to Kuramoto model specification
