+++ docs/examples/simulator_models/kuramoto_demo.py
@@ -1,0 +1,322 @@
+"""
+Kuramoto model: VBI simulator vs pure-NumPy reference.
+
+Verifies that the VBI KuramotoCoupling produces trajectories identical to a
+single self-contained NumPy implementation using the same Heun integrator.
+
+Both implementations solve:
+
+    dθ_i/dt = ω_i + (G/N) Σ_j W_{ij} sin(θ_j − θ_i)
+
+Run
+---
+    python kuramoto_demo.py                  # 5-node complete graph, 500 ms
+    python kuramoto_demo.py --n 10 --duration 1000
+    python kuramoto_demo.py --delayed        # add 5 ms axonal delay
+"""
+
+from __future__ import annotations
+
+import argparse
+from pathlib import Path
+import sys
+
+sys.dont_write_bytecode = True
+
+import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from helpers import ensure_repo_on_path, complete_graph_weights
+
+ensure_repo_on_path(__file__)
+
+from vbi.simulator import Simulator
+from vbi.simulator.models.kuramoto import kuramoto
+from vbi.simulator.spec.coupling import CouplingSpec
+from vbi.simulator.spec.integrator import IntegratorSpec
+from vbi.simulator.spec.monitor import MonitorSpec
+from vbi.simulator.spec.simulation import SimulationSpec
+
+
+# ---------------------------------------------------------------------------
+# Pure-NumPy reference (single function, no VBI dependency)
+# ---------------------------------------------------------------------------
+
+def kuramoto_heun(
+    theta0: np.ndarray,
+    omega: np.ndarray,
+    weights: np.ndarray,
+    G: float,
+    dt: float,
+    n_steps: int,
+    delays: np.ndarray | None = None,
+) -> tuple[np.ndarray, np.ndarray]:
+    """
+    Integrate the Kuramoto model with Heun's method (frozen-coupling convention).
+
+    This matches the VBI / TVB simulator convention: the coupling term
+    c_i = (G/N) Σ_j W_{ij} sin(θ_j − θ_i)  is computed once at the start
+    of each step from the current (or delayed) phases and held fixed for
+    both stages of Heun.  Because the Kuramoto dfun is entirely the coupling
+    term (dθ/dt = ω + c, no additional state-dependent nonlinearity), the
+    frozen-coupling Heun reduces to a first-order Euler step.
+
+    Parameters
+    ----------
+    theta0  : (N,)            initial phases [rad]
+    omega   : (N,)            natural frequencies [rad/ms]
+    weights : (N, N)          weight matrix, weights[tgt, src]
+    G       : float           global coupling strength
+    dt      : float           time step [ms]
+    n_steps : int             number of integration steps
+    delays  : (N, N) | None   axonal delays [ms]; None → zero delay
+
+    Returns
+    -------
+    t     : (n_steps,)    time vector
+    theta : (n_steps, N)  phase trajectories
+    """
+    N = len(theta0)
+    theta = np.empty((n_steps, N))
+    theta[0] = theta0.copy()
+
+    if delays is not None:
+        delay_steps = np.round(delays / dt).astype(int)
+        horizon = int(delay_steps.max()) + 2
+        # Fill all horizon slots with theta0 — matches VBI History.initialize()
+        buf = np.tile(theta0, (horizon, 1))  # (horizon, N)
+    else:
+        delay_steps = None
+        horizon = 1
[…]
