### 2026-05-21 97550e2
Add Kuramoto coupling support to simulation backends
- Enhanced C++ backend to support Kuramoto coupling alongside linear coupling.
- Updated JAX, Numba, and CUDA backends to implement Kuramoto coupling with phase frustration.
- Modified coupling specifications to include an alpha parameter for phase offset.
- Added tests to validate Kuramoto dynamics and ensure consistency across backends.
