### 2026-05-21 226bfe1
feat: Add StimSpec class for additive external stimuli in simulations
