+++ vbi/simulator/spec/__init__.py
@@ -3,6 +3,7 @@ from .integrator import IntegratorSpec
 from .coupling import CouplingSpec
 from .monitor import MonitorSpec
 from .simulation import SimulationSpec
+from .stimulus import StimSpec
 from .sweep import SweepSpec
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "CouplingSpec",
     "MonitorSpec",
     "SimulationSpec",
+    "StimSpec",
     "SweepSpec",
 ]
