### 2026-05-21 873b83c
feat: Add external stimulation demo for VBI NumPy backend
