### 2026-05-21 239e173
feat: Integrate stimulus handling into Numba backend simulations
