### 2026-05-21 173be63
feat: Implement stimulus handling in simulation backends with build_stim_data function
