+++ vbi/simulator/backend/jax_/sweeper.py
@@ -71,6 +71,7 @@ class JaxSweeper:
             coup_kind=spec.coupling.kind,
             coup_midpoint=float(getattr(spec.coupling, "midpoint", 0.0)),
             coup_sigma=float(getattr(spec.coupling, "sigma", 1.0)),
+            coup_alpha=float(getattr(spec.coupling, "alpha", 0.0)),
             use_heun=(spec.integrator.method == "heun"),
             stochastic=stochastic,
             noise_amp=self._noise_amp,
@@ -118,6 +119,7 @@ class JaxSweeper:
             coup_kind=spec.coupling.kind,
             coup_midpoint=float(getattr(spec.coupling, "midpoint", 0.0)),
             coup_sigma=float(getattr(spec.coupling, "sigma", 1.0)),
+            coup_alpha=float(getattr(spec.coupling, "alpha", 0.0)),
             use_heun=(spec.integrator.method == "heun"),
             stochastic=spec.integrator.stochastic,
             noise_amp=self._noise_amp,

[…]
