### 2026-05-21 58d5dc5
feat: Enhance cache key generation by incorporating template hash for C++ bindings
