### 2026-05-21 5f487ea
feat: Integrate stimulus data handling into JAX simulator and sweeper
