### 2026-05-21 3e8cf70
feat: Add comprehensive stimulation tests for backend consistency and stimulus effects
