+++ vbi/tests/validation/test_stimulation.py
@@ -1,0 +1,294 @@
+"""
+Stimulation tests across backends.
+
+Verifies that StimSpec injection works correctly and consistently:
+  - stimulus physically affects the target node
+  - onset/offset gating is respected
+  - pre-sampled waveform (array) and callable waveform give the same result
+  - numpy, numba, JAX and C++ backends all agree
+  - no-stimulus baseline is unaffected (zero stim = no change)
+  - frustrated Kuramoto (alpha) with stimulus
+"""
+import dataclasses
+
+import numpy as np
+import pytest
+
+from vbi.simulator import Simulator
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec, StimSpec,
+)
+from vbi.simulator.models.kuramoto import kuramoto
+from vbi.simulator.models.mpr import mpr
+
+pytestmark = pytest.mark.slow
+
+numba = pytest.importorskip("numba", reason="numba not installed")
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _weights(n, seed=0):
+    rng = np.random.default_rng(seed)
+    W = np.abs(rng.standard_normal((n, n)))
+    np.fill_diagonal(W, 0.0)
+    W /= W.sum(axis=1, keepdims=True).clip(min=1e-8)
+    return W
+
+
+def _kuramoto_spec(n, theta0, omega, G, stim=(), dt=0.01):
+    W = np.ones((n, n), dtype=float); np.fill_diagonal(W, 0.0)
+    sv  = (dataclasses.replace(kuramoto.state_variables[0], default_init=theta0),)
+    mdl = dataclasses.replace(kuramoto, state_variables=sv)
+    return SimulationSpec(
+        model=mdl,
+        integrator=IntegratorSpec(method="heun", dt=dt),
+        coupling=CouplingSpec("kuramoto"),
+        monitors=(MonitorSpec("raw"),),
+        weights=W,
+        node_params={"omega": np.full(n, float(omega)), "G": G},
+        stimuli=stim,
+    )
+
+
+def _mpr_spec(n, stim=(), dt=0.01):
+    W = _weights(n, seed=5)
+    return SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method="heun", dt=dt),
+        coupling=CouplingSpec("linear", a=0.1),
+        monitors=(MonitorSpec("raw"),),
+        weights=W,
+        stimuli=stim,
+    )
+
+
+def _run(spec, duration, backend="numpy"):
+    return Simulator(spec, backend=backend).run(duration)["raw"][1]
+
+
+# ---------------------------------------------------------------------------
+# Physics: stimulus has expected effect
+# ---------------------------------------------------------------------------
+
+class TestStimPhysics:
+
+    def test_stimulated_node_advances_faster(self):
+        """A stimulated Kuramoto node must accumulate more phase than un-stimulated ones."""
+        n = 4
+        theta0 = np.zeros(n)
+        stim = StimSpec(
+            sv_name="theta",
+            amplitude=np.array([1.0, 0.0, 0.0, 0.0]),
+            waveform=lambda t: 0.05,
+        )
+        spec = _kuramoto_spec(n, theta0, omega=1.0, G=0.0, stim=(stim,))
+        d = _run(spec, 100.0)
+        assert d[-1, 0, 0] > d[-1, 0, 1], \
+            "Stimulated node 0 should have larger phase than un-stimulated node 1"
+
+    def test_zero_amplitude_stim_no_effect(self):
[…]
