### 2026-05-21 0af6ab6
feat: Implement stimulus handling in CUDA backend and enhance backend consistency tests
