### 2026-05-26 039fd50
Add PR review notes and milestones for feature extraction and inference engine
- Created `PR_auto_generate_review.md` to document findings and recommendations for merging the `auto_generate` branch into `main`, highlighting migration expectations, backend parity risks, and testing notes.

- Added `milestones_feature_extraction.md` to outline the current state and future milestones for feature extraction across various backends, detailing existing functionalities, gaps, and planned enhancements.

- Introduced `milestones_inference.md` to establish a roadmap for developing a high-performance SBI
[…]
