+++ notes/milestones_inference.md
@@ -1,0 +1,685 @@
+# Inference Engine Milestones — `vbi/inference/`
+
+> **Goal:** A torch-free, high-performance SBI engine with a sbi-compatible
+> API.  Backends progress: numpy/autograd → numba → JAX.  The existing
+> `vbi/cde.py` is wrapped and gradually replaced; users never need to change
+> their code.
+
+---
+
+## Package target structure
+
+```
+vbi/
+  inference/
+    __init__.py          # public: SNPE, BoxUniform, Gaussian, Posterior
+    _api.py              # SNPE / SNLE classes (sbi-compatible interface)
+    _prior.py            # BoxUniform, Gaussian, CustomPrior
+    _posterior.py        # Posterior object
+    _estimators/
+      __init__.py
+      base.py            # ConditionalDensityEstimator ABC (from cde.py)
+      mdn.py             # MDNEstimator (from cde.py)
+      maf.py             # MAFEstimator (from cde.py)
+    _backends/
+      __init__.py
+      numpy_.py          # autograd/numpy (current, default)
+      numba_.py          # Numba JIT forward pass (MI-numba)
+      jax_.py            # JAX: jit + vmap + grad (MI1)
+  cde.py                 # deprecated shim → re-exports from vbi.inference
+```
+
+**Backend selector** (transparent to user):
+```python
+SNPE(prior, density_estimator='maf', backend='auto')
+# 'auto' → jax if installed, else numpy
+# 'numpy' → autograd/numpy (always available)
+# 'numba' → Numba JIT (install vbi[numba])
+# 'jax'   → JAX XLA (install vbi[jax])
+```
+
+---
+
+## API Design Decision: mirror `sbi`
+
+**Decision: yes, mirror the `sbi` API at the high level.**
+
+The `sbi` API is well-designed and already familiar to computational
+neuroscience users.  Mirroring it means:
+
+- Users can switch from `sbi` to `vbi.cde` by changing one import and
+  converting `torch.Tensor` → `np.ndarray`.
+- All `sbi` tutorials and examples translate almost verbatim.
+- `vbi.sbi_inference.Inference` (the current heavy wrapper) and the new
+  light CDE stack share the same mental model.
+
+### Two-layer architecture
+
+```
+┌──────────────────────────────────────────────────────────┐
+│  HIGH-LEVEL  sbi-compatible API  (vbi/cde_api.py)        │
+│  SNPE, SNLE  ←→  append_simulations → train →            │
+│              build_posterior → posterior.sample(...)      │
+│                                                          │
+│  Prior objects: BoxUniform, Gaussian, CustomPrior        │
+│  Posterior object: .sample((n,), x=x_obs)                │
+│                    .log_prob(theta, x=x_obs)              │
+│                    .map(x=x_obs)                          │
+└──────────────────┬───────────────────────────────────────┘
+                   │ delegates to
+┌──────────────────▼───────────────────────────────────────┐
+│  LOW-LEVEL  estimator building blocks  (vbi/cde.py)      │
+│  MDNEstimator, MAFEstimator                              │
+│  ConditionalDensityEstimator ABC                         │
+│  backends: autograd (default) → JAX (MI1) → torch (MI5) │
+└──────────────────────────────────────────────────────────┘
+```
+
+### API comparison: sbi 0.26 vs vbi.cde (target)
+
+Verified against sbi 0.26.1 exact signatures.
+
+| `sbi` call | `vbi.cde` call | Notes |
[…]
