+++ notes/milestones_feature_extraction.md
@@ -1,0 +1,255 @@
+# Feature Extraction Milestones — `vbi/feature_extraction/`
+
+> **Context:** The simulator backends (M0–M4) are complete including
+> stimulation support across all backends (numpy, numba, C++, JAX, CUDA).
+> This document tracks what remains for feature extraction to reach full
+> parity across backends and what new capabilities are needed for the
+> SBI pipeline.
+
+---
+
+## Current state (May 2026)
+
+### What exists
+
+| File | Status | Notes |
+|------|--------|-------|
+| `features_utils.py` | ✅ Complete | 68 functions, full NumPy, covers FC, FCD, power, coherence, etc. |
+| `features_utils_nb.py` | ⚠️ Partial | 7 `@njit` functions: `nb_mean`, `nb_std`, `nb_fc_flat`, `nb_fcd_mean`, `nb_extract`, `NbExtractorSpec` — used by Tier-2 numba sweep inline |
+| `features_utils_jax.py` | ⚠️ Partial | `get_fc`, `get_fcd` only — not vmap-compatible, not connected to sweeper |
+| `features_utils_cuda.py` | ❌ Missing | Not yet created |
+| `pipeline.py` | ⚠️ Partial | `FeaturePipeline` works for NumPy only; no backend dispatch; no inline sweep integration for JAX/C++/CUDA |
+| `calc_features.py` | ✅ Exists | Legacy wrapper, kept for backward compat |
+
+### What works end-to-end today
+
+- **NumPy sweeper Tier-1**: raw time series → `FeaturePipeline` in Python ✅
+- **Numba sweeper Tier-2 (inline)**: `nb_extract` inside `@njit` loop → `(mean, std, fc, fcd_ks)` only ✅
+- **All other backends**: Tier-1 only via Python post-processing ✅
+
+### What is missing
+
+- JAX sweeper does not yet feed into `FeaturePipeline` inline (vmap-compatible)
+- C++ sweeper has Tier-1 Python post-processing but no inline C++ feature loop
+- CUDA sweeper has no inline feature extraction
+- `FeaturePipeline` has no backend awareness — always uses NumPy
+- Stimulus is not yet supported inside feature pipelines (burn-in with stimulus)
+- No `features_utils_cuda.py`
+- `features_utils_jax.py` is not vmap-compatible and missing most features
+- G, alpha sweeping not yet unified in `CouplingSpec` (separate issue)
+
+---
+
+## MF1 — `FeaturePipeline` backend dispatch  *(numpy, numba)*
+
+**Goal:** `FeaturePipeline` dispatches to the right implementation based on
+the backend that produced the data, without changing user code.
+
+```python
+# Same call regardless of where data came from
+labels, values = pipeline.extract(monitor_result)
+```
+
+**Tasks:**
+
+- [ ] Add `backend: str = "numpy"` parameter to `FeaturePipeline`
+- [ ] Dispatch `fc`, `fcd_ks`, `mean`, `std` to `features_utils_nb.py` when
+      `backend="numba"` (already implemented, just needs wiring)
+- [ ] Validate that Numba Tier-1 matches NumPy to rtol=1e-4
+- [ ] Tests: `test_pipeline_backend_dispatch.py`
+
+**Effort:** Small — the numba functions already exist.
+
+---
+
+## MF2 — Extend `features_utils_nb.py`  *(numba inline coverage)*
+
+**Goal:** Numba inline sweep (Tier-2) covers more features beyond the current
+`(mean, std, fc, fcd_ks)`.
+
+**Tasks:**
+
+- [ ] `nb_band_power(ts, dt, band_lo, band_hi)` — `@njit`, Welch or AR
+- [ ] `nb_psd_peak(ts, dt)` — dominant frequency
[…]
