### 2026-05-26 03ce90f
Add Mixture Density Network and Posterior/Prior classes for SBI compatibility
- Implemented MDNEstimator for conditional density estimation using a Gaussian mixture model.
- Created Posterior class to wrap trained density estimators with a sbi-compatible API.
- Developed BoxUniform, Gaussian, and CustomPrior classes for prior distributions, mirroring sbi's interface.
- Added comprehensive tests for prior distributions, SNPE workflow, and posterior functionalities.
- Ensured compatibility with numpy arrays instead of tensors for all new classes.
