+++ vbi/inference/_estimators/maf.py
@@ -1,0 +1,485 @@
+"""Masked Autoregressive Flow conditional density estimator."""
+from __future__ import annotations
+from dataclasses import dataclass, field
+from typing import Optional
+
+import autograd.numpy as anp
+from autograd import grad
+from tqdm.auto import trange
+
+from .base import ConditionalDensityEstimator
+
+
+@dataclass
+class MAFEstimator(ConditionalDensityEstimator):
+    """
+    Masked Autoregressive Flow for conditional density estimation.
+
+    Implements MADE blocks with optional ActNorm, random permutations,
+    z-score normalisation, and PCA embedding of features.
+
+    Parameters
+    ----------
+    param_dim, feature_dim : int | None
+    n_flows : int
+        Number of MADE blocks (autoregressive transforms).
+    hidden_units : int
+        Hidden units per MADE block.
+    activation : 'tanh' | 'relu' | 'elu'
+    z_score_theta, z_score_x : bool
+        Standardise parameters / features internally.
+    use_actnorm : bool
+        Insert data-dependent ActNorm between flows.
+    embedding_dim : int | None
+        If set, PCA-compress features to this dimension before conditioning.
+    """
+
+    param_dim:     int | None = None
+    feature_dim:   int | None = None
+    n_flows:       int        = 4
+    hidden_units:  int        = 64
+    activation:    str        = "tanh"
+    z_score_theta: bool       = True
+    z_score_x:     bool       = True
+    use_actnorm:   bool       = True
+    embedding_dim: Optional[int] = None
+    actnorm_eps:   float      = 1e-6
+
+    def __post_init__(self):
+        super().__init__(self.param_dim, self.feature_dim)
+        self._dims_inferred         = False
+        self.model_constants        = None
+        self._actnorm_initialized   = [False] * self.n_flows
+        self.theta_mean = self.theta_std = None
+        self.x_mean     = self.x_std     = None
+        self._use_pca = False
+        self._pca_components = None
+        self.val_loss_history: list[float] = []
+        self.best_epoch: int = -1
+        self.best_val_loss: float | None = None
+
+    # ------------------------------------------------------------------
+    # Preprocessing
+    # ------------------------------------------------------------------
+
+    def _ctx_dim(self) -> int:
+        if self.feature_dim == 0:
+            return 0
+        if self._use_pca and self.embedding_dim is not None:
+            return self.embedding_dim
+        return self.feature_dim
+
+    def prepare_normalizers(self, features, params):
+        if self.z_score_theta:
+            self.theta_mean = anp.mean(params,   axis=0)
+            self.theta_std  = anp.std(params,    axis=0) + 1e-8
+        else:
+            self.theta_mean = anp.zeros(self.param_dim)
+            self.theta_std  = anp.ones(self.param_dim)
+
+        if self.feature_dim > 0:
+            if self.z_score_x:
+                self.x_mean = anp.mean(features, axis=0)
+                self.x_std  = anp.std(features,  axis=0) + 1e-8
+            else:
+                self.x_mean = anp.zeros(self.feature_dim)
+                self.x_std  = anp.ones(self.feature_dim)
+
+            if self.embedding_dim is not None and self.embedding_dim < self.feature_dim:
+                X = (features - self.x_mean) / self.x_std
+                _, _, Vt = anp.linalg.svd(X, full_matrices=False)
+                self._pca_components = Vt[:self.embedding_dim, :].T
+                self._use_pca = True
+            else:
+                self._use_pca = False
[…]
