+++ vbi/inference/_estimators/__init__.py
@@ -1,0 +1,10 @@
+from .base import ConditionalDensityEstimator
+from .mdn  import MDNEstimator
+from .maf  import MAFEstimator, MAFEstimator0
+
+__all__ = [
+    "ConditionalDensityEstimator",
+    "MDNEstimator",
+    "MAFEstimator",
+    "MAFEstimator0",
+]

[…]
