### 2026-05-26 1e802e5
feat: Expand feature comparison and milestone tracking for vbi.inference
