+++ notes/milestones_inference.md
@@ -7,6 +7,145 @@
 
 ---
 
+## Feature comparison: sbi 0.26 vs vbi.inference (current + planned)
+
+> Verified against sbi 0.26.1.  ✅ = done  ⚠️ = partial  ❌ = missing
+> Priority: 🔴 high (blocks real use)  🟡 medium  🟢 low / defer
+
+### Inference algorithms
+
+| Feature | sbi | vbi.inference | Priority | Milestone |
+|---------|-----|---------------|----------|-----------|
+| SNPE / NPE-C | ✅ | ✅ | — | done |
+| SNPE-A / SNPE-B (older) | ✅ | ❌ | 🟢 | not planned |
+| SNLE (neural likelihood) | ✅ | ❌ stub | 🟡 | MI-SNLE |
+| SNRE-A/B/C (neural ratio) | ✅ | ❌ | 🟢 | future |
+| FMPE (flow matching) | ✅ | ❌ | 🟢 | future |
+| MNPE / MNLE (mixed data) | ✅ | ❌ | 🟢 | future |
+| SMC-ABC / MC-ABC | ✅ | ❌ | 🟢 | future |
+
+### Density estimators
+
+| Feature | sbi | vbi.inference | Priority | Milestone |
+|---------|-----|---------------|----------|-----------|
+| MAF | ✅ | ✅ | — | done |
+| MDN | ✅ | ✅ | — | done |
+| NSF (neural spline flow) | ✅ | ❌ | 🔴 | MI0-NSF |
+| MADE (standalone) | ✅ | ❌ | 🟡 | future |
+| Custom nn via callable | ✅ | ❌ | 🟡 | MI-custom-net |
+| Embedding / summary network | ✅ | ❌ | 🔴 | MI0-embed |
+| z-score theta | ✅ | ✅ (MAF) | — | done |
+| z-score x | ✅ | ✅ (MAF) | — | done |
+| structured z-scoring | ✅ | ❌ | 🟢 | future |
+
+### SNPE training features
+
+| Feature | sbi | vbi.inference | Priority | Milestone |
+|---------|-----|---------------|----------|-----------|
+| Mini-batch training | ✅ | ✅ | — | done |
+| Train / val split | ✅ | ✅ | — | done |
+| Early stopping | ✅ | ✅ | — | done |
+| Gradient clipping | ✅ | ✅ | — | done |
+| Sequential rounds: append_simulations | ✅ | ✅ (data only) | ⚠️ | done |
+| APT importance weights (num_atoms) | ✅ | ❌ stored, ignored | 🔴 | MI3 |
+| Warm start (resume_training) | ✅ | ❌ | 🟡 | MI0-warm |
+| get_simulations() | ✅ | ❌ | 🟡 | MI0-warm |
+| Custom DataLoader kwargs | ✅ | ❌ | 🟢 | future |
+| Discard prior samples (round weighting) | ✅ | ❌ | 🟡 | MI3 |
+
+### Posterior sampling
+
+| Feature | sbi | vbi.inference | Priority | Milestone |
+|---------|-----|---------------|----------|-----------|
+| Direct / ancestral sampling | ✅ | ✅ | — | done |
+| sample_batched (vectorised over x) | ✅ | ❌ | 🟡 | MI0-batch |
+| Rejection sampling (against prior) | ✅ | ❌ | 🔴 | MI0-rejection |
+| reject_outside_prior in sample() | ✅ | ❌ | 🔴 | MI0-rejection |
+| MCMC: Metropolis-Hastings | ✅ | ❌ | 🔴 | MI4 |
+| MCMC: NUTS (pyro / pymc) | ✅ | ❌ | 🟡 | MI4-JAX |
[…]
