### 2026-05-26 d837d07
feat: Add inference demos for 1-D Gaussian, 2-D Gaussian, Two-moons, MDN vs MAF, and sequential rounds
