+++ docs/examples/inference/05_sequential_rounds.py
@@ -1,0 +1,116 @@
+"""
+Demo 5 — Sequential rounds: 1 round vs 2 rounds.
+
+Uses the 1-D Gaussian problem to show that a second round of simulation
+focused near the posterior concentrates samples closer to the true parameter.
+
+Round 1: simulate from prior theta ~ N(0, 2)  → broad coverage
+Round 2: simulate from the round-1 posterior  → focused near true theta
+
+We compare posterior mean error and std between the two strategies.
+
+Expected runtime: < 20 seconds.
+"""
+
+import numpy as np
+from vbi.inference import SNPE, Gaussian
+
+SIGMA_LIK   = 0.3
+SIGMA_PRIOR = 2.0
+TRUE_THETA  = 1.5     # the "true" parameter we're trying to infer
+N_ROUND1    = 1500
+N_ROUND2    = 500
+N_POST      = 1000
+SEED        = 0
+
+rng = np.random.default_rng(SEED)
+prior  = Gaussian(mean=np.array([0.0]), std=np.array([SIGMA_PRIOR]))
+x_obs  = np.array([[TRUE_THETA + rng.normal(0, SIGMA_LIK)]])
+
+# Analytical posterior
+prec_p   = 1 / SIGMA_PRIOR ** 2
+prec_l   = 1 / SIGMA_LIK   ** 2
+s_true   = np.sqrt(1 / (prec_p + prec_l))
+mu_true  = s_true ** 2 * prec_l * x_obs[0, 0]
+
+print("=" * 55)
+print("Demo 5 — Sequential rounds")
+print("=" * 55)
+print(f"  True theta = {TRUE_THETA},  x_obs = {x_obs[0,0]:.3f}")
+print(f"  Analytical posterior: mean={mu_true:.3f}  std={s_true:.3f}")
+
+
+def simulate_from(sampler, n, rng_seed):
+    """Sample theta from sampler, simulate x = theta + noise."""
+    rng_s  = np.random.default_rng(rng_seed)
+    theta  = sampler(n, rng_seed)
+    x      = theta + rng_s.normal(0, SIGMA_LIK, theta.shape)
+    return theta, x
+
+
+# ── 1-round strategy ───────────────────────────────────────────────────────────
+theta1, x1 = simulate_from(lambda n, s: prior.sample((n,), seed=s), N_ROUND1, SEED)
+
+inf1 = SNPE(prior=prior, density_estimator="maf")
+inf1.append_simulations(theta1, x1)
+est1 = inf1.train(
+    training_batch_size=256, learning_rate=5e-4,
+    stop_after_epochs=20, max_num_epochs=300, verbose=False,
+)
+post1   = inf1.build_posterior(est1)
+samp1   = post1.sample((N_POST,), x=x_obs, seed=0)
+me1     = abs(samp1[:, 0].mean() - mu_true)
+se1     = abs(samp1[:, 0].std()  - s_true)
+
+print(f"\n  Round 1 only  ({N_ROUND1} sims from prior):")
+print(f"    mean = {samp1[:,0].mean():.4f}  err = {me1:.4f}")
+print(f"    std  = {samp1[:,0].std():.4f}  err = {se1:.4f}")
+
+# ── 2-round strategy ───────────────────────────────────────────────────────────
+# Round 2: sample theta from the round-1 posterior, then simulate.
+#
+# WHY we use round-2 data ONLY (not round1 + round2):
+#   Round 1 was sampled from the prior; round 2 from the posterior.
+#   Mixing both proposals without importance-weighting (APT / num_atoms)
+#   biases the estimator — the prior-sampled round 1 pulls the posterior
+#   back toward the prior.
[…]
