+++ docs/examples/inference/02_gaussian_2d.py
@@ -1,0 +1,123 @@
+"""
+Demo 2 — 2-D correlated Gaussian: checks that covariance structure is captured.
+
+Simulator:   x = A @ theta + noise,   noise ~ N(0, I * sigma²)
+Prior:       theta ~ Uniform([-3,-3], [3,3])
+Posterior:   theta | x  — analytically derived (linear-Gaussian)
+
+With A = [[1,0],[0,1]] (identity) and BoxUniform prior, the posterior is
+approximately Gaussian near the observation, which lets us check that the
+estimator recovers the right mean AND that the two parameters are inferred
+independently (zero covariance expected).
+
+We also add a cross-correlation case: x[0] = theta[0] + 0.5*theta[1] + noise,
+x[1] = theta[1] + noise  — here the joint posterior shows correlation.
+
+Expected runtime: < 15 seconds.
+"""
+
+import numpy as np
+from vbi.inference import SNPE, BoxUniform
+
+SIGMA = 0.2
+N_TRAIN = 3000
+SEED = 7
+
+rng = np.random.default_rng(SEED)
+
+# ── Case A: identity simulator (independent parameters) ───────────────────────
+print("=" * 60)
+print("Demo 2 — 2-D Gaussian (covariance structure)")
+print("=" * 60)
+
+prior = BoxUniform(low=np.array([-3., -3.]), high=np.array([3., 3.]))
+theta_tr = prior.sample((N_TRAIN,), seed=SEED)
+x_tr     = theta_tr + rng.normal(0, SIGMA, theta_tr.shape)
+
+inf = SNPE(prior=prior, density_estimator="maf")
+inf.append_simulations(theta_tr, x_tr)
+est = inf.train(
+    training_batch_size=256, learning_rate=5e-4,
+    stop_after_epochs=20, max_num_epochs=300, verbose=False,
+)
+post = inf.build_posterior(est)
+
+x_obs = np.array([[1.0, -1.0]])
+samp  = post.sample((3000,), x=x_obs, seed=0)   # (3000, 2)
+
+mu_est  = samp.mean(axis=0)
+std_est = samp.std(axis=0)
+corr_est = np.corrcoef(samp[:, 0], samp[:, 1])[0, 1]
+
+print("\n  Case A — identity: x = theta + noise (params should be independent)")
+print(f"  x_obs = {x_obs[0]}")
+print(f"  Posterior mean : {mu_est}  (expected ≈ {x_obs[0]})")
+print(f"  Posterior std  : {std_est}  (expected ≈ [{SIGMA:.2f}, {SIGMA:.2f}])")
+print(f"  Correlation    : {corr_est:.3f}  (expected ≈ 0.0)")
+
+assert np.all(np.abs(mu_est - x_obs[0]) < 0.15), \
+    f"Posterior mean too far from truth: {mu_est}"
+assert abs(corr_est) < 0.25, \
+    f"Unexpected correlation in independent case: {corr_est:.3f}"
+print("  ✓ Independent case: OK")
+
+# ── Case B: correlated simulator ──────────────────────────────────────────────
+# x0 = t0 + 0.5*t1 + noise,  x1 = t1 + noise
+# The coupling of x0 to t1 creates posterior correlation between t0 and t1
+A = np.array([[1.0, 0.5], [0.0, 1.0]])
+
+theta_tr2 = prior.sample((N_TRAIN,), seed=SEED + 1)
+x_tr2     = (A @ theta_tr2.T).T + rng.normal(0, SIGMA, theta_tr2.shape)
+
+inf2 = SNPE(prior=prior, density_estimator="maf")
+inf2.append_simulations(theta_tr2, x_tr2)
+est2 = inf2.train(
+    training_batch_size=256, learning_rate=5e-4,
[…]
