### 2026-05-26 4b3aac4
feat: Enhance training stability by adding posterior collapse prevention strategies in MAF
