### 2026-05-26 6ed5948
feat: Implement posterior collapse prevention strategies and cosine learning rate scheduling in MAF
