+++ vbi/inference/_estimators/maf.py
@@ -1,6 +1,8 @@
 """Masked Autoregressive Flow conditional density estimator."""
 from __future__ import annotations
 from dataclasses import dataclass, field
+import logging
+import math
 from typing import Optional
 
 import autograd.numpy as anp
@@ -8,6 +10,17 @@ from autograd import grad
 from tqdm.auto import trange
 
 from .base import ConditionalDensityEstimator
+
+log = logging.getLogger(__name__)
+
+
+def _cosine_lr(lr_max: float, lr_min: float, epoch: int, period: int) -> float:
+    """Cosine annealing: lr_max → lr_min over ``period`` epochs, then stays at lr_min."""
+    if epoch >= period:
+        return lr_min
+    return lr_min + 0.5 * (lr_max - lr_min) * (
+        1.0 + math.cos(math.pi * epoch / max(period - 1, 1))
+    )
 
 
 @dataclass
@@ -254,10 +267,20 @@ class MAFEstimator(ConditionalDensityEstimator):
         stop_after_epochs: int = 20,
         early_stopping_delta: float = 0.0,
         clip_max_norm: float | None = 5.0,
+        # ── LR schedule ──────────────────────────────────────────────────
+        lr_schedule: str | None = "cosine",
+        lr_min: float = 1e-5,
+        lr_period: int = 500,
+        # ── Posterior collapse prevention ─────────────────────────────────
+        monitor_collapse: bool = True,
+        x_check=None,
+        collapse_threshold: float = 0.05,
+        check_every: int = 10,
+        n_check: int = 200,
     ):
         """
-        Train with Adam, optional mini-batches, train/val split, and
-        early stopping.
+        Train with Adam, optional mini-batches, train/val split, early stopping,
+        cosine LR annealing, and posterior-collapse monitoring.
 
         Parameters
         ----------
@@ -265,6 +288,26 @@ class MAFEstimator(ConditionalDensityEstimator):
             Mini-batch size per gradient step.  None → full training set.
         verbose : bool
             Show tqdm progress bar.
+        lr_schedule : 'cosine' | None
+            Learning-rate schedule.  'cosine' anneals from ``learning_rate``
+            down to ``lr_min`` over ``n_iter`` epochs.  None keeps lr fixed.
+        lr_min : float
+            Floor for cosine schedule.
+        monitor_collapse : bool
+            If True, sample from the posterior every ``check_every`` epochs
+            and stop training if the posterior std collapses below
+            ``collapse_threshold * data_std``.  Eliminates the need to tune
+            ``n_iter`` manually.
+        x_check : array (1, feature_dim) | None
+            Observation used for collapse monitoring.  Defaults to the first
+            training feature sample.
+        collapse_threshold : float
+            Stop if posterior_std < collapse_threshold * data_std on any dim.
+        check_every : int
+            Epochs between collapse checks.
+        n_check : int
+            Posterior samples drawn per collapse check.
+
         All other parameters mirror sbi's ``SNPE.train()`` kwarg names.
         """
         params   = anp.asarray(params)
@@ -324,8 +367,30 @@ class MAFEstimator(ConditionalDensityEstimator):
         self.best_epoch    = -1
         self.best_val_loss = None
 
+        # ── Collapse monitoring setup ─────────────────────────────────────────
+        # Two-checkpoint strategy:
+        #   best_val_weights    — best NLL on validation (may be collapsed)
+        #   last_healthy_weights — latest weights where posterior_std was OK
+        # When collapse is detected we restore last_healthy_weights, which
+        # avoids restoring an already-collapsed best_val checkpoint.
+        _data_std = p_tr.std(axis=0) + 1e-8   # (param_dim,)
[…]
