### 2026-05-26 a60365c
feat: Update error thresholds in sequential rounds demo and enhance verbose handling in SNPE and MAFEstimator
