### 2026-05-26 244b103
feat: Add demo for vbi.inference vs sbi comparison on 1-D and 2-D Gaussians
