+++ docs/examples/inference/06_vbi_vs_sbi.py
@@ -1,0 +1,318 @@
+"""
+Demo 6 — vbi.inference vs sbi: head-to-head on 1-D and 2-D Gaussians.
+
+Both implementations are trained on identical data with identical hyper-
+parameters.  Results are compared against the known analytical posterior.
+
+Metrics reported per method × density estimator × test observation:
+  - posterior mean error   |est_mean − true_mean|
+  - posterior std  error   |est_std  − true_std |
+  - 90 % CI coverage       fraction of test obs where true theta ∈ 90% CI
+  - training time          wall-clock seconds
+
+sbi is skipped gracefully when not installed.
+
+Expected runtime: < 60 seconds (both backends, 1-D + 2-D).
+"""
+
+import contextlib
+import io
+import sys
+import time
+import warnings
+import numpy as np
+
+warnings.filterwarnings("ignore")
+
+
+@contextlib.contextmanager
+def _silence():
+    """Suppress stdout/stderr noise from sbi's epoch logging."""
+    with open("/dev/null", "w") as devnull:
+        with contextlib.redirect_stdout(devnull), \
+             contextlib.redirect_stderr(devnull):
+            yield
+
+# ── helpers ───────────────────────────────────────────────────────────────────
+
+SIGMA_LIK   = 0.3
+SIGMA_PRIOR = 2.0
+N_TRAIN     = 2000
+N_POST      = 1000
+SEED        = 42
+# shared hyper-parameters (identical for both backends)
+TRAIN_KW = dict(
+    training_batch_size = 256,
+    learning_rate       = 5e-4,
+    stop_after_epochs   = 20,
+    max_num_epochs      = 300,
+)
+
+rng = np.random.default_rng(SEED)
+
+
+def analytical_posterior_1d(x_obs_val: float):
+    prec_p = 1 / SIGMA_PRIOR ** 2
+    prec_l = 1 / SIGMA_LIK   ** 2
+    s_post = np.sqrt(1 / (prec_p + prec_l))
+    m_post = s_post ** 2 * prec_l * x_obs_val
+    return m_post, s_post
+
+
+X_TEST = np.array([-2.0, -1.0, 0.0, 1.0, 2.0])
+
+
+# ── vbi.inference runner ──────────────────────────────────────────────────────
+
+def run_vbi(de: str, theta_tr, x_tr, x_test_vals, dims: int = 1):
+    from vbi.inference import SNPE
+    from vbi.inference import BoxUniform, Gaussian
+
+    if dims == 1:
+        prior = Gaussian(mean=np.zeros(1), std=np.full(1, SIGMA_PRIOR))
+    else:
+        prior = BoxUniform(low=np.full(dims, -5.), high=np.full(dims, 5.))
+
+    inf = SNPE(prior=prior, density_estimator=de)
+    inf.append_simulations(theta_tr, x_tr)
+
+    t0  = time.perf_counter()
+    est = inf.train(verbose=False, **TRAIN_KW)
+    train_sec = time.perf_counter() - t0
+
+    post = inf.build_posterior(est)
+    return post, train_sec
+
+
+def evaluate_1d(post, x_test_vals, backend_label: str):
+    mean_errs, std_errs, in_ci = [], [], []
+    for xv in x_test_vals:
+        x_obs = np.array([[xv]])
+        mu_t, s_t = analytical_posterior_1d(xv)
+        samp = post.sample((N_POST,), x=x_obs, seed=0)
+        # sbi returns a tensor; convert if needed
+        samp = np.array(samp).reshape(N_POST, -1)
+        lo, hi = np.percentile(samp[:, 0], [5, 95])
+        mean_errs.append(abs(samp[:, 0].mean() - mu_t))
[…]
