### 2026-05-26 7110230
feat: Add early stopping and adjust collapse monitoring in SNPE and MAFEstimator
