### 2026-05-26 0122d9d
feat: Enhance SNPE training with epoch-based early stopping and batch size adjustments
