### 2026-05-26 72cceb6
feat: Implement rejection sampling and leakage correction in Posterior class
