### 2026-05-26 94fc68f
feat: Add EmbeddingNet and utility functions; integrate embedding support in SNPE, MAF, and MDN estimators
