+++ vbi/inference/__init__.py
@@ -31,6 +31,8 @@ With::
 from ._api        import SNPE, SNLE
 from ._prior      import BoxUniform, Gaussian, CustomPrior
 from ._posterior  import Posterior
+from ._embedding  import EmbeddingNet
+from ._utils      import simulate_for_vbi, simulate_for_sbi
 from ._estimators import (
     ConditionalDensityEstimator,
     MDNEstimator,
@@ -48,6 +50,11 @@ __all__ = [
     "CustomPrior",
     # Posterior
     "Posterior",
+    # Embedding
+    "EmbeddingNet",
+    # Utilities
+    "simulate_for_vbi",
+    "simulate_for_sbi",
     # Low-level estimators (for direct use / subclassing)
     "ConditionalDensityEstimator",
     "MDNEstimator",
