+++ vbi/tests/test_inference_api.py
@@ -27,6 +27,7 @@ pytestmark = pytest.mark.skipif(
 from vbi.inference import (
     SNPE, BoxUniform, Gaussian, CustomPrior,
     Posterior, MDNEstimator, MAFEstimator,
+    EmbeddingNet, simulate_for_sbi, process_prior,
 )
 
 
@@ -240,6 +241,163 @@ class TestMiniBatch:
         assert len(est.loss_history) > 0
         # Loss should decrease overall
         assert est.loss_history[-1] < est.loss_history[0]
+
+
+# ---------------------------------------------------------------------------
+# Embedding network tests  (MI0-embed)
+# ---------------------------------------------------------------------------
+
+class TestEmbeddingNet:
+    """EmbeddingNet reduces feature dim; flow trains with it jointly."""
+
+    def test_forward_reduces_dim(self):
+        emb = EmbeddingNet(input_dim=10, output_dim=3)
+        rng = np.random.RandomState(0)
+        w   = emb.init_weights(rng)
+        x   = np.random.randn(50, 10).astype("f")
+        out = emb.forward(w, x)
+        assert out.shape == (50, 3)
+
+    def test_repr(self):
+        emb = EmbeddingNet(10, 4, hidden_sizes=(32,))
+        assert "10" in repr(emb) and "4" in repr(emb)
+
+    @pytest.mark.parametrize("de", ["maf", "mdn"])
+    def test_snpe_with_embedding_trains(self, de):
+        """SNPE with EmbeddingNet trains and samples the right shape."""
+        rng   = np.random.default_rng(42)
+        prior = BoxUniform(low=np.zeros(2), high=np.ones(2))
+        theta = prior.sample((200,), seed=42)
+        # High-dim x: 10-D, embedded to 4-D
+        x     = theta @ rng.normal(0, 1, (2, 10)) + rng.normal(0, 0.05, (200, 10))
+
+        emb = EmbeddingNet(input_dim=10, output_dim=4)
+        inf = SNPE(prior=prior, density_estimator=de, embedding_net=emb)
+        inf.append_simulations(theta, x)
+        est = inf.train(
+            training_batch_size=64,
+            max_num_epochs=10,
+            verbose=False,
+        )
+        post    = inf.build_posterior(est)
+        x_obs   = x[:1]   # single 10-D observation
+        samples = post.sample((100,), x=x_obs)
+        assert samples.shape == (100, 2), f"{de}: wrong sample shape"
+        assert np.all(np.isfinite(samples)), f"{de}: non-finite samples"
+
+    def test_estimator_feature_dim_set_to_embed_dim(self):
+        """Estimator's feature_dim is set to embedding output_dim after training."""
+        prior = BoxUniform(low=np.zeros(2), high=np.ones(2))
+        theta = prior.sample((100,), seed=0)
+        x     = np.random.randn(100, 8).astype("f")
+        emb   = EmbeddingNet(input_dim=8, output_dim=3)
+        inf   = SNPE(prior=prior, density_estimator='maf', embedding_net=emb)
+        inf.append_simulations(theta, x)
+        est = inf.train(max_num_epochs=3, verbose=False)
+        assert est.feature_dim == 3
+
+
+# ---------------------------------------------------------------------------
+# Utils tests  (MI0-utils)
+# ---------------------------------------------------------------------------
+
+class TestSimulateForSBI:
+    """simulate_for_sbi returns correct shapes and handles failures."""
+
+    def test_shape(self):
+        prior = BoxUniform(low=np.zeros(2), high=np.ones(2))
+        theta, x = simulate_for_sbi(
+            lambda th: th + np.random.randn(2) * 0.1,
+            prior,
+            num_simulations=50,
+            seed=0,
+        )
[…]
