### 2026-05-26 0a77ad7
feat: Add log probability computation for pre-embedded features in MDNEstimator
