### 2026-05-26 c74ed32
refactor: Update milestone priorities and feature status in inference documentation
