+++ notes/milestones_inference.md
@@ -894,7 +894,7 @@ samples = posterior.sample(5000)
 |-----------|---------|----------------|-----------|
 | **MI-API** | numpy/autograd | sbi-compatible API, full workflow | `numpy`, `autograd` |
 | **MI0** | numpy/autograd | mini-batch, save/load, logging | same |
-| **MI-numba** | Numba JIT | 5–10× faster training on CPU, no GPU needed | `numba` |
+| ~~MI-numba~~ | ~~Numba JIT~~ | ~~skipped — dead end without autograd~~ | — |
 | **MI1** | JAX | GPU, vmap, jit, gradient through posterior | `jax` |
 | **MI2** | all | prior integration, Posterior object (part of MI-API) | MI-API |
 | **MI3** | all | sequential rounds, multi-round SNPE | MI-API + MI2 |
@@ -904,88 +904,94 @@ samples = posterior.sample(5000)
 
 ## Priority order
 
-**Current focus: pure Python/NumPy — get feature parity with common sbi usage
-before moving to performance backends.**
+**Current focus: finish numpy-backend feature parity, then move to JAX.**
 
 ```
-★ DONE:  MI-API  sbi-compatible interface, MAF, MDN, BoxUniform, Gaussian,
-                  Posterior (sample/log_prob/map), mini-batch, save/load
+★ DONE:  MI-API        sbi-compatible interface, MAF, MDN, BoxUniform, Gaussian,
+                        Posterior (sample/log_prob/map), mini-batch, save/load
+         MI0-collapse  cosine LR, log_scale clamp, posterior-collapse monitor
+         MI0-rejection reject_outside_prior, leakage_correction, sample_with='rejection'
+         MI0-embed     EmbeddingNet (jointly trained MLP summary network)
+         MI0-utils     simulate_for_vbi, get_simulations, resume_training, process_prior
 
-  NEXT:   MI0-NSF        NSF density estimator (best model in sbi)
-          MI0-embed      Embedding/summary network
-          MI0-rejection  Rejection sampling + reject_outside_prior
-          MI0-priors     MultivariateNormal, MultipleIndependent, Beta, Gamma, ...
-          MI0-utils      simulate_for_sbi, get_simulations, resume_training
-          MI-diag        SBC, TARP, C2ST, pairplot
+  NEXT (numpy, remaining):
+          MI0-priors   MultivariateNormal, MultipleIndependent, Beta, Gamma, RestrictedPrior
+          MI-diag      SBC, TARP, C2ST, pairplot, plot_loss
+          MI0-NSF      NSF density estimator (optional — harder in autograd; may defer to MI1)
 
-  THEN:   MI3  Sequential rounds with APT weights (num_atoms properly)
-          MI4  MCMC posterior (MH first, NUTS after JAX)
-          MI6  End-to-end VBIInference API
+  THEN (after numpy is complete):
+          MI1   JAX backend (GPU, vmap, jit, PRNG refactor) — translate autograd → JAX
+          MI3   Sequential rounds with APT importance weights (num_atoms)
+          MI4   MCMC posterior (MH on numpy; NUTS requires JAX)
+          MI6   End-to-end VBIInference API
 
-  LATER:  MI-numba  Numba JIT backend
-          MI1       JAX backend (GPU, vmap, grad through posterior)
-          MI5       Torch backend (optional)
[…]
