### 2026-05-26 9d772a3
feat: Add new prior distributions (MultivariateNormal, LogNormal, Gamma, Beta, MultipleIndependent, RestrictedPrior) and update imports in inference module
