+++ vbi/tests/test_inference_api.py
@@ -14,20 +14,14 @@ import warnings
 import numpy as np
 import pytest
 
-try:
-    import autograd  # noqa: F401
-    AUTOGRAD_AVAILABLE = True
-except ImportError:
-    AUTOGRAD_AVAILABLE = False
-
-pytestmark = pytest.mark.skipif(
-    not AUTOGRAD_AVAILABLE, reason="autograd not installed"
-)
+pytest.importorskip("autograd", reason="autograd not installed")
 
 from vbi.inference import (
     SNPE, BoxUniform, Gaussian, CustomPrior,
     Posterior, MDNEstimator, MAFEstimator,
     EmbeddingNet, simulate_for_sbi, process_prior,
+    MultivariateNormal, LogNormal, Gamma, Beta,
+    MultipleIndependent, RestrictedPrior,
 )
 
 
@@ -100,6 +94,249 @@ class TestCustomPrior:
         lp = p.log_prob(s)
         assert s.shape == (10, 2)
         assert lp.shape == (10,)
+
+
+class TestSBICompatibility:
+    """Compare vbi prior behavior with sbi where matching sbi priors exist."""
+
+    @pytest.fixture
+    def sbi_boxuniform(self):
+        try:
+            from sbi.utils import BoxUniform as SBIBoxUniform
+        except Exception as exc:
+            pytest.skip(f"sbi BoxUniform not available: {exc}")
+        torch = pytest.importorskip("torch", reason="torch not installed")
+        return SBIBoxUniform, torch
+
+    def test_boxuniform_log_prob_matches_sbi(self, sbi_boxuniform):
+        SBIBoxUniform, torch = sbi_boxuniform
+        low  = np.array([-1.0, 0.0])
+        high = np.array([2.0, 4.0])
+        theta = np.array([
+            [0.0, 1.0],
+            [1.5, 3.5],
+            [-1.5, 1.0],
+            [0.0, 5.0],
+        ])
+
+        vbi_prior = BoxUniform(low=low, high=high)
+        sbi_prior = SBIBoxUniform(
+            low=torch.as_tensor(low, dtype=torch.float64),
+            high=torch.as_tensor(high, dtype=torch.float64),
+        )
+
+        got = vbi_prior.log_prob(theta)
+        expected = sbi_prior.log_prob(
+            torch.as_tensor(theta, dtype=torch.float64)
+        ).detach().cpu().numpy()
+        np.testing.assert_allclose(got, expected, rtol=1e-12, atol=0.0)
+
+    def test_boxuniform_sample_contract_matches_sbi(self, sbi_boxuniform):
+        SBIBoxUniform, torch = sbi_boxuniform
+        low  = np.array([-2.0, 0.5, 10.0])
+        high = np.array([2.0, 1.5, 20.0])
+
+        vbi_samples = BoxUniform(low=low, high=high).sample((64,), seed=0)
+        sbi_samples = SBIBoxUniform(
+            low=torch.as_tensor(low, dtype=torch.float64),
+            high=torch.as_tensor(high, dtype=torch.float64),
+        ).sample((64,)).detach().cpu().numpy()
+
+        assert vbi_samples.shape == sbi_samples.shape == (64, 3)
+        assert np.all(vbi_samples >= low)
+        assert np.all(vbi_samples <= high)
+        assert np.all(sbi_samples >= low)
+        assert np.all(sbi_samples <= high)
+
+
+# ---------------------------------------------------------------------------
+# MI0-priors: extended prior distributions
+# ---------------------------------------------------------------------------
+
+class TestMultivariateNormal:
+
[…]
