+++ vbi/inference/_prior.py
@@ -4,11 +4,12 @@ Prior distributions for SBI — torch-free, numpy-native.
 All priors expose the same interface as sbi's prior objects:
     prior.sample((n,))          → (n, d) ndarray
     prior.log_prob(theta)       → (n,)   ndarray
-    prior.support               → {'low': array, 'high': array} or None
+    prior.dim                   → int
 """
 from __future__ import annotations
 
 import numpy as np
+from scipy.special import gammaln
 
 
 class BoxUniform:
@@ -152,3 +153,332 @@ class CustomPrior:
 
     def __repr__(self):
         return f"CustomPrior(dim={self._dim})"
+
+
+# ---------------------------------------------------------------------------
+# MultivariateNormal — full covariance Gaussian
+# ---------------------------------------------------------------------------
+
+class MultivariateNormal:
+    """
+    Multivariate Gaussian with a full (dense) covariance matrix.
+
+    Parameters
+    ----------
+    mean : array-like  (d,)
+    cov  : array-like  (d, d)  — positive-definite covariance matrix
+    """
+
+    def __init__(self, mean, cov):
+        self.mean = np.asarray(mean, dtype=np.float64)
+        self.cov  = np.asarray(cov,  dtype=np.float64)
+        if self.mean.ndim != 1:
+            raise ValueError("mean must be 1-D.")
+        d = self.mean.size
+        if self.cov.shape != (d, d):
+            raise ValueError(f"cov must be ({d}, {d}).")
+        self._L       = np.linalg.cholesky(self.cov)          # lower triangular
+        self._L_inv   = np.linalg.inv(self._L)
+        self._log_det = 2.0 * np.sum(np.log(np.diag(self._L)))
+
+    @property
+    def dim(self) -> int:
+        return int(self.mean.size)
+
+    def sample(self, sample_shape, seed: int | None = None) -> np.ndarray:
+        rng  = np.random.default_rng(seed)
+        n    = int(np.prod(sample_shape))
+        z    = rng.standard_normal((n, self.dim))
+        samp = z @ self._L.T + self.mean
+        return samp.reshape(*sample_shape, self.dim)
+
+    def log_prob(self, theta) -> np.ndarray:
+        theta = np.atleast_2d(np.asarray(theta, dtype=np.float64))
+        diff  = theta - self.mean                         # (n, d)
+        z     = diff @ self._L_inv.T                      # (n, d): L_inv @ diff_i per row
+        quad  = np.sum(z ** 2, axis=1)                    # (n,)
+        return -0.5 * (self.dim * np.log(2.0 * np.pi) + self._log_det + quad)
+
+    def __repr__(self):
+        return f"MultivariateNormal(mean={self.mean}, cov=...)"
+
+
+# ---------------------------------------------------------------------------
+# LogNormal — log-normal marginals (independent per dimension)
+# ---------------------------------------------------------------------------
+
+class LogNormal:
+    """
+    Independent log-normal distribution per parameter dimension.
+
+    X ~ LogNormal(mean, std)  ⟺  log(X) ~ N(mean, std²)
+
+    Parameters
+    ----------
+    mean : array-like  (d,)  — mean of log(X)
+    std  : array-like  (d,)  — std of log(X)  (must be > 0)
+    """
+
+    def __init__(self, mean, std):
+        self.mean = np.asarray(mean, dtype=np.float64)
+        self.std  = np.asarray(std,  dtype=np.float64)
+        if np.any(self.std <= 0):
+            raise ValueError("std must be strictly positive.")
+
+    @property
+    def dim(self) -> int:
+        return int(self.mean.size)
+
+    def sample(self, sample_shape, seed: int | None = None) -> np.ndarray:
[…]
