### 2026-05-26 a912fc5
fix: Update test parameters in TestSBICompatibility and clear deprecated module warning in TestDeprecatedShim
