### 2026-05-26 d0a9a27
fix: Improve error message for untrained model in sample and log_prob methods
