### 2026-05-26 6eb5bb5
fix: Enhance input validation in get_fcd function for window length and overlap parameters
