### 2026-05-26 975890a
test: Add log_prob tests for MultivariateNormal, LogNormal, Gamma, Beta, and MultipleIndependent distributions
