### 2026-05-26 bbccdc9
feat: Add diagnostic tools for posterior validation and corresponding tests
