+++ vbi/tests/test_diagnostics.py
@@ -1,0 +1,447 @@
+"""
+Tests for vbi.inference diagnostic tools.
+
+Covers: run_sbc, check_sbc, sbc_rank_plot, run_tarp, check_tarp, plot_tarp,
+        c2st, pairplot, conditional_pairplot, plot_loss
+"""
+import numpy as np
+import pytest
+
+autograd = pytest.importorskip("autograd", reason="autograd not installed")
+matplotlib = pytest.importorskip("matplotlib", reason="matplotlib not installed")
+sklearn = pytest.importorskip("sklearn", reason="scikit-learn not installed")
+
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from vbi.inference import (
+    SNPE, BoxUniform,
+    run_sbc, check_sbc, sbc_rank_plot,
+    run_tarp, check_tarp, plot_tarp,
+    c2st, pairplot, conditional_pairplot, plot_loss,
+)
+
+
+# ---------------------------------------------------------------------------
+# Shared fixtures
+# ---------------------------------------------------------------------------
+
+def _make_1d_posterior(n_train=200, seed=42):
+    """Train a quick 1-D posterior on a Gaussian toy model."""
+    rng = np.random.RandomState(seed)
+    prior = BoxUniform(low=np.array([-3.0]), high=np.array([3.0]))
+    theta = prior.sample((n_train,))
+    x     = theta + 0.1 * rng.randn(*theta.shape)
+
+    inf = SNPE(prior=prior, density_estimator="mdn")
+    inf = inf.append_simulations(theta, x)
+    inf.train(max_num_epochs=20, verbose=False)
+    posterior = inf.build_posterior()
+    posterior.set_default_x(x[0])
+    return posterior, prior
+
+
+def _make_2d_posterior(n_train=300, seed=7):
+    rng = np.random.RandomState(seed)
+    prior = BoxUniform(low=np.array([-2.0, -2.0]), high=np.array([2.0, 2.0]))
+    theta = prior.sample((n_train,))
+    x     = theta + 0.1 * rng.randn(*theta.shape)
+
+    inf = SNPE(prior=prior, density_estimator="mdn")
+    inf = inf.append_simulations(theta, x)
+    inf.train(max_num_epochs=20, verbose=False)
+    posterior = inf.build_posterior()
+    return posterior, prior
+
+
+@pytest.fixture(scope="module")
+def posterior_1d():
+    return _make_1d_posterior()
+
+
+@pytest.fixture(scope="module")
+def posterior_2d():
+    return _make_2d_posterior()
+
+
+# ---------------------------------------------------------------------------
+# SBC
+# ---------------------------------------------------------------------------
+
+class TestRunSBC:
+    def test_returns_ranks_and_thetas(self, posterior_1d):
+        posterior, prior = posterior_1d
+        def sim(theta): return theta + 0.1 * np.random.randn(*theta.shape)
+
+        result = run_sbc(posterior, sim, prior,
+                         num_sbc_runs=20, num_posterior_samples=50, seed=0)
+        assert "ranks" in result
+        assert "thetas" in result
+
+    def test_ranks_shape(self, posterior_1d):
+        posterior, prior = posterior_1d
+        def sim(theta): return theta + 0.0
+
+        result = run_sbc(posterior, sim, prior,
+                         num_sbc_runs=15, num_posterior_samples=40, seed=1)
+        assert result["ranks"].shape == (15, 1)
+        assert result["thetas"].shape == (15, 1)
+
+    def test_ranks_in_valid_range(self, posterior_1d):
+        posterior, prior = posterior_1d
+        def sim(theta): return theta + 0.0
+
+        result = run_sbc(posterior, sim, prior,
[…]
