+++ vbi/inference/_diagnostics.py
@@ -1,0 +1,520 @@
+"""
+Diagnostic tools for posterior validation in simulation-based inference.
+
+Functions
+---------
+run_sbc          : Simulation-Based Calibration (rank statistics)
+check_sbc        : Uniformity tests on SBC ranks
+sbc_rank_plot    : Histogram plot of SBC ranks
+run_tarp         : Test of Accuracy with Random Points
+check_tarp       : Expected coverage probability error
+plot_tarp        : Coverage plot (ECP vs alpha)
+c2st             : Classifier Two-Sample Test accuracy
+pairplot         : Triangle scatter/KDE plot of samples
+conditional_pairplot : Pairplot conditioned on observed data
+plot_loss        : Training (and optional validation) loss curve
+"""
+import logging
+import warnings
+
+import numpy as np
+
+log = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# SBC
+# ---------------------------------------------------------------------------
+
+def run_sbc(posterior, simulator, prior, num_sbc_runs: int = 500,
+            num_posterior_samples: int = 1000, seed: int = 0) -> dict:
+    """
+    Run Simulation-Based Calibration.
+
+    For each of ``num_sbc_runs`` ground-truth parameters θ* drawn from the
+    prior, simulate x* = simulator(θ*), draw ``num_posterior_samples`` from
+    the posterior p(θ | x*), and record the rank of θ* among those samples
+    for each parameter dimension.
+
+    Parameters
+    ----------
+    posterior : Posterior
+        A trained posterior object with a ``sample((N,), x=...)`` method.
+    simulator : callable
+        ``simulator(theta_1d) -> x_1d``
+    prior : prior object
+        Must expose ``.sample((n,))`` and ``.log_prob(theta)``.
+    num_sbc_runs : int
+    num_posterior_samples : int
+    seed : int
+
+    Returns
+    -------
+    dict with keys
+        ``ranks``  : ndarray (num_sbc_runs, param_dim) — integer ranks in
+                     [0, num_posterior_samples]
+        ``thetas`` : ndarray (num_sbc_runs, param_dim) — ground-truth θ*
+    """
+    rng = np.random.RandomState(seed)
+    thetas_star = prior.sample((num_sbc_runs,), seed=int(rng.randint(1 << 31)))
+    thetas_star = np.asarray(thetas_star)
+    param_dim   = thetas_star.shape[1] if thetas_star.ndim > 1 else 1
+    if thetas_star.ndim == 1:
+        thetas_star = thetas_star[:, None]
+
+    ranks = np.empty((num_sbc_runs, param_dim), dtype=int)
+
+    for i in range(num_sbc_runs):
+        theta_i = thetas_star[i]
+        try:
+            x_i = np.asarray(simulator(theta_i))
+        except Exception as exc:
+            log.warning("SBC run %d: simulator failed (%s) — skipping.", i, exc)
+            ranks[i] = -1
+            continue
+
+        post_samples = np.asarray(
+            posterior.sample((num_posterior_samples,), x=x_i,
+                             seed=int(rng.randint(1 << 31)))
+        )
+        if post_samples.ndim == 1:
+            post_samples = post_samples[:, None]
+
+        for d in range(param_dim):
+            ranks[i, d] = int(np.sum(post_samples[:, d] < theta_i[d]))
+
+    return {"ranks": ranks, "thetas": thetas_star}
+
+
+def check_sbc(ranks: np.ndarray, num_posterior_samples: int) -> dict:
+    """
+    Test whether SBC ranks are uniform via Kolmogorov-Smirnov test.
+
+    Parameters
+    ----------
[…]
