### 2026-05-26 7b2935e
refactor: Update pairplot function to delegate to vbi.plot.pairplot and enhance documentation
