### 2026-05-26 1327f57
feat: Add Neural Spline Flow (NSF) estimator and corresponding batch sampling and log probability methods
