+++ vbi/inference/_estimators/__init__.py
@@ -1,10 +1,12 @@
 from .base import ConditionalDensityEstimator
 from .mdn  import MDNEstimator
 from .maf  import MAFEstimator, MAFEstimator0
+from .nsf  import NSFEstimator
 
 __all__ = [
     "ConditionalDensityEstimator",
     "MDNEstimator",
     "MAFEstimator",
     "MAFEstimator0",
+    "NSFEstimator",
 ]
