+++ vbi/tests/test_nsf_batch.py
@@ -1,0 +1,308 @@
+"""
+Tests for MI0-NSF (NSFEstimator) and MI0-batch (sample_batched / log_prob_batched).
+"""
+import numpy as np
+import pytest
+
+autograd = pytest.importorskip("autograd", reason="autograd not installed")
+
+from vbi.inference import (
+    SNPE, BoxUniform,
+    NSFEstimator, MAFEstimator,
+)
+from vbi.inference._estimators.nsf import (
+    _spline_params, _rq_forward_1d, _rq_inverse_1d,
+)
+
+
+# ---------------------------------------------------------------------------
+# Helpers
+# ---------------------------------------------------------------------------
+
+def _make_posterior(de="nsf", n_train=300, param_dim=2, seed=42):
+    rng   = np.random.RandomState(seed)
+    lo    = -2.0 * np.ones(param_dim)
+    hi    =  2.0 * np.ones(param_dim)
+    prior = BoxUniform(low=lo, high=hi)
+    theta = prior.sample((n_train,))
+    x     = theta + 0.1 * rng.randn(*theta.shape)
+    inf   = SNPE(prior=prior, density_estimator=de)
+    inf   = inf.append_simulations(theta, x)
+    inf.train(max_num_epochs=5, verbose=False)
+    return inf.build_posterior(), prior, theta, x
+
+
+# ---------------------------------------------------------------------------
+# RQ-spline unit tests
+# ---------------------------------------------------------------------------
+
+class TestSplineParams:
+    def test_widths_sum_to_2B(self):
+        rng = np.random.RandomState(0)
+        N, K, B = 10, 6, 5.0
+        uw = rng.randn(N, K).astype("f")
+        uh = rng.randn(N, K).astype("f")
+        ud = rng.randn(N, K - 1).astype("f")
+        cumw, cumh, widths, heights, derivs = _spline_params(uw, uh, ud, B, K)
+        np.testing.assert_allclose(np.sum(widths,  axis=1), 2 * B, atol=1e-4)
+        np.testing.assert_allclose(np.sum(heights, axis=1), 2 * B, atol=1e-4)
+
+    def test_derivatives_positive(self):
+        rng = np.random.RandomState(1)
+        N, K, B = 8, 4, 5.0
+        uw = rng.randn(N, K).astype("f")
+        uh = rng.randn(N, K).astype("f")
+        ud = rng.randn(N, K - 1).astype("f")
+        _, _, _, _, derivs = _spline_params(uw, uh, ud, B, K)
+        assert np.all(np.array(derivs) > 0)
+
+    def test_cumwidths_boundaries(self):
+        rng = np.random.RandomState(2)
+        N, K, B = 5, 8, 5.0
+        uw = rng.randn(N, K).astype("f")
+        uh = rng.randn(N, K).astype("f")
+        ud = rng.randn(N, K - 1).astype("f")
+        cumw, cumh, *_ = _spline_params(uw, uh, ud, B, K)
+        np.testing.assert_allclose(np.array(cumw[:, 0]),  -B, atol=1e-4)
+        np.testing.assert_allclose(np.array(cumw[:, -1]),  B, atol=1e-4)
+        np.testing.assert_allclose(np.array(cumh[:, 0]),  -B, atol=1e-4)
+        np.testing.assert_allclose(np.array(cumh[:, -1]),  B, atol=1e-4)
+
+
+class TestRQSplineForwardInverse:
+    def _make_params(self, N=20, K=6, B=5.0, seed=0):
+        rng = np.random.RandomState(seed)
+        return (rng.randn(N, K).astype("f"),
+                rng.randn(N, K).astype("f"),
+                rng.randn(N, K - 1).astype("f"))
+
[…]

[…]
