### 2026-05-26 67c925c
Add JAX backend implementations for MDN and NSF estimators, along with tests
- Implemented JaxMDNEstimator for Mixture Density Networks using JAX, including methods for weight initialization, forward pass, loss calculation, and sampling.
- Implemented JaxNSFEstimator for Neural Spline Flow using JAX, with methods for weight initialization, forward pass, log probability calculation, and sampling.
- Added comprehensive tests for JAX backend estimators, ensuring finite log probabilities, correct sample shapes, loss decrease during training, and numerical parity with autograd versions.
- Included 
[…]
