+++ vbi/inference/_backends/jax_/maf_jax.py
@@ -1,0 +1,496 @@
+"""
+JAX Masked Autoregressive Flow conditional density estimator.
+
+Translation of maf.py (autograd) to JAX.
+
+Key differences from the autograd version:
+  - jax.numpy instead of autograd.numpy
+  - jax.value_and_grad + jax.jit for the training step
+  - ActNorm: no data-dependent init (act_s=1, act_b=0 at start; learned
+    by gradient descent).  _actnorm_initialized is always True.
+  - PRNG in sample(): numpy RandomState is converted to a JAX PRNGKey
+  - weights dict holds numpy arrays at rest, converted to jnp inside jitted calls
+"""
+from __future__ import annotations
+
+import logging
+import math
+from dataclasses import dataclass, field
+from typing import Optional
+
+import numpy as np
+import jax
+import jax.numpy as jnp
+from tqdm.auto import trange
+
+from .base_jax import JaxConditionalDensityEstimator
+
+log = logging.getLogger(__name__)
+
+
+def _cosine_lr(lr_max: float, lr_min: float, epoch: int, period: int) -> float:
+    if epoch >= period:
+        return lr_min
+    return lr_min + 0.5 * (lr_max - lr_min) * (
+        1.0 + math.cos(math.pi * epoch / max(period - 1, 1))
+    )
+
+
+@dataclass
+class JaxMAFEstimator(JaxConditionalDensityEstimator):
+    """
+    Masked Autoregressive Flow — JAX backend.
+
+    Same interface as ``MAFEstimator`` (autograd version).
+    """
+
+    param_dim:     int | None    = None
+    feature_dim:   int | None    = None
+    n_flows:       int           = 4
+    hidden_units:  int           = 64
+    num_blocks:    int           = 2
+    activation:    str           = "tanh"
+    z_score_theta: bool          = True
+    z_score_x:     bool          = True
+    use_actnorm:   bool          = True
+    embedding_dim: Optional[int] = None
+    actnorm_eps:   float         = 1e-6
+
+    def __post_init__(self):
+        super().__init__(self.param_dim, self.feature_dim)
+        self._dims_inferred      = False
+        self.model_constants     = None
+        self.theta_mean = self.theta_std = None
+        self.x_mean     = self.x_std     = None
+        self._use_pca   = False
+        self._pca_components     = None
+        self.val_loss_history: list[float] = []
+        self.best_epoch: int     = -1
+        self.best_val_loss: float | None = None
+
+    # ------------------------------------------------------------------
+    # Preprocessing
+    # ------------------------------------------------------------------
+
+    def _ctx_dim(self) -> int:
+        if self.feature_dim == 0:
+            return 0
+        if self._use_pca and self.embedding_dim is not None:
+            return self.embedding_dim
+        return self.feature_dim
+
+    def prepare_normalizers(self, features, params):
+        if self.z_score_theta:
+            self.theta_mean = jnp.array(np.mean(params,   axis=0))
+            self.theta_std  = jnp.array(np.std(params,    axis=0) + 1e-8)
+        else:
+            self.theta_mean = jnp.zeros(self.param_dim)
+            self.theta_std  = jnp.ones(self.param_dim)
+
+        if self.feature_dim > 0:
+            if self.z_score_x:
+                self.x_mean = jnp.array(np.mean(features, axis=0))
+                self.x_std  = jnp.array(np.std(features,  axis=0) + 1e-8)
+            else:
+                self.x_mean = jnp.zeros(self.feature_dim)
+                self.x_std  = jnp.ones(self.feature_dim)
+
[…]
