+++ vbi/inference/_backends/jax_/base_jax.py
@@ -1,0 +1,247 @@
+"""
+Abstract base for JAX-native conditional density estimators.
+
+Drop-in replacement for the autograd ConditionalDensityEstimator.
+Key differences:
+  - weights dict holds jax.numpy arrays (not autograd arrays)
+  - grad computed via jax.grad / jax.value_and_grad
+  - training step is jax.jit-compiled for speed
+  - PRNG: sample() accepts numpy RandomState (converted internally to JAX key)
+"""
+from __future__ import annotations
+
+import abc
+import logging
+
+import numpy as np
+import jax
+import jax.numpy as jnp
+from tqdm.auto import trange
+
+log = logging.getLogger(__name__)
+
+
+class JaxConditionalDensityEstimator(abc.ABC):
+    """
+    Abstract base for JAX-native conditional density estimators.
+
+    Same public interface as ``ConditionalDensityEstimator`` so the
+    ``Posterior`` and ``SNPE`` layers work identically.
+    """
+
+    def __init__(self, param_dim: int = None, feature_dim: int = None):
+        self.param_dim   = param_dim
+        self.feature_dim = feature_dim
+        self._dims_inferred = False
+        self.weights: dict | None = None
+        self.loss_history: list[float] = []
+        self._emb = None
+
+    # ------------------------------------------------------------------
+    # Embedding support (mirrors base.py)
+    # ------------------------------------------------------------------
+
+    def set_embedding(self, emb) -> None:
+        self._emb = emb
+
+    def _wrapped_loss(self, weights, features, params) -> float:
+        if self._emb is not None:
+            features = self._emb.forward(weights, features)
+        return self._loss_function(weights, features, params)
+
+    # ------------------------------------------------------------------
+    # Abstract interface
+    # ------------------------------------------------------------------
+
+    @abc.abstractmethod
+    def _initialize_weights(self, seed: int) -> dict:
+        pass
+
+    @abc.abstractmethod
+    def _loss_function(self, weights: dict, features, params) -> float:
+        pass
+
+    @abc.abstractmethod
+    def sample(self, features, n_samples: int, rng):
+        if self.weights is None:
+            raise RuntimeError("Model has not been trained yet. Call train() first.")
+
+    @abc.abstractmethod
+    def log_prob(self, features, params):
+        if self.weights is None:
+            raise RuntimeError("Model has not been trained yet. Call train() first.")
+
+    # ------------------------------------------------------------------
+    # Dimension inference
+    # ------------------------------------------------------------------
+
+    def _infer_dimensions(self, params, features):
+        params   = np.asarray(params)
+        features = np.asarray(features)
+        if params.ndim   == 1: params   = params.reshape(-1, 1)
+        if features.ndim == 1: features = features.reshape(-1, 1)
+
+        self.param_dim   = params.shape[1]
+        self.feature_dim = features.shape[1]
+        self._dims_inferred = True
+
+        log.debug("Inferred dimensions: param_dim=%d  feature_dim=%d",
+                  self.param_dim, self.feature_dim)
+
+        if self.param_dim <= 0:
+            raise ValueError(f"param_dim must be positive, got {self.param_dim}")
+
+    # ------------------------------------------------------------------
+    # PRNG helper: accept numpy RandomState or int seed → JAX key
+    # ------------------------------------------------------------------
+
+    @staticmethod
+    def _to_jax_key(rng) -> jax.Array:
+        if isinstance(rng, np.random.RandomState):
+            seed = int(rng.randint(0, 2**31))
+        elif isinstance(rng, (int, np.integer)):
+            seed = int(rng)
+        else:
+            # Assume already a JAX key array
[…]
