### 2026-05-26 1bde9dc
feat: Enhance _tanh function to support JAX and refactor forward method for improved backend compatibility
