### 2026-05-26 898cfe1
feat: Update JAX vs NumPy demo to include NSF estimator and enhance README documentation
