+++ docs/examples/inference/07_jax_vs_numpy.py
@@ -10,26 +10,53 @@ The goal is a smoke-test style comparison, not a strict speed benchmark.  JAX
 includes compilation overhead on the first train/sample calls, so wall-clock
 times are most useful for checking that both backends run end-to-end.
 
+By default this runs MAF and MDN.  NumPy NSF is much slower because its spline
+transform is differentiated by autograd; include it explicitly with:
+
+    VBI_DEMO7_ESTIMATORS=maf,mdn,nsf python docs/examples/inference/07_jax_vs_numpy.py
+
 Expected runtime: < 30 seconds.
 """
 
+import contextlib
+import io
+import os
 import time
+import warnings
 
 import numpy as np
 
-from vbi.inference import SNPE, Gaussian
+os.environ.setdefault("MPLCONFIGDIR", "/tmp/matplotlib-vbi")
+# Demo runs on CPU; avoids cuDNN/XLA GPU init failures on machines where CUDA
+# is present but cuDNN is missing or mismatched.  Override with JAX_PLATFORMS=gpu.
+os.environ.setdefault("JAX_PLATFORMS", "cpu")
 
 # ── Problem definition ────────────────────────────────────────────────────────
 SIGMA_LIK   = 0.3
 SIGMA_PRIOR = 2.0
-N_TRAIN     = 1500
-N_POST      = 1000
+N_TRAIN     = 600
+N_POST      = 500
 SEED        = 42
 
-DENSITY_ESTIMATORS = ("maf", "mdn", "nsf")
+DENSITY_ESTIMATORS = tuple(
+    item.strip()
+    for item in os.environ.get("VBI_DEMO7_ESTIMATORS", "maf,mdn").split(",")
+    if item.strip()
+)
 BACKENDS           = ("numpy", "jax")
 
 rng = np.random.default_rng(SEED)
+
+
+def import_inference_api():
+    """Import through vbi while keeping optional dependency diagnostics quiet."""
+    stderr = io.StringIO()
+    with contextlib.redirect_stderr(stderr):
+        from vbi.inference import SNPE, Gaussian
+    return SNPE, Gaussian
+
+
+SNPE, Gaussian = import_inference_api()
 
 
 def analytical_posterior(x_obs_val: float):
@@ -45,16 +72,16 @@ def analytical_posterior(x_obs_val: float):
 prior    = Gaussian(mean=np.array([0.0]), std=np.array([SIGMA_PRIOR]))
 theta_tr = prior.sample((N_TRAIN,), seed=SEED)
 x_tr     = theta_tr + rng.normal(0, SIGMA_LIK, theta_tr.shape)
-x_test   = np.array([-2.0, -1.0, 0.0, 1.0, 2.0])
+x_test   = np.array([-1.0, 0.0, 1.0])
 
 
 def evaluate(density_estimator: str, backend: str):
     if density_estimator == "mdn":
-        max_epochs = 400
-        patience   = 30
+        max_epochs = 120
+        patience   = 15
     else:
-        max_epochs = 250
-        patience   = 25
+        max_epochs = 80
+        patience   = 12
 
     inference = SNPE(
         prior=prior,
@@ -66,7 +93,7 @@ def evaluate(density_estimator: str, backend: str):
 
     t0 = time.perf_counter()
     estimator = inference.train(
-        training_batch_size=256,
+        training_batch_size=512,
         learning_rate=5e-4,
         stop_after_epochs=patience,
         max_num_epochs=max_epochs,
@@ -103,19 +130,34 @@ def evaluate(density_estimator: str, backend: str):
     }
 
 
+def skip_message(exc: Exception) -> str:
+    msg = str(exc).splitlines()[0] if str(exc) else type(exc).__name__
[…]
