### 2026-05-26 8cda317
feat: Implement data-dependent initialization for ActNorm in JaxMAFEstimator
