### 2026-05-26 4aaaedf
feat: Add _forward_one_step method to JaxMAFEstimator and JaxNSFEstimator for improved flow step propagation
