### 2026-05-26 80fe268
feat: Add backend speed benchmark demo for sbi (torch), vbi numpy, and vbi jax
