+++ docs/examples/inference/08_backend_benchmark.py
@@ -1,0 +1,242 @@
+"""
+Demo 8 — Backend speed benchmark: sbi (torch) vs vbi numpy vs vbi jax.
+
+Measures three things separately for each backend:
+  1. Compilation / first-run overhead (JAX JIT; torch tracing for sbi).
+  2. Steady-state training throughput (gradient steps / second, after warmup).
+  3. Sampling throughput (posterior samples / second).
+  4. Posterior accuracy (mean error vs analytical solution).
+
+Problem: 1-D Gaussian, same as Demo 1.
+
+Expected runtimes (CPU):
+  sbi (torch) – varies by torch version / CPU
+  vbi numpy   – baseline
+  vbi jax     – first run slower (JIT); subsequent steps faster
+
+Run with:
+    python docs/examples/inference/08_backend_benchmark.py
+"""
+
+import os
+import time
+import warnings
+
+import numpy as np
+
+# Force JAX to CPU to avoid GPU cuDNN issues; override with JAX_PLATFORMS=gpu
+os.environ.setdefault("JAX_PLATFORMS", "cpu")
+os.environ.setdefault("MPLCONFIGDIR", "/tmp/matplotlib-vbi")
+
+# ── Problem ────────────────────────────────────────────────────────────────────
+SIGMA_LIK   = 0.3
+SIGMA_PRIOR = 2.0
+N_TRAIN     = 1000
+N_STEPS     = 300      # fixed gradient steps, no early stopping
+BATCH_SIZE  = 256
+SEED        = 0
+N_SAMPLES   = 1000     # posterior samples per timing call
+N_SAMPLE_REPS = 5      # repeat sampling to get stable timing
+
+X_OBS_VAL   = 1.0
+# Analytical posterior for x_obs=1.0
+_prec_p = 1.0 / SIGMA_PRIOR ** 2
+_prec_l = 1.0 / SIGMA_LIK   ** 2
+TRUE_STD  = float(np.sqrt(1.0 / (_prec_p + _prec_l)))
+TRUE_MEAN = float(TRUE_STD ** 2 * _prec_l * X_OBS_VAL)
+
+# ── Shared data ────────────────────────────────────────────────────────────────
+rng_np = np.random.default_rng(SEED)
+
+from vbi.inference import SNPE as vbiSNPE, Gaussian as vbiGaussian  # noqa: E402
+
+prior_vbi = vbiGaussian(mean=np.array([0.0]), std=np.array([SIGMA_PRIOR]))
+theta_tr  = prior_vbi.sample((N_TRAIN,), seed=SEED).astype("f")
+x_tr      = (theta_tr + rng_np.normal(0, SIGMA_LIK, theta_tr.shape)).astype("f")
+x_obs_vbi = np.array([[X_OBS_VAL]], dtype="f")
+
+
+# ── MAF architecture (kept small for CPU speed) ────────────────────────────────
+MAF_KWARGS = dict(n_flows=4, hidden_units=32, num_blocks=2)
+# Training uses fixed N_STEPS with no early stopping
+TRAIN_KWARGS = dict(
+    training_batch_size=BATCH_SIZE,
+    learning_rate=5e-4,
+    stop_after_epochs=N_STEPS,   # patience = total steps → no early stop
+    max_num_epochs=N_STEPS,
+    verbose=False,
+)
+
+
+def _posterior_accuracy(samples):
+    s = samples[:, 0] if samples.ndim == 2 else samples
+    return abs(s.mean() - TRUE_MEAN), abs(s.std() - TRUE_STD)
+
+
+def _print_result(name, compile_s, train_s, sample_s, mean_err, std_err, steps):
[…]
