### 2026-05-26 9801a57
feat: Set default JAX platform to CPU to prevent cuDNN initialization errors and enhance backend compatibility
