### 2026-05-26 c00ccf2
feat: Add set_jax_device function to configure JAX compute device and update backend initialization
