### 2026-05-26 6f540be
feat: Enhance backend benchmark demo with improved result formatting and error handling
