+++ docs/examples/inference/08_backend_benchmark.py
@@ -18,6 +18,8 @@ Run with:
     python docs/examples/inference/08_backend_benchmark.py
 """
 
+import contextlib
+import io
 import os
 import time
 import warnings
@@ -48,7 +50,9 @@ TRUE_MEAN = float(TRUE_STD ** 2 * _prec_l * X_OBS_VAL)
 # ── Shared data ────────────────────────────────────────────────────────────────
 rng_np = np.random.default_rng(SEED)
 
-from vbi.inference import SNPE as vbiSNPE, Gaussian as vbiGaussian  # noqa: E402
+_stderr = io.StringIO()
+with contextlib.redirect_stderr(_stderr):
+    from vbi.inference import SNPE as vbiSNPE, Gaussian as vbiGaussian  # noqa: E402
 
 prior_vbi = vbiGaussian(mean=np.array([0.0]), std=np.array([SIGMA_PRIOR]))
 theta_tr  = prior_vbi.sample((N_TRAIN,), seed=SEED).astype("f")
@@ -73,12 +77,17 @@ def _posterior_accuracy(samples):
     return abs(s.mean() - TRUE_MEAN), abs(s.std() - TRUE_STD)
 
 
-def _print_result(name, compile_s, train_s, sample_s, mean_err, std_err, steps):
+def _result_row(name, compile_s, train_s, sample_s, mean_err, std_err, steps):
     throughput = steps / train_s if train_s > 0 else float("nan")
-    print(f"  {name:<22}  compile={compile_s:6.2f}s  train={train_s:6.2f}s  "
-          f"({throughput:6.0f} steps/s)  "
-          f"sample={sample_s*1000:6.1f}ms  "
-          f"mean_err={mean_err:.3f}  std_err={std_err:.3f}")
+    return (
+        f"  {name:<14}"
+        f"{compile_s:>11.2f}"
+        f"{train_s:>10.2f}"
+        f"{throughput:>10.1f}"
+        f"{sample_s * 1000:>12.1f}"
+        f"{mean_err:>10.3f}"
+        f"{std_err:>9.3f}"
+    )
 
 
 results = []
@@ -209,49 +218,66 @@ def bench_sbi():
 print("=" * 76)
 print("Demo 8 — Backend benchmark: sbi/torch vs vbi/numpy vs vbi/jax")
 print("=" * 76)
-print(f"  Problem: 1-D Gaussian  N_train={N_TRAIN}  steps={N_STEPS}  "
-      f"batch={BATCH_SIZE}  MAF(flows=4,h=32)")
-print(f"  True posterior @ x_obs={X_OBS_VAL}: mean={TRUE_MEAN:.3f}  std={TRUE_STD:.3f}")
+print("  Problem")
+print(f"    simulator    : 1-D Gaussian")
+print(f"    N_train      : {N_TRAIN}")
+print(f"    train budget : {N_STEPS} epochs, batch={BATCH_SIZE}")
+print(f"    estimator    : MAF(flows=4, hidden_units=32)")
+print(f"    x_obs={X_OBS_VAL}: true mean={TRUE_MEAN:.3f}, true std={TRUE_STD:.3f}")
 print()
-print(f"  {'Backend':<22}  {'Compile':>8}  {'Train':>7}  {'Steps/s':>8}  "
-      f"{'Sample':>9}  {'Mean err':>8}  {'Std err':>7}")
-print("  " + "-" * 74)
 
-for backend in ("numpy", "jax"):
-    print(f"  running vbi/{backend} ...", flush=True)
+print("  Running")
+failures = []
+jobs = [("vbi/numpy", lambda: bench_vbi("numpy")),
+        ("vbi/jax",   lambda: bench_vbi("jax")),
+        ("sbi/torch", bench_sbi)]
+
[…]
