### 2026-05-27 #10612 @editorialbot
[REVIEW]: Virtual Brain Inference (VBI) toolkit on EBRAINS
**Submitting author:** @ziaeemehr (<a href="http://orcid.org/0000-0002-4696-9947">Abolfazl Ziaeemehr</a>)
**Repository:** https://github.com/ins-amu/vbi
**Branch with paper.md** (empty if default branch): main
**Version:** 0.4.4
