### 2026-05-27 539b4d2
feat: Optimize training step in JaxMAFEstimator with jitted full Adam step to reduce overhead
