### 2026-05-27 424f8ce
feat: Update milestones for vbi.inference with completed features and backend improvements
