+++ notes/milestones_inference.md
@@ -1,7 +1,7 @@
 # Inference Engine Milestones — `vbi/inference/`
 
 > **Goal:** A torch-free, high-performance SBI engine with a sbi-compatible
-> API.  Backends progress: numpy/autograd → numba → JAX.  The existing
+> API.  Backends progress: numpy/autograd → JAX.  The existing
 > `vbi/cde.py` is wrapped and gradually replaced; users never need to change
 > their code.
 
@@ -30,10 +30,10 @@
 |---------|-----|---------------|----------|-----------|
 | MAF | ✅ | ✅ | — | done |
 | MDN | ✅ | ✅ | — | done |
-| NSF (neural spline flow) | ✅ | ❌ | 🔴 | MI0-NSF |
+| NSF (neural spline flow) | ✅ | ✅ | — | done |
 | MADE (standalone) | ✅ | ❌ | 🟡 | future |
 | Custom nn via callable | ✅ | ❌ | 🟡 | MI-custom-net |
-| Embedding / summary network | ✅ | ❌ | 🔴 | MI0-embed |
+| Embedding / summary network | ✅ | ✅ (EmbeddingNet) | — | done |
 | z-score theta | ✅ | ✅ (MAF) | — | done |
 | z-score x | ✅ | ✅ (MAF) | — | done |
 | structured z-scoring | ✅ | ❌ | 🟢 | future |
@@ -48,8 +48,8 @@
 | Gradient clipping | ✅ | ✅ | — | done |
 | Sequential rounds: append_simulations | ✅ | ✅ (data only) | ⚠️ | done |
 | APT importance weights (num_atoms) | ✅ | ❌ stored, ignored | 🔴 | MI3 |
-| Warm start (resume_training) | ✅ | ❌ | 🟡 | MI0-warm |
-| get_simulations() | ✅ | ❌ | 🟡 | MI0-warm |
+| Warm start (resume_training) | ✅ | ✅ | — | done |
+| get_simulations() | ✅ | ✅ | — | done |
 | Custom DataLoader kwargs | ✅ | ❌ | 🟢 | future |
 | Discard prior samples (round weighting) | ✅ | ❌ | 🟡 | MI3 |
 
@@ -58,9 +58,9 @@
 | Feature | sbi | vbi.inference | Priority | Milestone |
 |---------|-----|---------------|----------|-----------|
 | Direct / ancestral sampling | ✅ | ✅ | — | done |
-| sample_batched (vectorised over x) | ✅ | ❌ | 🟡 | MI0-batch |
-| Rejection sampling (against prior) | ✅ | ❌ | 🔴 | MI0-rejection |
-| reject_outside_prior in sample() | ✅ | ❌ | 🔴 | MI0-rejection |
+| sample_batched (vectorised over x) | ✅ | ✅ | — | done |
+| Rejection sampling (against prior) | ✅ | ✅ | — | done |
+| reject_outside_prior in sample() | ✅ | ✅ | — | done |
 | MCMC: Metropolis-Hastings | ✅ | ❌ | 🔴 | MI4 |
 | MCMC: NUTS (pyro / pymc) | ✅ | ❌ | 🟡 | MI4-JAX |
 | MCMC: slice sampling | ✅ | ❌ | 🟢 | MI4 |
@@ -73,10 +73,10 @@
 |---------|-----|---------------|----------|-----------|
 | sample((n,), x=x_obs) | ✅ | ✅ | — | done |
 | log_prob(theta, x=x_obs) | ✅ | ✅ | — | done |
-| log_prob_batched | ✅ | ❌ | 🟡 | MI0-batch |
[…]
