### 2026-05-27 b34fc47
feat: Add new backend selection demo for JAX with minimal API smoke test
