+++ vbi/tests/test_jax_backend.py
@@ -325,3 +325,218 @@ class TestSNPEJaxBackend:
         """backend='auto' should resolve to 'jax' when JAX is installed."""
         from vbi.inference._backends import resolve_backend
         assert resolve_backend("auto") == "jax"
+
+
+# ---------------------------------------------------------------------------
+# jax.vmap over observations (batch log_prob)
+# ---------------------------------------------------------------------------
+
+class TestJaxVmap:
+    """_get_log_prob must be vmappable over the features (x_obs) axis."""
+
+    def test_maf_vmap_over_features(self, linear_gaussian_data):
+        theta, x, _ = linear_gaussian_data
+        est = _train_jax(JaxMAFEstimator, theta, x,
+                         n_flows=2, hidden_units=16, num_blocks=1)
+        w = {k: jnp.array(v) for k, v in est.weights.items()}
+        f5 = jnp.array(x[:5], dtype="f")
+        p5 = jnp.array(theta[:5], dtype="f")
+
+        lp_vmapped = jax.vmap(
+            lambda fi: est._get_log_prob(w, fi[None], p5)
+        )(f5)
+        assert lp_vmapped.shape == (5, 5), f"unexpected shape {lp_vmapped.shape}"
+        assert jnp.all(jnp.isfinite(lp_vmapped))
+
+    def test_nsf_vmap_over_features(self, linear_gaussian_data):
+        theta, x, _ = linear_gaussian_data
+        est = _train_jax(JaxNSFEstimator, theta, x,
+                         n_flows=2, hidden_units=16, num_blocks=1, num_bins=4)
+        w = {k: jnp.array(v) for k, v in est.weights.items()}
+        f5 = jnp.array(x[:5], dtype="f")
+        p5 = jnp.array(theta[:5], dtype="f")
+
+        lp_vmapped = jax.vmap(
+            lambda fi: est._get_log_prob(w, fi[None], p5)
+        )(f5)
+        assert lp_vmapped.shape == (5, 5)
+        assert jnp.all(jnp.isfinite(lp_vmapped))
+
+    def test_mdn_vmap_over_features(self, linear_gaussian_data):
+        theta, x, _ = linear_gaussian_data
+        est = _train_jax(JaxMDNEstimator, theta, x,
+                         n_components=3, hidden_sizes=(16,))
+        w = {k: jnp.array(v) for k, v in est.weights.items()}
+        f5 = jnp.array(x[:5], dtype="f")
+        p5 = jnp.array(theta[:5], dtype="f")
+
+        lp_vmapped = jax.vmap(
+            lambda fi: est._log_prob_core(w, fi[None], p5)
+        )(f5)
+        assert lp_vmapped.shape == (5, 5)
+        assert jnp.all(jnp.isfinite(lp_vmapped))
+
+
+# ---------------------------------------------------------------------------
+# jax.grad through log_prob w.r.t. theta
+# ---------------------------------------------------------------------------
+
+class TestJaxGrad:
+    """jax.grad through _get_log_prob w.r.t. params must be finite."""
+
+    def test_maf_grad_wrt_theta(self, linear_gaussian_data):
+        theta, x, _ = linear_gaussian_data
+        est = _train_jax(JaxMAFEstimator, theta, x,
+                         n_flows=2, hidden_units=16, num_blocks=1)
+        w = {k: jnp.array(v) for k, v in est.weights.items()}
+        f = jnp.array(x[:5], dtype="f")
+        p = jnp.array(theta[:5], dtype="f")
+
+        grad_fn = jax.grad(lambda p_: jnp.mean(est._get_log_prob(w, f, p_)))
+        g = grad_fn(p)
+        assert g.shape == p.shape
+        assert jnp.all(jnp.isfinite(g)), "MAF gradient has non-finite values"
+
[…]
