### 2026-05-27 eb5fec4
feat: Add backend selection demo and enhance posterior_1d function with xlim parameter
