### 2026-05-27 a449035
Implement MCMC sampling methods and integrate into posterior sampling
- Added MetropolisHastings and NUTS classes for MCMC posterior sampling.
- Enhanced Posterior class to support MCMC sampling with new parameters.
- Updated MDNEstimator to separate log probability calculation from loss function.
- Introduced r_hat and effective_sample_size diagnostics for MCMC chains.
- Added tests for MCMC functionality, including MetropolisHastings and NUTS.
- Ensured compatibility with both numpy and JAX backends.
- Updated existing tests to validate new MCMC features and ensure correct behavior.
