+++ vbi/inference/_mcmc.py
@@ -1,0 +1,413 @@
+"""
+MCMC posterior samplers for simulation-based inference (MI4).
+
+Classes
+-------
+MetropolisHastings
+    Random-walk MH using the trained neural posterior as a log-likelihood
+    surrogate.  Works with any backend (numpy / JAX).
+NUTS
+    No-U-Turn Sampler using JAX gradients.  Requires a JAX-backend estimator.
+
+Diagnostics
+-----------
+r_hat(chains)            Gelman-Rubin convergence statistic.
+effective_sample_size    Per-parameter ESS.
+"""
+from __future__ import annotations
+
+import logging
+import math
+from typing import TYPE_CHECKING
+
+import numpy as np
+
+if TYPE_CHECKING:
+    pass
+
+log = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Diagnostics
+# ---------------------------------------------------------------------------
+
+def r_hat(chains: np.ndarray) -> np.ndarray:
+    """
+    Gelman-Rubin R-hat convergence diagnostic.
+
+    Parameters
+    ----------
+    chains : ndarray (n_chains, n_samples, param_dim)
+
+    Returns
+    -------
+    ndarray (param_dim,)  — values near 1.0 indicate convergence.
+    """
+    chains = np.asarray(chains)
+    if chains.ndim == 2:
+        chains = chains[None]   # treat as single chain
+    M, N, D = chains.shape
+    if M <= 1 or N < 2:
+        return np.full(D, np.nan)
+
+    chain_means = chains.mean(axis=1)          # (M, D)
+    grand_mean  = chain_means.mean(axis=0)     # (D,)
+
+    # Between-chain variance
+    B = N / (M - 1) * np.sum((chain_means - grand_mean) ** 2, axis=0)
+
+    # Within-chain variance
+    W = np.mean(np.var(chains, axis=1, ddof=1), axis=0)
+
+    var_hat = (N - 1) / N * W + B / N
+    return np.sqrt(var_hat / (W + 1e-30))
+
+
+def effective_sample_size(chain: np.ndarray) -> np.ndarray:
+    """
+    Bulk effective sample size via autocorrelation sum.
+
+    Parameters
+    ----------
+    chain : ndarray (n_samples, param_dim)
+
+    Returns
+    -------
+    ndarray (param_dim,)
+    """
+    chain = np.asarray(chain)
+    if chain.ndim == 1:
+        chain = chain[:, None]
+    N, D = chain.shape
+    ess = np.empty(D)
+    for d in range(D):
+        x = chain[:, d]
+        x = x - x.mean()
+        # Autocorrelation via FFT
+        f  = np.fft.fft(x, n=2 * N)
+        ac = np.real(np.fft.ifft(f * np.conj(f)))[:N] / (N - np.arange(N))
+        ac = ac / (ac[0] + 1e-30)
+        # Sum pairs until first non-positive pair
+        rho_sum = 1.0
+        for t in range(1, N // 2):
+            pair = ac[2 * t - 1] + ac[2 * t]
+            if pair <= 0:
+                break
+            rho_sum += 2 * pair
+        ess[d] = N / max(rho_sum, 1.0)
+    return ess
+
+
+# ---------------------------------------------------------------------------
+# MetropolisHastings
+# ---------------------------------------------------------------------------
+
+class MetropolisHastings:
+    """
+    Random-walk Metropolis-Hastings sampler.
+
+    Uses ``log_posterior = estimator.log_prob(x_obs, theta) + prior.log_prob(theta)``
+    as the target density.
+
+    Parameters
+    ----------
+    estimator : ConditionalDensityEstimator
+        Trained density estimator (any backend).
[…]

[…]
