### 2026-05-27 7172dcd
feat: Rename NUTS to HMC and update related documentation; add backward-compatible alias
