### 2026-05-27 d575a7c
feat: Update MCMC references from NUTS to HMC and add backward-compatible alias
