+++ vbi/inference/_mcmc.py
@@ -188,7 +188,7 @@ class MetropolisHastings:
             theta_cur = np.asarray(init_theta, dtype="f")
         elif self._prior is not None:
             try:
-                theta_cur = np.array(self._prior.sample((1,)), dtype="f").flatten()
+                theta_cur = np.array(self._prior.sample((1,), seed=seed), dtype="f").flatten()
             except Exception:
                 theta_cur = np.zeros(param_dim, dtype="f")
         else:
@@ -239,16 +239,16 @@ class HMC:
     """
     Hamiltonian Monte Carlo sampler using JAX automatic differentiation.
 
-    Uses fixed-length leapfrog trajectories (2^min(max_tree_depth, 4) steps)
-    with dual-averaging step-size adaptation during warmup.
+    Uses fixed-length leapfrog trajectories with dual-averaging step-size
+    adaptation during warmup.
 
     Requires a JAX-backend estimator (JaxMAFEstimator / JaxNSFEstimator /
     JaxMDNEstimator) so that ``jax.grad`` can flow through ``log_prob``.
 
     The MCMC target is ``estimator.log_prob(x_obs, theta)`` — the trained SNPE
     posterior — so direct sampling and HMC target the same distribution.  If a
-    prior is supplied it is used only as a **support mask** for bounded priors
-    (e.g. BoxUniform); unbounded priors are not added to the target.
+    prior is supplied it is used only as a **support mask**: samples outside the
+    prior support are rejected; no density is added.
 
     Parameters
     ----------
@@ -256,10 +256,15 @@ class HMC:
         Trained JAX-backend estimator.
     prior : prior object | None
         Prior with ``.log_prob(theta)`` method.  Used for support masking only.
+        ``RestrictedPrior`` is not supported (constraint function is not
+        JAX-traceable); pass it to ``MetropolisHastings`` instead.
     step_size : float
         Initial leapfrog step size.  Adapted during warmup.
+    num_leapfrog_steps : int | None
+        Number of leapfrog steps per sample.  When ``None``, falls back to
+        ``2 ** min(max_tree_depth, 4)`` for backward compatibility.
     max_tree_depth : int
-        Controls leapfrog steps per sample: 2^min(max_tree_depth, 4) steps.
+        Deprecated legacy parameter.  Use ``num_leapfrog_steps`` instead.
     """
 
     def __init__(
@@ -267,6 +272,7 @@ class HMC:
         estimator,
         prior=None,
         step_size: float = 0.1,
+        num_leapfrog_steps: int | None = None,
         max_tree_depth: int = 10,
     ):
         try:
@@ -282,32 +288,60 @@ class HMC:
                 "Use MetropolisHastings for non-JAX estimators."
             )
 
-        self._est            = estimator
-        self._prior          = prior
-        self._step_size      = step_size
-        self._max_tree_depth = max_tree_depth
+        from ._prior import RestrictedPrior
+        if prior is not None and isinstance(prior, RestrictedPrior):
+            raise TypeError(
+                "HMC does not support RestrictedPrior: the constraint function "
+                "is not JAX-traceable. Use MetropolisHastings instead."
+            )
+
+        self._est               = estimator
+        self._prior             = prior
+        self._step_size         = step_size
+        self._num_leapfrog_steps = (
+            num_leapfrog_steps if num_leapfrog_steps is not None
+            else 2 ** min(max_tree_depth, 4)
+        )
+        self._max_tree_depth    = max_tree_depth
 
     @staticmethod
[…]
