### 2026-05-27 d2a278a
feat: Enhance monitors with sample validation and reshape output; update simulation spec validation
