### 2026-05-27 76236c5
feat: Enhance error handling and validation across various components; add new contract tests for simulator and sweeper
