### 2026-05-27 1128b89
feat: Update bold signal shape in JAX simulator and adjust related tests for consistency
