### 2026-05-27 057540b
refactor: Update parameter names in VEP simulation tests for consistency
