### 2026-05-27 78a92fa
feat: Add input validation for parameters in C++ bindings and sweeper; enhance error handling in codegen
