+++ vbi/tests/validation/test_cpp_sweep.py
@@ -177,8 +177,8 @@ def test_stochastic_sweep_unique_trajectories():
         monitors=(MonitorSpec(kind="raw"),),
         weights=_weights(5),
     )
-    # All eta identical — only noise differs between runs
-    sweep_spec = SweepSpec(params={"eta": np.full(4, -4.6)})
+    # All eta identical — only noise differs between runs (same_noise=False)
+    sweep_spec = SweepSpec(params={"eta": np.full(4, -4.6)}, same_noise=False)
     sweeper = CppSweeper(spec, sweep_spec)
     results = sweeper.run_serial(30.0)
 
