### 2026-05-27 a2d71fb
fix: Update import statement for vbi.cde module to ensure proper accessibility
