### 2026-05-27 981409b
feat: Update CMake configuration and error handling in build process; enhance parameter handling in CppSweeper
