### 2026-05-27 2c6e9fc
feat: Enhance CUDA backend with improved cache directory handling and noise generation; validate sweep parameters in CudaSweeperGPU
