### 2026-05-27 4c4095f
feat: Add CUDA tests for same_noise behavior and G sweep validation; include input validation for connectivity and sweep parameters
