+++ vbi/tests/validation/test_mpr_cuda.py
@@ -417,3 +417,139 @@ def test_cuda_sweep_throughput(n_nodes, n_samples, duration, capsys):
     with capsys.disabled():
         print(f"\nCUDA sweep: {rate:.1f} samples/s  "
               f"(n_nodes={n_nodes}, duration={duration}ms, n_samples={n_samples})")
+
+
+# ---------------------------------------------------------------------------
+# same_noise semantics
+# ---------------------------------------------------------------------------
+
+@cuda_only
+@cuda_fast
+def test_cuda_same_noise_true_identical_trajectories():
+    """same_noise=True: duplicate parameter entries must produce identical output."""
+    n_noise = len(mpr.noise_indices)
+    spec = SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method="heun", dt=0.1, stochastic=True,
+                                  noise_nsig=np.full(n_noise, 1e-4),
+                                  noise_seed=42),
+        coupling=CouplingSpec("linear", a=0.1),
+        monitors=(MonitorSpec("raw"),),
+        weights=_weights(6),
+    )
+    # Two samples with the same parameter — shared noise must make them identical.
+    sweep_spec = SweepSpec(params={"eta": np.array([-4.6, -4.6])}, same_noise=True)
+    results = Sweeper(spec, sweep_spec, backend="cuda").run(20.0)
+    _, d0 = results[0]["raw"]
+    _, d1 = results[1]["raw"]
+    np.testing.assert_array_equal(
+        d0, d1,
+        err_msg="same_noise=True: identical param entries produced different trajectories",
+    )
+
+
+@cuda_only
+@cuda_fast
+def test_cuda_same_noise_false_independent_trajectories():
+    """same_noise=False: identical parameter entries must produce different output."""
+    n_noise = len(mpr.noise_indices)
+    spec = SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method="heun", dt=0.1, stochastic=True,
+                                  noise_nsig=np.full(n_noise, 1e-4),
+                                  noise_seed=42),
+        coupling=CouplingSpec("linear", a=0.1),
+        monitors=(MonitorSpec("raw"),),
+        weights=_weights(6),
+    )
+    sweep_spec = SweepSpec(params={"eta": np.array([-4.6, -4.6])}, same_noise=False)
+    results = Sweeper(spec, sweep_spec, backend="cuda").run(20.0)
+    _, d0 = results[0]["raw"]
+    _, d1 = results[1]["raw"]
+    assert not np.array_equal(d0, d1), \
+        "same_noise=False: identical param entries produced identical trajectories"
+
+
+# ---------------------------------------------------------------------------
+# G sweep
+# ---------------------------------------------------------------------------
+
+@cuda_only
+@cuda_fast
+def test_cuda_sweep_G_outputs_differ():
+    """Sweeping G must produce different deterministic outputs for each value."""
+    G_values = np.array([0.0, 0.05, 0.1, 0.2])
+    spec = SimulationSpec(
+        model=mpr,
+        integrator=IntegratorSpec(method="heun", dt=0.1),
+        coupling=CouplingSpec("linear", a=1.0),
+        monitors=(MonitorSpec("raw"),),
+        weights=_weights(6, seed=7),
+    )
+    sweep_spec = SweepSpec(params={"G": G_values})
+    results = Sweeper(spec, sweep_spec, backend="cuda").run(20.0)
+
+    for i, res in enumerate(results):
+        _, d = res["raw"]
[…]
