### 2026-05-27 e171a66
feat: Add CUDA sweep demo for G sweep and same_noise control; implement optional plotting and input validation
