+++ docs/examples/simulator_models/cuda_sweep_demo.py
@@ -1,0 +1,185 @@
+"""
+CUDA sweep demo — G sweep and same_noise control.
+
+Demonstrates two features added to the CUDA backend:
+
+1. Sweeping the global coupling strength G.
+   G scales the coupling term for every sample without re-compiling the kernel.
+
+2. same_noise semantics.
+   same_noise=True  — all sweep samples share the same noise realisation.
+                      Differences in output reflect only the swept parameter.
+   same_noise=False — each sample gets an independent noise seed.
+                      More realistic for SBI training data.
+
+Run:
+    python cuda_sweep_demo.py
+    python cuda_sweep_demo.py --no-plot
+"""
+
+from __future__ import annotations
+
+import argparse
+import sys
+from pathlib import Path
+
+import numpy as np
+
+sys.dont_write_bytecode = True
+sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+
+try:
+    from numba import cuda as _cuda
+    if not _cuda.is_available():
+        print("No CUDA device found. Exiting.")
+        sys.exit(0)
+except Exception:
+    print("numba[cuda] not installed. Exiting.")
+    sys.exit(0)
+
+from vbi.simulator import Sweeper
+from vbi.simulator.models.mpr import mpr
+from vbi.simulator.spec import (
+    CouplingSpec,
+    IntegratorSpec,
+    MonitorSpec,
+    SimulationSpec,
+)
+from vbi.simulator.spec.sweep import SweepSpec
+
+
+# ---------------------------------------------------------------------------
+# Shared setup
+# ---------------------------------------------------------------------------
+
+N_NODES  = 8
+DT       = 0.1     # ms
+DURATION = 500.0   # ms
+
+rng = np.random.default_rng(0)
+W   = np.abs(rng.standard_normal((N_NODES, N_NODES))).astype(np.float64)
+np.fill_diagonal(W, 0.0)
+W  /= W.sum(axis=1, keepdims=True).clip(1e-8)
+
+N_NOISE  = len(mpr.noise_indices)
+NSIG     = np.full(N_NOISE, 5e-4)
+
+base_spec = SimulationSpec(
+    model=mpr,
+    integrator=IntegratorSpec(
+        method="heun", dt=DT,
+        stochastic=True, noise_nsig=NSIG, noise_seed=0,
+    ),
+    coupling=CouplingSpec("linear", a=1.0),  # a=1 so that G is the only scale
+    monitors=(MonitorSpec("tavg", period=1.0),),
+    weights=W,
+)
+
+
+# ---------------------------------------------------------------------------
+# 1. G sweep
+# ---------------------------------------------------------------------------
+
+print("Running G sweep …")
+
+G_values   = np.linspace(0.0, 0.5, 8)
+sweep_G    = SweepSpec(params={"G": G_values}, same_noise=True)
+results_G  = Sweeper(base_spec, sweep_G, backend="cuda").run(DURATION)
+
+# Collect mean activity (averaged across nodes and time) for each G
+mean_activity = []
+for res in results_G:
+    t, d = res["tavg"]          # d: (n_time, n_sv, n_nodes)
+    mean_activity.append(float(d[:, 0, :].mean()))
+
+print(f"  G values  : {np.round(G_values, 3)}")
+print(f"  mean |r|  : {np.round(mean_activity, 4)}")
+
+
+# ---------------------------------------------------------------------------
+# 2. same_noise comparison — eta sweep
+# ---------------------------------------------------------------------------
+
[…]
