### 2026-05-27 4e22214
fix: Improve code formatting and readability in CUDA sweep demo; adjust noise parameter and add CUDA availability check
