+++ docs/examples/simulator_models/cuda_sweep_demo.py
@@ -30,6 +30,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
 
 try:
     from numba import cuda as _cuda
+
     if not _cuda.is_available():
         print("No CUDA device found. Exiting.")
         sys.exit(0)
@@ -47,28 +48,30 @@ from vbi.simulator.spec import (
 )
 from vbi.simulator.spec.sweep import SweepSpec
 
-
 # ---------------------------------------------------------------------------
 # Shared setup
 # ---------------------------------------------------------------------------
 
-N_NODES  = 8
-DT       = 0.1     # ms
-DURATION = 500.0   # ms
+N_NODES = 8
+DT = 0.1  # ms
+DURATION = 500.0  # ms
 
 rng = np.random.default_rng(0)
-W   = np.abs(rng.standard_normal((N_NODES, N_NODES))).astype(np.float64)
+W = np.abs(rng.standard_normal((N_NODES, N_NODES))).astype(np.float64)
 np.fill_diagonal(W, 0.0)
-W  /= W.sum(axis=1, keepdims=True).clip(1e-8)
+W /= W.sum(axis=1, keepdims=True).clip(1e-8)
 
-N_NOISE  = len(mpr.noise_indices)
-NSIG     = np.full(N_NOISE, 5e-4)
+N_NOISE = len(mpr.noise_indices)
+NSIG = np.full(N_NOISE, 0.1)
 
 base_spec = SimulationSpec(
     model=mpr,
     integrator=IntegratorSpec(
-        method="heun", dt=DT,
-        stochastic=True, noise_nsig=NSIG, noise_seed=0,
+        method="heun",
+        dt=DT,
+        stochastic=True,
+        noise_nsig=NSIG,
+        noise_seed=0,
     ),
     coupling=CouplingSpec("linear", a=1.0),  # a=1 so that G is the only scale
     monitors=(MonitorSpec("tavg", period=1.0),),
@@ -82,14 +85,14 @@ base_spec = SimulationSpec(
 
 print("Running G sweep …")
 
-G_values   = np.linspace(0.0, 0.5, 8)
-sweep_G    = SweepSpec(params={"G": G_values}, same_noise=True)
-results_G  = Sweeper(base_spec, sweep_G, backend="cuda").run(DURATION)
+G_values = np.linspace(0.0, 0.5, 8)
+sweep_G = SweepSpec(params={"G": G_values}, same_noise=True)
+results_G = Sweeper(base_spec, sweep_G, backend="cuda").run(DURATION)
 
 # Collect mean activity (averaged across nodes and time) for each G
 mean_activity = []
 for res in results_G:
-    t, d = res["tavg"]          # d: (n_time, n_sv, n_nodes)
+    t, d = res["tavg"]  # d: (n_time, n_sv, n_nodes)
     mean_activity.append(float(d[:, 0, :].mean()))
 
 print(f"  G values  : {np.round(G_values, 3)}")
@@ -112,29 +115,29 @@ results_indep = Sweeper(base_spec, sweep_indep, backend="cuda").run(DURATION)
 
 # Variance of the mean activity across sweep samples (lower with same_noise=True
 # when parameters are close together — noise no longer adds sample-to-sample spread)
-means_same  = [float(res["tavg"][1][:, 0, :].mean()) for res in results_same]
+means_same = [float(res["tavg"][1][:, 0, :].mean()) for res in results_same]
 means_indep = [float(res["tavg"][1][:, 0, :].mean()) for res in results_indep]
 mean_delta = np.asarray(means_indep) - np.asarray(means_same)
 traj_delta = []
 for res_a, res_b in zip(results_same, results_indep):
[…]
