### 2026-05-27 6460995
fix: Update .dockerignore and Dockerfile to streamline build process; remove unnecessary entries and dependencies
