+++ notes/PR_numba_simulator_backend_review.md
@@ -1,0 +1,280 @@
+# PR Review: `vbi.simulator` Numba Backend
+
+Reviewed: 2026-05-27
+
+Scope:
+- `vbi/simulator/backend/numba_/simulator.py`
+- `vbi/simulator/backend/numba_/sweeper.py`
+- `vbi/simulator/backend/numba_/codegen.py`
+- `vbi/simulator/backend/numba_/_nb_sim.py`
+- Numba validation tests under `vbi/tests/validation/`
+
+## Summary
+
+The Numba backend has a solid core for deterministic/stochastic single-run simulation and parallel sweeps. The generated `dfun`, static JIT kernels, delay ring buffer, bounds handling, stimulation injection, and raw/tavg single-run parity are covered by useful validation tests. The existing Numba validation suites pass.
+
+The remaining concerns are mostly API and semantic parity issues with the NumPy backend. The highest-risk items are BOLD monitor mismatch, sweep monitor post-processing being skipped, and stochastic sweep seed semantics ignoring `same_noise`. These can produce silently different scientific outputs while still returning plausible shapes.
+
+## Verification Run
+
+Commands run:
+
+```bash
+python3 -m pytest vbi/tests/validation/test_mpr_numba.py -q
+python3 -m pytest vbi/tests/validation/test_new_models_numba.py -q
+python3 -m pytest vbi/tests/validation/test_wilson_cowan_numba.py vbi/tests/validation/test_jansen_rit_numba.py -q
+```
+
+Results:
+- `test_mpr_numba.py`: 19 passed, 1 warning
+- `test_new_models_numba.py`: 54 passed, 1 warning
+- `test_wilson_cowan_numba.py` + `test_jansen_rit_numba.py`: 24 passed, 1 warning
+
+These tests confirm broad execution coverage, but they do not currently cover the monitor and sweep parity cases listed below.
+
+## Findings
+
+### 1. High: Numba BOLD monitor advances Balloon-Windkessel dynamics in milliseconds instead of seconds
+
+Location: `vbi/simulator/backend/numba_/simulator.py:74`
+
+`_apply_monitor("bold", ...)` computes `dt_raw` from `raw_times`, where times are in milliseconds, then calls:
+
+```python
+bw_state = _bw_step(bw_state, neural, dt_raw, _BW_DEFAULTS)
+```
+
+The NumPy `BoldMonitor` explicitly passes seconds:
+
+```python
+self._bw = _bw_step(self._bw, neural, self.dt * 1e-3, self._bw_p)
+```
+
+Because `_bw_step()` models BW ODE time constants in seconds, the Numba BOLD path currently integrates about 1000x too fast. This will distort BOLD amplitude and timing while still returning arrays of the expected shape.
+
+Recommendation:
+- Change the Numba BOLD call to pass `dt_raw * 1e-3`.
+- Add a Numba-vs-NumPy BOLD parity test with the same deterministic simulation, same `tr`, and a duration long enough to produce at least one BOLD sample.
+
+Suggested test:
+- Build the same model/spec with `backend="numpy"` and `backend="numba"`.
+- Use `MonitorSpec(kind="bold", tr=...)`.
+- Compare time arrays exactly or with `np.testing.assert_allclose`.
+- Compare data with a tolerance appropriate for numerical ordering differences.
+
+### 2. High: Numba BOLD output shape differs from the NumPy backend
+
+Location: `vbi/simulator/backend/numba_/simulator.py:88`
+
+The Numba BOLD path returns:
+
+```python
+np.stack(data_bold)
+```
+
+Each `data_bold` item is shape `(n_nodes,)`, so the final output is `(n_samples, n_nodes)`. The NumPy `BoldMonitor` appends `bold[np.newaxis, :]`, so its output is `(n_samples, 1, n_nodes)`.
+
[…]
