+++ notes/PR_cpp_simulator_backend_review.md
@@ -1,0 +1,413 @@
+# PR Review: `vbi.simulator` C++ Backend
+
+Reviewed: 2026-05-27
+
+Scope:
+- `vbi/simulator/backend/cpp/simulator.py`
+- `vbi/simulator/backend/cpp/sweeper.py`
+- `vbi/simulator/backend/cpp/codegen.py`
+- `vbi/simulator/backend/cpp/build.py`
+- `vbi/simulator/backend/cpp/_src/*.mako`
+- `vbi/simulator/backend/cpp/_src/runtime.hpp`
+- C++ validation tests under `vbi/tests/validation/`
+
+## Current Status
+
+Follow-up review after the committed corrections is included at the end of this file. In short:
+
+- Most original high/medium findings have been addressed.
+- Focused C++ validation passes:
+  - `test_cpp_sweep.py`: 21 passed
+  - `test_mpr_cpp.py` + `test_cpp_models.py`: 19 passed
+- Remaining items are narrower:
+  - `raw` can still be decimated in mixed-monitor sweeps when a decimated monitor comes first.
+  - Pipeline `t_cut` is still applied before monitor post-processing.
+  - Default cache writes can fail in read-only home directories.
+  - CMake still hardcodes native/fast-math flags.
+  - Build cache writes are still not concurrency-safe.
+
+## Summary
+
+The C++ backend has a strong deterministic core. It compiles model-specific code, validates against NumPy across several models, releases the GIL during simulation, and has useful sweep coverage for serial, parallel, batching, and pipeline paths. The focused C++ validation suites pass locally.
+
+The main risks are not in the generated C++ arithmetic itself; they are in the Python wrapper/sweeper contract around monitors, stochastic sweep seed semantics, parameter validation, and build/cache robustness. The highest-priority item is sweep monitor correctness: in several C++ sweep paths raw or decimated raw arrays are labeled as processed monitor outputs.
+
+## Verification Run
+
+Commands run:
+
+```bash
+python3 -m pytest vbi/tests/validation/test_mpr_cpp.py -q
+python3 -m pytest vbi/tests/validation/test_cpp_models.py -q
+python3 -m pytest vbi/tests/validation/test_cpp_sweep.py -q
+```
+
+Results:
+- `test_mpr_cpp.py`: 3 passed, 1 warning
+- `test_cpp_models.py`: 16 passed, 1 warning
+- `test_cpp_sweep.py`: 21 passed, 1 warning
+
+These tests give good coverage for deterministic C++ vs NumPy raw trajectories, stochastic smoke tests, delayed MPR, serial/parallel sweep consistency, and some pipeline feature checks. They do not fully cover monitor semantic parity in C++ sweeps.
+
+## What Looks Good
+
+- The generated deterministic integration loop is simple and easy to audit: coupling, stimulus injection, integration, bounds, history update, and recording happen in a clear order.
+- The C++ extension releases the GIL while running the simulation, so `ThreadPoolExecutor` can provide real parallelism for sweep workloads.
+- The cache key includes a template hash, which avoids reusing stale binaries after template changes.
+- The backend has helpful prerequisite checks and error messages for missing `mako`, `pybind11`, or a compiler.
+- Existing validation covers several equation structures: MPR, Jansen-Rit, Wilson-Cowan, Reduced Wong-Wang, and Wong-Wang Exc/Inh.
+- Delayed coupling is covered against NumPy for MPR.
+
+## Findings
+
+### 1. High: C++ sweeps can return processed monitor names backed by unprocessed raw data
+
+Locations:
+- `vbi/simulator/backend/cpp/sweeper.py:137`
+- `vbi/simulator/backend/cpp/sweeper.py:191`
+- `vbi/simulator/backend/cpp/sweeper.py:270`
+
[…]

[…]
