### 2026-05-28 a1154f6
feat: Add VBI Docker Image Builder and Container Manager script; implement various commands for container management and image handling
