+++ vbi-docker.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-# VBI Docker Container Startup Script
-# This script provides easy ways to run the VBI Docker container
+# VBI Docker Image Builder and Container Manager
 
 set -e
 
@@ -17,7 +16,6 @@ IMAGE_NAME="vbi:latest"
 CONTAINER_NAME="vbi-workspace"
 DEFAULT_PORT=8888
 
-# Function to print colored output
 print_info() {
     echo -e "${BLUE}[INFO]${NC} $1"
 }
@@ -34,21 +32,20 @@ print_error() {
     echo -e "${RED}[ERROR]${NC} $1"
 }
 
-# Function to ask for confirmation
 confirm_action() {
     local message="$1"
     local default="${2:-n}"
-    
+
     if [ "$default" = "y" ]; then
         prompt="[Y/n]"
     else
         prompt="[y/N]"
     fi
-    
+
     echo -e "${YELLOW}[WARNING]${NC} $message"
     read -p "Do you want to continue? $prompt: " -n 1 -r
-    echo    # Move to a new line
-    
+    echo
+
     if [ "$default" = "y" ]; then
         [[ $REPLY =~ ^[Nn]$ ]] && return 1 || return 0
     else
@@ -56,7 +53,6 @@ confirm_action() {
     fi
 }
 
-# Function to check if Docker is running
 check_docker() {
     if ! docker info > /dev/null 2>&1; then
         print_error "Docker is not running. Please start Docker and try again."
@@ -64,32 +60,33 @@ check_docker() {
     fi
 }
 
-# Function to check if image exists
+# Returns 0 if host has nvidia-smi and Docker has the nvidia runtime registered
+check_gpu_support() {
+    command -v nvidia-smi > /dev/null 2>&1 && \
+        docker info 2>/dev/null | grep -qi "nvidia"
+}
+
 check_image() {
-    if ! docker image inspect $IMAGE_NAME > /dev/null 2>&1; then
+    if ! docker image inspect "$IMAGE_NAME" > /dev/null 2>&1; then
         print_error "VBI Docker image '$IMAGE_NAME' not found."
         print_info "Please build the image first with: docker build -t $IMAGE_NAME ."
         exit 1
     fi
 }
 
-# Function to check if image exists (returns true/false without exiting)
 image_exists() {
-    docker image inspect $IMAGE_NAME > /dev/null 2>&1
+    docker image inspect "$IMAGE_NAME" > /dev/null 2>&1
 }
 
-# Function to build VBI Docker image
 build_image() {
     local force_rebuild="$1"
-    
-    # Check if Dockerfile exists
+
     if [ ! -f "Dockerfile" ]; then
         print_error "Dockerfile not found in current directory: $(pwd)"
         print_info "Please run this script from the directory containing the Dockerfile."
         exit 1
     fi
-    
-    # Check if image already exists
+
     if image_exists && [ "$force_rebuild" != "--force" ] && [ "$force_rebuild" != "-f" ]; then
         print_warning "VBI Docker image '$IMAGE_NAME' already exists."
         echo ""
@@ -98,31 +95,26 @@ build_image() {
             exit 0
         fi
     fi
-    
+
     print_info "Building VBI Docker image..."
     print_info "This may take several minutes..."
-    
-    # Build the image
-    if docker build -t $IMAGE_NAME .; then
+
+    if docker build -t "$IMAGE_NAME" .; then
         print_success "VBI Docker image '$IMAGE_NAME' built successfully!"
-        
-        # Show image info
         print_info "Image details:"
-        docker images $IMAGE_NAME --format "table {{.Repository}}\t{{.Tag}}\t{{.ID}}\t{{.CreatedAt}}\t{{.Size}}"
[…]
