### 2026-05-28 e5f97c4
fix: Update Dockerfile to ensure compatibility with CUDA 12 runtime; streamline installation process
