### 2026-05-28 669fdcf
docs: Add section on upcoming unified API transition; outline changes and documentation plans
