### 2026-05-28 6869d75
feat: Add GitHub Actions workflow for publishing to PyPI; automate package build and release process
