### 2026-05-28 fdf9a9b
fix: Comment out environment variable in GitHub Actions workflow for clarity
