### 2026-05-28 9054bd7
feat: Implement end-to-end VBIInference API; streamline workflow from prior sampling to posterior building
