+++ notes/milestones_inference.md
@@ -808,55 +808,354 @@ mdn.train(theta, x, device='cuda')
 
 ---
 
-## MI6 — End-to-end VBI integration
+## MI6 — End-to-end `VBIInference` API
 
-**Goal:** Seamless workflow from VBI simulator → features → posterior.
+**Goal:** One object owns the full workflow: prior sampling → sweep simulation
+→ feature extraction → SNPE training → posterior.  No manual glue code.
+
+### Target API
 
 ```python
-from vbi.simulator import Sweeper
-from vbi.cde import MAFEstimator, BoxUniformPrior
-from vbi.inference_api import VBIInference
+from vbi.inference import VBIInference, BoxUniform
+import numpy as np
 
-# Fully integrated — single API, no manual glue
+prior = BoxUniform(low=np.array([0.5, -5.5]), high=np.array([5.0, -3.0]))
+
 inf = VBIInference(
-    spec=sim_spec,
-    pipeline=feature_pipeline,
-    estimator=MAFEstimator(n_flows=4, backend='jax'),
-    prior=BoxUniformPrior(low=[0.0, -5.0], high=[2.0, 0.0]),
+    sim_spec   = sim_spec,          # SimulationSpec (model + connectivity)
+    prior      = prior,
+    pipeline   = feature_pipeline,  # FeaturePipeline instance
+    density_estimator = "maf",      # "maf" | "mdn" | "nsf"
+    sim_backend = "numba",          # backend for Sweeper
+    backend     = "auto",           # backend for density estimator
 )
 
-# Round 1
-inf.simulate(n_sims=2000, backend='numba')
-inf.train()
+# --- Round 1: simulate from prior, train ---
+theta, x = inf.simulate(num_simulations=2000, duration=5000.0, seed=0)
+estimator = inf.train(training_batch_size=256, stop_after_epochs=30)
+posterior  = inf.build_posterior(estimator)
 
-# Round 2 (simulate from current posterior)
-inf.simulate(n_sims=1000, round=2)
-inf.train()
+# --- Round 2: simulate from posterior (sequential) ---
+theta2, x2 = inf.simulate(num_simulations=500, duration=5000.0,
+                           proposal=posterior, x_obs=x_obs)
+estimator2 = inf.train()
+posterior2  = inf.build_posterior(estimator2)
 
-# Posterior
-posterior = inf.build_posterior(x_obs=observed_features)
-samples = posterior.sample(5000)
+# --- Use the posterior ---
+samples    = posterior2.sample((2000,), x=x_obs)
+inf.save("run_mpr_g.npz")
+
+# --- Reload and continue ---
+inf2 = VBIInference.load("run_mpr_g.npz", sim_spec=sim_spec, pipeline=feature_pipeline)
+```
+
+### File layout
+
+```
+vbi/inference/
+  _vbi_inference.py    NEW — VBIInference class
+  _utils.py            ADD simulate_for_vbi_sweep() helper
+  __init__.py          ADD VBIInference export
+```
+
+---
+
+### Step 1 — Simulation bridge in `_utils.py`
+
+**New function:** `simulate_for_vbi_sweep`
+
+Bridges `prior.sample(n)` → `SweepSpec(params=theta_array)` → `Sweeper.run()`
+→ `FeaturePipeline.extract()` → `(theta, x)` numpy arrays.
+
+```python
+def simulate_for_vbi_sweep(
+    sim_spec,
+    prior,
+    pipeline,
+    num_simulations: int,
+    duration: float,
+    sim_backend: str = "numba",
+    seed: int | None = None,
+    proposal=None,          # None = sample from prior; Posterior = sample from posterior
+    x_obs=None,             # required when proposal is a Posterior
[…]
