### 2026-05-28 418e56b
feat: Enhance VBIInference with caching and parameter naming; streamline simulation workflow
