+++ notes/milestones_inference.md
@@ -812,6 +812,28 @@ mdn.train(theta, x, device='cuda')
 
 **Goal:** One object owns the full workflow: prior sampling → sweep simulation
 → feature extraction → SNPE training → posterior.  No manual glue code.
+Expensive simulations can be cached to disk so features can be recomputed
+without re-running the sweeper.
+
+### Resolved design decisions
+
+**Q1 — Sweeper return format with pipeline:**
+`NumpySweeper.run(dur)` returns `(labels, values)` where
+`values.shape == (n_samples, n_params + n_features)`, params first.
+`simulate_for_vbi_sweep` therefore calls `sweeper.run()` and slices:
+`theta = values[:, :n_params]`, `x = values[:, n_params:]`.
+
+**Q2 — `SimulationSpec.from_dict`:** Does not exist.
+Step 4 must add a minimal `SimulationSpec.from_dict(d)` to
+`vbi/simulator/spec/simulation.py`.
+
+**Q3 — `param_names` location:** Add `param_names: list[str] | None = None`
+to all prior constructors in `_prior.py` (see Step 0).
+Rationale: the prior semantically owns its parameter names; this keeps
+`VBIInference` clean, makes pairplots automatic, and matches how sbi
+users think about priors.
+
+---
 
 ### Target API
 
@@ -819,33 +841,52 @@ mdn.train(theta, x, device='cuda')
 from vbi.inference import VBIInference, BoxUniform
 import numpy as np
 
-prior = BoxUniform(low=np.array([0.5, -5.5]), high=np.array([5.0, -3.0]))
-
-inf = VBIInference(
-    sim_spec   = sim_spec,          # SimulationSpec (model + connectivity)
-    prior      = prior,
-    pipeline   = feature_pipeline,  # FeaturePipeline instance
-    density_estimator = "maf",      # "maf" | "mdn" | "nsf"
-    sim_backend = "numba",          # backend for Sweeper
-    backend     = "auto",           # backend for density estimator
+prior = BoxUniform(
+    low=np.array([0.5, -5.5]),
+    high=np.array([5.0, -3.0]),
+    param_names=["G", "eta"],       # NEW — see Step 0
 )
 
-# --- Round 1: simulate from prior, train ---
-theta, x = inf.simulate(num_simulations=2000, duration=5000.0, seed=0)
+inf = VBIInference(
+    sim_spec          = sim_spec,
+    prior             = prior,
+    pipeline          = feature_pipeline,
+    density_estimator = "maf",
+    sim_backend       = "numba",
+    backend           = "auto",
+)
+
+# --- Round 1: simulate → cache raw → extract features → train ---
+theta, x = inf.simulate(
+    num_simulations = 2000,
+    duration        = 5000.0,
+    seed            = 0,
+    cache_dir       = "./sim_cache/r1/",   # optional; omit to skip caching
+    chunk_size      = 500,                 # simulations per chunk (GPU memory control)
+)
 estimator = inf.train(training_batch_size=256, stop_after_epochs=30)
 posterior  = inf.build_posterior(estimator)
 
-# --- Round 2: simulate from posterior (sequential) ---
-theta2, x2 = inf.simulate(num_simulations=500, duration=5000.0,
-                           proposal=posterior, x_obs=x_obs)
+# --- Re-extract with different features, no re-simulation ---
+new_pipeline = FeaturePipeline(new_cfg, signal="tavg", t_cut=1000.0)
+theta2, x2 = VBIInference.extract_from_cache("./sim_cache/r1/", new_pipeline)
+
[…]
