### 2026-05-28 7bfbd42
feat: Add simulation caching functionality and tests for VBI sweep; enhance performance and usability Step 0 - param_names + _resolved_param_names on all 8 prior classes, with eager validation Step 1 - simulate_for_vbi_sweep in _utils.py (5 tests) Step 2 - simulate_for_vbi_sweep_cached + extract_from_cache (9 tests)
