+++ vbi/inference/_prior.py
@@ -12,6 +12,17 @@ import numpy as np
 from scipy.special import gammaln
 
 
+def _resolve_param_names(param_names, dim: int) -> list[str]:
+    if param_names is not None:
+        names = list(param_names)
+        if len(names) != dim:
+            raise ValueError(
+                f"param_names has {len(names)} entries but prior has dim={dim}."
+            )
+        return names
+    return [f"p{i}" for i in range(dim)]
+
+
 class BoxUniform:
     """
     Independent uniform distribution on a hyperrectangle.
@@ -31,7 +42,7 @@ class BoxUniform:
     (1000, 2)
     """
 
-    def __init__(self, low, high):
+    def __init__(self, low, high, param_names=None):
         self.low  = np.asarray(low,  dtype=np.float64)
         self.high = np.asarray(high, dtype=np.float64)
         if self.low.shape != self.high.shape:
@@ -39,6 +50,9 @@ class BoxUniform:
         if np.any(self.low >= self.high):
             raise ValueError("low must be strictly less than high for every dimension.")
         self._log_volume = np.sum(np.log(self.high - self.low))
+        self.param_names = param_names
+        if param_names is not None:
+            _resolve_param_names(param_names, self.dim)
 
     @property
     def event_shape(self):
@@ -47,6 +61,10 @@ class BoxUniform:
     @property
     def dim(self) -> int:
         return int(self.low.size)
+
+    @property
+    def _resolved_param_names(self) -> list[str]:
+        return _resolve_param_names(self.param_names, self.dim)
 
     def sample(self, sample_shape, seed: int | None = None) -> np.ndarray:
         """
@@ -98,15 +116,22 @@ class Gaussian:
     std  : array-like  shape (d,)  — standard deviations (not variances)
     """
 
-    def __init__(self, mean, std):
+    def __init__(self, mean, std, param_names=None):
         self.mean = np.asarray(mean, dtype=np.float64)
         self.std  = np.asarray(std,  dtype=np.float64)
         if np.any(self.std <= 0):
             raise ValueError("std must be strictly positive.")
+        self.param_names = param_names
+        if param_names is not None:
+            _resolve_param_names(param_names, self.dim)
 
     @property
     def dim(self) -> int:
         return int(self.mean.size)
+
+    @property
+    def _resolved_param_names(self) -> list[str]:
+        return _resolve_param_names(self.param_names, self.dim)
 
     def sample(self, sample_shape, seed: int | None = None) -> np.ndarray:
         rng  = np.random.default_rng(seed)
@@ -136,14 +161,21 @@ class CustomPrior:
     dim         : int  dimensionality of the parameter space
     """
 
-    def __init__(self, sample_fn, log_prob_fn, dim: int):
+    def __init__(self, sample_fn, log_prob_fn, dim: int, param_names=None):
         self._sample_fn   = sample_fn
         self._log_prob_fn = log_prob_fn
         self._dim         = dim
+        self.param_names  = param_names
+        if param_names is not None:
+            _resolve_param_names(param_names, self.dim)
 
     @property
     def dim(self) -> int:
         return self._dim
+
+    @property
+    def _resolved_param_names(self) -> list[str]:
+        return _resolve_param_names(self.param_names, self.dim)
 
     def sample(self, sample_shape, seed: int | None = None) -> np.ndarray:
         return np.asarray(self._sample_fn(sample_shape))
@@ -169,7 +201,7 @@ class MultivariateNormal:
     cov  : array-like  (d, d)  — positive-definite covariance matrix
     """
 
-    def __init__(self, mean, cov):
+    def __init__(self, mean, cov, param_names=None):
         self.mean = np.asarray(mean, dtype=np.float64)
[…]
