### 2026-05-28 e8e9316
feat: Implement VBIInference class and associated tests; establish end-to-end SBI workflow
- Add param_names and _resolved_param_names to all 8 prior classes
- Implement simulate_for_vbi_sweep in _utils.py for batch simulation
- Add simulate_for_vbi_sweep_cached and extract_from_cache for simulation reuse
- Implement VBIInference core with __init__, simulate, train, build_posterior, get_simulations, and extract_from_cache methods

Tests: 5/5 for simulate_for_vbi_sweep, 9/9 for caching functions, 13/13 for VBIInference core
