+++ vbi/inference/_vbi_inference.py
@@ -1,0 +1,281 @@
+"""
+VBIInference — end-to-end SBI workflow object.  MI6.
+
+Owns the full loop: prior sampling → sweep simulation → feature extraction
+→ SNPE training → posterior.  Raw recordings can optionally be cached to
+disk so features can be recomputed with different pipeline settings later.
+"""
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import numpy as np
+
+from ._api import SNPE
+from ._utils import (
+    simulate_for_vbi_sweep,
+    simulate_for_vbi_sweep_cached,
+    extract_from_cache as _extract_from_cache,
+    _extract_from_cache_impl,
+)
+
+if TYPE_CHECKING:
+    from ._posterior import Posterior
+    from ._estimators import ConditionalDensityEstimator
+
+log = logging.getLogger(__name__)
+
+
+class VBIInference:
+    """
+    End-to-end VBI inference workflow.
+
+    Owns the full loop: prior sampling → sweep simulation → feature
+    extraction → SNPE training → posterior.
+
+    Parameters
+    ----------
+    sim_spec : SimulationSpec
+    prior : prior object
+        Must expose ``._resolved_param_names``, ``.sample()``, ``.log_prob()``.
+    pipeline : FeaturePipeline
+    density_estimator : 'maf' | 'mdn' | 'nsf'
+    sim_backend : str
+        Backend for the Sweeper ('numpy' | 'numba' | 'cuda' | 'jax').
+    backend : str
+        Backend for the density estimator ('auto' | 'numpy' | 'jax').
+    show_progress_bars : bool
+    embedding_net : EmbeddingNet | None
+
+    Examples
+    --------
+    >>> inf = VBIInference(sim_spec, prior, pipeline)
+    >>> theta, x = inf.simulate(num_simulations=2000, duration=5000.0)
+    >>> est = inf.train(training_batch_size=256, stop_after_epochs=30)
+    >>> post = inf.build_posterior(est)
+    >>> samples = post.sample((1000,), x=x_obs)
+    >>> inf.save("run.npz")
+    """
+
+    def __init__(
+        self,
+        sim_spec,
+        prior,
+        pipeline,
+        density_estimator: str = "maf",
+        sim_backend: str = "numba",
+        backend: str = "auto",
+        show_progress_bars: bool = True,
+        embedding_net=None,
+    ):
+        self._sim_spec          = sim_spec
+        self._prior             = prior
+        self._pipeline          = pipeline
+        self._sim_backend       = sim_backend
+        self._de_type           = density_estimator
+        self._snpe              = SNPE(
+            prior               = prior,
+            density_estimator   = density_estimator,
+            backend             = backend,
+            show_progress_bars  = show_progress_bars,
+            embedding_net       = embedding_net,
+        )
+        self._feature_labels: list[str] | None = None
+        self._param_names:    list[str] | None = None
+        self._default_train_kwargs: dict       = {}
+        self._last_estimator                   = None
+
+    # ------------------------------------------------------------------
+    # Core workflow
+    # ------------------------------------------------------------------
+
+    def simulate(
+        self,
+        num_simulations: int,
+        duration: float,
+        seed: int | None = None,
+        proposal=None,
+        x_obs=None,
+        cache_dir=None,
+        chunk_size: int = 500,
+    ) -> tuple[np.ndarray, np.ndarray]:
+        """
+        Sample parameters and run the sweep simulator.
+
+        Appends the resulting ``(theta, x)`` pairs to the internal SNPE
+        object, ready for training.
+
+        Parameters
+        ----------
+        num_simulations : int
+        duration : float  simulation length in ms
[…]
