### 2026-05-28 5a47a8c
feat: Add save/load functionality to VBIInference; implement estimator serialization and deserialization
