+++ vbi/inference/_vbi_inference.py
@@ -28,6 +28,93 @@ if TYPE_CHECKING:
 log = logging.getLogger(__name__)
 
 
+# ---------------------------------------------------------------------------
+# Estimator serialisation helpers
+# ---------------------------------------------------------------------------
+
+def _pack_estimator(est, data: dict) -> None:
+    """
+    Write all state needed to reconstruct ``est`` for inference (not training)
+    into the flat numpy array dict ``data`` using an ``est__`` prefix.
+
+    Saved beyond the basic ``weights / param_dim / feature_dim``:
+      - z-score normalizers  (theta_mean/std, x_mean/std)
+      - autoregressive masks (model_constants per flow)
+      - ActNorm init flags
+      - PCA components (if used)
+      - n_flows (so load can set up the right list sizes)
+    """
+    for k, v in est.weights.items():
+        data[f"est__w__{k}"] = np.asarray(v)
+
+    data["est__param_dim"]   = np.array(est.param_dim)
+    data["est__feature_dim"] = np.array(est.feature_dim)
+    data["est__n_flows"]     = np.array(getattr(est, "n_flows", 0))
+
+    for attr in ("theta_mean", "theta_std", "x_mean", "x_std"):
+        val = getattr(est, attr, None)
+        if val is not None:
+            data[f"est__{attr}"] = np.asarray(val)
+
+    mc = getattr(est, "model_constants", None)
+    if mc is not None:
+        for k_flow, layer in enumerate(mc["layers"]):
+            for key in ("M1", "M2", "Mm", "perm", "inv_perm"):
+                if key in layer:
+                    data[f"est__mc__{k_flow}__{key}"] = np.asarray(layer[key])
+
+    actnorm = getattr(est, "_actnorm_initialized", None)
+    if actnorm is not None:
+        data["est__actnorm_init"] = np.array(actnorm, dtype=bool)
+
+    if getattr(est, "_use_pca", False) and getattr(est, "_pca_components", None) is not None:
+        data["est__pca_components"] = np.asarray(est._pca_components)
+
+
+def _unpack_estimator(d, EstCls):
+    """
+    Reconstruct a fully functional estimator from the flat array dict ``d``
+    (keys produced by ``_pack_estimator``).
+
+    Uses ``EstCls(n_flows=n)`` so ``__post_init__`` initialises the correct
+    list sizes, then overrides all attributes with the restored values.
+    """
+    n_flows = int(d["est__n_flows"]) if "est__n_flows" in d else 4
+    est = EstCls(n_flows=n_flows)
+
+    est.weights        = {k[len("est__w__"):]: d[k] for k in d if k.startswith("est__w__")}
+    est.param_dim      = int(d["est__param_dim"])
+    est.feature_dim    = int(d["est__feature_dim"])
+    est._dims_inferred = True
+    est.loss_history   = []
+
+    for attr in ("theta_mean", "theta_std", "x_mean", "x_std"):
+        if f"est__{attr}" in d:
+            setattr(est, attr, d[f"est__{attr}"])
+
+    mc_keys = [k for k in d if k.startswith("est__mc__")]
+    if mc_keys:
+        n_mc_flows = max(int(k.split("__")[2]) for k in mc_keys) + 1
+        layers = []
+        for k_flow in range(n_mc_flows):
+            layer = {
+                key: d[f"est__mc__{k_flow}__{key}"]
+                for key in ("M1", "M2", "Mm", "perm", "inv_perm")
+                if f"est__mc__{k_flow}__{key}" in d
+            }
+            layers.append(layer)
+        est.model_constants = {"layers": layers}
+
+    if "est__actnorm_init" in d:
[…]
