### 2026-05-28 f4b77a0
feat: Implement configuration loading and parsing for VBIInference; add tests for from_config method
refactor(vbi): complete VBIInference serialization and config workflow

Completed implementation and test coverage for:
- param_names support across all 8 prior classes
- simulate_for_vbi_sweep pipeline
- simulation cache with chunked write + re-extract
- VBIInference core functionality
- estimator save/load with full state restoration
  (weights, normalizers, masks)
- from_config support for YAML/JSON configs
- SimulationSpec.from_dict integration

Test summary:
- simulate_for_vbi_sweep: 5 tes
[…]
