+++ vbi/simulator/spec/simulation.py
@@ -92,6 +92,125 @@ class SimulationSpec:
         d = self.delay_steps(dt)
         return int(d.max()) + 1 if d.size > 0 else 1
 
+    @classmethod
+    def from_dict(cls, d: dict) -> "SimulationSpec":
+        """
+        Build a SimulationSpec from a plain dictionary (e.g. parsed from YAML).
+
+        Required keys
+        -------------
+        ``model`` : str  model name, e.g. ``"mpr"``
+        ``connectivity`` : str | dict
+            Path to a ``.npz`` file containing ``weights`` (and optionally
+            ``tract_lengths``), **or** a dict with ``weights`` and optionally
+            ``tract_lengths`` as lists/arrays.
+
+        Optional keys
+        -------------
+        ``dt``, ``method``, ``stochastic``, ``noise_seed``,
+        ``monitors``  (list of ``{kind, period}`` dicts),
+        ``coupling``  (``{kind, a, b}`` dict),
+        ``node_params``  (``{name: value}`` dict),
+        ``speed``, ``noise_nsig``
+        """
+        from vbi.simulator.models import (
+            mpr, jansen_rit, wilson_cowan, reduced_wong_wang,
+            wong_wang_exc_inh, generic_2d_oscillator, kuramoto,
+            sup_hopf, linear, larter_breakspear,
+            coombes_byrne_2d, gast_sd, gast_sf, vep, ghb, sl,
+            damped_oscillator,
+        )
+        _all_models = (
+            mpr, jansen_rit, wilson_cowan, reduced_wong_wang,
+            wong_wang_exc_inh, generic_2d_oscillator, kuramoto,
+            sup_hopf, linear, larter_breakspear,
+            coombes_byrne_2d, gast_sd, gast_sf, vep, ghb, sl,
+            damped_oscillator,
+        )
+        # Register by long model name and by short Python variable alias
+        _aliases = {
+            "mpr": mpr, "jansen_rit": jansen_rit, "jansenrit": jansen_rit,
+            "wilson_cowan": wilson_cowan, "wilsoncowan": wilson_cowan,
+            "reduced_wong_wang": reduced_wong_wang,
+            "wong_wang_exc_inh": wong_wang_exc_inh,
+            "generic_2d_oscillator": generic_2d_oscillator,
+            "generic2doscillator": generic_2d_oscillator,
+            "sup_hopf": sup_hopf, "suphopf": sup_hopf,
+            "larter_breakspear": larter_breakspear,
+            "coombes_byrne_2d": coombes_byrne_2d,
+            "gast_sd": gast_sd, "gast_sf": gast_sf,
+            "ghb": ghb, "sl": sl,
+            "damped_oscillator": damped_oscillator,
+            "kuramoto": kuramoto, "linear": linear, "vep": vep,
+        }
+        _MODELS = {m.name.lower(): m for m in _all_models}
+        _MODELS.update(_aliases)
+
+        # Model
+        model_key = str(d["model"]).lower()
+        if model_key not in _MODELS:
+            raise ValueError(
+                f"Unknown model {d['model']!r}. "
+                f"Available: {sorted(_MODELS)}"
+            )
+        model = _MODELS[model_key]
+
+        # Connectivity
+        conn = d["connectivity"]
+        if isinstance(conn, str):
+            npz = np.load(conn)
+            weights      = npz["weights"]
+            tract_lengths = npz["tract_lengths"] if "tract_lengths" in npz else None
+        else:
+            weights       = np.asarray(conn["weights"])
+            tract_lengths = np.asarray(conn["tract_lengths"]) if "tract_lengths" in conn else None
+
+        # Integrator
+        integrator = IntegratorSpec(
+            method     = d.get("method", "heun"),
+            dt         = float(d.get("dt", 0.01)),
[…]
