### 2026-05-28 cb1fe2e
feat: Implement diagnostic helpers for VBIInference; add tests for plot_loss, pairplot, and run_sbc
