### 2026-05-28 90dc146
feat: Add connectivity utilities for loading and saving SC matrices; integrate with SimulationSpec
