+++ vbi/simulator/spec/__init__.py
@@ -5,6 +5,7 @@ from .monitor import MonitorSpec
 from .simulation import SimulationSpec
 from .stimulus import StimSpec, build_stim_data
 from .sweep import SweepSpec
+from .connectivity import prepare_connectivity, prepare_connectivity_from_npz, save_connectivity
 
 __all__ = [
     "ModelSpec", "StateVar", "Parameter",
@@ -15,4 +16,7 @@ __all__ = [
     "StimSpec",
     "build_stim_data",
     "SweepSpec",
+    "prepare_connectivity",
+    "prepare_connectivity_from_npz",
+    "save_connectivity",
 ]
