+++ vbi/simulator/spec/connectivity.py
@@ -1,0 +1,180 @@
+"""
+Connectivity utilities — load, normalise, and save SC matrices.
+
+Users bring connectivity data in many formats; this module provides a
+single entry point that accepts numpy arrays, plain text files (.txt),
+CSV files, NPZ archives, and NumPy `.npy` files.
+
+Quick start
+-----------
+>>> W, D = prepare_connectivity('weights.txt', 'tract_lengths.txt')
+>>> spec = SimulationSpec(..., weights=W, tract_lengths=D)
+
+Or save to .npz for use with ``VBIInference.from_config``::
+
+>>> save_connectivity(W, D, 'connectivity.npz')
+
+Config YAML then just needs::
+
+    sim:
+      connectivity: connectivity.npz
+"""
+from __future__ import annotations
+
+from pathlib import Path
+
+import numpy as np
+
+
+def _load_array(src) -> np.ndarray:
+    """Load a matrix from a path (txt / csv / npz / npy) or return as-is."""
+    if isinstance(src, np.ndarray):
+        return src.astype(np.float64)
+
+    path = Path(src)
+    suffix = path.suffix.lower()
+
+    if suffix in (".txt", ".csv"):
+        return np.loadtxt(path, dtype=np.float64)
+    if suffix == ".npy":
+        return np.load(path).astype(np.float64)
+    if suffix == ".npz":
+        d = np.load(path)
+        # Accept first array if only one, else require 'weights' key
+        keys = list(d.keys())
+        if "weights" in keys:
+            return d["weights"].astype(np.float64)
+        if len(keys) == 1:
+            return d[keys[0]].astype(np.float64)
+        raise ValueError(
+            f"{path}: .npz has multiple arrays {keys}; pass the key explicitly "
+            f"or use prepare_connectivity_from_npz()."
+        )
+    raise ValueError(
+        f"Unsupported file format {suffix!r} for {path}. "
+        f"Supported: .txt, .csv, .npy, .npz"
+    )
+
+
+def prepare_connectivity(
+    weights,
+    tract_lengths=None,
+    normalize: bool = True,
+    normalise: bool | None = None,   # British spelling alias
+) -> tuple[np.ndarray, np.ndarray]:
+    """
+    Load and validate SC connectivity matrices from various sources.
+
+    Parameters
+    ----------
+    weights : array-like | str | Path
+        (n_nodes, n_nodes) weight matrix, **or** path to a file:
+
+        * ``.txt`` / ``.csv`` — space- or comma-separated text
+        * ``.npy`` — NumPy binary
+        * ``.npz`` — NumPy archive with a ``weights`` key (or a single key)
+
+    tract_lengths : array-like | str | Path | None
+        (n_nodes, n_nodes) tract-length matrix in **mm**.  Same formats as
+        ``weights``.  ``None`` → zero delays (pure ODE / SDE).
+
+    normalize : bool
+        Row-sum normalise ``weights`` so each row sums to 1.  Recommended
+        for most VBI models (MPR, VEP, GHB) where coupling strength is
+        controlled separately by a ``G`` parameter.
+
+    Returns
+    -------
+    weights       : (n, n) float64
+    tract_lengths : (n, n) float64   (zeros if None was supplied)
+
+    Examples
+    --------
+    >>> W, D = prepare_connectivity('weights.txt', 'tract_lengths.txt')
+    >>> W, D = prepare_connectivity(my_array)               # already in memory
+    >>> W, D = prepare_connectivity('connectivity.npz')     # npz has 'weights' key
+    """
+    if normalise is not None:
+        normalize = normalise
+
+    W = _load_array(weights)
+    if W.ndim != 2 or W.shape[0] != W.shape[1]:
+        raise ValueError(
[…]
