### 2026-05-28 18d12e8
feat: Enhance VBIInference with SBI backend support; add conversion and training functionalities
