+++ vbi/inference/_vbi_inference.py
@@ -230,6 +230,119 @@ def _make_simulator_fn(sim_spec, prior, pipeline, sim_backend, duration):
     return simulator_fn
 
 
+# ---------------------------------------------------------------------------
+# SBI inference-backend helpers  (Step 8)
+# ---------------------------------------------------------------------------
+
+def _prior_to_sbi(prior):
+    """Convert a vbi prior to a sbi/torch-compatible prior."""
+    import torch
+    from vbi.inference._prior import BoxUniform as _VBIBoxUniform
+
+    if isinstance(prior, _VBIBoxUniform):
+        from sbi.utils import BoxUniform as _SBIBoxUniform
+        return _SBIBoxUniform(
+            low  = torch.tensor(prior.low,  dtype=torch.float32),
+            high = torch.tensor(prior.high, dtype=torch.float32),
+        )
+
+    # Generic wrapper: expose torch Distribution interface
+    class _TorchWrapper(torch.distributions.Distribution):
+        arg_constraints: dict = {}
+        has_rsample = False
+
+        def __init__(self, vbi_prior):
+            self._p = vbi_prior
+            super().__init__(
+                batch_shape=torch.Size([]),
+                event_shape=torch.Size([vbi_prior.dim]),
+                validate_args=False,
+            )
+
+        def sample(self, sample_shape=torch.Size()):
+            n = sample_shape[0] if len(sample_shape) > 0 else 1
+            arr = self._p.sample((n,))
+            return torch.tensor(arr, dtype=torch.float32)
+
+        def log_prob(self, theta):
+            arr = theta.detach().cpu().numpy()
+            lp  = self._p.log_prob(arr)
+            return torch.tensor(lp, dtype=torch.float32)
+
+    return _TorchWrapper(prior)
+
+
+def _setup_sbi_engine(prior, density_estimator: str, inference_engine, show_progress_bars: bool):
+    """Initialise (or return) an sbi SNPE engine."""
+    try:
+        from sbi.inference import SNPE as _SBI_SNPE
+    except ImportError as e:
+        raise ImportError(
+            "inference_backend='sbi' requires the 'sbi' package. "
+            "Install with:  pip install sbi"
+        ) from e
+
+    if inference_engine is not None:
+        return inference_engine
+
+    torch_prior = _prior_to_sbi(prior)
+    return _SBI_SNPE(
+        prior             = torch_prior,
+        density_estimator = density_estimator,
+        show_progress_bars = show_progress_bars,
+    )
+
+
+def _filter_sbi_train_kwargs(kwargs: dict) -> dict:
+    """
+    Keep only kwargs that sbi's SNPE.train() accepts.
+    sbi uses the same kwarg names as vbi for the common ones.
+    """
+    _SBI_TRAIN_KEYS = {
+        "training_batch_size", "learning_rate", "validation_fraction",
+        "stop_after_epochs", "max_num_epochs", "clip_max_norm",
+        "num_atoms", "resume_training", "show_train_summary",
+    }
+    return {k: v for k, v in kwargs.items() if k in _SBI_TRAIN_KEYS}
+
+
+class _NumpyPosteriorWrapper:
+    """
+    Wraps an sbi Posterior so that ``sample`` and ``log_prob`` return numpy
+    arrays instead of torch tensors.  The underlying sbi posterior is
+    accessible via ``.sbi_posterior``.
+    """
+
+    def __init__(self, sbi_posterior):
+        self.sbi_posterior = sbi_posterior
+
+    def sample(self, sample_shape, x=None, seed=None):
+        import torch
+        n = sample_shape[0] if isinstance(sample_shape, tuple) else int(sample_shape)
[…]
