### 2026-05-29 3c23738
feat: Add error handling for missing PyTorch in _prior_to_sbi; update simulation methods to use torch tensors
