+++ vbi/simulator/spec/simulation.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 import hashlib
 import json
 from dataclasses import dataclass, field
+from pathlib import Path
 import numpy as np
 
 from .model import ModelSpec
