### 2026-05-29 63c7197
feat: Add validation for feature labels in cache extraction; raise error for inconsistent labels feat: Enhance estimator packing and unpacking for sbi backend; save and load estimator state feat: Add error handling for unsupported stimuli configuration in SimulationSpec.from_dict()
