+++ vbi/inference/_vbi_inference.py
@@ -18,7 +18,6 @@ from ._utils import (
     simulate_for_vbi_sweep,
     simulate_for_vbi_sweep_cached,
     extract_from_cache as _extract_from_cache,
-    _extract_from_cache_impl,
 )
 
 if TYPE_CHECKING:
@@ -50,6 +49,15 @@ def _pack_estimator(est, data: dict) -> None:
     data["est__param_dim"]   = np.array(est.param_dim)
     data["est__feature_dim"] = np.array(est.feature_dim)
     data["est__n_flows"]     = np.array(getattr(est, "n_flows", 0))
+    data["est__type"]        = np.array(type(est).__name__)
+
+    # MDN-specific hyperparams needed to reconstruct the estimator
+    n_comp = getattr(est, "n_components", None)
+    if n_comp is not None:
+        data["est__n_components"] = np.array(n_comp)
+    hidden = getattr(est, "hidden_sizes", None)
+    if hidden is not None:
+        data["est__hidden_sizes"] = np.array(list(hidden))
 
     for attr in ("theta_mean", "theta_std", "x_mean", "x_std"):
         val = getattr(est, attr, None)
@@ -76,11 +84,22 @@ def _unpack_estimator(d, EstCls):
     Reconstruct a fully functional estimator from the flat array dict ``d``
     (keys produced by ``_pack_estimator``).
 
-    Uses ``EstCls(n_flows=n)`` so ``__post_init__`` initialises the correct
-    list sizes, then overrides all attributes with the restored values.
+    Branches on whether the estimator class uses ``n_flows`` (MAF/NSF) or
+    not (MDN), reconstructing with the saved hyperparams.
     """
-    n_flows = int(d["est__n_flows"]) if "est__n_flows" in d else 4
-    est = EstCls(n_flows=n_flows)
+    import dataclasses
+    _fields = {f.name for f in dataclasses.fields(EstCls)} if dataclasses.is_dataclass(EstCls) else set()
+    if "n_flows" in _fields:
+        n_flows = int(d["est__n_flows"]) if "est__n_flows" in d else 4
+        est = EstCls(n_flows=n_flows)
+    else:
+        # MDN-like estimator: reconstruct with saved hyperparams
+        kwargs = {}
+        if "est__n_components" in d:
+            kwargs["n_components"] = int(d["est__n_components"])
+        if "est__hidden_sizes" in d:
+            kwargs["hidden_sizes"] = tuple(int(x) for x in d["est__hidden_sizes"])
+        est = EstCls(**kwargs)
 
     est.weights        = {k[len("est__w__"):]: d[k] for k in d if k.startswith("est__w__")}
     est.param_dim      = int(d["est__param_dim"])
@@ -208,7 +227,6 @@ def _make_simulator_fn(sim_spec, prior, pipeline, sim_backend, duration):
     Each call patches the SimulationSpec with the given parameter values,
     runs one simulation, and extracts features via the pipeline.
     """
-    from vbi.simulator.api import Simulator
     from vbi.simulator.spec.sweep import SweepSpec
     from vbi.simulator.api import Sweeper
 
@@ -508,10 +526,12 @@ class VBIInference:
             self._snpe.append_simulations(theta, x, proposal=proposal)
         else:
             import torch
-            self._sbi_engine.append_simulations(
-                torch.tensor(theta, dtype=torch.float32),
-                torch.tensor(x,     dtype=torch.float32),
-            )
+            theta_t = torch.tensor(theta, dtype=torch.float32)
+            x_t     = torch.tensor(x,     dtype=torch.float32)
+            if proposal is not None:
+                self._sbi_engine.append_simulations(theta_t, x_t, proposal=proposal)
[…]
