### 2026-05-29 e4f0bef
feat: Update CudaSimulator to use device array for coup_a; modify test to ensure unique outputs with same_noise=False
