### 2026-05-29 14a21af
feat: Add custom model demo for FitzHugh-Nagumo; implement model registration and simulation execution
