+++ docs/examples/simulator_models/custom_model_demo.py
@@ -1,0 +1,203 @@
+"""
+Custom model demo — adding a user-defined model to the VBI simulator.
+
+This script shows the full workflow:
+
+  1. Define a custom ModelSpec (FitzHugh-Nagumo as a concrete example).
+  2. Register it so SimulationSpec.from_dict() / from_config() can find it.
+  3. Run it directly with Simulator and with a parameter sweep.
+  4. Use it via a config dict (as would come from a YAML file).
+
+Run
+---
+    python custom_model_demo.py
+    python custom_model_demo.py --duration 200 --backend numba
+"""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+sys.dont_write_bytecode = True
+
+# Make sure the package root is importable when running from the examples dir.
+_ROOT = Path(__file__).resolve().parents[3]
+if str(_ROOT) not in sys.path:
+    sys.path.insert(0, str(_ROOT))
+
+import numpy as np
+
+# ---------------------------------------------------------------------------
+# Step 1 — Define the custom model
+# ---------------------------------------------------------------------------
+#
+# ModelSpec is the backend-agnostic description understood by every VBI
+# simulator backend (numpy, numba, CUDA, JAX, C++).
+#
+# dfun_str uses plain math with these symbols:
+#   - state variable names (defined in state_variables)
+#   - parameter names      (defined in parameters)
+#   - 'c'                  the net coupling input at this node
+#   - exp, log, sin, cos, tanh, sqrt, abs, pi
+#
+# No 'np.' prefix — the code generator handles the namespace.
+
+from vbi.simulator.spec.model import ModelSpec, StateVar, Parameter
+
+fitzhugh_nagumo = ModelSpec(
+    name="FitzHughNagumo",
+    state_variables=(
+        # v: fast membrane variable; bounded to avoid runaway
+        StateVar("v", default_init=0.0, noise=True,
+                 lower_bound=-2.5, upper_bound=2.5),
+        # w: slow recovery variable
+        StateVar("w", default_init=0.0, noise=False),
+    ),
+    parameters=(
+        Parameter("a",   0.7,   "threshold parameter"),
+        Parameter("b",   0.8,   "recovery rate"),
+        Parameter("tau", 12.5,  "time-scale separation (1/tau multiplies dw/dt)"),
+        Parameter("I",   0.5,   "external current"),
+    ),
+    # v is the coupling variable — it enters other nodes' equations via c.
+    cvar=("v",),
+    dfun_str={
+        "v": "v - (v**3) / 3.0 - w + I + c",
+        "w": "(v + a - b * w) / tau",
+    },
+    noise_variables=("v",),
+    reference="FitzHugh (1961) / Nagumo et al. (1962)",
+    dfun_latex={
+        "v": r"\dot{v} = v - \frac{v^3}{3} - w + I + c^{\rm net}",
+        "w": r"\dot{w} = \frac{v + a - b\,w}{\tau}",
+    },
+    latex_notes=(
+        r"$c^{\rm net} = G\sum_j W_{ij}\,v_j$."
+        r" Oscillatory regime: $|I| > (a - 1)$ approximately."
+    ),
+)
+
+# ---------------------------------------------------------------------------
+# Step 2 — Register the model
+# ---------------------------------------------------------------------------
+#
+# register_model() adds the model to the global lookup table used by
+# SimulationSpec.from_dict() and VBIInference.from_config().
+# Call it once, before any from_dict()/from_config() that references this model.
+#
+# The first alias you supply becomes the string to use in the config file:
+#
+#   sim:
+#     model: fitzhugh_nagumo
+#
[…]
