+++ vbi/simulator/spec/__init__.py
@@ -2,7 +2,7 @@ from .model import ModelSpec, StateVar, Parameter
 from .integrator import IntegratorSpec
 from .coupling import CouplingSpec
 from .monitor import MonitorSpec
-from .simulation import SimulationSpec
+from .simulation import SimulationSpec, register_model
 from .stimulus import StimSpec, build_stim_data
 from .sweep import SweepSpec
 from .connectivity import prepare_connectivity, prepare_connectivity_from_npz, save_connectivity
@@ -13,6 +13,7 @@ __all__ = [
     "CouplingSpec",
     "MonitorSpec",
     "SimulationSpec",
+    "register_model",
     "StimSpec",
     "build_stim_data",
     "SweepSpec",
