### 2026-05-29 d1e0fb7
feat: Add visualization step for FitzHugh-Nagumo model demo; include plotting of time series, phase portrait, and bifurcation summary
