### 2026-05-29 9f083cb
feat: Add comprehensive code review notes for MI6 VBIInference; document findings, follow-up status, and recommended next steps
