+++ notes/PR_MI6_vbi_inference_review.md
@@ -1,0 +1,406 @@
+# MI6 VBIInference Code Review
+
+Reviewed commits:
+
+- `418e56b` through `18d12e8`
+- Main files: `vbi/inference/_vbi_inference.py`, `vbi/inference/_utils.py`,
+  `vbi/inference/_prior.py`, `vbi/simulator/spec/simulation.py`,
+  `vbi/simulator/spec/connectivity.py`, and MI6 validation tests.
+
+Review stance: API correctness, end-to-end workflow reliability, checkpoint
+recoverability, and gaps against the MI6 milestone.
+
+## Summary
+
+The implementation establishes the intended end-to-end shape: `VBIInference`
+owns simulation, feature extraction, SNPE training, posterior construction,
+checkpointing, config loading, cache extraction, and diagnostic helpers. The
+NumPy path with `maf` is covered by useful smoke/integration tests.
+
+The largest risks are around persistence and backend coverage. Checkpointing is
+currently only safe for a narrow estimator shape, the optional `sbi` backend has
+several broken or misleading paths, and sequential/proposal metadata is lost
+outside the immediate in-memory SNPE object. These should be fixed before
+calling MI6 complete.
+
+## Follow-up Status: 2026-05-29
+
+Most original findings have been partially or fully addressed. The focused MI6
+validation subset now passes:
+
+```bash
+python3 -m pytest \
+  vbi/tests/validation/test_vbi_inference.py \
+  vbi/tests/validation/test_vbi_inference_config.py \
+  vbi/tests/validation/test_vbi_inference_diag.py \
+  vbi/tests/validation/test_sim_cache.py -q
+```
+
+Result:
+
+```text
+51 passed, 1 warning in 926.19s
+```
+
+`pytest-xdist` is available (`xdist 3.8.0`), so future broad runs can use:
+
+```bash
+OMP_NUM_THREADS=1 MKL_NUM_THREADS=1 OPENBLAS_NUM_THREADS=1 \
+python3 -m pytest vbi -n 8 --lf
+```
+
+Current remaining actionable items:
+
+- High: MDN checkpoint load is still broken after unpack because
+  `_offdiag_basis` is not rebuilt. A direct pack/unpack reproduction raises
+  `AttributeError: 'MDNEstimator' object has no attribute '_offdiag_basis'`.
+- High: sequential/proposal metadata is still lost in `VBIInference`;
+  `_sim_rounds` stores only `(theta, x)`, `get_simulations()` returns an empty
+  proposals list, and save/load flattens all rounds into one.
+- Medium: non-default MAF/NSF checkpoint config is still incomplete; load
+  preserves `n_flows` but not the rest of the architecture/config fields.
+- Medium: `from_config()` still does not pass through
+  `inference_backend="sbi"`.
+- Medium: cache extraction now validates labels, but still materializes all
+  rows in Python lists before stacking.
+
+Resolved or materially improved since the original review:
+
+- `Path` import/config connectivity failure is fixed.
+- `sbi` save/load no longer crashes on the old `_snpe.n_simulations` log path,
+  and trained `sbi` estimators are saved to a companion `<stem>_sbi.pt` file.
+- `sbi` simulation append now converts `theta`/`x` to `torch.float32` and passes
+  `proposal` when provided.
+- `stimuli` in config is explicitly rejected instead of silently ignored.
+- cache re-extraction validates feature-label consistency.
+- `pairplot()` now tells users to call `train()` first.
+
+## Findings
+
+### Resolved High: `VBIInference.from_config()` was broken for path-based connectivity
+
+Location: `vbi/simulator/spec/simulation.py`, `SimulationSpec.from_dict()`
+
+`from_dict()` checks `isinstance(conn, (str, Path))`, but `Path` is not
[…]
