### 2026-05-29 c717f8e
Refactor comments to standardize punctuation and formatting
