### 2026-05-29 6b2f98f
feat: Add PyYAML as a dependency for enhanced YAML support
