### 2026-05-29 87dc13a
feat: Add autograd as a dependency for enhanced automatic differentiation support
