+++ pyproject.toml
@@ -35,6 +35,7 @@ classifiers = [
 dependencies = [
     "numpy>=1.23",
     "scipy>=1.9",
+    "autograd>=1.6",
     "numba>=0.57",
     "h5py",
     "pandas",
