### 2026-05-29 410df6f
feat: Add milestones documentation for sphinx-gallery and RTD versioning strategy
