+++ notes/milestones_documentation.md
@@ -1,0 +1,316 @@
+# Documentation Milestone — sphinx-gallery + RTD Versioning
+
+## Goal
+
+Provide MNE-style documentation where every demo/tutorial:
+- Renders executed outputs (plots, tables) inline in the HTML docs
+- Offers a "Download `.py`" and "Download `.ipynb`" button per page
+- Has a "Open in Colab" badge that works without committing notebooks to git
+- Lives in clean `.py` scripts that produce readable git diffs
+
+Reference: https://mne.tools/stable/auto_tutorials/inverse/10_stc_class.html
+
+---
+
+## Chosen Stack
+
+| Tool | Role |
+|---|---|
+| `sphinx-gallery` | Executes `.py` scripts during docs build, renders outputs inline, auto-generates `.ipynb` and download buttons |
+| `ju2py` (`pip install ju2py`) | Converts existing `.ipynb` notebooks to clean sphinx-gallery-compatible `.py` scripts |
+| `pydata-sphinx-theme` | HTML theme — same as MNE, NumPy, SciPy, pandas. Replaces `furo`. Has built-in version switcher dropdown (latest/stable), sphinx-gallery thumbnail grid, and top navigation bar. |
+| ReadTheDocs | Hosts `latest` (main branch) and `stable` (latest git tag) versions |
+
+Jupytext was considered and rejected: sphinx-gallery solves discoverability,
+Colab, docs rendering, and git cleanliness all at once, matching the target
+reference exactly.
+
+---
+
+## RTD Versioning Strategy
+
+| RTD version | Source | Notebooks executed? |
+|---|---|---|
+| `latest` | `main` branch | Yes (during RTD build) |
+| `stable` | latest git tag | Yes (during RTD build) |
+| feature branches | any other branch | Not built / not published |
+
+No `dev` docs version. `latest` already serves that role.
+
+### How RTD knows which versions to build
+
+Configured in the RTD project dashboard (not in `.readthedocs.yaml`):
+- Enable "Automation Rules" → build tags matching `v*` as new versions
+- RTD auto-designates the newest tag as `stable`
+- Default branch (`main`) is always `latest`
+
+---
+
+## Steps
+
+### Step 1 — Add root-level symlink for discoverability
+
+Users expect `examples/` at repo root. Sphinx resolves files at `docs/examples/`.
+A symlink satisfies both without moving anything.
+
+```bash
+ln -s docs/examples examples
+git add examples
+git commit -m "docs: add examples/ symlink for discoverability"
+```
+
+### Step 2 — Switch to pydata-sphinx-theme
+
+Add to `docs/requirements.txt`:
+```
+pydata-sphinx-theme
+```
+
+Remove `furo` from `docs/requirements.txt`.
+
+In `docs/conf.py` replace:
+```python
+html_theme = "furo"
+html_theme_options = {
+    "default_mode": "light",
+}
+```
+with:
+```python
+html_theme = "pydata_sphinx_theme"
+html_theme_options = {
+    "navbar_start": ["navbar-logo"],
+    "navbar_center": ["navbar-nav"],
+    "navbar_end": ["version-switcher", "navbar-icon-links"],
+    "switcher": {
+        "json_url": "https://<your-rtd-project>.readthedocs.io/en/latest/_static/switcher.json",
+        "version_match": version,
+    },
+    "icon_links": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/<org>/vbi",
+            "icon": "fa-brands fa-github",
+        },
+    ],
+    "logo": {
+        "text": "vbi",
+    },
+}
+```
+
+The `switcher.json` file (placed in `docs/_static/`) lists available doc versions
+so users can switch between `latest` and `stable` from a dropdown — identical to
[…]
