### 2026-05-29 60d5eac
feat: Update Python setup and install autograd dependency for CI workflow
