+++ .github/workflows/tests.yml
@@ -24,12 +24,16 @@ jobs:
         with:
           python-version: "3.12"
           cache: pip
+          cache-dependency-path: pyproject.toml
 
       - name: Install core + CI dependencies
         run: |
           pip install --upgrade pip
           # Core simulation deps only - no mako/pybind11 (no C++ backend needed)
           pip install ".[ci]"
+          # Explicit install guards against setup.py/pyproject.toml metadata merge
+          # issues where setup.py's missing install_requires can shadow pyproject.toml deps.
+          pip install "autograd>=1.6"
 
       - name: Fast tests  (not slow, not cuda)
         run: |
