+++ vbi/inference/_diagnostics.py
@@ -6,10 +6,12 @@ Functions
 run_sbc          : Simulation-Based Calibration (rank statistics)
 check_sbc        : Uniformity tests on SBC ranks
 sbc_rank_plot    : Histogram plot of SBC ranks
+pp_plot          : Probability-Probability plot from SBC ranks
 run_tarp         : Test of Accuracy with Random Points
 check_tarp       : Expected coverage probability error
 plot_tarp        : Coverage plot (ECP vs alpha)
 c2st             : Classifier Two-Sample Test accuracy
+lc2st            : Local C2ST at a specific observation
 pairplot         : Wrapper around vbi.plot.pairplot
 conditional_pairplot : Pairplot conditioned on observed data
 plot_loss        : Training (and optional validation) loss curve
@@ -461,3 +463,182 @@ def plot_loss(loss_history, val_loss_history=None, ax=None,
     ax.legend()
     fig.tight_layout()
     return fig
+
+
+# ---------------------------------------------------------------------------
+# PP plot (Probability-Probability plot from SBC ranks)
+# ---------------------------------------------------------------------------
+
+def pp_plot(ranks: np.ndarray, labels=None, fig=None, axes=None):
+    """
+    Probability-Probability plot from SBC ranks.
+
+    For each parameter dimension, plots the empirical CDF of SBC ranks
+    against the theoretical uniform CDF.  A well-calibrated posterior
+    lies close to the diagonal.
+
+    Parameters
+    ----------
+    ranks  : ndarray (num_sbc_runs, d)   - ranks from :func:`run_sbc`
+    labels : list[str] | None            - parameter names
+    fig, axes : existing figure / axes (optional)
+
+    Returns
+    -------
+    matplotlib.figure.Figure
+    """
+    import matplotlib.pyplot as plt
+
+    ranks = np.asarray(ranks)
+    if ranks.ndim == 1:
+        ranks = ranks[:, None]
+    n_runs, d = ranks.shape
+    n_bins = n_runs + 1   # ranks are integers in [0, n_posterior_samples]
+
+    if labels is None:
+        labels = [f"$\\theta_{i}$" for i in range(d)]
+
+    ncols = min(d, 4)
+    nrows = int(np.ceil(d / ncols))
+    if fig is None or axes is None:
+        fig, axes = plt.subplots(
+            nrows, ncols,
+            figsize=(3.2 * ncols, 3.2 * nrows),
+            squeeze=False,
+        )
+
+    axes_flat = np.array(axes).ravel()
+
+    for i in range(d):
+        ax = axes_flat[i]
+        r = np.sort(ranks[:, i])
+        # Empirical CDF: fraction of ranks <= each unique value
+        ecdf_x = np.concatenate([[0], r / (n_bins - 1), [1]])
+        ecdf_y = np.concatenate([[0], np.arange(1, n_runs + 1) / n_runs, [1]])
+        ax.plot(ecdf_x, ecdf_y, color="steelblue", linewidth=1.5)
+        ax.plot([0, 1], [0, 1], "k--", linewidth=1.0, alpha=0.6)
+        ax.set_xlabel("Theoretical quantile")
+        ax.set_ylabel("Empirical CDF")
+        ax.set_title(labels[i] if i < len(labels) else f"$\\theta_{i}$")
+        ax.set_xlim(0, 1)
+        ax.set_ylim(0, 1)
+
+    # Hide unused axes
+    for j in range(d, len(axes_flat)):
+        axes_flat[j].set_visible(False)
+
+    fig.suptitle("PP plot (SBC calibration)", fontsize=12)
+    fig.tight_layout()
+    return fig
+
+
+# ---------------------------------------------------------------------------
[…]
