### 2026-05-29 8157e52
feat: Update pp_plot function to infer number of bins from data and enhance test coverage for pp_plot and lc2st
