+++ vbi/tests/test_inference_api.py
@@ -813,3 +813,169 @@ class TestDeprecatedShim:
             warnings.simplefilter("ignore", DeprecationWarning)
             from vbi.cde import MDNEstimator, MAFEstimator, ConditionalDensityEstimator
             assert MDNEstimator is not None
+
+
+# ---------------------------------------------------------------------------
+# discard_prior_samples
+# ---------------------------------------------------------------------------
+
+class TestDiscardPriorSamples:
+
+    def test_discards_first_round_on_second_append(self):
+        prior, theta, x = _linear_gaussian(n=200, d=2)
+        inf = SNPE(prior=prior)
+        inf.append_simulations(theta[:100], x[:100])
+        assert inf.n_rounds == 1
+        # Second round with discard: prior round is removed
+        inf.append_simulations(theta[100:], x[100:], discard_prior_samples=True)
+        assert inf.n_rounds == 1
+        assert inf.n_simulations == 100   # only round-1 data remains
+
+    def test_does_not_discard_on_first_append(self):
+        prior, theta, x = _linear_gaussian(n=100, d=2)
+        inf = SNPE(prior=prior)
+        # discard_prior_samples=True on round 0 should have no effect
+        inf.append_simulations(theta, x, discard_prior_samples=True)
+        assert inf.n_rounds == 1
+        assert inf.n_simulations == 100
+
+    def test_false_keeps_all_rounds(self):
+        prior, theta, x = _linear_gaussian(n=200, d=2)
+        inf = SNPE(prior=prior)
+        inf.append_simulations(theta[:100], x[:100])
+        inf.append_simulations(theta[100:], x[100:], discard_prior_samples=False)
+        assert inf.n_rounds == 2
+        assert inf.n_simulations == 200
+
+    def test_training_works_after_discard(self):
+        prior, theta, x = _linear_gaussian(n=200, d=2)
+        inf = SNPE(prior=prior, density_estimator="mdn")
+        inf.append_simulations(theta[:100], x[:100])
+        inf.append_simulations(theta[100:], x[100:], discard_prior_samples=True)
+        est  = inf.train(max_num_epochs=5, verbose=False)
+        post = inf.build_posterior(est)
+        s = post.sample((50,), x=np.array([[0.5, 0.5]]))
+        assert s.shape == (50, 2)
+
+
+# ---------------------------------------------------------------------------
+# Custom density_estimator callable
+# ---------------------------------------------------------------------------
+
+class TestCustomDensityEstimatorCallable:
+
+    def test_lambda_factory_trains_and_samples(self):
+        prior, theta, x = _linear_gaussian(n=200, d=2)
+        inf = SNPE(prior=prior, density_estimator=lambda: MAFEstimator(n_flows=2))
+        inf.append_simulations(theta, x)
+        est  = inf.train(max_num_epochs=5, verbose=False)
+        post = inf.build_posterior(est)
+        s = post.sample((100,), x=np.array([[0.5, 0.5]]))
+        assert s.shape == (100, 2)
+        assert np.all(np.isfinite(s))
+
+    def test_class_factory_trains_and_samples(self):
+        prior, theta, x = _linear_gaussian(n=200, d=2)
+        # Pass the class itself as the factory (zero-arg default constructor)
+        inf = SNPE(prior=prior, density_estimator=MDNEstimator)
+        inf.append_simulations(theta, x)
+        est  = inf.train(max_num_epochs=5, verbose=False)
+        post = inf.build_posterior(est)
+        s = post.sample((50,), x=np.array([[0.5, 0.5]]))
+        assert s.shape == (50, 2)
+
+    def test_invalid_string_still_raises(self):
[…]
