+++ vbi/tests/test_diagnostics.py
@@ -1,8 +1,9 @@
 """
 Tests for vbi.inference diagnostic tools.
 
-Covers: run_sbc, check_sbc, sbc_rank_plot, run_tarp, check_tarp, plot_tarp,
-        c2st, pairplot, conditional_pairplot, plot_loss
+Covers: run_sbc, check_sbc, sbc_rank_plot, pp_plot,
+        run_tarp, check_tarp, plot_tarp,
+        c2st, lc2st, pairplot, conditional_pairplot, plot_loss
 """
 import numpy as np
 import pytest
@@ -17,9 +18,9 @@ import matplotlib.pyplot as plt
 
 from vbi.inference import (
     SNPE, BoxUniform,
-    run_sbc, check_sbc, sbc_rank_plot,
+    run_sbc, check_sbc, sbc_rank_plot, pp_plot,
     run_tarp, check_tarp, plot_tarp,
-    c2st, pairplot, conditional_pairplot, plot_loss,
+    c2st, lc2st, pairplot, conditional_pairplot, plot_loss,
 )
 
 
@@ -436,3 +437,123 @@ class TestPlotLoss:
         fig    = plot_loss(losses)
         assert hasattr(fig, "savefig")
         plt.close("all")
+
+
+# ---------------------------------------------------------------------------
+# pp_plot
+# ---------------------------------------------------------------------------
+
+class TestPPPlot:
+
+    def test_returns_figure_1d(self):
+        ranks = np.random.randint(0, 100, (50, 1))
+        fig   = pp_plot(ranks, labels=["theta"])
+        assert hasattr(fig, "savefig")
+        plt.close("all")
+
+    def test_returns_figure_2d(self):
+        ranks = np.random.randint(0, 100, (60, 2))
+        fig   = pp_plot(ranks, labels=["G", "eta"])
+        assert hasattr(fig, "savefig")
+        plt.close("all")
+
+    def test_1d_ranks_as_vector(self):
+        """1-D rank array (not 2-D) is accepted."""
+        ranks = np.random.randint(0, 50, 40)
+        fig   = pp_plot(ranks)
+        assert hasattr(fig, "savefig")
+        plt.close("all")
+
+    def test_auto_labels_when_none(self):
+        ranks = np.random.randint(0, 80, (30, 3))
+        fig   = pp_plot(ranks)          # labels=None → auto θ_i labels
+        assert hasattr(fig, "savefig")
+        plt.close("all")
+
+    def test_many_parameters(self):
+        """More than 4 parameters wraps to multiple rows."""
+        ranks = np.random.randint(0, 60, (40, 6))
+        fig   = pp_plot(ranks)
+        assert hasattr(fig, "savefig")
+        plt.close("all")
+
+    def test_calibrated_ranks_near_diagonal(self):
+        """Uniform ranks → empirical CDF close to the diagonal."""
+        rng   = np.random.default_rng(0)
+        n_posterior = 200
+        # ranks span [0, n_posterior]; n_runs can differ
+        ranks = rng.integers(0, n_posterior + 1, (500, 1))
+        fig   = pp_plot(ranks)
+        ax    = fig.axes[0]
+        line  = ax.lines[0]         # first line = empirical CDF
+        ecdf_x = line.get_xdata()
+        ecdf_y = line.get_ydata()
+        # Empirical CDF y should be close to x (the diagonal) for uniform ranks
+        assert float(np.max(np.abs(ecdf_y - ecdf_x))) < 0.15
+        plt.close("all")
+
+    def test_from_sbc_output(self, posterior_1d):
+        """pp_plot works end-to-end with real run_sbc output."""
+        posterior, prior = posterior_1d
+        def sim(theta): return theta + 0.0
+
+        result = run_sbc(posterior, sim, prior,
[…]
