### 2026-05-29 42c99e3
feat: Refactor visualization helpers and add damped oscillator workflow example
