+++ docs/examples/workflows/damp_oscillator.py
@@ -1,0 +1,225 @@
+"""
+Damped Oscillator - SBI with vbi.inference.SNPE
+================================================
+
+Reproduce the ``damp_oscillator_cde.ipynb`` workflow using the new
+``vbi.inference`` API.  No torch, no sbi dependency.
+
+The model
+---------
+A nonlinear damped oscillator::
+
+    dx/dt = x - xy - a*x²
+    dy/dt = xy - y  - b*y²
+
+Parameters to infer: ``a`` (x-damping) and ``b`` (y-damping).
+
+Workflow
+--------
+1. Run the model at the true parameters to get the "observed" data.
+2. Define a simulator callable ``theta_1d → x_features``.
+3. Use ``simulate_for_sbi`` to batch-generate 2000 training samples.
+4. Train a MAF density estimator with ``SNPE``.
+5. Draw posterior samples and print shrinkage / z-score diagnostics.
+6. Plot: time-series, training loss, posterior pairplot.
+7. (Optional) Run SBC + PP plot for calibration check.
+
+Note
+----
+``VBIInference`` is designed for large brain models (MPR, JR, …) that use the
+``SimulationSpec`` / ``Sweeper`` pipeline.  The DampedOscillator is a
+standalone ODE without a connectome, so we use ``SNPE`` directly with a
+custom simulator function instead.
+
+Expected runtime: ~2–3 min on a modern CPU (numba backend, 2000 simulations).
+"""
+
+import sys
+from pathlib import Path
+import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from vbi.models.numba.damp_oscillator import DO
+from vbi.inference import (
+    SNPE, BoxUniform,
+    simulate_for_sbi,
+    pairplot,
+    plot_loss,
+    run_sbc, pp_plot,
+)
+from vbi.utils import posterior_shrinkage_numpy, posterior_zscore_numpy
+
+OUT = Path(__file__).parent / "outputs"
+OUT.mkdir(exist_ok=True)
+
+SEED = 2
+rng  = np.random.default_rng(SEED)
+
+# ── Model parameters (fixed, not inferred) ────────────────────────────────────
+
+BASE_PAR = {
+    "dt":           0.01,
+    "t_start":      0.0,
+    "method":       "rk4",
+    "t_end":        100.0,
+    "t_cut":        20.0,
+    "output":       "output",
+    "initial_state": [0.5, 1.0],
+}
+
+THETA_TRUE = np.array([0.1, 0.05])   # a=0.1, b=0.05
+PRIOR_LOW  = np.array([0.0, 0.0])
+PRIOR_HIGH = np.array([1.0, 1.0])
+
+# ── 1 - Run the model at the true parameters ──────────────────────────────────
+
+print("=" * 60)
+print("Damped Oscillator  —  SBI with vbi.inference.SNPE")
+print("=" * 60)
+
+ode_true = DO({**BASE_PAR, "a": THETA_TRUE[0], "b": THETA_TRUE[1]})
+sol      = ode_true.run()
+t_obs    = sol["t"]
+x_obs_ts = sol["x"]   # (T, 2)
+
+fig, ax = plt.subplots(figsize=(6, 3))
+ax.plot(t_obs, x_obs_ts[:, 0], label="x  (state 0)", lw=1.2)
+ax.plot(t_obs, x_obs_ts[:, 1], label="y  (state 1)", lw=1.2)
+ax.set_xlabel("t")
+ax.set_ylabel("state")
+ax.set_title(f"True dynamics  a={THETA_TRUE[0]}, b={THETA_TRUE[1]}")
+ax.legend(fontsize=9)
+fig.tight_layout()
[…]

[…]
