### 2026-05-29 ac663ef
feat: Update damped oscillator workflow to use vbi.inference API and replace simulate_for_sbi with simulate_for_vbi
