### 2026-05-29 fd580cf
feat: Add damped oscillator VBIInference workflow and enhance feature extraction pipeline
