+++ docs/examples/workflows/damp_oscillator_vbi.py
@@ -1,0 +1,284 @@
+"""
+Damped Oscillator — end-to-end SBI with VBIInference
+=====================================================
+
+Uses the ``damped_oscillator`` ModelSpec registered in the new VBI simulator
+pipeline, so the full ``VBIInference`` workflow applies — the same object
+that drives MPR, JR, WC, etc.
+
+The model
+---------
+A nonlinear damped oscillator (Lotka-Volterra type)::
+
+    dx/dt = x - xy - a*x²
+    dy/dt = xy - y  - b*y²
+
+Parameters to infer: ``a`` (x-damping) and ``b`` (y-damping).
+
+Single-node setup (N=1, no network coupling).
+
+Monitor
+-------
+``subsample`` with period=0.1 ms records every 10 integration steps.
+Use ``MonitorSpec('raw')`` for the full dt-resolution signal (larger memory).
+
+Features
+--------
+Pipeline uses ``voi=None`` to expose **both** state variables (x and y) as
+separate channels.  Select whichever statistical / spectral features suit
+the problem — the user should customise the feature list in section 3.
+
+Workflow
+--------
+1. Build SimulationSpec (single node, no coupling).
+2. Define prior and feature pipeline.
+3. Build VBIInference.
+4. Round 1: simulate + train MAF + build posterior.
+5. Plot loss and pairplot.
+6. Save / load checkpoint.
+7. Round 2: posterior-focused simulation (sequential SNPE).
+
+Expected runtime: ~2–4 min  (numba backend, 2000 simulations × 200 ms).
+"""
+
+from pathlib import Path
+import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from vbi.simulator.models.damped_oscillator import damped_oscillator
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+)
+from vbi.simulator import Simulator
+from vbi.feature_extraction import (
+    FeaturePipeline, get_features_by_domain, get_features_by_given_names,
+)
+from vbi.inference import (
+    VBIInference, BoxUniform,
+    pairplot, plot_loss,
+)
+
+OUT = Path(__file__).parent / "outputs"
+OUT.mkdir(exist_ok=True)
+
+print("=" * 62)
+print("Damped Oscillator  —  VBIInference end-to-end pipeline")
+print("=" * 62)
+
+# ── 1 - SimulationSpec (single node, no coupling) ─────────────────────────────
+
+N = 1
+W = np.zeros((N, N))
+D = np.zeros((N, N))
+
+sim_spec = SimulationSpec(
+    model         = damped_oscillator,
+    integrator    = IntegratorSpec(method="heun", dt=0.01),
+    coupling      = CouplingSpec("linear", a=0.0),
+    monitors      = (MonitorSpec("subsample", period=0.1),),
+    # Use MonitorSpec("raw") for full dt-resolution signal instead.
+    weights       = W,
+    tract_lengths = D,
+    speed         = 4.0,
+)
+
+# ── 2 - True parameters and observed data ────────────────────────────────────
+
+THETA_TRUE = np.array([0.1, 0.05])   # a=0.1, b=0.05
+PRIOR_LOW  = np.array([0.0, 0.0])
+PRIOR_HIGH = np.array([1.0, 1.0])
[…]
