### 2026-05-29 9d7b929
feat: Enhance feature extraction and simulation processes with n_voi_feat support
