### 2026-05-29 e009e28
feat: Implement JIT warm-up for numba kernels and enable caching for simulation functions
