### 2026-05-29 d12f889
fix: Adjust simulation duration and integrator time step for damped oscillator workflow
