### 2026-05-30 4f9eb9b
feat: Refactor simulation specifications to use variables for dt, monitor, and coupling
