### 2026-05-30 68a143a
fix: Update simulation backend to use variable for sim_backend in VBIInference
