### 2026-05-30 ea150c6
fix: Correct spectrum_moments return values to match number of moments
