### 2026-05-30 2777b8f
fix: Update workflow steps and expected runtime in damped oscillator example
