### 2026-05-30 9c7e546
feat: Add Jansen-Rit end-to-end pipeline and helper functions for spectral analysis
