+++ vbi/simulator/models/jansen_rit.py
@@ -1,7 +1,9 @@
 from vbi.simulator.spec.model import ModelSpec, StateVar, Parameter
 
-# Defaults match TVB JansenRit exactly.
-# cvar=[y1, y2]: TVB cvar=[1, 2]; long-range coupling enters y4 as c (=c_y1=coupling[0]).
+# Local JR equations and defaults follow TVB JansenRit.
+# Long-range coupling intentionally follows the VBI JR convention:
+# c_y1 - c_y2 = G * sum_j W_ij * (y1_j - y2_j), matching the recorded PSP output.
+# This differs from TVB's built-in JR coupling, which uses only coupling[0] in y4.
 # Sigmoid: S(v) = 2*nu_max / (1 + exp(r*(v0 - v)))
 # State ordering: y0=pyramidal soma, y1=exc dendritic, y2=inh dendritic,
 #                 y3=d(y0)/dt,       y4=d(y1)/dt,      y5=d(y2)/dt
@@ -36,7 +38,7 @@ jansen_rit = ModelSpec(
         "y1": "y4",
         "y2": "y5",
         "y3": "A * a * (2*nu_max / (1 + exp(r*(v0 - (y1 - y2))))) - 2*a*y3 - a**2*y0",
-        "y4": "A * a * (mu + a_2*J*(2*nu_max/(1+exp(r*(v0 - a_1*J*y0)))) + c) - 2*a*y4 - a**2*y1",
+        "y4": "A * a * (mu + a_2*J*(2*nu_max/(1+exp(r*(v0 - a_1*J*y0)))) + c_y1 - c_y2) - 2*a*y4 - a**2*y1",
         "y5": "B * b * (a_4*J*(2*nu_max/(1+exp(r*(v0 - a_3*J*y0))))) - 2*b*y5 - b**2*y2",
     },
     noise_variables=("y4",),
@@ -46,7 +48,7 @@ jansen_rit = ModelSpec(
         "y1": r"y_4",
         "y2": r"y_5",
         "y3": r"Aa\,S(y_1 - y_2) - 2a\,y_3 - a^2 y_0",
-        "y4": r"Aa\!\left(\mu + \alpha_2 J\,S(\alpha_1 J y_0) + c^{\rm net}\right) - 2a\,y_4 - a^2 y_1",
+        "y4": r"Aa\!\left(\mu + \alpha_2 J\,S(\alpha_1 J y_0) + c_{y_1} - c_{y_2}\right) - 2a\,y_4 - a^2 y_1",
         "y5": r"Bb\,\alpha_4 J\,S(\alpha_3 J y_0) - 2b\,y_5 - b^2 y_2",
     },
     latex_notes=(
