+++ vbi/tests/validation/test_jansen_rit_numpy.py
@@ -131,6 +131,12 @@ class TestJansenRitNumpy:
         assert data.shape == (t.shape[0], 6, 2)
         assert np.isfinite(data).all()
 
+    @pytest.mark.skip(
+        reason=(
+            "VBI Jansen-Rit intentionally couples W @ (y1 - y2); "
+            "TVB's built-in JansenRit uses only coupling[0]."
+        )
+    )
     @pytest.mark.parametrize("method", ["euler", "heun"])
     def test_raw_trajectory_matches_tvb_without_delays(self, method):
         weights = np.array([
