+++ docs/examples/workflows/jr_vbi.py
@@ -1,0 +1,256 @@
+"""
+Jansen-Rit whole-brain — end-to-end SBI with VBIInference
+==========================================================
+
+Uses the ``jansen_rit`` ModelSpec with the VBI simulator pipeline.
+Reproduces the inference setup from ``jansen_rit_sde_numba_cde.ipynb``
+using the modern ``VBIInference`` API.
+
+The model
+---------
+Six-dimensional JR neural-mass model per node (Jansen & Rit 1995).
+Stochastic (additive noise on y4, the excitatory interneuron velocity).
+Long-range coupling enters the excitatory drive of each node (linear
+coupling on the y1-y2 cvar, as in TVB).
+
+Parameters to infer
+-------------------
+* ``G``   — global coupling strength (scales all white-matter weights).
+* ``a_2`` — slow-excitatory synaptic contact probability; maps to C1 in
+  classic JR notation via ``C1 = J * a_2`` where ``J = 135``.
+  Notebook reference: ``TRUE_THETA = [G=1.5, C1=135.0]``, i.e. ``a_2=1.0``.
+  Prior equivalent to C1 ∈ [130, 300].
+
+Observable
+----------
+``y1`` (excitatory dendritic potential, VOI index 1) via temporal-average
+monitor at 1 ms resolution (1 kHz).  Spectral features are averaged
+across all 84 cortical nodes, matching the notebook's ``average=True``.
+
+Features
+--------
+Spectral domain (Welch PSD): ``spectrum_stats``, ``spectrum_auc``,
+``spectrum_moments``.  Built by ``helpers.build_jr_spectral_pipeline``.
+
+Simulation parameters (matching jansen_rit_sde_numba_cde.ipynb)
+----------------------------------------------------------------
+* dt          = 0.05 ms
+* noise_amp   = 0.1  (noise_nsig on y4)
+* mu          = 0.24
+* t_cut       = 500  ms
+* duration    = 2501 ms
+
+Workflow
+--------
+1. Load 84-node Hagmann SC and build ``SimulationSpec`` (stochastic Heun).
+2. Run one simulation at the true parameters for visual inspection.
+3. Define prior and feature pipeline.
+4. Build ``VBIInference``.
+5. Single-round: simulate → train MAF → build posterior.
+6. Plot loss, pairplot.
+7. Save checkpoint.
+
+Expected runtime: ~8–15 min  (numba backend, 84 nodes, 1000 sims × 2501 ms).
+"""
+
+from pathlib import Path
+import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from vbi.simulator import Simulator
+from vbi.simulator.spec import (
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+    prepare_connectivity,
+)
+from vbi.simulator.models.jansen_rit import jansen_rit
+from vbi.inference import (
+    VBIInference, BoxUniform,
+    pairplot, plot_loss,
+)
+
+from helpers import build_jr_spectral_pipeline, plot_jr_timeseries_psd
+
+# ── Paths ──────────────────────────────────────────────────────────────────────
+
+DATA_DIR = Path(__file__).parent.parent / "data"
+OUT_DIR  = Path(__file__).parent / "outputs"
+OUT_DIR.mkdir(exist_ok=True)
+
[…]

[…]
