### 2026-05-30 209d9e2
feat: Implement difference coupling and update Jansen-Rit model to use it
