### 2026-05-30 6bf9b5d
feat: Add difference coupling support in NumbaSimulator and NumbaSweeperCPU
