### 2026-05-31 79f1abf
feat: Introduce JRSigmoidalCoupling and update coupling specifications in Jansen-Rit model
