+++ vbi/simulator/models/jansen_rit.py
@@ -5,10 +5,11 @@ from vbi.simulator.spec.model import ModelSpec, StateVar, Parameter
 # State ordering: y0=pyramidal soma, y1=exc dendritic, y2=inh dendritic,
 #                 y3=d(y0)/dt,       y4=d(y1)/dt,      y5=d(y2)/dt
 #
-# Long-range coupling: use CouplingSpec(kind="difference") so that
-#   c_y1 = W @ (y1 - y2)  (raw weighted PSP difference, no gain scaling).
-# The model dfun then applies G * S(c_y1), matching the legacy numba convention:
-#   G * S(sum_j W_ij * (y1_j - y2_j)).
+# Long-range coupling: use CouplingSpec(kind="jr_sigmoidal") so that
+#   c_y1 = W @ S(y1 - y2)  (weighted firing-rate drive; sigmoid applied per
+#                            source node before structural weighting).
+# The model dfun then applies G * c_y1, giving the TVB/JR convention:
+#   G * sum_j W_ij * S(y1_j(t - tau_ji) - y2_j(t - tau_ji))
 jansen_rit = ModelSpec(
     name="JansenRit",
     state_variables=(
@@ -41,7 +42,7 @@ jansen_rit = ModelSpec(
         "y1": "y4",
         "y2": "y5",
         "y3": "A * a * (2*nu_max / (1 + exp(r*(v0 - (y1 - y2))))) - 2*a*y3 - a**2*y0",
-        "y4": "A * a * (mu + a_2*J*(2*nu_max/(1+exp(r*(v0 - a_1*J*y0)))) + G*(2*nu_max/(1+exp(r*(v0 - c_y1))))) - 2*a*y4 - a**2*y1",
+        "y4": "A * a * (mu + a_2*J*(2*nu_max/(1+exp(r*(v0 - a_1*J*y0)))) + G*c_y1) - 2*a*y4 - a**2*y1",
         "y5": "B * b * (a_4*J*(2*nu_max/(1+exp(r*(v0 - a_3*J*y0))))) - 2*b*y5 - b**2*y2",
     },
     noise_variables=("y4",),
@@ -51,14 +52,16 @@ jansen_rit = ModelSpec(
         "y1": r"y_4",
         "y2": r"y_5",
         "y3": r"Aa\,S(y_1 - y_2) - 2a\,y_3 - a^2 y_0",
-        "y4": r"Aa\!\left(\mu + \alpha_2 J\,S(\alpha_1 J y_0) + G\,S(c_{y_1})\right) - 2a\,y_4 - a^2 y_1",
+        "y4": r"Aa\!\left(\mu + \alpha_2 J\,S(\alpha_1 J y_0) + G\,c_{y_1}\right) - 2a\,y_4 - a^2 y_1",
         "y5": r"Bb\,\alpha_4 J\,S(\alpha_3 J y_0) - 2b\,y_5 - b^2 y_2",
     },
     latex_notes=(
         r"Sigmoid: $S(v) = \dfrac{2\nu_{\max}}{1 + e^{r(v_0 - v)}}$. "
         r"State variables: $y_0$ pyramidal soma, $y_1$ exc dendritic, $y_2$ inh dendritic; "
         r"$y_3{=}\dot{y}_0,\; y_4{=}\dot{y}_1$ (noisy), $y_5{=}\dot{y}_2$. "
-        r"Long-range: $c_{y_1} = \sum_j W_{ij}(y_{1,j}-y_{2,j})$ via \texttt{CouplingSpec(kind='difference')}; "
-        r"gain $G$ and sigmoid applied in $\dot{y}_4$."
+        r"Long-range: $c_{y_1} = \sum_j W_{ij}\,S(y_{1,j}-y_{2,j})$ via "
[…]
