+++ vbi/simulator/spec/coupling.py
@@ -5,7 +5,7 @@ from typing import Literal
 
 @dataclass(frozen=True)
 class CouplingSpec:
-    kind: Literal["linear", "sigmoidal", "kuramoto", "difference"] = "linear"
+    kind: Literal["linear", "sigmoidal", "kuramoto", "difference", "jr_sigmoidal"] = "linear"
     # linear: c_i = G * a * sum_j(w_ij * x_j(t-tau)) + b
     a: float = 1.0
     b: float = 0.0

[…]
