### 2026-05-31 0b184e1
feat: Implement JRSigmoidalCoupling and update coupling specifications across Numba and JAX backends
