+++ vbi/simulator/backend/numba_/_nb_sim.py
@@ -123,32 +123,43 @@ def _coup_linear_delayed(srcbuf, step, horizon, weights, delay_steps, G, a, b):
 
 
 # ---------------------------------------------------------------------------
-# Coupling - Difference (instant, no delays)
+# Coupling - JR sigmoidal (instant, no delays)
 # ---------------------------------------------------------------------------
 
 @njit(cache=True)
-def _coup_difference_instant(cvar_state, weights):
+def _coup_jr_instant(cvar_state, weights, nu_max, r, v0):
     """
-    c[0, tgt] = sum_src W[tgt, src] * (cvar[0, src] - cvar[1, src])
-    c[1, tgt] = 0
-    No G or sigmoid applied here.
+    c[0, tgt] = sum_src W[tgt,src] * S(y1[src] - y2[src])
+    S(v) = 2*nu_max / (1 + exp(r*(v0 - v)))
+
+    TVB/JR convention: sigmoid applied per source node before weighting.
+    cvar_state : (2, n_nodes)  [y1, y2]
+    Returns    : (2, n_nodes)  c[1] = 0
     """
     n_nodes = cvar_state.shape[1]
     out = np.zeros((2, n_nodes))
     for tgt in range(n_nodes):
         s = 0.0
         for src in range(n_nodes):
-            s += weights[tgt, src] * (cvar_state[0, src] - cvar_state[1, src])
+            diff = cvar_state[0, src] - cvar_state[1, src]
+            s += weights[tgt, src] * (2.0 * nu_max / (1.0 + np.exp(r * (v0 - diff))))
         out[0, tgt] = s
     return out
 
 
+# ---------------------------------------------------------------------------
+# Coupling - JR sigmoidal (delayed, ring-buffer read)
+# ---------------------------------------------------------------------------
+
 @njit(cache=True)
-def _coup_difference_delayed(srcbuf, step, horizon, weights, delay_steps):
+def _coup_jr_delayed(srcbuf, step, horizon, weights, delay_steps, nu_max, r, v0):
     """
-    c[0, tgt] = sum_src W[tgt, src] * (y1[src, t-tau] - y2[src, t-tau])
-    c[1, tgt] = 0
-    srcbuf : (2, n_nodes, horizon)
+    c[0, tgt] = sum_src W[tgt,src] * S(y1[src,t-tau] - y2[src,t-tau])
+    S(v) = 2*nu_max / (1 + exp(r*(v0 - v)))
+
+    srcbuf      : (2, n_nodes, horizon)  [y1, y2]
+    delay_steps : (n_nodes, n_nodes) int32   delay_steps[src, tgt]
+    Returns     : (2, n_nodes)  c[1] = 0
     """
     n_nodes = srcbuf.shape[1]
     out = np.zeros((2, n_nodes))
@@ -158,7 +169,8 @@ def _coup_difference_delayed(srcbuf, step, horizon, weights, delay_steps):
         for src in range(n_nodes):
             d = delay_steps[src, tgt]
             idx = (t - d + horizon) % horizon
-            s += weights[tgt, src] * (srcbuf[0, src, idx] - srcbuf[1, src, idx])
+            diff = srcbuf[0, src, idx] - srcbuf[1, src, idx]
+            s += weights[tgt, src] * (2.0 * nu_max / (1.0 + np.exp(r * (v0 - diff))))
         out[0, tgt] = s
     return out
 
@@ -225,16 +237,18 @@ def nb_simulate_det(
     has_delays, cvar_indices,
     lower_bounds, has_lower, upper_bounds, has_upper,
     record_period, t_cut_step, n_voi, voi_indices,
[…]
