### 2026-05-31 319dbd5
feat: Update Jansen-Rit spectral pipeline to support state-variable difference and adjust documentation
