### 2026-06-01 7d73905
feat: Add support for VOI difference extraction in FeaturePipeline and related components
