### 2026-06-01 631ca90
feat: Introduce FeaturePruner for pruning uninformative features in FeaturePipeline
