+++ vbi/feature_extraction/__init__.py
@@ -1,6 +1,7 @@
 # from .tests.all_tests import tests
 
 from .pipeline import FeaturePipeline
+from .pruner import FeaturePruner
 from .features_settings import (
     get_features_by_domain,
     get_features_by_given_names,
@@ -10,6 +11,7 @@ from .features_settings import (
 
 __all__ = [
     "FeaturePipeline",
+    "FeaturePruner",
     "get_features_by_domain",
     "get_features_by_given_names",
     "update_cfg",
