### 2026-06-01 2762ae4
feat: Update build_jr_spectral_pipeline to accept dynamic signal parameter and adjust PSD plotting
