### 2026-06-01 005fea0
feat: Integrate FeaturePruner into Jansen-Rit spectral pipeline and update pruning logic in VBIInference
