### 2026-06-03 e3f4987
feat: Enhance FeaturePipeline and VBIInference to support dynamic worker threading for improved performance
