### 2026-06-03 a144dc5
feat: Update damped oscillator workflow to use voi="all" for state variable exposure
