+++ docs/examples/workflows/damp_oscillator_vbi.py
@@ -24,7 +24,7 @@ Use ``MonitorSpec('raw')`` for the full dt-resolution signal (larger memory).
 
 Features
 --------
-Pipeline uses ``voi=None`` to expose **both** state variables (x and y) as
+Pipeline uses ``voi="all"`` to expose **both** state variables (x and y) as
 separate channels.  Select whichever statistical / spectral features suit
 the problem — the user should customise the feature list in section 3.
 
@@ -42,20 +42,28 @@ Expected runtime: ~2–4 min  (numba backend, 2000 simulations × 250 ms).
 from pathlib import Path
 import numpy as np
 import matplotlib
+
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 
 from vbi.simulator.models.damped_oscillator import damped_oscillator
 from vbi.simulator.spec import (
-    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+    SimulationSpec,
+    IntegratorSpec,
+    CouplingSpec,
+    MonitorSpec,
 )
 from vbi.simulator import Simulator
 from vbi.feature_extraction import (
-    FeaturePipeline, get_features_by_domain, get_features_by_given_names,
+    FeaturePipeline,
+    get_features_by_domain,
+    get_features_by_given_names,
 )
 from vbi.inference import (
-    VBIInference, BoxUniform,
-    pairplot, plot_loss,
+    VBIInference,
+    BoxUniform,
+    pairplot,
+    plot_loss,
 )
 
 OUT = Path(__file__).parent / "outputs"
@@ -73,41 +81,40 @@ D = np.zeros((N, N))
 
 dt = 0.1
 monitor = MonitorSpec("raw")
-coupling = CouplingSpec("linear", a=0.0)   # no coupling (a=0.0)
+coupling = CouplingSpec("linear", a=0.0)  # no coupling (a=0.0)
 sim_backend = "numba"
 
 sim_spec = SimulationSpec(
-    model         = damped_oscillator,
-    integrator    = IntegratorSpec(method="heun", dt=dt),
-    coupling      = coupling,
-    monitors      = (monitor,),
-    weights       = W,
-    tract_lengths = D,
-    speed         = 4.0,
+    model=damped_oscillator,
+    integrator=IntegratorSpec(method="heun", dt=dt),
+    coupling=coupling,
+    monitors=(monitor,),
+    weights=W,
+    tract_lengths=D,
+    speed=4.0,
 )
 
 # ── 2 - True parameters and observed data ────────────────────────────────────
 
-THETA_TRUE = np.array([0.1, 0.05])   # a=0.1, b=0.05
-PRIOR_LOW  = np.array([0.0, 0.0])
+THETA_TRUE = np.array([0.1, 0.05])  # a=0.1, b=0.05
+PRIOR_LOW = np.array([0.0, 0.0])
 PRIOR_HIGH = np.array([1.0, 1.0])
-DURATION   = 250.0   # ms  (system reaches steady state by ~100 ms)
-T_CUT      = 20.0    # ms  (discard initial transient)
+DURATION = 250.0  # ms  (system reaches steady state by ~100 ms)
+T_CUT = 20.0  # ms  (discard initial transient)
 
 # Run a single simulation at the true parameters for visual check
 sim_spec_true = SimulationSpec(
-    model         = damped_oscillator,
-    integrator    = IntegratorSpec(method="heun", dt=dt),
-    coupling      = coupling,
-    monitors      = (monitor,),
-    weights       = W,
-    tract_lengths = D,
[…]
