### 2026-06-03 7a0589b
feat: Add feature pruner state persistence and restoration in VBIInference
