### 2026-06-03 f62e41a
feat: Update default pruner in build_jr_spectral_pipeline to None
