### 2026-06-03 30b3abc
feat: Add plot_feature_scatter function for visualizing training features against a_2
