### 2026-06-03 e985005
feat: Add append_simulations method to VBIInference for integrating precomputed simulation data
