### 2026-06-03 1506592
feat: Enhance jr_vbi workflow with inference_backend specification and simulation caching
