+++ docs/examples/workflows/jr_vbi.py
@@ -4,7 +4,8 @@ Jansen-Rit whole-brain — end-to-end SBI with VBIInference
 
 Uses the ``jansen_rit`` ModelSpec with the VBI simulator pipeline.
 Reproduces the inference setup from ``jansen_rit_sde_numba_cde.ipynb``
-using the modern ``VBIInference`` API.
+using the modern ``VBIInference`` API with upstream ``sbi`` as the inference
+backend.
 
 The model
 ---------
@@ -13,13 +14,13 @@ Stochastic (additive noise on y4, the excitatory interneuron velocity).
 Long-range coupling enters the excitatory drive of each node via
 G * W @ S(y1-y2), using CouplingSpec(kind="jr_sigmoidal").
 
-Parameters to infer
--------------------
-* ``G``   — global coupling strength (scales all white-matter weights).
+Parameter to infer
+------------------
 * ``a_2`` — slow-excitatory synaptic contact probability; maps to C1 in
-  classic JR notation via ``C1 = J * a_2`` where ``J = 135``.
-  Notebook reference: ``TRUE_THETA = [G=1.5, C1=135.0]``, i.e. ``a_2=1.0``.
-  Prior equivalent to C1 ∈ [130, 300].
+  classic JR notation via ``C1 = J * a_2``.
+  This script uses ``G=1.5`` and ``a_2=1.0`` for the reference run.
+  Prior equivalent to C1 ∈ [130, 300], provided J=135.
+  ``G`` is fixed at 1.5.
 
 Observable
 ----------
@@ -39,213 +40,319 @@ Simulation parameters (matching jansen_rit_sde_numba_cde.ipynb)
 * noise_amp   = 0.1  (noise_nsig on y4)
 * mu          = 0.24
 * t_cut       = 500  ms
-* duration    = 2501 ms
+* duration    = 2500 ms
 
 Workflow
 --------
 1. Load 84-node Hagmann SC and build ``SimulationSpec`` (stochastic Heun).
 2. Run one simulation at the true parameters for visual inspection.
 3. Define prior and feature pipeline.
-4. Build ``VBIInference``.
+4. Build ``VBIInference`` with ``inference_backend="sbi"``.
 5. Single-round: simulate → train MAF → build posterior.
 6. Plot loss, pairplot.
 7. Save checkpoint.
 
-Expected runtime: ~8–15 min  (numba backend, 84 nodes, 1000 sims × 2501 ms).
+Expected runtime: ~8–15 min  (numba backend, 84 nodes, 1000 sims × 2500 ms).
 """
 
+import argparse
 from pathlib import Path
 import numpy as np
 import matplotlib
+
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 
 from vbi.simulator import Simulator
 from vbi.simulator.spec import (
-    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+    SimulationSpec,
+    IntegratorSpec,
+    CouplingSpec,
+    MonitorSpec,
     prepare_connectivity,
 )
 from vbi.simulator.models.jansen_rit import jansen_rit
 from vbi.inference import (
-    VBIInference, BoxUniform,
-    pairplot, plot_loss,
+    VBIInference,
+    BoxUniform,
+    pairplot,
 )
 
-from helpers import build_jr_spectral_pipeline, plot_jr_timeseries_psd
+from vbi.feature_extraction import FeaturePruner
+from helpers import (
[…]
