### 2026-06-03 b26cf80
feat: Add test for append_simulations method in VBIInference
