### 2026-06-03 efb6bbf
feat: Update inference backend to 'vbi' and adjust training parameters in jr_vbi workflow
