### 2026-06-03 30e643e
feat: Update default training parameters in SNPE and MAFEstimator, and add test for SNPE defaults
