+++ vbi/inference/_backends/jax_/maf_jax.py
@@ -339,10 +339,10 @@ class JaxMAFEstimator(JaxConditionalDensityEstimator):
         verbose: bool = True,
         validation_fraction: float = 0.1,
         stop_after_epochs: int = 20,
-        early_stopping_delta: float | None = None,
+        early_stopping_delta: float | None = 0.0,
         clip_max_norm: float | None = 5.0,
         resume_training: bool = False,
-        lr_schedule: str | None = "cosine",
+        lr_schedule: str | None = None,
         lr_min: float = 1e-5,
         lr_period: int = 500,
         monitor_collapse: bool = False,
