### 2026-06-03 2f957fa
feat: Refactor connectivity handling in SimulationSpec and Connectivity classes
