+++ vbi/simulator/spec/simulation.py
@@ -10,6 +10,7 @@ from .integrator import IntegratorSpec
 from .coupling import CouplingSpec
 from .monitor import MonitorSpec
 from .stimulus import StimSpec
+from .connectivity import Connectivity
 
 
 _MODEL_REGISTRY: "dict | None" = None
@@ -90,70 +91,49 @@ class SimulationSpec:
     integrator : IntegratorSpec
     coupling : CouplingSpec
     monitors : tuple[MonitorSpec, ...]
-    weights : np.ndarray
-        (n_nodes, n_nodes) structural connectivity; weights[tgt, src].
-    tract_lengths : np.ndarray
-        (n_nodes, n_nodes) delay distances in mm; tract_lengths[src, tgt].
-    speed : float
-        Conduction velocity in mm/ms (default 4.0 mm/ms = 4 m/s).
+    connectivity : Connectivity
+        Structural connectivity (weights, tract_lengths, speed, optional metadata).
     node_params : dict[str, np.ndarray]
         Per-node parameter overrides, e.g. {"eta": np.full(80, -4.6)}.
         Overrides ModelSpec.parameters defaults for named entries.
     """
-    model: ModelSpec
-    integrator: IntegratorSpec
-    coupling: CouplingSpec
-    monitors: tuple[MonitorSpec, ...]
-    weights: np.ndarray
-    tract_lengths: np.ndarray | None = None  # None → zero delays (pure ODE/SDE)
-    speed: float = 4.0
-    node_params: dict[str, np.ndarray] = field(default_factory=dict)
-    stimuli: tuple[StimSpec, ...] = field(default_factory=tuple)
+    model:        ModelSpec
+    integrator:   IntegratorSpec
+    coupling:     CouplingSpec
+    monitors:     tuple[MonitorSpec, ...]
+    connectivity: Connectivity
+    node_params:  dict[str, np.ndarray] = field(default_factory=dict)
+    stimuli:      tuple[StimSpec, ...] = field(default_factory=tuple)
 
-    def __post_init__(self):
-        # Coerce weights and tract_lengths to float64 ndarray (accepts lists etc.)
-        object.__setattr__(self, "weights",
-                           np.asarray(self.weights, dtype=np.float64))
-        if self.tract_lengths is None:
-            object.__setattr__(self, "tract_lengths",
-                               np.zeros_like(self.weights))
-        else:
-            object.__setattr__(self, "tract_lengths",
-                               np.asarray(self.tract_lengths, dtype=np.float64))
-        # Shape / value validation
-        w = self.weights
-        tl = self.tract_lengths
-        if w.ndim != 2 or w.shape[0] != w.shape[1]:
-            raise ValueError(
-                f"weights must be a square 2-D array; got shape {w.shape}."
-            )
-        if tl.shape != w.shape:
-            raise ValueError(
-                f"tract_lengths shape {tl.shape} must match weights shape {w.shape}."
-            )
-        if np.any(tl < 0):
-            raise ValueError(
-                "tract_lengths must be non-negative; found negative values."
-            )
-        if self.speed <= 0:
-            raise ValueError(
-                f"speed must be > 0 mm/ms; got speed={self.speed!r}."
-            )
+    # ------------------------------------------------------------------
+    # Delegation properties — backends use these without change
+    # ------------------------------------------------------------------
+
+    @property
+    def weights(self) -> np.ndarray:
+        return self.connectivity.weights
+
+    @property
+    def tract_lengths(self) -> np.ndarray:
+        return self.connectivity.tract_lengths
+
+    @property
+    def speed(self) -> float:
+        return self.connectivity.speed
 
     @property
     def n_nodes(self) -> int:
-        return self.weights.shape[0]
+        return self.connectivity.n_nodes
 
     @property
     def has_delays(self) -> bool:
[…]
