+++ vbi/simulator/spec/connectivity.py
@@ -1,33 +1,35 @@
 """
-Connectivity utilities - load, normalise, and save SC matrices.
-
-Users bring connectivity data in many formats; this module provides a
-single entry point that accepts numpy arrays, plain text files (.txt),
-CSV files, NPZ archives, and NumPy `.npy` files.
+Structural connectivity for brain network models.
 
 Quick start
 -----------
->>> W, D = prepare_connectivity('weights.txt', 'tract_lengths.txt')
->>> spec = SimulationSpec(..., weights=W, tract_lengths=D)
+>>> conn = Connectivity.from_file('weights.txt', 'tract_lengths.txt')
+>>> spec = SimulationSpec(..., connectivity=conn)
 
-Or save to .npz for use with ``VBIInference.from_config``::
+Load from a TVB Connectivity object (TVB optional)::
 
->>> save_connectivity(W, D, 'connectivity.npz')
+>>> conn = Connectivity.from_tvb(tvb_conn)
 
-Config YAML then just needs::
+Save / load round-trip::
+
+>>> conn.save('connectivity.npz')
+>>> conn = Connectivity.load('connectivity.npz')
+
+Config YAML just needs::
 
     sim:
       connectivity: connectivity.npz
 """
 from __future__ import annotations
 
+from dataclasses import dataclass
 from pathlib import Path
 
 import numpy as np
 
 
 def _load_array(src) -> np.ndarray:
-    """Load a matrix from a path (txt / csv / npz / npy) or return as-is."""
+    """Load a matrix from a path (txt/csv/npy/npz) or return as-is."""
     if isinstance(src, np.ndarray):
         return src.astype(np.float64)
 
@@ -40,15 +42,14 @@ def _load_array(src) -> np.ndarray:
         return np.load(path).astype(np.float64)
     if suffix == ".npz":
         d = np.load(path)
-        # Accept first array if only one, else require 'weights' key
         keys = list(d.keys())
         if "weights" in keys:
             return d["weights"].astype(np.float64)
         if len(keys) == 1:
             return d[keys[0]].astype(np.float64)
         raise ValueError(
-            f"{path}: .npz has multiple arrays {keys}; pass the key explicitly "
-            f"or use prepare_connectivity_from_npz()."
+            f"{path}: .npz has multiple arrays {keys}; "
+            f"use Connectivity.load() for full archives."
         )
     raise ValueError(
         f"Unsupported file format {suffix!r} for {path}. "
@@ -56,125 +57,205 @@ def _load_array(src) -> np.ndarray:
     )
 
 
-def prepare_connectivity(
-    weights,
-    tract_lengths=None,
-    normalize: bool = True,
-    normalise: bool | None = None,   # British spelling alias
-) -> tuple[np.ndarray, np.ndarray]:
+@dataclass
+class Connectivity:
     """
-    Load and validate SC connectivity matrices from various sources.
+    Structural connectivity for a brain network model.
 
     Parameters
     ----------
-    weights : array-like | str | Path
-        (n_nodes, n_nodes) weight matrix, **or** path to a file:
-
-        * ``.txt`` / ``.csv`` - space- or comma-separated text
-        * ``.npy`` - NumPy binary
-        * ``.npz`` - NumPy archive with a ``weights`` key (or a single key)
-
-    tract_lengths : array-like | str | Path | None
-        (n_nodes, n_nodes) tract-length matrix in **mm**.  Same formats as
-        ``weights``.  ``None`` → zero delays (pure ODE / SDE).
-
-    normalize : bool
-        Row-sum normalise ``weights`` so each row sums to 1.  Recommended
-        for most VBI models (MPR, VEP, GHB) where coupling strength is
[…]
