+++ vbi/simulator/spec/__init__.py
@@ -5,7 +5,7 @@ from .monitor import MonitorSpec
 from .simulation import SimulationSpec, register_model
 from .stimulus import StimSpec, build_stim_data
 from .sweep import SweepSpec
-from .connectivity import prepare_connectivity, prepare_connectivity_from_npz, save_connectivity
+from .connectivity import Connectivity
 
 __all__ = [
     "ModelSpec", "StateVar", "Parameter",
@@ -17,7 +17,5 @@ __all__ = [
     "StimSpec",
     "build_stim_data",
     "SweepSpec",
-    "prepare_connectivity",
-    "prepare_connectivity_from_npz",
-    "save_connectivity",
+    "Connectivity",
 ]
