### 2026-06-03 972c313
feat: Integrate Connectivity class into simulation specifications across multiple demo scripts
