+++ docs/examples/simulator_models/jansen_rit_demo.py
@@ -30,6 +30,7 @@ with quiet_optional_imports():
     from vbi.simulator.spec.integrator import IntegratorSpec
     from vbi.simulator.spec.monitor import MonitorSpec
     from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator.spec.connectivity import Connectivity
 
 
 JR_PARAMS = {
@@ -94,9 +95,7 @@ def build_vbi_spec(method: str) -> SimulationSpec:
         integrator=IntegratorSpec(method=method, dt=JR_PARAMS["dt"]),
         coupling=CouplingSpec(kind="linear", a=JR_PARAMS["coupling_strength"]),
         monitors=(MonitorSpec(kind="raw"),),
-        weights=weights,
-        tract_lengths=np.zeros_like(weights),
-        speed=1.0,
+        connectivity=Connectivity(weights, speed=1.0),
         node_params=node_params,
     )
 
