### 2026-06-03 8bd8410
feat: Add scatter plot function for visualizing features against parameters
