### 2026-06-03 48f5ab0
feat: Remove redundant plot_feature_scatter function and re-export from vbi.plot
