### 2026-06-03 4ff7ff4
feat: Update A2_TRUE parameter and enhance progress bar for simulation chunks
