### 2026-06-03 9bebae4
feat: Enhance feature extraction by adding multiprocessing support and caching mechanisms
