### 2026-06-08 ec39a08
feat: Add n_workers parameter to Sweeper for backend worker control and implement related tests
