+++ vbi/inference/_vbi_inference.py
@@ -301,7 +301,7 @@ def _unpack_pruner(d, pipeline) -> None:
 # SBC simulator helper  (Step 6)
 # ---------------------------------------------------------------------------
 
-def _make_simulator_fn(sim_spec, prior, pipeline, sim_backend, duration):
+def _make_simulator_fn(sim_spec, prior, pipeline, integrator_backend, duration):
     """
     Return a callable ``theta_1d -> x_1d`` suitable for :func:`run_sbc`.
 
@@ -320,7 +320,7 @@ def _make_simulator_fn(sim_spec, prior, pipeline, sim_backend, duration):
             param_names = tuple(param_names),
             pipeline    = pipeline,
         )
-        sweeper = Sweeper(sim_spec, sweep_spec, backend=sim_backend)
+        sweeper = Sweeper(sim_spec, sweep_spec, backend=integrator_backend)
         labels, values = sweeper.run(duration)
         n_params = len(param_names)
         x_1d = values[0, n_params:].astype(np.float64)
@@ -462,9 +462,9 @@ class VBIInference:
         Must expose ``._resolved_param_names``, ``.sample()``, ``.log_prob()``.
     pipeline : FeaturePipeline
     density_estimator : 'maf' | 'mdn' | 'nsf'
-    sim_backend : str
+    integrator_backend : str
         Backend for the Sweeper ('numpy' | 'numba' | 'cuda' | 'jax').
-    backend : str
+    estimator_backend : str
         Backend for the density estimator ('auto' | 'numpy' | 'jax').
     show_progress_bars : bool
     embedding_net : EmbeddingNet | None
@@ -485,8 +485,8 @@ class VBIInference:
         prior,
         pipeline,
         density_estimator: str = "maf",
-        sim_backend: str = "numba",
-        backend: str = "auto",
+        integrator_backend: str = "numba",
+        estimator_backend: str = "auto",
         show_progress_bars: bool = True,
         embedding_net=None,
         inference_backend: str = "vbi",
@@ -508,7 +508,7 @@ class VBIInference:
         self._sim_spec            = sim_spec
         self._prior               = prior
         self._pipeline            = pipeline
-        self._sim_backend         = sim_backend
+        self._integrator_backend  = integrator_backend
         self._de_type             = density_estimator
         self._show_progress_bars  = show_progress_bars
         self._feature_labels: list[str] | None = None
@@ -529,7 +529,7 @@ class VBIInference:
             self._snpe              = SNPE(
                 prior               = prior,
                 density_estimator   = density_estimator,
-                backend             = backend,
+                backend             = estimator_backend,
                 show_progress_bars  = show_progress_bars,
                 embedding_net       = embedding_net,
             )
@@ -587,7 +587,7 @@ class VBIInference:
             pipeline            = self._pipeline,
             num_simulations     = num_simulations,
             duration            = duration,
-            sim_backend         = self._sim_backend,
+            sim_backend         = self._integrator_backend,
             seed                = seed,
             proposal            = proposal,
             x_obs               = x_obs,
@@ -832,8 +832,8 @@ class VBIInference:
         data["meta__feature_labels"]    = np.array("|".join(self._feature_labels or []))
         data["meta__param_names"]       = np.array("|".join(self._param_names or []))
         data["meta__density_estimator"] = np.array(self._de_type)
-        data["meta__sim_backend"]       = np.array(self._sim_backend)
-        data["meta__inference_backend"] = np.array(self._inference_backend)
+        data["meta__integrator_backend"] = np.array(self._integrator_backend)
+        data["meta__inference_backend"]  = np.array(self._inference_backend)
         data["meta__n_rounds"]          = np.array(len(self._sim_rounds))
         # VBI-only: persist the resolved estimator backend string
[…]
