### 2026-06-08 ee3d232
feat: Rename sim_backend to integrator_backend for clarity in damp oscillator example
