### 2026-06-08 7ecdfa8
feat: Rename sim_backend and backend to integrator_backend and estimator_backend for consistency across tests
