### 2026-06-08 32ed9f8
feat: Add extract_features_after parameter to simulate_for_vbi_sweep_cached for optional feature extraction
