### 2026-06-25 ins-amu/vbi#70 @sampottinger
Credit third party libraries
As part of https://github.com/openjournals/joss-reviews/issues/10612, just checking in on third party credit and license compliance. It’s possible I missed it but do you already have a location where you give credit to the third party libraries you use along with links to or copies of their licenses? For example, I see scipy, matplotlib, etc. I wonder if you might need to list attribution in your docs to comply with their licenses. For example, see [scipy license](https://github.com/scipy/scipy/blob/main/LICENSE.txt). Thanks!
