### 2026-06-25 ins-amu/vbi#71 @sampottinger
Add statement of need in main docs
As part of https://github.com/openjournals/joss-reviews/issues/10612, I think we may need a statement of need in the project documentation. I believe this is in addition to the statement in the paper. Are you able to add that, ideally to the README? For more information see https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need.
