### 2026-06-25 ins-amu/vbi#72 @sampottinger
Expand C++ documentation
You have a lot of really excellent docstrings on your Python side. However, I notice that the C++ side is not quite as well documented. For example, please see https://github.com/ins-amu/vbi/blob/main/vbi/models/cpp/_src/jr_sdde.hpp. Are you able to extend your documentation practice to this side of the codebase as part of https://github.com/openjournals/joss-reviews/issues/10612. Really appreciate your efforts thank you!
