### 2026-06-25 ins-amu/vbi#73 @sampottinger
Clarify Java components
Hello! As part of https://github.com/openjournals/joss-reviews/issues/10612 , I was looking in `features_utils.py`. I apologize I got a little confused about where `infodynamics.jar` came from. I wonder if it may be https://github.com/jlizier/jidt? I also don’t see Java or JDK information in [install instructions](https://vbi.readthedocs.io/latest/installation.html). Can you please clarify here or add a little more documentation on this? Thanks for clarifying!
