### 2026-06-25 ins-amu/vbi#74 @sampottinger
Consider adding top of file docstrings
Your docstrings on public members is great on the Python side. However, if you don’t mind, please consider adding top of file docstrings. This can be helpful for a number of IDEs and could potentially be useful for outside developers given the size of the codebase. This may also help some users leveraging things like `help` and may also assist those using their computers through different modalities like by using adaptive technologies where something like `help` may be easier to navigate than the HTML documents. Part of https://github.com/openjournals/joss-reviews/issues/10612. Note that, at minimum, modules like https://github.com/ins-amu/vbi/blob/main/vbi/__init__.py are expected to have docstrings per [PEP 257](https://peps.python.org/pep-0257/#what-is-a-docstring).
