### 2026-06-25 ins-amu/vbi#75 @sampottinger
JOSS paper feedback
Thank you for your project and your great paper. I have a few questions and suggestions to further improve this part of your work!

 - The statement of need and software design sections are giving me a lot of great information about your project. However, I wonder if a few elements required by JOSS can be made more clear. First, I’m getting pieces of this but the “state of the field” section is required by JOSS. It may help to break this out into its own labeled section (giving you a chance to talk further about The Virtual Brain, for example). Can you please add? Additionally, I think the statement of need may want to be a bit more clear on the audience is.
 - While optional, your Figure 1 is very dense. I wonder if it may make sense to add a labels or part G and list parts A - F in the diagram (in non-bold) of G to show which parts of the codebase are used where. I think this overview is useful and having a system diagram is helpful but tying G to the rest of the figure may be useful.
 - You mention that your work has been used in other studies in the Research Impact Statement. Please directly cite if possible.
 - Please specify which AI models were used if possible in your AI Usage Disclosure. Some models have special licensing requirements and, even if you aren’t subject, the citation helps clarify.
 - I think you may be missing DOIs in your citations.

Part of https://github.com/openjournals/joss-reviews/issues/10612.
