### 2026-06-29 f2e26cb
Enhance documentation by adding third-party licenses section and updating requirements for pypandoc_binary
