+++ docs/third_party_licenses.rst
@@ -1,0 +1,137 @@
+Third-Party Libraries & Licenses
+=================================
+
+VBI builds on a number of open-source libraries. We gratefully acknowledge their
+authors and list each library, its license type, and a link to the full license
+text below.
+
+Core Dependencies
+-----------------
+
+.. list-table::
+   :header-rows: 1
+   :widths: 20 15 65
+
+   * - Library
+     - License
+     - License URL
+   * - `NumPy <https://numpy.org>`_
+     - BSD-3-Clause
+     - https://github.com/numpy/numpy/blob/main/LICENSE.txt
+   * - `SciPy <https://scipy.org>`_
+     - BSD-3-Clause
+     - https://github.com/scipy/scipy/blob/main/LICENSE.txt
+   * - `Matplotlib <https://matplotlib.org>`_
+     - PSF / BSD-compatible
+     - https://github.com/matplotlib/matplotlib/blob/main/LICENSE/LICENSE
+   * - `Numba <https://numba.pydata.org>`_
+     - BSD-2-Clause
+     - https://github.com/numba/numba/blob/main/LICENSE
+   * - `pandas <https://pandas.pydata.org>`_
+     - BSD-3-Clause
+     - https://github.com/pandas-dev/pandas/blob/main/LICENSE
+   * - `h5py <https://www.h5py.org>`_
+     - BSD-3-Clause
+     - https://github.com/h5py/h5py/blob/master/LICENSE
+   * - `NetworkX <https://networkx.org>`_
+     - BSD-3-Clause
+     - https://github.com/networkx/networkx/blob/main/LICENSE.txt
+   * - `scikit-learn <https://scikit-learn.org>`_
+     - BSD-3-Clause
+     - https://github.com/scikit-learn/scikit-learn/blob/main/COPYING
+   * - `autograd <https://github.com/HIPS/autograd>`_
+     - MIT
+     - https://github.com/HIPS/autograd/blob/master/license.txt
+   * - `tqdm <https://tqdm.github.io>`_
+     - MIT / MPL-2.0
+     - https://github.com/tqdm/tqdm/blob/master/LICENCE
+   * - `Rich <https://rich.readthedocs.io>`_
+     - MIT
+     - https://github.com/Textualize/rich/blob/master/LICENSE
+   * - `corner <https://corner.readthedocs.io>`_
+     - MIT
+     - https://github.com/dfm/corner.py/blob/main/LICENSE
+   * - `nbconvert <https://nbconvert.readthedocs.io>`_
+     - BSD-3-Clause
+     - https://github.com/jupyter/nbconvert/blob/main/LICENSE
+   * - `parameterized <https://github.com/wolever/parameterized>`_
+     - BSD-2-Clause
+     - https://github.com/wolever/parameterized/blob/master/LICENSE.txt
+   * - `pytest <https://pytest.org>`_
+     - MIT
+     - https://github.com/pytest-dev/pytest/blob/main/LICENSE
+
+Optional / Inference Dependencies
+----------------------------------
+
+.. list-table::
+   :header-rows: 1
+   :widths: 20 15 65
+
+   * - Library
+     - License
+     - License URL
+   * - `PyTorch <https://pytorch.org>`_
+     - BSD-3-Clause
+     - https://github.com/pytorch/pytorch/blob/main/LICENSE
+   * - `sbi <https://sbi-dev.github.io/sbi>`_
+     - Apache-2.0
+     - https://github.com/sbi-dev/sbi/blob/main/LICENSE.txt
+   * - `JAX <https://jax.readthedocs.io>`_
+     - Apache-2.0
[…]
