### 2026-06-29 ins-amu/vbi#70 @Ziaeemehr
Thank you for raising this! You are right that we were missing a dedicated attribution page. We have now added a Third-Party Libraries & Licenses page to our documentation (`docs/third_party_licenses.rst`), which lists all dependencies, core, optional/inference, documentation, and build tools, along with their license types and direct links to each upstream license file. The page is also linked from the main documentation index.

The libraries covered include NumPy, SciPy, Matplotlib, Numba, pandas, h5py, NetworkX, scikit-learn, autograd, tqdm, Rich, corner, PyTorch, sbi, JAX, CuPy, pycatch22, Sphinx, SWIG, setuptools, and Hatchling.
