### 2026-06-29 1238050
Add Statement of Need to README, update license to Apache-2.0 in pyproject.toml, and include pypandoc in third-party licenses
