### 2026-06-29 ins-amu/vbi#71 @Ziaeemehr
Thank you for the suggestion. We have added a Statement of Need section to the README, covering the scientific problem VBI addresses (probabilistic model inversion for whole-brain network models), the target audience (computational neuroscientists and clinical researchers working with virtual brain twins), and how VBI fills a gap not covered by existing tools such as the sbi library or The Virtual Brain framework individually.
