### 2026-06-29 9a82593
Enhance model documentation and improve code clarity
- Added comprehensive documentation comments to the MPR, VEP, WC_ODE, and utility modules, detailing model descriptions, state-vector layouts, equations, and parameter explanations.
- Improved variable naming and added comments for clarity in the MPR_sde and VEP classes.
- Enhanced the BoldParams structure with detailed parameter descriptions and constructors.
- Updated the utility functions with additional documentation for better usability.
- Refactored the Wilson-Cowan model to include detailed comments on the dynamics and integration met
[…]
