+++ vbi/models/cpp/_src/utility.hpp
@@ -1,3 +1,17 @@
+/**
+ * @file utility.hpp
+ * @brief Common type aliases, utility functions, and RNG helpers shared by all
+ *        VBI C++ model backends.
+ *
+ * Provides:
+ *   - 1-D / 2-D vector type aliases (dim1, dim2, …)
+ *   - Adjacency-matrix to adjacency-list conversion
+ *   - Mathematical helpers (moving average, matrix–vector multiply)
+ *   - Cross-platform memory and wall-clock timing
+ *   - A seeded/random Mersenne-Twister RNG (rng())
+ *   - File / folder existence checks and matrix I/O
+ */
+
 #ifndef UTILITY_HPP
 #define UTILITY_HPP
 
@@ -32,11 +46,20 @@
 using std::string;
 using std::vector;
 
+/** @brief 1-D vector of doubles. */
 typedef std::vector<double> dim1;
+/** @brief 1-D vector of unsigned integers. */
 typedef std::vector<unsigned> dim1I;
+/** @brief 2-D vector of doubles (vector of dim1). */
 typedef std::vector<std::vector<double>> dim2;
+/** @brief 2-D vector of unsigned integers. */
 typedef std::vector<std::vector<unsigned>> dim2I;
 
+/**
+ * @brief Check whether a file exists on disk.
+ * @param filename Path to the file.
+ * @return true if the file exists, false otherwise.
+ */
 bool fileExists(const std::string &filename)
 {
     struct stat buf;
@@ -47,6 +70,11 @@ bool fileExists(const std::string &filename)
     return false;
 }
 
+/**
+ * @brief Check whether a directory exists on disk.
+ * @param path Path to the directory.
+ * @return true if the path exists and is a directory, false otherwise.
+ */
 bool folderExists(const std::string &path)
 {
     struct stat st;
@@ -61,6 +89,15 @@ bool folderExists(const std::string &path)
         return false;
 }
 
+/**
+ * @brief Convert a weighted adjacency matrix to a sparse adjacency list.
+ *
+ * Only edges with |A[i][j]| > 1e-8 are retained, which avoids iterating over
+ * zero-weight entries in the inner simulation loops.
+ *
+ * @param A  Square N×N adjacency / weight matrix.
+ * @return   adjlist[i] lists the indices j of all neighbours of node i.
+ */
 std::vector<std::vector<unsigned>> adjmat_to_adjlist(const dim2 &A)
 {
     size_t n = A.size();
@@ -99,6 +136,16 @@ std::vector<std::vector<unsigned>> adjmat_to_adjlist(const dim2 &A)
 //     return adjlist;
 // }
 
+/**
+ * @brief Compute a non-overlapping block moving average.
+ *
+ * Partitions @p vec into consecutive blocks of size @p window and returns the
+ * mean of each block.  The output length is floor(size / window).
+ *
+ * @param vec    Input time series.
+ * @param window Block size (number of samples to average).
+ * @return       Downsampled signal of length floor(vec.size() / window).
+ */
 dim1 moving_average(const dim1 &vec, const size_t window)
 {
     size_t size = vec.size();
@@ -118,7 +165,12 @@ dim1 moving_average(const dim1 &vec, const size_t window)
     return vec_out;
 }
 
-// adding vectors
+/**
+ * @brief Element-wise addition of two double vectors, writing the result into @p c.
+ * @param a  First operand.
+ * @param b  Second operand (same size as @p a).
+ * @param c  Output vector (must be pre-allocated to the same size).
+ */
 void add(const vector<double> &a, const vector<double> &b, vector<double> &c)
 {
     // c need to be allocated memory.
[…]
