+++ vbi/models/cpp/_src/jr_sde.hpp
@@ -1,3 +1,32 @@
+/**
+ * @file jr_sde.hpp
+ * @brief Jansen-Rit neural mass model with stochastic noise and no axonal
+ *        delays (SDE — Stochastic Differential Equations).
+ *
+ * Implements the Jansen-Rit (JR) model for a network of N instantaneously
+ * coupled cortical columns.  Each column is described by three second-order
+ * linear operators (excitatory pyramidal, excitatory interneuron, inhibitory
+ * interneuron) driven by a sigmoidal firing-rate function.
+ *
+ * **Model reference:**
+ * Jansen, B.H. & Rit, V.G. (1995). Electroencephalogram and visual evoked
+ * potential generation in a mathematical model of coupled cortical columns.
+ * *Biological Cybernetics*, 73(4), 357–366.
+ * https://doi.org/10.1007/BF00199471
+ *
+ * **State-vector layout** (flat vector of length 6×N):
+ * | Index range | Variable | Description |
+ * |-------------|----------|-------------|
+ * | [0,   N)    | x0       | Slow IPSP — inhibitory interneuron output |
+ * | [N,  2N)    | x1       | EPSP — pyramidal cell output |
+ * | [2N, 3N)    | x2       | Fast IPSP — inhibitory interneuron output |
+ * | [3N, 4N)    | ẋ0       | Velocity of x0 |
+ * | [4N, 5N)    | ẋ1       | Velocity of x1 — **noise applied here** |
+ * | [5N, 6N)    | ẋ2       | Velocity of x2 |
+ *
+ * The EEG proxy (pyramidal-cell PSP) is: `x1 - x2` = `y[i+N] - y[i+2N]`.
+ */
+
 #ifndef jr_sde_HPP
 #define jr_sde_HPP
 
@@ -17,46 +46,78 @@ using std::vector;
 typedef std::vector<double> dim1;
 typedef std::vector<dim1> dim2;
 
+/**
+ * @brief Jansen-Rit SDE — N instantaneously coupled cortical columns.
+ *
+ * Noise enters only through the ẋ1 (excitatory pyramidal velocity) component,
+ * modelling stochastic afferent input.  Integration uses either the
+ * Euler-Maruyama or Heun (order-1.5 strong) scheme.
+ */
 class JR_sde
 {
 private:
-    size_t N;
-    double dt;
-    double t_initial;
-    double t_final;
-    double t_transition;
-    size_t dimension;
-    size_t num_steps;
-    size_t index_transition;
-    vector<vector<unsigned>> adjlist;
+    size_t N;               ///< Number of network nodes
+    double dt;              ///< Integration time step [s]
+    double t_initial;       ///< Simulation start time [s]
+    double t_final;         ///< Total simulation time [s]
+    double t_transition;    ///< Transient period to discard [s]
+    size_t dimension;       ///< Total state dimension (= 6*N)
+    size_t num_steps;       ///< Total number of integration steps
+    size_t index_transition;///< Step index corresponding to t_transition
+    vector<vector<unsigned>> adjlist; ///< Sparse adjacency list
 
-    dim1 A;
-    double par_a;
-    double par_B;
-    double par_b;
-    double par_r;
-    double par_v0;
-    double par_vmax;
-    double coupling;
-    double noise_mu;
-    double noise_sigma;
+    dim1 A;         ///< Per-node excitatory synaptic gain [mV] (length N)
+    double par_B;   ///< Inhibitory synaptic gain [mV]
+    double par_a;   ///< Excitatory synaptic time constant [1/s]
+    double par_b;   ///< Inhibitory synaptic time constant [1/s]
+    double par_r;   ///< Sigmoid slope parameter [1/mV]
[…]

[…]
