### 2026-06-29 ins-amu/vbi#72 @Ziaeemehr
Thank you for the kind words and for pointing this out, it is a fair observation. We have now extended our documentation practice to the entire C++ backend (commit 9a8259313d0f96e0e66c613653374df5e776fc0f). All 8 model headers and the shared utility header now include Doxygen-style documentation covering:

File-level docblocks with a description of the model, the governing equations, the state-vector layout, and citations to the original publications (Jansen & Rit 1995; Wilson & Cowan 1972; Montbrió et al. 2015; Jirsa et al. 2014; Kuramoto 1984; Friston et al. 2000).
Class-level overviews describing the role of each class and its integration scheme.
Constructor `@param` documentation for every parameter.
Method docblocks with `@param` / `@return` annotations on all public methods, including the mathematical equations where the code would otherwise be opaque (e.g., sigmoid function, RHS terms, BOLD output formula).
