### 2026-06-29 ins-amu/vbi#73 @Ziaeemehr
Thank you, you are correct. The bundled `infodynamics.jar` is the **Java Information Dynamics Toolkit (JIDT)** by Joseph T. Lizier (https://github.com/jlizier/jidt), distributed under GPL-3.0. We have now:

Added a dedicated **"Bundled Binary Dependencies"** section to our third-party licenses page explicitly crediting JIDT.
Added a **"Java JDK (for Information-Theoretic Features)"** section to the installation guide, explaining that a Java JDK ≥ 8 and `JPype1` are required for transfer-entropy and mutual-information features, with install instructions for Linux, macOS, and Windows.
Updated the `get_jar_location()` docstring to name JIDT, link to its repository, and note the licence and JDK dependency.
