### 2026-06-29 d2c8c98
Add Java JDK installation instructions and JIDT library details to documentation
