### 2026-06-29 6fb01ad
Enhance documentation across VBI modules
- Added module-level docstrings to provide clear descriptions of functionality and usage for:
  - vbi/__init__.py
  - vbi/dataset/__init__.py
  - vbi/feature_extraction/__init__.py
  - vbi/feature_extraction/calc_features.py
  - vbi/feature_extraction/features.py
  - vbi/feature_extraction/features_settings.py
  - vbi/feature_extraction/features_utils.py
  - vbi/feature_extraction/utility.py
  - vbi/models/__init__.py
  - vbi/models/cpp/damp_oscillator.py
  - vbi/models/cpp/jansen_rit.py
  - vbi/models/cpp/km.py
  - vbi/models/cpp/mpr.py
  - vbi/models/
[…]
