+++ vbi/dataset/__init__.py
@@ -1,0 +1,5 @@
+"""Dataset utilities for VBI.
+
+Provides helpers for loading, saving, and managing simulation datasets
+used in the VBI inference pipeline.
+"""
