+++ vbi/tests/__init__.py
@@ -1,0 +1,6 @@
+"""Test suite for VBI.
+
+Collects unit and integration tests covering the simulation backends,
+feature-extraction routines, inference pipeline, and utility functions.
+Run via ``vbi.tests()`` or directly with pytest.
+"""
