+++ vbi/papers/__init__.py
@@ -1,0 +1,5 @@
+"""Reproducibility sub-package for VBI.
+
+Contains scripts and utilities for reproducing results from papers that
+used VBI, serving as worked examples for end-to-end inference workflows.
+"""
