### 2026-06-29 0fde950
Enhance documentation by adding module docstrings for version, C++ model wrapper, Pavlides reproducibility script, and MPR Numba backend tests
