+++ vbi/models/cpp/_src/__init__.py
@@ -1,0 +1,1 @@
+"""Auto-generated SWIG wrapper package for VBI C++ model binaries."""
