+++ vbi/_version.py
@@ -1,1 +1,2 @@
+"""Version string for the VBI package."""
 __version__ = "0.4.3"
