### 2026-06-29 ins-amu/vbi#74 @Ziaeemehr
Thank you for this feedback and for highlighting the accessibility angle — that is a very important point we had not considered. We have now added PEP 257-compliant module-level docstrings to all Python files in the package (commit pending). This covers:

- **Top-level modules** (`vbi/__init__.py`, `vbi/utils.py`, `vbi/plot.py`, `vbi/sbi_inference.py`, etc.) with descriptions of purpose and key exports
-  **All sub-package** __init__.py files (`models`, `feature_extraction`, `dataset`, `tests`, and all backend sub-packages)
- **All model implementation modules** across the Numba, CuPy, C++, PyTorch, and TVBK backends, including citations to the underlying model references
- **Feature extraction modules** and test files
- **Auto-generated SWIG wrapper files**, marked clearly as such
Users can now use `help(vbi)`, `help(vbi.models.numba.mpr)`, etc. to get meaningful descriptions directly in the Python REPL or via adaptive technology.
