### 2026-06-29 2334517
Refactor setup.py to streamline C++ extension compilation and improve dependency installation steps in tests.yml
