+++ setup.py
@@ -1,6 +1,4 @@
 import os
-import sys
-import platform
 import subprocess
 from setuptools import setup, Extension, find_packages
 from setuptools.command.build_ext import build_ext
@@ -14,213 +12,60 @@ def get_version():
                 return line.split('=')[1].strip().strip('"\'')
 
 
-class OptionalBuildExt(build_ext):
-    """
-    Build C++ extensions with graceful fallback on failure.
-    Supports environment variables for controlling compilation.
-    """
-    
+# Check for environment variable to control C++ compilation
+SKIP_CPP = os.environ.get('SKIP_CPP', '').lower() in ('1', 'true', 'yes')
+
+SRC_DIR = "vbi/models/cpp/_src"
+
+
+class CustomBuildExtCommand(build_ext):
     def run(self):
-        # Automatically skip C++ compilation on Windows
-        if platform.system() == "Windows":
-            print("Detected Windows system - skipping C++ extensions automatically")
-            print("VBI will work with Python/NumPy/Numba models only.")
+        if SKIP_CPP:
+            print("Skipping C++ compilation due to SKIP_CPP environment variable")
             return
-        
-        # Check for skip flags on non-Windows systems
-        skip_reasons = []
-        
-        # Check multiple environment variables for flexibility
-        skip_env_vars = [
-            'SKIP_CPP', 'VBI_NO_CPP', 'VBI_SKIP_CPP', 
-            'NO_CPP', 'DISABLE_CPP', 'CPP_DISABLE'
-        ]
-        
-        for env_var in skip_env_vars:
-            if os.environ.get(env_var, '').lower() in ('1', 'true', 'yes', 'on'):
-                skip_reasons.append(f"{env_var} environment variable set")
-                break  # Only report the first one found
-        
-        # Check for special marker file (alternative to env vars)
-        skip_file = os.path.join(os.path.dirname(__file__), '.skip_cpp')
-        if os.path.exists(skip_file):
-            skip_reasons.append(".skip_cpp file found")
-        
-        # Check for required tools
-        if not self._check_swig():
-            skip_reasons.append("SWIG not found")
-        
-        if not self._check_compiler():
-            skip_reasons.append("C++ compiler not found or incompatible")
-        
-        if skip_reasons:
-            print(f"Skipping C++ extensions: {', '.join(skip_reasons)}")
-            print("VBI will work with Python/NumPy/Numba models only.")
-            print("To force skipping C++ compilation, set SKIP_CPP=1 or VBI_NO_CPP=1")
-            return
-        
-        try:
-            self._compile_swig_interfaces()
-            super().run()
-            print("C++ extensions compiled successfully!")
-        except Exception as e:
-            print(f"Failed to compile C++ extensions: {e}")
-            print("VBI will work with Python/NumPy/Numba models only.")
-            print("To force skipping C++ compilation, set SKIP_CPP=1 or VBI_NO_CPP=1")
-    
-    def _check_swig(self):
-        try:
-            subprocess.run(['swig', '-version'], capture_output=True, check=True)
-            return True
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            return False
-    
-    def _check_compiler(self):
-        try:
-            # Check for gcc/g++ (non-Windows systems only)
-            subprocess.run(['g++', '--version'], capture_output=True, check=True)
-            return True
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            return False
-    
-    def _compile_swig_interfaces(self):
-        src_dir = "vbi/models/cpp/_src"
[…]
