+++ .github/workflows/tests.yml
@@ -11,13 +11,13 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: "3.10"
-    - name: Install dependencies
+    - name: Install system dependencies
       run: |
-        # Install system dependencies for C++ compilation
         sudo apt-get update
         sudo apt-get install -y build-essential swig
-        
-        # Install Python package with C++ extensions (setup.py handles compilation)
+
+    - name: Install Python package (compiles C++ extensions)
+      run: |
         python -m pip install .[inference]
 
     - name: Run tests
