### 2026-06-29 0e5e005
chore: Remove connectivity_88 dataset pending license verification
The AIBL-derived connectivity_88 dataset has been removed until redistribution
rights under the Australian Imaging, Biomarker & Lifestyle (AIBL) dataset
license are confirmed.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
