+++ vbi/dataset/connectivity_88/tract_lengths.txt
@@ -1,88 +1,0 @@
-0.000000e+00 3.889547e+01 6.556364e+01 3.384829e+01 6.376058e+01 2.708219e+01 3.921660e+01 5.351403e+01 2.723527e+01 3.055607e+01 5.422733e+01 5.706426e+01 7.154111e+01 6.743060e+01 3.558664e+01 5.400185e+01 3.077446e+01 4.807200e+01 4.737366e+01 5.408355e+01 5.071699e+01 7.693065e+01 7.233260e+01 6.994548e+01 7.222960e+01 6.959044e+01 7.643642e+01 5.912071e+01 1.513408e+01 4.812854e+01 3.506794e+01 3.195384e+01 4.853169e+01 5.115046e+01 3.448797e+01 3.935962e+01 3.744568e+01 4.027565e+01 4.072441e+01 7.941389e+01 5.294514e+01 4.746089e+01 6.131373e+01 5.720124e+01 6.960516e+01 6.201017e+01 8.105168e+01 7.531684e+01 9.058978e+01 8.197029e+01 8.738718e+01 8.854783e+01 8.315885e+01 4.170676e+01 6.324666e+01 6.432092e+01 7.512570e+01 7.253312e+01 7.654765e+01 6.016558e+01 4.153667e+01 5.508346e+01 7.479358e+01 7.620609e+01 7.616945e+01 8.133895e+01 7.926292e+01 8.161943e+01 8.589210e+01 9.431914e+01 1.027193e+02 8.614801e+01 7.148239e+01 7.254098e+01 8.171576e+01 8.710800e+01 8.672232e+01 6.049025e+01 4.710911e+01 5.726148e+01 6.827234e+01 6.316169e+01 5.633284e+01 9.096977e+01 9.084230e+01 9.577137e+01 9.482561e+01 9.837419e+01
-3.889547e+01 0.000000e+00 4.014208e+01 1.362891e+01 4.024238e+01 3.519254e+01 3.124402e+01 4.145467e+01 4.920029e+01 3.033502e+01 4.368435e+01 1.910361e+01 4.041441e+01 4.460625e+01 4.039082e+01 2.482506e+01 4.451866e+01 6.883252e+01 6.109017e+01 6.343070e+01 5.329038e+01 1.020994e+02 9.939825e+01 9.398702e+01 1.020221e+02 1.013543e+02 1.053428e+02 7.867211e+01 5.323929e+01 8.089303e+01 7.201389e+01 6.770232e+01 8.427693e+01 7.797076e+01 5.501355e+01 3.177584e+01 3.972214e+01 4.252029e+01 5.193738e+01 7.422514e+01 5.094158e+01 7.470849e+01 5.946887e+01 7.729559e+01 6.302338e+01 3.477775e+01 5.231687e+01 4.892478e+01 6.032181e+01 6.322398e+01 6.301445e+01 6.453572e+01 7.304544e+01 3.871140e+01 4.796492e+01 2.795633e+01 4.430466e+01 4.918036e+01 6.207847e+01 2.971678e+01 4.362179e+01 7.025729e+01 7.309476e+01 7.319773e+01 6.589438e+01 1.002137e+02 1.012771e+02 9.820850e+01 1.055807e+02 1.100318e+02 1.166466e+02 8.999413e+01 7.237918e+01 8.886013e+01 8.853573e+01 8.626347e+01 9.734789e+01 8.063948e+01 6.236311e+01 4.222720e+01 5.580773e+01 5.364162e+01 5.855123e+01 8.548936e+01 7.328941e+01 9.517723e+01 7.814084e+01 9.593464e+01
[…]
