### 2026-06-29 820a60a
fix: Change Python package installation to editable mode for C++ extensions
