### 2026-06-29 fc7bd6d
fix: Update parameter names for backward compatibility in VEP model
