### 2026-06-30 586e90d
Cherry-pick JOSS paper from joss branch onto main
Brings paper.md, paper.bib, figures, and run.sh from the joss branch's
joss/ directory so the paper can be edited alongside the current
codebase for the review response. Merged DOI fields from a stale
duplicate paper/paper.bib copy on the joss branch into joss/paper.bib
where missing.
