+++ joss/paper.bib
@@ -1,0 +1,192 @@
+@Article{Betancourt2014,
+author="M. Betancourt  and S. Byrne and S. Livingstone and M. Girolami",
+title="The geometric foundations of Hamiltonian Monte Carlo",
+journal="arXiv:1410.5110",
+year="2014",
+}
+
+@Article{Betancourt2013b,
+author="M. Betancourt  and M. Girolami",
+title="Hamiltonian Monte Carlo for hierarchical models",
+journal="arXiv:1312.0906",
+year="2013",
+doi={10.48550/arXiv.1312.0906},
+}
+@article{Hashemi2020,
+  title={The {Bayesian} Virtual Epileptic Patient: A probabilistic framework designed to infer the spatial map of epileptogenicity in a personalized large-scale brain model of epilepsy spread},
+  author={Hashemi, Meysam and Vattikonda, AN and Sip, Viktor and Guye, Maxime and Bartolomei, Fabrice and Woodman, Michael Marmaduke and Jirsa, Viktor K},
+  journal={NeuroImage},
+  volume={217},
+  pages={116839},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2020.116839},
+}
+@article{Sisson2007,
+  title={Sequential monte carlo without likelihoods},
+  author={Sisson, Scott A and Fan, Yanan and Tanaka, Mark M},
+  journal={Proceedings of the National Academy of Sciences},
+  volume={104},
+  number={6},
+  pages={1760--1765},
+  year={2007},
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.0607208104},
+}
+@article{ghosh2008noise,
+  title={Noise during rest enables the exploration of the brain's dynamic repertoire},
+  author={Ghosh, Anandamohan and Rho, Y and McIntosh, Anthony Randal and K{\"o}tter, Rolf and Jirsa, Viktor K},
+  journal={PLoS computational biology},
+  volume={4},
+  number={10},
+  pages={e1000196},
+  year={2008},
+  publisher={Public Library of Science San Francisco, USA},
+  doi={10.1371/journal.pcbi.1000196},
+}
+@article{petkoski2019transmission,
+  title={Transmission time delays organize the brain network synchronization},
+  author={Petkoski, Spase and Jirsa, Viktor K},
+  journal={Philosophical Transactions of the Royal Society A},
+  volume={377},
+  number={2153},
+  pages={20180132},
+  year={2019},
+  publisher={The Royal Society Publishing},
+  doi={10.1098/rsta.2018.0132},
+}
+@article{sanz2015mathematical,
+  title={Mathematical framework for large-scale brain network modeling in The Virtual Brain},
+  author={Sanz-Leon, Paula and Knock, Stuart A and Spiegler, Andreas and Jirsa, Viktor K},
+  journal={Neuroimage},
+  volume={111},
+  pages={385--430},
+  year={2015},
+  publisher={Elsevier}
+}
+
+@article{ziaeemehr2020frequency,
+  title={Frequency-dependent organization of the brain’s functional network through delayed-interactions},
+  author={Ziaeemehr, Abolfazl and Zarei, Mina and Valizadeh, Alireza and Mirasso, Claudio R},
+  journal={Neural Networks},
+  volume={132},
+  pages={155--165},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.neunet.2020.08.003},
+}
+@article{Papamakarios2017,
+  title={Masked autoregressive flow for density estimation},
+  author={Papamakarios, George and Pavlakou, Theo and Murray, Iain},
+  journal={Advances in Neural Information Processing Systems},
[…]
