+++ joss/paper.md
@@ -1,0 +1,135 @@
+---
+title: 'Virtual Brain Inference (VBI) toolkit on EBRAINS'
+date: 19 September 2025
+tags:
+  - Python
+  - Neuroscience
+  - Virtual Brain
+  - Simulation-Based Inference
+  - Probabilistic Inference
+  - fMRI
+  - EEG
+  - MEG
+authors:
+  - name: Abolfazl Ziaeemehr
+    orcid: 0000-0002-4696-9947
+    affiliation: 1
+    corresponding: true
+  - name: Lia Domide
+    orcid: 0000-0002-4822-2046
+    affiliation: 2
+  - name: Marmaduke Woodman
+    orcid: 0000-0002-8410-4581
+    affiliation: 1
+  - name: Spase Petkoski
+    orcid: 0000-0003-4540-6293
+    affiliation: 1
+  - name: Viktor Jirsa
+    orcid: 0000-0002-8251-8860
+    corresponding: true
+    affiliation: 1
+  - name: Meysam Hashemi
+    orcid: 0000-0001-5289-9837
+    affiliation: 1
+    corresponding: true
+affiliations:
+  - name: Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France
+    index: 1
+  - name: Codemart, Cluj-Napoca, Romania
+    index: 2
+bibliography: paper.bib
+
+
+
+---
+
+# Summary
+
+Understanding the principles and causal mechanisms underlying complex brain (dys)function is essential for advances in precision medicine. The virtual brain modeling represents a major step forward, combining computational models of brain dynamics with personalized anatomical data to simulate at scales commensurate with neuroimaging modalities. However, there is a lack of widely available tools to address the inverse problem of inferring neural mechanisms from observed brain activity. This work introduces Virtual Brain Inference (VBI), a flexible, scalable, and integrative Python toolkit designed for efficient probabilistic inference on virtual brain models. VBI is now available as an open-source service on the cloud research platform EBRAINS (ebrains.eu). It offers fast simulations of popular whole-brain models, a taxonomy of feature extraction methods, efficient data storage, and advanced probabilistic machine learning algorithms for causal inference from non-invasive and invasive recordings. VBI cloud services on EBRAINS, including demo applications, enable reproducible online collaboration, seamless integration with scalable and secure workflows, fostering research on large cohort datasets, improving generalizability, and potentially clinical translation.
+
+# Statement of Need
+
+VBI (https://github.com/ins-amu/vbi) is a Python-based toolkit tailored for probabilistic inference at the whole-brain scale. A lightweight version with minimal dependencies is available to run online in JupyterLab. It also leverages the flexibility of Python while harnessing the high performance of C++ for optimized computation and massive parallelization using CUDA on GPUs. VBI seamlessly integrates structural and functional neuroimaging data, providing space-efficient storage and memory-optimized batch processing. With its user-friendly API supporting:
+
+- **Whole-brain models**: Wilson-Cowan, Jansen-Rit, Stuart-Landau, Montbrió, Wong-Wang, and Epileptor.
+- **Fast simulation**: Just-in-time compilation of models across Python/C++ and CPU/GPU devices.
+- **Feature extraction**: Functional connectivity (FC), functional connectivity dynamics (FCD), and power spectral density (PSD).
+- **Deep neural density estimators**: Masked autoregressive flows (MAFs; @Papamakarios2017), and neural spline flows (NSFs; @Durkan2019).
+
[…]
