### 2026-06-30 1287cef
fix: Add DOI references and enhance documentation in paper.bib and paper.md
