+++ joss/paper.bib
@@ -3,6 +3,7 @@ author="M. Betancourt  and S. Byrne and S. Livingstone and M. Girolami",
 title="The geometric foundations of Hamiltonian Monte Carlo",
 journal="arXiv:1410.5110",
 year="2014",
+doi={10.48550/arXiv.1410.5110},
 }
 
 @Article{Betancourt2013b,
@@ -62,7 +63,8 @@ doi={10.48550/arXiv.1312.0906},
   volume={111},
   pages={385--430},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2015.01.002},
 }
 
 @article{ziaeemehr2020frequency,
@@ -108,7 +110,8 @@ doi={10.48550/arXiv.1312.0906},
   volume={163},
   pages={178--194},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neunet.2023.03.040},
 }
 
 
@@ -176,7 +179,8 @@ doi={10.1109/RBME.2025.3562951}
   volume={14},
   pages={RP106194},
   year={2025},
-  publisher={eLife Sciences Publications Limited}
+  publisher={eLife Sciences Publications Limited},
+  doi={10.7554/eLife.106194.1},
 }
 
 
@@ -188,5 +192,74 @@ doi={10.1109/RBME.2025.3562951}
   number={3},
   pages={035016},
   year={2022},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/2632-2153/ac9037},
+}
+
+@article{Sorrentino2024,
+  title={The virtual multiple sclerosis patient},
+  author={Sorrentino, P and Pathak, A and Ziaeemehr, A and Troisi Lopez, E and Cipriano, L and Romano, A and Sparaco, M and Quarantelli, M and Banerjee, A and Sorrentino, G and others},
+  journal={iScience},
+  volume={27},
+  number={7},
+  pages={110101},
+  year={2024},
+  publisher={Elsevier},
+  doi={10.1016/j.isci.2024.110101},
+}
+
+@article{Angiolelli2025,
+  title={The Virtual Parkinsonian patient},
+  author={Angiolelli, Marianna and Depannemaecker, Damien and Agouram, Hasnae and R{\'e}gis, Jean and Carron, Romain and Woodman, Marmaduke and Chiodo, Letizia and Triebkorn, Paul and Ziaeemehr, Abolfazl and Hashemi, Meysam and others},
+  journal={npj Systems Biology and Applications},
+  volume={11},
+  number={1},
+  pages={40},
+  year={2025},
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41540-025-00516-y},
+}
+
+@article{Mazzara2025,
+  title={Mapping Brain Lesions to Conduction Delays: The Next Step for Personalized Brain Models in Multiple Sclerosis},
+  author={Mazzara, Camille and Ziaeemehr, Abolfazl and Troisi Lopez, E and Cipriano, Lorenzo and Angiolelli, Marianna and Sparaco, Maddalena and Quarantelli, Mario and Granata, Carmine and Sorrentino, Giuseppe and Hashemi, Meysam and others},
+  journal={Human Brain Mapping},
+  volume={46},
+  number={7},
+  pages={e70219},
+  year={2025},
+  publisher={Wiley Online Library},
+  doi={10.1002/hbm.70219},
+}
+
+@article{Rabuffo2025,
+  title={Mapping global brain reconfigurations following local targeted manipulations},
[…]
