+++ joss/paper.md
@@ -60,23 +60,27 @@ Traditional methods such as Markov Chain Monte Carlo (MCMC) and Approximate Baye
 
 Designed for researchers and clinical applications, VBI enables personalized simulations of normal and pathological brain activity, aiding in distinguishing healthy from diseased states, and potentially informing clinical decision making. By addressing the efficient and scalable probabilistic inverse problem, VBI leverages high-performance computing for parallel processing of large-scale datasets [@Ziaeemehr2025].
 
+VBI is intended for computational neuroscientists, clinical researchers, and method developers who need to fit whole-brain network models to neuroimaging data (fMRI, EEG, MEG) at scale, without having to implement custom simulators, feature extractors, or deep-learning-based inference pipelines themselves.
+
+# State of the Field
+
+No single existing tool combines the components needed for scalable, amortized inference on whole-brain models. The Virtual Brain (TVB) [@sanz2013virtual] provides mature forward simulators that many of VBI's models build upon, but it lacks simulation-based inference with deep neural density estimators or amortized posteriors. General-purpose SBI libraries [@gonccalves2020training; @cranmer2020frontier] offer flexible density estimation but lack domain-specific whole-brain simulators, neuroimaging feature extractors, and GPU/C++ acceleration. Related model-specific Bayesian approaches, such as self-tuning Hamiltonian Monte Carlo [@Jha2022] and earlier amortized SBI for epilepsy models [@hashemi2023amortized; @hashemi2024simulation], demonstrate the need but are not general, reusable toolkits. VBI fills this gap, integrating whole-brain simulators, neuroimaging-specific feature extraction, and amortized inference in one GPU-accelerated, open-source package interoperable with TVB, rather than duplicating existing efforts.
+
 # Software Design
 
 **Design goals and trade-offs.** VBI was designed to solve the inverse problem for whole-brain models under realistic computational constraints. The primary goals were scalability to large simulation budgets, support for amortized probabilistic inference, and modularity across models, features, and inference algorithms. A key trade-off concerned flexibility versus performance: while pure Python implementations simplify extensibility, they are insufficient for large-scale simulation-based inference. VBI therefore adopts a hybrid design, exposing a Python API while relying on just-in-time compilation and GPU acceleration for performance-critical components. This choice enables rapid prototyping while achieving up to 100× speedups compared to CPU-based single core implementations.
 
 **Architecture.** The toolkit is organized around three decoupled layers: (1) simulators for whole-brain neural mass models, (2) feature-extraction operators for neuroimaging-derived statistics, and (3) inference backends based on simulation-based inference. Clear interfaces between these layers allow users to independently exchange models, features, or density estimators. Deployment on EBRAINS further motivated containerized, dependency-light components and a JupyterLab-based execution model, which was later extended with a graphical workflow editor for non-programmatic interaction.
 
-**Build vs. contribute.** Existing platforms such as The Virtual Brain provide mature forward simulators but do not support scalable simulation-based inference with deep neural density estimators or amortized posteriors. Conversely, generic SBI libraries lack domain-specific simulators, neuroimaging features, and data-handling workflows. VBI was therefore developed as a new toolkit to bridge these gaps while remaining interoperable with existing TVB components, rather than duplicating or fragmenting current efforts.
-
 # Research Impact Statement
 
-VBI enables a class of large-scale, amortized inference workflows for whole-brain models that were previously impractical using MCMC or ABC approaches. The methodology and core software have already supported peer-reviewed research demonstrating reliable parameter recovery and mechanistic insights in whole-brain dynamics, including a full validation study published in eLife. This constitutes realized impact beyond a purely methodological contribution.
[…]
