### 2026-06-30 ins-amu/vbi#75 @Ziaeemehr
Thank you for the helpful comments, responses below.

**1. State of the field**
We agree this is required as a distinct section. We have added a "State of
the Field" section to the paper, discussing The Virtual Brain and
general-purpose simulation-based inference libraries, with a build-vs-contribute
justification for why VBI was needed. We also added a sentence to the
Statement of Need clarifying the intended audience (computational
neuroscientists, clinical researchers, and method developers working with
whole-brain models).

**2. Figure 1 density**
We agree Figure 1 was dense. We added a new panel **G** (codebase/module
diagram) with non-bold labels linking the modules back to panels A–F:
Simulation → A–C, Features → D, Inference → E–F.

**3. Citing studies in the Research Impact Statement**
Done. We now directly cite the supporting studies: the full validation study
in eLife [@Ziaeemehr2025], methodological work on simulation-based inference
for virtual brain models [@hashemi2024simulation], applications to
personalized disease modeling and brain-network reconfiguration
[@Sorrentino2024; @Angiolelli2025; @Mazzara2025; @Rabuffo2025], and broader
virtual-brain workflows in probabilistic modeling and clinical neuroscience
[@Hashemi2024vbt; @Baldy2025; @Monteverdi2026].

**4. AI model disclosure**
Claude and ChatGPT/GPT models were used mainly for language polishing and
documentation drafting. All scientific content, software design,
implementation, and validation were performed and verified by the authors.
We clarify that the majority of the code was written by the authors and was
not produced through autonomous or "vibe coding" workflows.

**5. Missing DOIs**
We reviewed the bibliography and added missing DOIs where available.

We thank the reviewer again for these constructive suggestions, which we believe have meaningfully strengthened the clarity and completeness of the manuscript.
