### 2026-07-01 d076e94
feat: Enhance Kuramoto model demo with improved documentation and structure
