+++ docs/examples/simulator_models/kuramoto_demo.py
@@ -1,5 +1,6 @@
 """
-Kuramoto model: VBI simulator vs pure-NumPy reference.
+Kuramoto Model Demo
+====================
 
 Verifies that the VBI KuramotoCoupling produces trajectories identical to a
 single self-contained NumPy implementation using the same Heun integrator.
@@ -12,11 +13,19 @@ where α (--alpha) is the frustration angle. α=0 gives the standard model.
 
 Run
 ---
+::
+
     python kuramoto_demo.py                        # 5-node complete graph
     python kuramoto_demo.py --n 10 --duration 1000
     python kuramoto_demo.py --delayed              # add 5 ms axonal delay
     python kuramoto_demo.py --alpha 0.5236         # frustrated  (α = π/6)
 """
+
+# %%
+# Setup
+# -----
+# Imports and path setup so the demo runs both as a standalone script and as
+# a sphinx-gallery example.
 
 from __future__ import annotations
 
@@ -33,7 +42,14 @@ import matplotlib.pyplot as plt
 
 from helpers import ensure_repo_on_path, complete_graph_weights
 
-ensure_repo_on_path(__file__)
+try:
+    _SCRIPT_PATH = Path(__file__)
+except NameError:
+    # sphinx-gallery execs this file without setting __file__; it already
+    # chdirs into the script's own directory first, so cwd is equivalent.
+    _SCRIPT_PATH = Path.cwd() / "kuramoto_demo.py"
+
+ensure_repo_on_path(_SCRIPT_PATH)
 
 from vbi.simulator import Simulator
 from vbi.simulator.models.kuramoto import kuramoto
@@ -44,9 +60,11 @@ from vbi.simulator.spec.simulation import SimulationSpec
 from vbi.simulator.spec.connectivity import Connectivity
 
 
-# ---------------------------------------------------------------------------
-# Pure-NumPy reference (single function, no VBI dependency)
-# ---------------------------------------------------------------------------
+# %%
+# Reference implementation (pure NumPy)
+# --------------------------------------
+# A single, dependency-free function that integrates the (frustrated)
+# Kuramoto model with Heun's method, including optional axonal delays.
 
 def kuramoto_heun(
     theta0: np.ndarray,
@@ -129,9 +147,11 @@ def kuramoto_heun(
     return t, theta
 
 
-# ---------------------------------------------------------------------------
-# VBI runner
-# ---------------------------------------------------------------------------
+# %%
+# VBI simulator
+# -------------
+# Builds a :class:`~vbi.simulator.Simulator` with a Kuramoto coupling and
+# Heun integrator, mirroring the reference implementation's setup.
 
 def run_vbi(
     theta0: np.ndarray,
@@ -144,13 +164,8 @@ def run_vbi(
     speed: float,
     alpha: float = 0.0,
 ) -> tuple[np.ndarray, np.ndarray]:
-    import dataclasses
-
-    sv_with_init = (dataclasses.replace(kuramoto.state_variables[0], default_init=theta0),)
-    model_with_init = dataclasses.replace(kuramoto, state_variables=sv_with_init)
-
     spec = SimulationSpec(
-        model=model_with_init,
+        model=kuramoto.with_init(theta=theta0),
         integrator=IntegratorSpec(method="heun", dt=dt),
         coupling=CouplingSpec(kind="kuramoto", alpha=alpha),
         monitors=(MonitorSpec(kind="raw"),),
@@ -167,9 +182,11 @@ def run_vbi(
     return t, theta
 
 
-# ---------------------------------------------------------------------------
-# Plotting
-# ---------------------------------------------------------------------------
+# %%
+# Comparison plot
+# ---------------
+# Plots per-node ``sin(θ)``, unwrapped phase, the Kuramoto order parameter
[…]
