### 2026-07-01 0839466
feat: Update documentation for simulator models with sphinx-gallery integration and new examples
