+++ docs/conf.py
@@ -14,8 +14,27 @@ extensions = [
     "numpydoc",
     "sphinx.ext.graphviz",
     "sphinx.ext.viewcode",
-    'nbsphinx'
+    'nbsphinx',
+    "sphinx_gallery.gen_gallery",
 ]
+
+# Scoped to kuramoto_demo.py only: it is the only simulator_models script
+# fixed so far to run without a real __file__ (see docs/examples/simulator_models
+# helper pattern). Widen filename_pattern as more scripts are migrated
+# (notes/milestones_documentation.md, Step 4/5).
+sphinx_gallery_conf = {
+    "examples_dirs": ["examples/simulator_models"],
+    "gallery_dirs": ["auto_examples/simulator_models"],
+    "filename_pattern": r"kuramoto_demo\.py",
+    "ignore_pattern": r"helpers\.py",
+    "plot_gallery": True,
+    "download_all_examples": False,
+    "first_notebook_cell": (
+        "# This notebook is auto-generated from vbi documentation.\n"
+        "# Install: pip install vbi\n"
+    ),
+    "thumbnail_size": (400, 280),
+}
 
 source_suffix = ".rst"
 master_doc = "index"
@@ -34,7 +53,15 @@ pygments_style = "colorful"
 add_function_parentheses = True
 html_static_path = ['_static']
 
-exclude_patterns = ["_build", "**.ipynb_checkpoints"]
+# auto_examples/**/*.ipynb are sphinx-gallery-generated by-products; excluding
+# them from source discovery stops nbsphinx from also trying to build them
+# (it registers .ipynb as a source suffix and would otherwise pick them up
+# independently of the .rst page sphinx-gallery already generates for them).
+exclude_patterns = [
+    "_build",
+    "**.ipynb_checkpoints",
+    "auto_examples/**/*.ipynb",
+]
 
 numpydoc_show_class_members = False
 autodoc_member_order = "bysource"
