+++ docs/examples/simulator_models/GALLERY_HEADER.rst
@@ -1,0 +1,6 @@
+Simulator Model Demos
+=====================
+
+Demos for the new-API simulator pipeline, executed and rendered inline via
+sphinx-gallery. Each page offers "Download Python source code" and
+"Download Jupyter notebook" buttons.
