### 2026-07-01 66b106f
feat: Add livehtml target for live-reloading local preview in Sphinx documentation
