### 2026-07-01 71f9395
feat: Enhance ModelSpec with rich HTML representation and update Sphinx configuration for improved documentation
