### 2026-07-01 8d99e3f
feat: Revise MPR demo documentation and streamline simulation parameters for clarity
