+++ docs/examples/simulator_models/mpr_demo.py
@@ -1,6 +1,23 @@
 """
-Compare VBI simulator and TVB MPR trajectories.
+MPR Model Demo
+==============
+
+Compares the VBI simulator's Montbrio-Pazo-Roxin (MPR) mean-field model
+against The Virtual Brain (TVB)'s own MPR implementation, using the same
+Heun integrator and connectivity. Requires the optional ``tvb-library``
+package.
+
+Run
+---
+::
+
+    python mpr_demo.py                  # 6-node complete graph, 40 ms
+    python mpr_demo.py --duration 100
 """
+
+# %%
+# Setup
+# -----
 
 from __future__ import annotations
 
@@ -11,186 +28,172 @@ import sys
 sys.dont_write_bytecode = True
 
 import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
 
-from helpers import (
-    comparison_metrics,
-    complete_graph_weights,
-    constant_tract_lengths,
-    ensure_repo_on_path,
-    homogeneous_node_params,
-    make_tvb_connectivity,
-    quiet_optional_imports,
-    quiet_tvb,
-    save_state_comparison_plot,
-)
+try:
+    _SCRIPT_PATH = Path(__file__)
+except NameError:
+    # sphinx-gallery execs this file without setting __file__; it already
+    # chdirs into the script's own directory first, so cwd is equivalent.
+    _SCRIPT_PATH = Path.cwd() / "mpr_demo.py"
 
-ensure_repo_on_path(__file__)
+# Prefer the vbi living in this checkout over any other version already
+# installed (e.g. a different vbi checkout installed editable elsewhere).
+# Downloaded standalone copies of this script should `pip install vbi`
+# instead - see the first notebook cell.
+_repo_root = _SCRIPT_PATH.resolve().parents[3]
+sys.path.insert(0, str(_repo_root))
 
-# Importing the package root may probe optional dependencies; keep this example's
-# output focused on the simulator run.
-with quiet_optional_imports():
-    from vbi.simulator import Simulator
-    from vbi.simulator.models.mpr import mpr
-    from vbi.simulator.spec.coupling import CouplingSpec
-    from vbi.simulator.spec.integrator import IntegratorSpec
-    from vbi.simulator.spec.monitor import MonitorSpec
-    from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator import Simulator
+from vbi.simulator.models.mpr import mpr
+from vbi.simulator.spec.coupling import CouplingSpec
+from vbi.simulator.spec.integrator import IntegratorSpec
+from vbi.simulator.spec.monitor import MonitorSpec
+from vbi.simulator.spec.simulation import SimulationSpec
 from vbi.simulator.spec.connectivity import Connectivity
 
-
-NOTEBOOK_PARAMS = {
-    "nn": 6,
-    "G": 0.33,
-    "dt": 0.01,
-    "tau": 1.0,
-    "eta": -4.6,
-    "Delta": 0.7,
-    "J": 14.5,
-    "I": 2.0,
-    "Gamma": 0.0,
-    "cr": 1.0,
-    "cv": 0.0,
-    "rv_decimate": 10,
-    "tract_length": 4.0,
-    "speed": 4.0,
-}
+N_NODES = 6
+G = 0.33
+DT = 0.01
+TRACT_LENGTH = 4.0
+SPEED = 4.0
 
 
-def build_vbi_spec(method: str) -> SimulationSpec:
-    weights = complete_graph_weights(NOTEBOOK_PARAMS["nn"])
-    tract_lengths = constant_tract_lengths(weights, NOTEBOOK_PARAMS["tract_length"])
[…]
