### 2026-07-01 6ee1af7
feat: Update Sphinx configuration to include mpr_demo.py in filename pattern and add model equations reference documentation
