### 2026-07-01 66808f4
feat: Update Wilson-Cowan demo to include additional filename pattern and enhance documentation
