+++ docs/examples/simulator_models/wilson_cowan_demo.py
@@ -1,4 +1,23 @@
-"""Reproduce the first Wilson-Cowan SDE notebook visualization."""
+"""
+Wilson-Cowan Model Demo
+=========================
+
+Reproduces the first Wilson-Cowan SDE notebook visualization: a two-node
+excitatory-inhibitory oscillator driven into a limit cycle by external
+drive ``P``, integrated with additive noise. Plots the stochastic
+trajectory, its power spectrum, and the E-I phase plane.
+
+Run
+---
+::
+
+    python wilson_cowan_demo.py
+    python wilson_cowan_demo.py --duration 4000
+"""
+
+# %%
+# Setup
+# -----
 
 from __future__ import annotations
 
@@ -9,149 +28,105 @@ import sys
 sys.dont_write_bytecode = True
 
 import numpy as np
+import matplotlib
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
 
-from helpers import (
-    ensure_repo_on_path,
-    homogeneous_node_params,
-    quiet_optional_imports,
-)
+try:
+    _SCRIPT_PATH = Path(__file__)
+except NameError:
+    # sphinx-gallery execs this file without setting __file__; it already
+    # chdirs into the script's own directory first, so cwd is equivalent.
+    _SCRIPT_PATH = Path.cwd() / "wilson_cowan_demo.py"
 
-ensure_repo_on_path(__file__)
+# Prefer the vbi living in this checkout over any other version already
+# installed (e.g. a different vbi checkout installed editable elsewhere).
+# Downloaded standalone copies of this script should `pip install vbi`
+# instead - see the first notebook cell.
+_repo_root = _SCRIPT_PATH.resolve().parents[3]
+sys.path.insert(0, str(_repo_root))
 
-with quiet_optional_imports():
-    from vbi.simulator import Simulator
-    from vbi.simulator.models.wilson_cowan import wilson_cowan
-    from vbi.simulator.spec.coupling import CouplingSpec
-    from vbi.simulator.spec.integrator import IntegratorSpec
-    from vbi.simulator.spec.monitor import MonitorSpec
-    from vbi.simulator.spec.simulation import SimulationSpec
+from vbi.simulator import Simulator
+from vbi.simulator.models.wilson_cowan import wilson_cowan
+from vbi.simulator.spec.coupling import CouplingSpec
+from vbi.simulator.spec.integrator import IntegratorSpec
+from vbi.simulator.spec.monitor import MonitorSpec
+from vbi.simulator.spec.simulation import SimulationSpec
 from vbi.simulator.spec.connectivity import Connectivity
 
+DT = 0.05
+NOISE_AMP = 0.0005
+NOISE_SEED = 42
+T_CUT = 101.0  # drop this much initial transient before analysis/plotting
 
-WC_PARAMS = {
-    "nn": 2,
-    "coupling_strength": 0.0,
-    "dt": 0.05,
-    "t_cut": 101.0,
-    "noise_amp": 0.0005,
-    "noise_seed": 42,
-    "c_ee": 16.0,
-    "c_ei": 12.0,
-    "c_ie": 15.0,
-    "c_ii": 3.0,
-    "tau_e": 8.0,
-    "tau_i": 8.0,
-    "a_e": 1.3,
-    "b_e": 4.0,
-    "c_e": 1.0,
-    "theta_e": 0.0,
-    "a_i": 2.0,
-    "b_i": 3.7,
-    "c_i": 1.0,
-    "theta_i": 0.0,
-    "r_e": 1.0,
-    "r_i": 1.0,
-    "k_e": 0.994,
-    "k_i": 0.999,
[…]
