### 2026-07-01 7d2e18e
Refactor simulator model demos and backend integration
- Updated Jansen-Rit demo to improve documentation and streamline imports.
- Enhanced JAX demo by removing unnecessary repo path helper and improving path handling.
- Revised Reduced Wong-Wang demo for clarity and updated parameter handling.
- Improved Stimulation demo with better structure and output handling.
- Refactored validation script to use connectivity instead of weights for backend compatibility.
- Updated Wong-Wang demo to clarify model parameters and improve output handling.
- Adjusted numpy backend sweeper to use connectivity 
[…]
