+++ docs/examples/simulator_models/wong_wang_demo.py
@@ -1,28 +1,49 @@
-"""Simulate full excitatory-inhibitory Wong-Wang BOLD and FCD.
-
-This mirrors rww_bold_fcd.py, but uses the full Wong-Wang excitatory-
-inhibitory model. The BOLD monitor is driven by the excitatory gating
-variable S_e, which is the full model's long-range coupled variable.
 """
+Wong-Wang Exc-Inh BOLD + FCD Demo
+===================================
+
+Simulates the full excitatory-inhibitory Wong-Wang model on an 84-node
+structural connectome and computes BOLD + FCD (functional connectivity
+dynamics), driven by the excitatory gating variable ``S_e`` - the model's
+long-range coupled variable. Compare with :doc:`reduced_wong_wang_demo`,
+which uses the reduced single-population model on the same connectome.
+
+Run
+---
+::
+
+    python wong_wang_demo.py
+"""
+
+# %%
+# Setup
+# -----
 
 from __future__ import annotations
 
-import sys
-import warnings
 from pathlib import Path
+import sys
 
 sys.dont_write_bytecode = True
-warnings.filterwarnings("ignore")
 
-from helpers import ensure_repo_on_path
-
-ensure_repo_on_path(__file__)
-
-import matplotlib
 import numpy as np
-
+import matplotlib
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
+
+try:
+    _SCRIPT_PATH = Path(__file__)
+except NameError:
+    # sphinx-gallery execs this file without setting __file__; it already
+    # chdirs into the script's own directory first, so cwd is equivalent.
+    _SCRIPT_PATH = Path.cwd() / "wong_wang_demo.py"
+
+# Prefer the vbi living in this checkout over any other version already
+# installed (e.g. a different vbi checkout installed editable elsewhere).
+# Downloaded standalone copies of this script should `pip install vbi`
+# instead - see the first notebook cell.
+_repo_root = _SCRIPT_PATH.resolve().parents[3]
+sys.path.insert(0, str(_repo_root))
 
 import vbi
 from vbi.feature_extraction.features import get_fcd
@@ -34,96 +55,72 @@ from vbi.simulator.spec.monitor import MonitorSpec
 from vbi.simulator.spec.simulation import SimulationSpec
 from vbi.simulator.spec.connectivity import Connectivity
 
+SEED = 42
+G = 1.91                    # global coupling (model default is 2.0)
+DT = 1.0                     # ms
+DURATION = 5 * 60 * 1000.0   # ms (5 min)
+T_CUT = 1 * 60 * 1000.0      # ms (1 min burn-in)
+TR = 300.0                   # ms BOLD repetition time
+SIGMA = 0.005                # noise amplitude for S_e and S_i
 
-D = vbi.LoadSample(nn=84)
-weights = D.get_weights()
-nn = weights.shape[0]
 
-seed = 42
-dt = 1.0                  # ms
-T = 5 * 60 * 1000.0       # ms  (5 min)
-t_cut = 1 * 60 * 1000.0   # ms  (1 min burn-in)
-tr = 300.0                # ms  BOLD repetition time
-sigma = 0.005             # noise amplitude for S_e and S_i
+# %%
+# VBI simulator
+# -------------
+# Every model parameter besides ``G`` uses wong_wang_exc_inh's own default.
 
[…]
