+++ docs/examples/simulator_models/benchmark_all_backends.py
@@ -31,7 +31,27 @@ Usage
     python benchmark_all_backends.py --no-jax           # skip JAX
     python benchmark_all_backends.py --jax-sizes 16 32 64 128 256 512 1024
     python benchmark_all_backends.py --n-workers 8 --cpu-sizes 4 8 16 32 64 128
+
+This script is not executed on RTD (needs CUDA/JAX/a compiled C++ extension,
+and bakes in wall-clock timings that would be meaningless on shared build
+hardware). The gallery page shows its source plus a real, pre-rendered
+benchmark plot generated by manually running it on real hardware - see
+docs/examples/simulator_models/reference_images/ and the
+install_reference_images() hook in docs/conf.py.
 """
+# sphinx_gallery_dummy_images = 1
+
+# %%
+# Example output
+# ---------------
+# Generated by manually running this script on real hardware (not
+# re-executed on every docs build - see the module docstring).
+#
+# .. figure:: /examples/simulator_models/reference_images/benchmark_all_backends_001.png
+#    :alt: Backend sweep benchmark - wall-clock time, CPU speedup, throughput
+#
+#    SupHopf, 10 nodes, 3 CPU sweep sizes (4/8/16), NumPy/Numba/C++ only.
+
 from __future__ import annotations
 
 import argparse
@@ -50,19 +70,15 @@ warnings.filterwarnings(
     message="Grid size.*will likely result in GPU under-utilization",
 )
 
+# Prefer the vbi living in this checkout over any other version already
+# installed (e.g. a different vbi checkout installed editable elsewhere).
 sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
-from helpers import (
-    ensure_repo_on_path,
-    complete_graph_weights,
-    plot_sweep_benchmark,
-)
-ensure_repo_on_path(__file__)
 
 import numba
 
 from vbi.simulator import Sweeper
 from vbi.simulator.spec import (
-    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec,
+    SimulationSpec, IntegratorSpec, CouplingSpec, MonitorSpec, Connectivity,
 )
 from vbi.simulator.spec.sweep import SweepSpec
 from vbi.simulator.backend.cpp.sweeper import CppSweeper
@@ -106,6 +122,25 @@ N_REPEATS = 2
 # Helpers
 # ---------------------------------------------------------------------------
 
+def complete_graph_weights(n_nodes: int) -> np.ndarray:
+    """Return a dense complete-graph weight matrix with zero diagonal."""
+    weights = np.ones((n_nodes, n_nodes), dtype=np.float64)
+    np.fill_diagonal(weights, 0.0)
+    return weights
+
+
+def extrapolate_wall(ns_measured: list[int], times: list[float],
+                     ns_extrap: list[int]) -> tuple[np.ndarray, np.ndarray]:
+    """
+    Fit  t = slope * n  on the last <=3 measured CPU points and evaluate at
+    ns_extrap.  CPU sweep time is linear in n_samples after JIT warmup.
+    """
+    ns    = np.array(ns_measured[-3:], dtype=float)
+    ts    = np.array(times[-3:],       dtype=float)
+    slope = float(np.mean(ts / ns))
+    return np.array(ns_extrap, dtype=float), slope * np.array(ns_extrap, dtype=float)
+
+
 def _weights(n: int, seed: int = 0) -> np.ndarray:
     rng = np.random.default_rng(seed)
     W = complete_graph_weights(n).astype(np.float64)
[…]
