### 2026-07-01 fe483cc
feat: Update documentation headers for simulator model benchmarks and demos for clarity and consistency
