+++ docs/examples/simulator_models/benchmark_cpp_vs_numpy_numba.py
@@ -1,4 +1,7 @@
 """
+NumPy vs Numba vs C++ Sweep Benchmark
+=======================================
+
 Benchmark: NumPy vs Numba vs C++ (serial and parallel) for parameter sweeps.
 
 NumPy runs sequentially in Python.
