### 2026-07-01 1d976b5
feat: Add Connectivity import to custom model demo for enhanced simulation capabilities
