### 2026-07-01 ffa7624
feat: Enhance Sphinx configuration and custom model demo for improved execution and clarity
