+++ docs/conf.py
@@ -18,20 +18,46 @@ extensions = [
     "sphinx_gallery.gen_gallery",
 ]
 
-# Scoped to these scripts: they are the only simulator_models scripts
-# fixed so far to run without a real __file__ and without a helpers.py
-# dependency (see docs/examples/simulator_models helper pattern). Widen
-# filename_pattern as more scripts are migrated
-# (notes/milestones_documentation.md, Step 4/5).
-# NOTE: mpr_demo.py imports tvb-library, which is NOT in docs/requirements.txt -
-# this works for a local build (tvb installed in this dev venv) but will break
-# the real RTD build until that's resolved (see conversation with user).
+# Execute every script by default; only the ones below are permanently
+# excluded, each for a real technical reason (not "not yet migrated" - all
+# scripts in this directory are now self-contained, no helpers.py needed):
+#   - jax_demo.py: `jax` is an optional vbi extra (pip install vbi[jax]),
+#     not in docs/requirements.txt, and it calls sys.exit(1) if missing -
+#     which could crash the whole sphinx-gallery build, not just one page.
+#   - benchmark_all_backends.py, benchmark_all_models.py,
+#     benchmark_cpp_vs_numpy_numba.py, benchmark_numba_vs_numpy.py: need
+#     CUDA/JAX/a compiled C++ extension, and bake in wall-clock timings
+#     that would be meaningless on shared RTD build hardware. Show a real,
+#     pre-rendered plot instead - see reference_images/ and
+#     install_reference_images() below.
+#   - validate_sweep_backends.py: a pass/fail QA tool, not doc content.
+#   - cuda_sweep_demo.py: needs an actual CUDA GPU (RTD has none).
+# NOTE: mpr_demo.py and jansen_rit_demo.py import tvb-library, which is NOT
+# in docs/requirements.txt - this works for a local build (tvb installed in
+# this dev venv) but will break the real RTD build until that's resolved
+# (see conversation with user).
+_GALLERY_NOT_EXECUTED = (
+    "jax_demo", "benchmark_all_backends", "benchmark_all_models",
+    "benchmark_cpp_vs_numpy_numba", "benchmark_numba_vs_numpy",
+    "validate_sweep_backends", "cuda_sweep_demo",
+)
 sphinx_gallery_conf = {
     "examples_dirs": ["examples/simulator_models"],
     "gallery_dirs": ["auto_examples/simulator_models"],
-    "filename_pattern": r"kuramoto_demo\.py|model_equations\.py|mpr_demo\.py|wilson_cowan_demo\.py",
+    "filename_pattern": (
+        r"^(?!.*(?:" + "|".join(_GALLERY_NOT_EXECUTED) + r")\.py$).*\.py$"
+    ),
     "ignore_pattern": r"helpers\.py",
     "plot_gallery": True,
+    # Default (True) gives every `# %%` section its own toctree/sidebar
+    # entry - with our scripts reusing generic section names (Usage, Setup,
+    # Run the demo, Example output) across many pages, the sidebar fills up
+    # with ambiguous flat entries. One entry per page is clearer here.
+    "nested_sections": False,
+    # Hide `# sphinx_gallery_x = y` config-comment lines (e.g. dummy_images)
+    # from the rendered page - they're directives for the builder, not
+    # content for readers.
+    "remove_config_comments": True,
     "download_all_examples": False,
     "first_notebook_cell": (
         "# This notebook is auto-generated from vbi documentation.\n"
@@ -120,10 +146,39 @@ def on_missing_reference(app, env, node, contnode):
         return None
 
 
+def install_reference_images(app):
+    """
+    Copy pre-rendered reference plots into the exact path sphinx-gallery
[…]
