### 2026-07-01 c769fb8
feat: Enhance LaTeX rendering for ModelSpec to support horizontal scrolling for wide equations
