### 2026-07-01 e071150
feat: Document workflow bottleneck in feature extraction across backends
