### 2026-07-07 5fec988
feat: Add limitations section to documentation for clarity on VBI usage
