### 2026-07-07 ins-amu/vbi#70 @sampottinger
Perfect! https://github.com/ins-amu/vbi/blob/main/docs/third_party_licenses.rst looks great thank you!
