### 2026-07-07 ins-amu/vbi#71 @sampottinger
Thanks! I took a look at https://github.com/ins-amu/vbi/blob/main/README.md and the new section looks fantastic. Appreciate it!
