### 2026-07-07 ins-amu/vbi#72 @sampottinger
Incredible work. Thank you for expanding all of this out!
