### 2026-07-07 ins-amu/vbi#73 @sampottinger
Perfect sounds good. I’d just be mindful about the GPL license requirements on https://github.com/jlizier/jidt. Following up on using this via JPype... If you don’t mind, can you please confirm you are not subject to the copyleft provisions given how you structured the interaction with the jar?  If you have confirmed already, please go ahead and close this issue with comment explaining compliance. Thanks!
