### 2026-07-07 ins-amu/vbi#74 @sampottinger
Thanks for this! The [expansions](https://github.com/ins-amu/vbi/blob/main/vbi/__init__.py) look great!
