### 2026-07-07 ins-amu/vbi#76 @sampottinger
Datasets README
Hello! Thank you for your continued work on https://github.com/openjournals/joss-reviews/issues/10612#issuecomment-4902340874. You [mentioned in the other thread](https://github.com/openjournals/joss-reviews/issues/10612#issuecomment-4834259304) that you added `vbi/dataset/README.md` but I don’t see it at https://github.com/ins-amu/vbi/tree/main/vbi/dataset on `main`. Is it still waiting on another branch? Thanks! Also, apologies if I missed something obvious on my end.
