### 2026-07-07 ins-amu/vbi#75 @sampottinger
Thank you! This looks great. Just a few comments on another read here mostly related to citations:

 - The [updated figure](https://github.com/ins-amu/vbi/blob/main/joss/Fig1.png) looks great! For some reason, the [editorial bot build](https://github.com/openjournals/joss-reviews/issues/10612#issuecomment-4905486850) of your paper is not updating though with new changes. Can you take a quick look at that?
 - I love how you handled the `Deep neural density estimators` bullet. Would it be possible to add similar formal citations in the `Whole-brain models` bullet point in your Statement of Need?
 - Thank you for your expanded AI Usage Disclosure. Would you be comfortable adding citations for the models used? For example, you can use https://www.anthropic.com/system-cards and https://openai.com/index/gpt-5-system-card/ or similar.
