### 2026-07-09 81ec4dd
feat: Add bundled datasets section to documentation with details on connectivity_84
