+++ vbi/dataset/README.md
@@ -1,0 +1,21 @@
+# Bundled datasets
+
+These datasets are included as minimal working examples so users can run the
+VBI tutorials and examples without preparing their own data.
+
+## connectivity_84
+
+- **Description**: Structural connectivity matrix (weights and tract
+  lengths) and region centres/labels for 84 cortical regions.
+- **Source**: Hagmann et al. (2008), derived from diffusion-weighted MRI
+  (DW-MRI) of healthy adult subjects.
+- **Provenance**: Distributed as part of
+  [The Virtual Brain (TVB)](https://github.com/the-virtual-brain/tvb-data)
+  demo datasets and redistributed here unmodified.
+- **License**: [CC BY 3.0](https://creativecommons.org/licenses/by/3.0/)
+- **Citation**:
+  Hagmann, P., Cammoun, L., Gigandet, X., Meuli, R., Honey, C. J., Wedeen,
+  V. J., & Sporns, O. (2008). Mapping the structural core of human cerebral
+  cortex. *PLoS Biology*, 6(7), e159.
+  https://doi.org/10.1371/journal.pbio.0060159
+
