### 2026-07-09 ins-amu/vbi#76 @Ziaeemehr
Hi, thanks for catching that, apologies, the earlier reply was premature; the README hadn't actually been pushed yet. It's now committed and live on `main`:

- https://github.com/ins-amu/vbi/blob/main/vbi/dataset/README.md
- Commit: https://github.com/ins-amu/vbi/commit/81ec4ddfc5a8be88e27f916ccb91a5cae994ab46

This adds provenance/license details for `connectivity_84` (Hagmann et al. 2008, CC BY 3.0, via The Virtual Brain) and notes that `connectivity_88` has been removed pending confirmation of its redistribution license. A corresponding entry has also been added to `docs/third_party_licenses.rst`.
