### 2026-07-09 ins-amu/vbi#75 @Ziaeemehr
thank you
- **Editorial bot build**:  the review issue is building from the `joss` branch, which is behind `main` (missing the State of the Field section, updated citations, etc.). the reviewer-requested changes are on the main branch; could you re-trigger the bot considering that?
- **Whole-brain models citations**: added, matching the style used for the density estimators bullet, Wilson-Cowan [@WilsonCowan1972], Jansen-Rit [@JansenRit1995], Stuart-Landau [@Cabral2011], Montbrió [@Montbrio2015], Wong-Wang [@Deco2013], and Epileptor [@Jirsa2014].
- **AI Usage Disclosure**: added citations to the Anthropic and OpenAI system cards you linked.

All changes are pushed to `main`. Thanks again for the careful read!
