### 2026-07-09 22030e9
fix: Add references to new articles in bibliography and update model names in documentation
