+++ joss/paper.bib
@@ -263,3 +263,89 @@ doi={10.1109/RBME.2025.3562951}
   publisher={Springer},
   doi={10.1186/s13195-026-02114-4},
 }
+
+@article{WilsonCowan1972,
+  title={Excitatory and inhibitory interactions in localized populations of model neurons},
+  author={Wilson, Hugh R and Cowan, Jack D},
+  journal={Biophysical Journal},
+  volume={12},
+  number={1},
+  pages={1--24},
+  year={1972},
+  publisher={Elsevier},
+  doi={10.1016/S0006-3495(72)86068-5},
+}
+
+@article{JansenRit1995,
+  title={Electroencephalogram and visual evoked potential generation in a mathematical model of coupled cortical columns},
+  author={Jansen, Ben H and Rit, Vincent G},
+  journal={Biological Cybernetics},
+  volume={73},
+  number={4},
+  pages={357--366},
+  year={1995},
+  publisher={Springer},
+  doi={10.1007/BF00199471},
+}
+
+@article{Cabral2011,
+  title={Role of local network oscillations in resting-state functional connectivity},
+  author={Cabral, Joana and Hugues, Etienne and Sporns, Olaf and Deco, Gustavo},
+  journal={NeuroImage},
+  volume={57},
+  number={1},
+  pages={130--139},
+  year={2011},
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2011.04.010},
+}
+
+@article{Montbrio2015,
+  title={Macroscopic description for networks of spiking neurons},
+  author={Montbri{\'o}, Ernest and Paz{\'o}, Diego and Roxin, Alex},
+  journal={Physical Review X},
+  volume={5},
+  number={2},
+  pages={021028},
+  year={2015},
+  publisher={American Physical Society},
+  doi={10.1103/PhysRevX.5.021028},
+}
+
+@article{Deco2013,
+  title={Resting-state functional connectivity emerges from structurally and dynamically shaped slow linear fluctuations},
+  author={Deco, Gustavo and Ponce-Alvarez, Adri{\'a}n and Hagmann, Patric and Romani, Gian Luca and Mantini, Dante and Corbetta, Maurizio},
+  journal={Journal of Neuroscience},
+  volume={33},
+  number={27},
+  pages={11239--11252},
+  year={2013},
+  publisher={Society for Neuroscience},
+  doi={10.1523/JNEUROSCI.1091-13.2013},
+}
+
+@article{Jirsa2014,
+  title={On the nature of seizure dynamics},
+  author={Jirsa, Viktor K and Stacey, William C and Quilichini, Pascale P and Ivanov, Anton I and Bernard, Christophe},
+  journal={Brain},
+  volume={137},
+  number={8},
+  pages={2210--2230},
+  year={2014},
+  publisher={Oxford University Press},
+  doi={10.1093/brain/awu133},
+}
+
+@misc{AnthropicSystemCard2025,
+  title={Claude's system card},
+  author={{Anthropic}},
+  year={2025},
+  howpublished={\url{https://www.anthropic.com/system-cards}},
+}
+
+@misc{OpenAIGPT5SystemCard2025,
+  title={GPT-5 system card},
+  author={{OpenAI}},
+  year={2025},
+  howpublished={\url{https://openai.com/index/gpt-5-system-card/}},
+}
