### 2026-07-09 9e35f4b
feat: Implement JIDT subprocess bridge for information theory calculations and add tests for correctness and process isolation
