### 2026-07-09 ins-amu/vbi#73 @Ziaeemehr
Thanks for pushing on this, you were right to ask us to look closer rather than take the previous setup for granted.

**Short version:** we don't think the old architecture supported a clean "not subject to copyleft" claim, so instead of asserting compliance, we changed the architecture.

**What changed:** previously, JPype embedded the JIDT JVM directly inside VBI's own Python process, and VBI's Apache-2.0 code called JIDT's Java classes in-process (`jp.startJVM` + `JClass`/`JPackage`). That's linking, not arm's-length use.

Now, `infodynamics.jar` (GPL-3.0) runs in its own independent OS process (`vbi/feature_extraction/jidt_worker.py`), launched on demand by the main VBI process. The two processes communicate *only* through a line-delimited JSON protocol over stdin/stdout, plain numeric arrays in, plain numeric results out. VBI's own process never embeds the JIDT JVM or links against its classes; it just invokes a separate program and exchanges data with it, which is the standard "mere aggregation" pattern rather than a combined/linked work.

This is documented in `docs/third_party_licenses.rst` under "GPL-3.0 isolation."

We verified the three affected features (mutual information, transfer entropy, differential entropy) against the new subprocess bridge — same numerical results, and clean process teardown with no lingering JVMs.

Thanks again for catching this, happy to close the issue, or keep it open if you'd like more detail on the isolation boundary.
