### 2026-07-16 4f8d3b5
feat: Add legacy examples documentation and update index structure for clarity
