### 2026-07-16 baf7c68
Move examples/ out of docs/ to the project root
Examples are user-facing content, not documentation-build machinery,
so they belong at the top level rather than nested under docs/.
Updates sphinx-gallery's examples_dirs, the reference-image copy hook,
docker_build.rst, and README links accordingly, and fixes the
repo-root path math in the simulator_models gallery scripts (now one
directory shallower).

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
