+++ examples/simulator_models/benchmark_cpp_vs_numpy_numba.py
@@ -18,7 +18,7 @@ This script is not executed on RTD (needs a compiled C++ extension and bakes
 in wall-clock timings that would be meaningless on shared build hardware).
 The gallery page shows its source plus a real, pre-rendered benchmark plot
 generated by manually running it on real hardware - see
-docs/examples/simulator_models/reference_images/ and the
+examples/simulator_models/reference_images/ and the
 install_reference_images() hook in docs/conf.py.
 """
 # sphinx_gallery_dummy_images = 1
@@ -49,7 +49,7 @@ import numba
 
 # Prefer the vbi living in this checkout over any other version already
 # installed (e.g. a different vbi checkout installed editable elsewhere).
-sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
 from vbi.simulator.api import Sweeper
 from vbi.simulator.models.mpr import mpr
