+++ examples/simulator_models/benchmark_numba_vs_numpy.py
@@ -19,7 +19,7 @@ Run:
 This script is not executed on RTD (bakes in wall-clock timings that would
 be meaningless on shared build hardware). The gallery page shows its source
 plus a real, pre-rendered benchmark plot generated by manually running it on
-real hardware - see docs/examples/simulator_models/reference_images/ and the
+real hardware - see examples/simulator_models/reference_images/ and the
 install_reference_images() hook in docs/conf.py.
 """
 # sphinx_gallery_dummy_images = 1
@@ -48,7 +48,7 @@ import numpy as np
 
 # Prefer the vbi living in this checkout over any other version already
 # installed (e.g. a different vbi checkout installed editable elsewhere).
-sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
 from vbi.simulator import Sweeper
 from vbi.simulator.models.mpr import mpr
