+++ examples/simulator_models/benchmark_all_models.py
@@ -29,7 +29,7 @@ This script is not executed on RTD (needs CUDA/JAX/a compiled C++ extension,
 and bakes in wall-clock timings that would be meaningless on shared build
 hardware). The gallery page shows its source plus real, pre-rendered
 benchmark plots generated by manually running it on real hardware - see
-docs/examples/simulator_models/reference_images/ and the
+examples/simulator_models/reference_images/ and the
 install_reference_images() hook in docs/conf.py.
 """
 # sphinx_gallery_dummy_images = 2
@@ -64,7 +64,7 @@ import numpy as np
 
 # Prefer the vbi living in this checkout over any other version already
 # installed (e.g. a different vbi checkout installed editable elsewhere).
-sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 # This script's own directory must be on sys.path too, for the sibling
 # benchmark_all_backends import below (sphinx-gallery execs files with cwd
 # already set to this directory, but plain `python` invocation needs this).
