### 2026-07-16 a473808
feat: Add milestone documentation for TVB-Style Jansen-Rit Coupling and TVB Hybrid Simulator Backend Integration
