+++ notes/milestone_tvb_hybrid_backend_integration.md
@@ -1,0 +1,194 @@
+# Milestone: TVB Hybrid Simulator Backend Integration
+
+## Status
+
+Planning only. No implementation started. This note captures the design
+so the work can be picked up later without re-deriving context.
+
+## Goal
+
+Let `vbi.simulator.Simulator` / `Sweeper` dispatch to the TVB **hybrid**
+simulator (from the local fork at
+`/home/ziaee/git/tvb-root-hybrid-cpp/tvb_library`, package `tvb-library`,
+importable as `tvb`) as an additional backend, instead of only vbi's own
+`numpy` / `numba` / `cpp` / `cuda` / `jax` backends.
+
+The user-facing result should stay consistent with the existing pattern:
+
+```python
+sim = Simulator(spec, backend="tvb-numba")
+result = sim.run(duration=1000.0)
+```
+
+Changing backend should still mean changing one string, matching the
+project's [[feedback_api_simplicity]] principle — the TVB path should not
+require the caller to learn TVB's own object model (`NetworkSet`,
+`Subnetwork`, `Stim`, …).
+
+## Why
+
+TVB's hybrid fork already re-implements several models/cfuns with
+multiple execution backends of its own:
+
+- `ReferenceBackend` — plain Python/NumPy loop (`tvb/simulator/backend/ref.py`)
+- `NbHybridBackend` — Numba-compiled (`tvb/simulator/backend/nb_hybrid.py`)
+- `CppHybridBackend` — generated C++ extension (`tvb/simulator/backend_cpp/backend.py`)
+- a CUDA sweep backend (`tvb/simulator/backend/nb_hybrid_cuda_sweep_backend.py`)
+
+This overlaps with vbi's own backend set, but is a separate implementation
+with its own validated numerics (coupling functions, connectivity/delay
+conventions, monitors). Wiring it in as a backend gives:
+
+- a way to cross-validate vbi models against TVB's reference implementation
+  without hand-rolled comparison scripts (see `notes/milestones.md` TVB
+  parity section, currently deferred),
+- a path for users who already have TVB `Connectivity`/model definitions to
+  reuse them directly inside vbi inference workflows.
+
+## Current State
+
+- `vbi.simulator.backend.base.load_backend(name)` / `load_sweep_backend(name)`
+  are simple string-keyed registries returning a class that satisfies
+  `AbstractBackend` (`build(spec)`, `run(duration)`). See
+  `vbi/simulator/backend/base.py:13`. Adding a backend is a matter of adding
+  another branch plus a new implementation module.
+- `SimulationSpec` (`vbi/simulator/spec/simulation.py`) bundles model,
+  integrator, coupling, monitors, weights, node params — this is vbi's own
+  spec object, distinct from TVB's `Connectivity` / `Model` / `Coupling`
+  classes.
+- `vbi/models/tvbk/` already wraps a *different* TVB-adjacent project
+  (`tvbk`, a compiled MPR kernel) behind a `DeprecationWarning`-guarded
+  optional import — it is not the same code path as the hybrid fork and
+  should not be conflated with it.
+- `tvb-library` is currently installed as an editable dev dependency
+  pointing at the fork, not a pinned/released version. The fork is a
+  personal branch (`tvb-root-hybrid-cpp`), not upstream TVB.
+
+## Design Options
+
+1. **Adapter backend module** (recommended): add
+   `vbi/simulator/backend/tvb_/` with `build(spec)` translating
+   `SimulationSpec` → TVB's `Subnetwork` + `NetworkSet`, and `run(duration)`
+   dispatching to the requested TVB execution backend. Register under
+   backend names `"tvb-reference"`, `"tvb-numba"`, `"tvb-cpp"` (mirrors TVB's
[…]
