### 2026-07-16 4cd7d92
Remove tvbk backend and its example/docs references
vbi/models/tvbk/ wrapped an external TVB-Kernel library that isn't a
project dependency and had no test coverage or public re-export in
vbi/models/__init__.py. Drops the module, the mpr_tvbk.ipynb example,
and its API/toctree entries in docs/.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
