### 2026-07-16 d14ddfd
Remove leftover tvbk references from docs toctrees/API page
Follow-up to 4cd7d92 - these edits were dropped from that commit
because git add aborted on an already-removed pathspec.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
