### 2026-07-16 37654a5
Add cache=True to all numba @njit model kernels
Every multiprocessing-based batch workflow (e.g. the batch_run pattern
in examples/wilson_cowan_sde_numba.ipynb) forks/spawns fresh worker
processes that each recompile every @njit kernel from scratch on
first use, since none of the numba backends persisted compiled code
to disk. With many workers compiling simultaneously this can look
like a hang and cost minutes before any simulation work starts.

cache=True persists compiled machine code under __pycache__ so only
the very first run in the repo ever pays the compile cost. Verified:
vbi.models.n
[…]
