### 2026-07-16 d3b2732
Sync examples directory to docs/examples for nbsphinx compatibility
