### 2026-07-16 92eb93a
Remove deprecated workflow scripts for Jansen-Rit and MPR models
- Deleted `jr_vbi.py` and `jr_vbi_sweep.py` as they are no longer needed.
- Removed `mpr_sbi.py` which is outdated and not aligned with current practices.
- These changes streamline the codebase by eliminating unused files and improving maintainability.
