+++ examples/workflows/jr_vbi.py
@@ -1,278 +1,0 @@
-import argparse
-import os
-from pathlib import Path
-from time import perf_counter
-import numpy as np
-import matplotlib
-
-matplotlib.use("Agg")
-import matplotlib.pyplot as plt
-
-from vbi.simulator import Simulator, Sweeper
-from vbi.simulator.spec import (
-    SimulationSpec,
-    IntegratorSpec,
-    CouplingSpec,
-    MonitorSpec,
-    Connectivity,
-    SweepSpec,
-)
-from vbi.simulator.models.jansen_rit import jansen_rit
-from vbi.inference import (
-    VBIInference,
-    BoxUniform,
-    pairplot,
-    simulate_for_vbi_sweep_cached,
-)
-
-from vbi.feature_extraction import FeaturePruner
-from helpers import (
-    build_jr_spectral_pipeline,
-    plot_jr_timeseries_psd,
-    plot_feature_scatter,
-)
-
-
-def parse_args():
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument(
-        "--reuse-simulations",
-        action="store_true",
-        help="Load saved theta/x from the checkpoint and skip the training sweep.",
-    )
-    parser.add_argument(
-        "--simulate-only",
-        action="store_true",
-        help=(
-            "Run the cached simulation sweep, skip feature extraction/training, "
-            "and exit. Useful for timing simulation plus cache writing."
-        ),
-    )
-    return parser.parse_args()
-
-
-ARGS = parse_args()
-
-# ── Paths ──────────────────────────────────────────────────────────────────────
-
-DATA_DIR = Path(__file__).parent.parent / "data"
-OUT_DIR = Path(__file__).parent / "outputs"
-OUT_DIR.mkdir(exist_ok=True)
-CKPT = OUT_DIR / "jr_vbi.npz"
-CACHE_DIR = OUT_DIR / "jr_vbi_sim_cache"
-
-print("=" * 62)
-print("Jansen-Rit  —  VBIInference + sbi end-to-end pipeline")
-print("=" * 62)
-
-# ── 1 - Connectivity ───────────────────────────────────────────────────────────
-
-conn = Connectivity.from_file(
-    weights=DATA_DIR / "weights.txt",
-    tract_lengths=None,  # delays ignored for JR
-    normalize=True,
-)
-N_NODES = conn.n_nodes
-print(
-    f"\n  Connectivity : {N_NODES} nodes   "
-    f"W ∈ [{conn.weights.min():.4f}, {conn.weights.max():.4f}]   delays ignored"
-)
-
-# ── 2 - Simulation settings (match jansen_rit_sde_numba_cde.ipynb) ─────────────
-
-DT = 0.1  # ms
-SUBSAMPLE_PERIOD = 1.0  # ms
-FS_HZ = 1000.0 / DT
-DURATION = 2500.0  # ms
-T_CUT = 500.0  # ms
-INTEGRATOR_BACKEND = "numba"
-INFERENCE_BACKEND = "vbi"
-if INTEGRATOR_BACKEND == "numba":
-    os.environ.setdefault("VBI_NB_CACHE", str(OUT_DIR / "numba_cache"))
-
-# True parameters for the reference run: G=1.5, a_2=1.0 (C1=135)
-G_TRUE = 1.5
-A2_TRUE = 1.5  # a_2 = C1 / J
-J = 135.0
-
-# Prior: C1 ∈ [130, 300] → a_2 ∈ [130/135, 300/135]
-A2_LOW, A2_HIGH = 130.0 / J, 300.0 / J  # ≈ [0.963, 2.222]
-MU = 0.24
-NOISE_AMP = 0.1
-
-N_SIM = 2  # number of simulations
-N_WORKERS = 1
-
[…]
