### 2026-07-16 d4f6026
Update documentation and examples for workflows: add new gallery header, include workflows in Sphinx configuration, and adjust output path in damp oscillator demo.
