### 2026-07-16 1335dd5
Set JAX_PLATFORMS to "cpu" to avoid GPU-related issues in RTD environment
