### 2026-07-16 4484498
Refactor VBIInference to InferencePipeline and introduce TrainingOptions
- Created a new TrainingOptions dataclass to encapsulate training parameters for the InferencePipeline.
- Updated all references from VBIInference to InferencePipeline across the test suite.
- Adjusted test descriptions to reflect the new InferencePipeline class.
- Ensured compatibility of the new training options with existing functionality.
