+++ vbi/inference/_inference_pipeline.py
@@ -1,5 +1,5 @@
 """
-VBIInference - end-to-end SBI workflow object.  MI6.
+InferencePipeline - end-to-end SBI workflow object.  MI6.
 
 Owns the full loop: prior sampling → sweep simulation → feature extraction
 → SNPE training → posterior.  Raw recordings can optionally be cached to
@@ -162,6 +162,27 @@ def _unpack_estimator(d, EstCls):
     return est
 
 
+def _snpe_de_key(snpe) -> str:
+    """
+    Best-effort short key ('maf' | 'mdn' | 'nsf') for a vbi SNPE engine's
+    density estimator, used for checkpoint metadata / reconstruction on
+    ``load()``.  Works regardless of whether ``density_estimator`` was
+    given to ``SNPE`` as a string, factory, or instance.
+    """
+    from ._backends import get_estimator_map
+    from ._api import _is_estimator_instance
+
+    de_map = get_estimator_map(snpe._backend)
+    est = snpe._estimator
+    if est is None and _is_estimator_instance(snpe._de_type):
+        est = snpe._de_type
+    if est is not None:
+        for key, cls in de_map.items():
+            if isinstance(est, cls):
+                return key
+    return snpe._de_type if isinstance(snpe._de_type, str) else "maf"
+
+
 # ---------------------------------------------------------------------------
 # Config-loading helpers  (Step 5)
 # ---------------------------------------------------------------------------
@@ -301,12 +322,12 @@ def _unpack_pruner(d, pipeline) -> None:
 # SBC simulator helper  (Step 6)
 # ---------------------------------------------------------------------------
 
-def _make_simulator_fn(sim_spec, prior, pipeline, integrator_backend, duration):
+def _make_simulator_fn(sim_spec, prior, feature_pipeline, integrator_backend, duration):
     """
     Return a callable ``theta_1d -> x_1d`` suitable for :func:`run_sbc`.
 
     Each call patches the SimulationSpec with the given parameter values,
-    runs one simulation, and extracts features via the pipeline.
+    runs one simulation, and extracts features via the feature pipeline.
     """
     from vbi.simulator.spec.sweep import SweepSpec
     from vbi.simulator.api import Sweeper
@@ -318,7 +339,7 @@ def _make_simulator_fn(sim_spec, prior, pipeline, integrator_backend, duration):
         sweep_spec = SweepSpec(
             params      = theta_1d[None, :],   # (1, d_theta)
             param_names = tuple(param_names),
-            pipeline    = pipeline,
+            pipeline    = feature_pipeline,
         )
         sweeper = Sweeper(sim_spec, sweep_spec, backend=integrator_backend)
         labels, values = sweeper.run(duration)
@@ -339,7 +360,7 @@ def _prior_to_sbi(prior):
         import torch
     except ImportError as e:
         raise ImportError(
-            "inference_backend='sbi' requires PyTorch. "
+            "the sbi engine requires PyTorch. "
             "Install with:  pip install torch"
         ) from e
     from vbi.inference._prior import BoxUniform as _VBIBoxUniform
@@ -377,18 +398,29 @@ def _prior_to_sbi(prior):
     return _TorchWrapper(prior)
 
 
-def _setup_sbi_engine(prior, density_estimator: str, inference_engine, show_progress_bars: bool):
-    """Initialise (or return) an sbi SNPE engine."""
+def use_sbi(prior, density_estimator: str = "maf", show_progress_bars: bool = True):
+    """
+    Build an ``sbi.inference.SNPE`` engine wired up with a vbi prior, for use
+    as ``InferencePipeline(..., engine=use_sbi(prior))``.
+
+    Parameters
+    ----------
[…]
