### 2026-07-16 3ca59fd
Rename pipeline variable to feature_pipeline and extract features accordingly; introduce TrainingOptions for model training configuration.
