### 2026-07-16 2056c14
Refactor connectivity handling in C++ model tests: replace weights with Connectivity specification for improved clarity and structure.
