### 2026-07-16 b4ec44d
Refactor connectivity handling in simulation specs
- Updated multiple test files to replace direct weight and tract length parameters with a new Connectivity class.
- Ensured all instances of weights and tract lengths in simulation specifications are encapsulated within the Connectivity class for better structure and maintainability.
- Adjusted imports to include the Connectivity class from the appropriate module.
- Verified that all tests continue to function correctly after the refactor.
