+++ vbi/tests/validation/test_jansen_rit_numpy.py
@@ -2,6 +2,7 @@
 M0 validation: NumPy backend - deterministic Jansen-Rit.
 """
 
+from vbi.simulator.spec.connectivity import Connectivity
 import numpy as np
 import pytest
 
@@ -46,9 +47,9 @@ def make_jr_spec(
         integrator=IntegratorSpec(method=method, dt=dt),
         coupling=CouplingSpec("linear", a=coupling_strength),
         monitors=(MonitorSpec("raw"),),
-        weights=weights,
-        tract_lengths=tract_lengths,
-        speed=1.0,
+        connectivity=Connectivity(weights=weights, tract_lengths=tract_lengths, speed=1.0),
+
+
         node_params={
             name: np.full(weights.shape[0], value)
             for name, value in JR_PARAMS.items()
