### 2026-07-16 65f56ed
Replace weights, tract_lengths, and speed with connectivity in CppSweeper for improved parameter handling
