### 2026-07-16 1f52342
Refactor TestSweep to replace weights and tract_lengths with connectivity for improved parameter handling
