### 2026-07-16 40138e4
Refactor InferencePipeline to replace 'pipeline' with 'feature_pipeline' for improved clarity and consistency across tests
