+++ vbi/tests/validation/test_vbi_inference_diag.py
@@ -7,7 +7,7 @@ from vbi.simulator.spec import MonitorSpec
 from vbi.feature_extraction import (
     FeaturePipeline, get_features_by_domain, get_features_by_given_names,
 )
-from vbi.inference import InferencePipeline, BoxUniform
+from vbi.inference import InferencePipeline, BoxUniform, SNPE
 from .conftest import make_mpr_spec
 
 
@@ -26,10 +26,14 @@ N_SIM    = 20
 DURATION = 200.0
 
 
+def _numpy_engine():
+    return SNPE(PRIOR, density_estimator="maf", backend="numpy")
+
+
 def _trained_inf():
     inf = InferencePipeline(
-        sim_spec=_make_spec(), prior=PRIOR, pipeline=_stat_pipeline(),
-        integrator_backend="numpy", estimator_backend="numpy", show_progress_bars=False,
+        sim_spec=_make_spec(), prior=PRIOR, feature_pipeline=_stat_pipeline(),
+        integrator_backend="numpy", engine=_numpy_engine(), show_progress_bars=False,
     )
     inf.simulate(N_SIM, DURATION, seed=0)
     inf.train(stop_after_epochs=3, max_num_epochs=5)
@@ -46,8 +50,8 @@ class TestPlotLoss:
 
     def test_raises_before_train(self):
         inf = InferencePipeline(
-            sim_spec=_make_spec(), prior=PRIOR, pipeline=_stat_pipeline(),
-            integrator_backend="numpy", estimator_backend="numpy", show_progress_bars=False,
+            sim_spec=_make_spec(), prior=PRIOR, feature_pipeline=_stat_pipeline(),
+            integrator_backend="numpy", engine=_numpy_engine(), show_progress_bars=False,
         )
         with pytest.raises(RuntimeError, match="train"):
             inf.plot_loss()
@@ -64,8 +68,8 @@ class TestPairplot:
 
     def test_raises_before_train(self):
         inf = InferencePipeline(
-            sim_spec=_make_spec(), prior=PRIOR, pipeline=_stat_pipeline(),
-            integrator_backend="numpy", estimator_backend="numpy", show_progress_bars=False,
+            sim_spec=_make_spec(), prior=PRIOR, feature_pipeline=_stat_pipeline(),
+            integrator_backend="numpy", engine=_numpy_engine(), show_progress_bars=False,
         )
         with pytest.raises(RuntimeError, match="train"):
             inf.pairplot(np.zeros(4))
@@ -80,8 +84,8 @@ class TestRunSBC:
 
     def test_raises_before_train(self):
         inf = InferencePipeline(
-            sim_spec=_make_spec(), prior=PRIOR, pipeline=_stat_pipeline(),
-            integrator_backend="numpy", estimator_backend="numpy", show_progress_bars=False,
+            sim_spec=_make_spec(), prior=PRIOR, feature_pipeline=_stat_pipeline(),
+            integrator_backend="numpy", engine=_numpy_engine(), show_progress_bars=False,
         )
         inf.simulate(N_SIM, DURATION, seed=1)
         with pytest.raises(RuntimeError, match="train"):
