### 2026-07-17 0b45c95
Update test_repr to assert "vbi:SNPE" instead of "maf" for improved clarity in representation
