### 2026-07-17 ins-amu/vbi#75 @sampottinger
Thank you! The build on main worked. Please close this issue after considering the following final optional feedback:

 - The final sentence of summary was just a tiny bit difficult to parse. You may consider breaking into two sentences.
 - I'd probably still cite JupyterLab (see [Kluyver et al](https://eprints.soton.ac.uk/403913/)) but I leave it to your discretion as I believe you credit them elsewhere in the project.
 - You may consider breaking up the sentence at line 40 into two (Gradient-free MCMC struggles in...)

Appreciate your hard work on this @Ziaeemehr!
