### 2026-07-17 ins-amu/vbi#77 @sampottinger
Add link to an example from quickstart
An optional part of https://github.com/openjournals/joss-reviews/issues/10612. Quickstarts often include example code not just installation information. Reference to https://github.com/ins-amu/vbi#quick-start. Given the robustness of your other documentation, I recommend just adding a link to some example code like https://vbi.readthedocs.io/latest/examples/intro_feature.html at the bottom of your quickstart in your README so users have a place to try out a working example immediately.
