### 2026-07-19 dbb687e
feat: Add Jupyter Notebooks reference to bibliography and update paper summary for clarity
