+++ joss/paper.bib
@@ -349,3 +349,12 @@ doi={10.1109/RBME.2025.3562951}
   year={2025},
   howpublished={\url{https://openai.com/index/gpt-5-system-card/}},
 }
+
+@inproceedings{Kluyver2016jupyter,
+  title={Jupyter Notebooks-a publishing format for reproducible computational workflows},
+  author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
+  booktitle={Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+  pages={87--90},
+  year={2016},
+  organization={IOS Press}
+}
