### 2026-07-19 ins-amu/vbi#75 @Ziaeemehr
Thanks for the careful final pass! I've addressed all three points:

Split the final summary sentence into two.
Added the JupyterLab citation (Kluyver et al., 2016) at its first mention.
Split the "Gradient-free MCMC struggles..." sentence at line 40 into two.
Closing this issue now, thanks again for the thorough review throughout, @sampottinger !
