### 2026-07-19 9771a09
feat: Add introductory example link to README for improved user guidance
