### 2026-07-19 ins-amu/vbi#77 @Ziaeemehr
Good suggestion, added a direct link to a working example ([Introduction & Feature Extraction](https://vbi.readthedocs.io/latest/examples/intro_feature.html)) right after the install commands in the Quick Start section, so users can try VBI immediately after installing.
