### 2026-07-19 6f33225
refactor: Remove installation options section from README for clarity
