+++ README.md
@@ -62,12 +62,6 @@ pip install vbi[inference-gpu]     # Full functionality with GPU support
 
 👉 Ready to try it out? Start with a working example: [Introduction & Feature Extraction](https://vbi.readthedocs.io/latest/examples/intro_feature.html).
 
-### Installation Options
-
-- **`pip install vbi`** - Light version with CPU simulation only  
-- **`pip install vbi[inference]`** - Adds parameter inference capabilities  
-- **`pip install vbi[inference-gpu]`** - Full functionality with GPU acceleration  
-
 ### Using Docker
 
 ```bash
