### 2026-07-19 661241f
fix: Update package version in README from v0.2.1 to v0.4.3
