### 2026-08-04 53ca3cc
chore: Release v0.4.4 with CITATION.cff and PyPI publish workflow
Adds CITATION.cff with the full JOSS paper author list and ORCIDs so
the Zenodo GitHub-integration archive picks up matching metadata, and
a trusted-publishing workflow to push sdists to PyPI on v* tags.

Co-Authored-By: Claude Sonnet 5 <noreply@anthropic.com>
