+++ CITATION.cff
@@ -1,0 +1,33 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it using the metadata below."
+title: "Virtual Brain Inference (VBI) toolkit on EBRAINS"
+version: 0.4.4
+date-released: "2026-08-04"
+license: Apache-2.0
+repository-code: "https://github.com/ins-amu/vbi"
+url: "https://vbi.readthedocs.io/"
+authors:
+  - family-names: Ziaeemehr
+    given-names: Abolfazl
+    orcid: "https://orcid.org/0000-0002-4696-9947"
+    affiliation: "Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France"
+  - family-names: Domide
+    given-names: Lia
+    orcid: "https://orcid.org/0000-0002-4822-2046"
+    affiliation: "Codemart, Cluj-Napoca, Romania"
+  - family-names: Woodman
+    given-names: Marmaduke
+    orcid: "https://orcid.org/0000-0002-8410-4581"
+    affiliation: "Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France"
+  - family-names: Petkoski
+    given-names: Spase
+    orcid: "https://orcid.org/0000-0003-4540-6293"
+    affiliation: "Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France"
+  - family-names: Jirsa
+    given-names: Viktor
+    orcid: "https://orcid.org/0000-0002-8251-8860"
+    affiliation: "Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France"
+  - family-names: Hashemi
+    given-names: Meysam
+    orcid: "https://orcid.org/0000-0001-5289-9837"
+    affiliation: "Aix Marseille Univ, INSERM, INS, Institut de Neurosciences des Systèmes, Marseille, France"
