### 2026-08-27 #10612 @crvernon
@Ziaeemehr - I just need you to address the following before I accept for publication.

In your paper;

- [ ] LINE 24: The citation “JupyterLab (Kluyver et al. (2016))” uses nested parentheses. Change to “JupyterLab (Kluyver et al., 2016)” or “JupyterLab, as described by Kluyver et al. (2016).”
- [ ] LINES 29-31: The citations for Wilson-Cowan, Jansen-Rit, Stuart-Landau, Montbrió, Wong-Wang, and Epileptor use nested parentheses, e.g., “Wilson-Cowan (Wilson & Cowan (1972)).” Change these to forms such as “Wilson-Cowan (Wilson & Cowan, 1972).”
- [ ] LINES 36-37: The MAF and NSF citations also use nested parentheses. Change “Papamakarios et al. (2017)” and “Durkan et al. (2019)” inside the parenthetical list to “Papamakarios et al., 2017” and “Durkan et al., 2019.”
- [ ] LINE 42: The citation renders as “M. Betancourt & Girolami (2013),” while the later Betancourt citation renders with the full first name. Normalize the author metadata in the bibliography so citations render consistently as “Betancourt & Girolami (2013)” and “Betancourt et al. (2017).”
- [ ] LINES 80-81: The claim of “up to 100× speedups” is a quantitative performance claim but has no supporting citation or benchmark reference at this point. JOSS explicitly expects performance claims to be verifiable; cite the benchmark or point readers to reproducible benchmark material.
- [ ] LINES 108-109: The AI-tool citations are duplicated: “Claude Sonnet models (Anthropic; Anthropic (2025))” and “GPT-5 models (OpenAI; OpenAI (2025)).” Use forms such as “Claude Sonnet models (Anthropic, 2025) and GPT-5 models (OpenAI, 2025).”
- [ ] LINES 108-113: JOSS requires the AI disclosure to identify the tools/models and versions used. “Claude Sonnet models” and “GPT-5 models” are not sufficiently version-specific. Give the exact model/version identifiers used.
- [ ] LINES 134-136: Capitalization following the list labels is inconsistent: “generating,” “Computing,” and “Amortized.” Use parallel capitalization throughout, preferably lowercase after each colon.
- [ ] LINES 138-139: The citation renders as “Michael Betancourt et al. (2017).” Normalize the BibTeX author metadata so the author is represented consistently with the Betancourt & Girolami reference.
- [ ] LINE 145: For consistency with the other section headings, capitalize “Integration”: “EBRAINS Integration.”
- [ ] LINE 149: The paper otherwise calls VBI a “toolkit”; “toolbox” here should be changed to “toolkit” for terminology consistency.
- [ ] LINES 159-161: “and government grant managed by…” is missing an article. Change to “and a government grant managed by the Agence Nationale de la Recherche….”
- [ ] LINE 165: “The virtual parkinsonian patient” should capitalize the proper name: “The virtual Parkinsonian patient.” You can maintain the formatting of characters in your references by placing curly brackets around those characters in the bib file.
- [ ] LINE 170: The journal title should be “Journal of the Royal Society Interface,” not “Journal of The Royal Society Interface.”
- [ ] LINES 172-176: Capitalize the proper algorithm name “Hamiltonian Monte Carlo” in both Betancourt reference titles.
- [ ] LINE 195: “Elife” should use the journal’s proper capitalization: “eLife.”
- [ ] LINES 202-203: Capitalize “Bayesian” in “Amortized Bayesian inference….”
- [ ] LINES 215-217: Capitalize the proper terms in the title: “Fully Bayesian estimation … with self-tuning Hamiltonian Monte Carlo.”
- [ ] LINES 250-252: Capitalize “Monte Carlo” in “Sequential Monte Carlo without likelihoods.”
- [ ] LINES 259-262: “Virtual brain inference (VBI)” should preserve the software name as “Virtual Brain Inference (VBI),” and “Elife” should be “eLife.”

Let me know when these have been addressed.  Thank you.
