### 2026-08-27 c128445
fix: Address JOSS editor's pre-acceptance revision comments on paper
- Replace nested-parenthesis citations with pandoc bracket-citation
  syntax throughout (JupyterLab, whole-brain model list, SBI mention);
  same fix applied to an unflagged instance of the same defect.
- Normalize Betancourt author-name formatting across the two 2013/2017
  entries so citeproc no longer disambiguates with extra initials.
- Protect proper nouns (Bayesian, Hamiltonian Monte Carlo, Parkinsonian,
  Virtual Brain Inference, eLife) from CSL sentence-casing via braces.
- Fix "Journal of The Royal Society Interface"
[…]
