+++ joss/paper.bib
@@ -7,8 +7,8 @@ doi={10.48550/arXiv.1410.5110},
 }
 
 @Article{Betancourt2013b,
-author="M. Betancourt  and M. Girolami",
-title="Hamiltonian Monte Carlo for hierarchical models",
+author={Betancourt, Michael and Girolami, Mark},
+title={Hamiltonian {Monte Carlo} for hierarchical models},
 journal="arXiv:1312.0906",
 year="2013",
 doi={10.48550/arXiv.1312.0906},
@@ -24,7 +24,7 @@ doi={10.48550/arXiv.1312.0906},
   doi={10.1016/j.neuroimage.2020.116839},
 }
 @article{Sisson2007,
-  title={Sequential monte carlo without likelihoods},
+  title={Sequential {Monte Carlo} without likelihoods},
   author={Sisson, Scott A and Fan, Yanan and Tanaka, Mark M},
   journal={Proceedings of the National Academy of Sciences},
   volume={104},
@@ -96,7 +96,7 @@ doi={10.48550/arXiv.1312.0906},
 @article{gonccalves2020training,
   title={Training deep neural density estimators to identify mechanistic models of neural dynamics},
   author={Gon{\c{c}}alves, Pedro J and Lueckmann, Jan-Matthis and Deistler, Michael and Nonnenmacher, Marcel and {\"O}cal, Kaan and Bassetto, Giacomo and Chintaluri, Chaitanya and Podlaski, William F and Haddad, Sara A and Vogels, Tim P and others},
-  journal={elife},
+  journal={{eLife}},
   volume={9},
   pages={e56261},
   year={2020},
@@ -104,7 +104,7 @@ doi={10.48550/arXiv.1312.0906},
   doi={10.7554/eLife.56261},
 }
 @article{hashemi2023amortized,
-  title={Amortized Bayesian inference on generative dynamical network models of epilepsy using deep neural density estimators},
+  title={Amortized {Bayesian} inference on generative dynamical network models of epilepsy using deep neural density estimators},
   author={Hashemi, Meysam and Vattikonda, Anirudh N and Jha, Jayant and Sip, Viktor and Woodman, Marmaduke M and Bartolomei, Fabrice and Jirsa, Viktor K},
   journal={Neural Networks},
   volume={163},
@@ -128,7 +128,7 @@ doi={10.48550/arXiv.1312.0906},
   doi={10.1088/2632-2153/ad6230},
 }
 @article{betancourt2017geometric,
-  title={The geometric foundations of hamiltonian monte carlo},
+  title={The geometric foundations of {Hamiltonian Monte Carlo}},
   author={Betancourt, Michael and Byrne, Simon and Livingstone, Sam and Girolami, Mark},
   journal={Bernoulli},
   volume={23},
@@ -173,9 +173,9 @@ doi={10.48550/arXiv.1312.0906},
 doi={10.1109/RBME.2025.3562951}
  }
 @article{Ziaeemehr2025,
-  title={Virtual Brain Inference (VBI), a flexible and integrative toolkit for efficient probabilistic inference on whole-brain models},
+  title={{Virtual Brain Inference} (VBI), a flexible and integrative toolkit for efficient probabilistic inference on whole-brain models},
   author={Ziaeemehr, Abolfazl and Woodman, Marmaduke and Domide, Lia and Petkoski, Spase and Jirsa, Viktor and Hashemi, Meysam},
-  journal={Elife},
+  journal={{eLife}},
   volume={14},
   pages={RP106194},
   year={2025},
@@ -185,7 +185,7 @@ doi={10.1109/RBME.2025.3562951}
 
 
 @article{Jha2022,
[…]
