+++ joss/paper.md
@@ -49,14 +49,14 @@ Understanding the principles and causal mechanisms underlying complex brain (dys
 
 # Statement of Need
 
-VBI (https://github.com/ins-amu/vbi) is a Python-based toolkit tailored for probabilistic inference at the whole-brain scale. A lightweight version with minimal dependencies is available to run online in JupyterLab (@Kluyver2016jupyter). It also leverages the flexibility of Python while harnessing the high performance of C++ for optimized computation and massive parallelization using CUDA on GPUs. VBI seamlessly integrates structural and functional neuroimaging data, providing space-efficient storage and memory-optimized batch processing. With its user-friendly API supporting:
+VBI (https://github.com/ins-amu/vbi) is a Python-based toolkit tailored for probabilistic inference at the whole-brain scale. A lightweight version with minimal dependencies is available to run online in JupyterLab [@Kluyver2016jupyter]. It also leverages the flexibility of Python while harnessing the high performance of C++ for optimized computation and massive parallelization using CUDA on GPUs. VBI seamlessly integrates structural and functional neuroimaging data, providing space-efficient storage and memory-optimized batch processing. With its user-friendly API supporting:
 
-- **Whole-brain models**: Wilson-Cowan (@WilsonCowan1972), Jansen-Rit (@JansenRit1995), Stuart-Landau (@Cabral2011), Montbrió (@Montbrio2015), Wong-Wang (@Deco2013), and Epileptor (@Jirsa2014).
+- **Whole-brain models**: Wilson-Cowan [@WilsonCowan1972], Jansen-Rit [@JansenRit1995], Stuart-Landau [@Cabral2011], Montbrió [@Montbrio2015], Wong-Wang [@Deco2013], and Epileptor [@Jirsa2014].
 - **Fast simulation**: Just-in-time compilation of models across Python/C++ and CPU/GPU devices.
 - **Feature extraction**: Functional connectivity (FC), functional connectivity dynamics (FCD), and power spectral density (PSD).
-- **Deep neural density estimators**: Masked autoregressive flows (MAFs; @Papamakarios2017), and neural spline flows (NSFs; @Durkan2019).
+- **Deep neural density estimators**: Masked autoregressive flows [MAFs\; @Papamakarios2017], and neural spline flows [NSFs\; @Durkan2019].
 
-Traditional methods such as Markov Chain Monte Carlo (MCMC) and Approximate Bayesian Computation (ABC) face significant challenges in this context [@Sisson2007]. Gradient-free MCMC struggles in high-dimensional settings. Gradient-based approaches, meanwhile, often have difficulty converging for bistable systems with switching dynamics, often requiring extensive tuning [@Betancourt2013b] and computational resources [@Hashemi2020]. ABC suffers from the curse of dimensionality and relies on predefined thresholds for sample acceptance, leading to inefficiencies and potential biases when rejecting samples that fall outside narrow criteria [@Jha2022]. VBI leverages advanced Simulation-based Inference (SBI; @cranmer2020frontier), which sidesteps these issues by using only forward simulations and training deep neural density estimators to directly approximate posterior distributions [@gonccalves2020training ; @hashemi2023amortized]. This approach enhances efficiency, scalability, and robustness for inverting complex whole-brain models [@hashemi2024simulation].
[…]
